### 2020-06-13 #2335 @whedon
[PRE REVIEW]: HPX - The C++ Standard Library for Parallelism and Concurrency
**Submitting author:** @diehlpk (<a href="http://orcid.org/0000-0003-3922-8419">Patrick Diehl</a>)
**Repository:** <a href="https://github.com/STEllAR-GROUP/hpx" target ="_blank">https://github.com/STEllAR-GROUP/hpx</a>
**Version:** 1.4.1
