### 2020-06-13 3c035aa
Revert "Create coverage reports and upload them to codecov.io"
