+++ .circleci/config.yml
@@ -34,17 +34,6 @@ move_debug_log: &move_debug_log
         mkdir -p ${CIRCLE_JOB}
         cp debug-log.txt ${CIRCLE_JOB} || true
 
-collect_coverage: &collect_coverage
-    name: Collecting coverage reports
-    command: |
-        mkdir -p /hpx/coverage
-        lcov -q --gcov-tool llvm-cov-wrapper --no-external --capture --base-directory /hpx/source/ --directory /hpx/build/ --output-file /hpx/coverage/${CIRCLE_JOB}.info
-
-persist_coverage: &persist_coverage
-    root: /hpx
-    paths:
-      - ./coverage
-
 gh_pages_filter: &gh_pages_filter
     filters:
       branches:
@@ -202,25 +191,24 @@ jobs:
                 /hpx/source \
                 -G "Ninja" \
                 -DCMAKE_BUILD_TYPE=Debug \
-                -DCMAKE_CXX_CLANG_TIDY=clang-tidy \
-                -DCMAKE_CXX_FLAGS="--coverage -fcolor-diagnostics" \
-                -DCMAKE_EXE_LINKER_FLAGS="--coverage" \
-                -DCMAKE_EXPORT_COMPILE_COMMANDS=On \
                 -DHPX_WITH_MALLOC=system \
                 -DHPX_WITH_GIT_COMMIT=${CIRCLE_SHA1} \
                 -DHPX_WITH_GIT_BRANCH="${CIRCLE_BRANCH}" \
                 -DHPX_WITH_GIT_TAG="${CIRCLE_TAG}" \
                 -DHPX_WITH_TOOLS=On \
+                -DCMAKE_CXX_FLAGS="-fcolor-diagnostics" \
                 -DHPX_WITH_TESTS_HEADERS=On \
                 -DHPX_WITH_COMPILER_WARNINGS=On \
                 -DHPX_WITH_COMPILER_WARNINGS_AS_ERRORS=On \
                 -DHPX_WITH_DEPRECATION_WARNINGS=On \
+                -DCMAKE_CXX_CLANG_TIDY=clang-tidy \
                 -DHPX_WITH_THREAD_LOCAL_STORAGE=On \
                 -DHPX_WITH_STACKTRACES_STATIC_SYMBOLS=On \
                 -DHPX_WITH_STACKTRACES_DEMANGLE_SYMBOLS=Off \
                 -DHPX_WITH_TESTS_DEBUG_LOG=On \
                 -DHPX_WITH_TESTS_DEBUG_LOG_DESTINATION=/hpx/build/debug-log.txt \
                 -DHPX_WITH_SPINLOCK_DEADLOCK_DETECTION=On \
+                -DCMAKE_EXPORT_COMPILE_COMMANDS=On \
                 -DHPX_WITH_DOCUMENTATION=On \
                 -DHPX_WITH_DOCUMENTATION_OUTPUT_FORMATS="${DOCUMENTATION_OUTPUT_FORMATS}"
       - persist_to_workspace:
@@ -384,14 +372,10 @@ jobs:
           <<: *move_core_dump
       - run:
           <<: *move_debug_log
-      - run:
-          <<: *collect_coverage
       - store_test_results:
           path: tests.examples
       - store_artifacts:
           path: tests.examples
-      - persist_to_workspace:
-          <<: *persist_coverage
 
   tests.unit.actions:
     <<: *defaults
@@ -414,14 +398,10 @@ jobs:
           <<: *move_core_dump
       - run:
           <<: *move_debug_log
-      - run:
-          <<: *collect_coverage
       - store_test_results:
           path: tests.unit.actions
       - store_artifacts:
           path: tests.unit.actions
-      - persist_to_workspace:
-          <<: *persist_coverage
 
[…]
