### 2020-06-13 cd3609a
move libcds setup back
