### 2020-06-14 4ce22cb
Re-adding API functions that were lost during merges
