### 2020-06-14 cae4240
Adding hpx::threads::get_idle_core_mask() API returning a bitmask representing the idle cores
- flyby: remove old (leftover) file
