+++ libs/threadmanager/include/hpx/modules/threadmanager.hpp
@@ -20,6 +20,7 @@
 #include <hpx/threading_base/thread_init_data.hpp>
 #include <hpx/threading_base/thread_num_tss.hpp>
 #include <hpx/threading_base/thread_pool_base.hpp>
+#include <hpx/topology/cpu_mask.hpp>
 
 #include <atomic>
 #include <cstddef>
@@ -164,6 +165,8 @@ namespace hpx { namespace threads {
             std::size_t num_thread = std::size_t(-1), bool reset = false);
 
         std::int64_t get_idle_core_count();
+
+        mask_type get_idle_core_mask();
 
         std::int64_t get_background_thread_count();
 
