+++ libs/runtime_local/include/hpx/runtime_local/thread_pool_helpers.hpp
@@ -10,6 +10,7 @@
 
 #include <hpx/config.hpp>
 #include <hpx/threading_base/thread_pool_base.hpp>
+#include <hpx/topology/cpu_mask.hpp>
 
 #include <cstddef>
 #include <cstdint>
@@ -85,6 +86,10 @@ namespace hpx { namespace threads {
     /// The function \a get_idle_core_count returns the number of currently
     /// idling threads (cores).
     HPX_EXPORT std::int64_t get_idle_core_count();
+
+    /// The function \a get_idle_core_mask returns a bit-mask representing the
+    /// currently idling threads (cores).
+    HPX_EXPORT mask_type get_idle_core_mask();
 
     /// The function \a enumerate_threads will invoke the given function \a f
     /// for each thread with a matching thread state.
