### 2020-06-15 fd7fff4
Add PMC mailing list address to SUPPORT.md
