### 2020-06-15 08a143f
Fix PAPI compilation
