### 2020-06-15 #2335 @diehlpk
We have the paper in the following branch https://github.com/STEllAR-GROUP/hpx/tree/JOSS, since we want to keep the paper aligned with the latest stable release.
