### 2020-06-15 5f508af
integrating Hartmuts changes to splittable executor
