+++ hpx/parallel/executors/splittable_executor.hpp
@@ -9,9 +9,11 @@
 #define HPX_SPLITTABLE_EXECUTOR_HPP
 
 #include <hpx/config.hpp>
-#include <hpx/error.hpp>
-#include <hpx/parallel/util/detail/splittable_task.hpp>
-#include <hpx/traits/is_executor.hpp>
+#include <hpx/execution/traits/is_executor.hpp>
+#include <hpx/executors.hpp>
+#include <hpx/iterator_support/iterator_range.hpp>
+#include <hpx/serialization/serialize.hpp>
+#include <hpx/timing.hpp>
 
 #include <algorithm>
 #include <cstddef>
@@ -27,8 +29,11 @@ namespace hpx { namespace parallel { namespace execution {
     /// This executor parameters type makes sure that as many loop iterations
     /// are combined as necessary to run for the amount of time specified.
     ///
-    struct splittable_executor : parallel_executor
+    struct splittable_executor
+      : parallel_policy_executor<hpx::launch::async_policy>
     {
+        using base_type = parallel_policy_executor<hpx::launch::async_policy>;
+
     public:
         /// Construct an \a splittable_executor executor parameters object
         ///
@@ -36,16 +41,9 @@ namespace hpx { namespace parallel { namespace execution {
         ///       types will use 80 microseconds as the minimal time for which
         ///       any of the scheduled chunks should run.
         ///
-        splittable_executor() {}
-
-	splittable_executor(std::string split_type)
+        splittable_executor()
+          : split_type_(splittable_mode::all)
         {
-            if (split_type != "all" && split_type != "idle")
-            {
-                HPX_THROW_EXCEPTION(hpx::bad_parameter, "throw_hpx_exception",
-                            "unknwn type, type should be either all or idle");
-            }
-            split_type_ = split_type;
         }
 
         /// Construct an \a splittable_executor executor parameters object
@@ -54,13 +52,32 @@ namespace hpx { namespace parallel { namespace execution {
         ///                     to decide how many loop iterations should be
         ///                     combined.
 
+        splittable_executor(splittable_mode split_type)
+          : split_type_(split_type)
+        {
+            if (split_type != splittable_mode::all &&
+                split_type != splittable_mode::idle &&
+                split_type != splittable_mode::idle_mask)
+            {
+                HPX_THROW_EXCEPTION(hpx::bad_parameter,
+                    "splittable_executor::splittable_executor",
+                    "unknown type, type should be either all, idle, or "
+                    "idle_mask");
+            }
+        }
+
         /// \cond NOINTERNAL
         // Estimate a chunk size based on number of cores used.
-        template <typename Executor, typename F>
-        std::size_t get_chunk_size(
-            Executor&& exec, F&& f, std::size_t cores, std::size_t count)
+        template <typename Parameters, typename F>
+        static std::size_t get_chunk_size(
+            Parameters&&, F&&, std::size_t, std::size_t count)
         {
             return count;
+        }
+
+        HPX_FORCEINLINE static std::size_t processing_units_count()
+        {
+            return hpx::get_os_thread_count();
         }
         /// \endcond
 
@@ -76,8 +93,9 @@ namespace hpx { namespace parallel { namespace execution {
 
             for (auto const& elem : shape)
             {
-                results.push_back(hpx::async(make_splittable_task(
-                    std::forward<F>(f), elem, split_type_)));
+                results.push_back(hpx::async(
+                    make_splittable_task(static_cast<base_type&>(*this),
+                        std::forward<F>(f), elem, split_type_)));
             }
 
             return results;
[…]
