+++ hpx/parallel/executors.hpp
@@ -8,6 +8,7 @@
 
 #include <hpx/config.hpp>
 
+#include <hpx/parallel/executors/util/splittable_task.hpp>
 #include <hpx/parallel/executors/default_executor.hpp>
 #include <hpx/parallel/executors/distribution_policy_executor.hpp>
 #include <hpx/parallel/executors/parallel_executor.hpp>
