### 2020-06-16 2f927bc
Fix missing pragma once in compatibility headers
