+++ libs/execution_base/include_compatibility/hpx/basic_execution/register_locks.hpp
@@ -4,6 +4,8 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
+#pragma once
+
 #include <hpx/config.hpp>
 #include <hpx/execution_base/config/defines.hpp>
 #include <hpx/execution_base/register_locks.hpp>
