+++ libs/lcos_local/include_compatibility/hpx/local_lcos/receive_buffer.hpp
@@ -4,6 +4,8 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
+#pragma once
+
 #include <hpx/config.hpp>
 #include <hpx/lcos_local/config/defines.hpp>
 #include <hpx/lcos_local/receive_buffer.hpp>
