+++ libs/lcos_local/include_compatibility/hpx/local_lcos/force_linking.hpp
@@ -1,19 +1,0 @@
-//  Copyright (c) 2019 Ste||ar Group
-//
-//  SPDX-License-Identifier: BSL-1.0
-//  Distributed under the Boost Software License, Version 1.0. (See accompanying
-//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-#include <hpx/config.hpp>
-#include <hpx/lcos_local/config/defines.hpp>
-#include <hpx/lcos_local/force_linking.hpp>
-
-#if defined(HPX_LCOS_LOCAL_HAVE_DEPRECATION_WARNINGS)
-#if defined(HPX_MSVC)
-#pragma message("The header hpx/local_lcos/force_linking.hpp is deprecated, \
-    please include hpx/lcos_local/force_linking.hpp instead")
-#else
-#warning "The header hpx/local_lcos/force_linking.hpp is deprecated, \
-    please include hpx/lcos_local/force_linking.hpp instead"
-#endif
-#endif
