### 2020-06-16 db5f3e7
Add sized_sentiel_for
