### 2020-06-16 0ec293f
Fix format errors intrduced after latest conflicts
