+++ libs/async_cuda/include/hpx/async_cuda/cuda_executor.hpp
@@ -11,9 +11,9 @@
 #include <hpx/async_cuda/cuda_exception.hpp>
 #include <hpx/async_cuda/cuda_future.hpp>
 #include <hpx/async_cuda/target.hpp>
-#include <hpx/basic_execution/execution.hpp>
 #include <hpx/errors/exception.hpp>
 #include <hpx/execution/traits/is_executor.hpp>
+#include <hpx/execution_base/execution.hpp>
 #include <hpx/futures/future.hpp>
 #include <hpx/futures/traits/future_access.hpp>
 
