+++ libs/async_cuda/include/hpx/async_cuda/cublas_executor.hpp
@@ -12,8 +12,8 @@
 #include <hpx/async_cuda/cuda_executor.hpp>
 #include <hpx/async_cuda/cuda_future.hpp>
 #include <hpx/async_cuda/target.hpp>
-#include <hpx/basic_execution/execution.hpp>
 #include <hpx/execution/traits/is_executor.hpp>
+#include <hpx/execution_base/execution.hpp>
 #include <hpx/futures/future.hpp>
 
 // CUDA runtime
