### 2020-06-16 7465579
Reduce memory requirements for our main shared state
- flyby: add missing headers to CMakeLists.txt
