### 2020-06-16 4df188a
Storing hpx_options in global property to speed up summary report
