### 2020-06-17 4929be1
Add disable_sized_sentinel_for concept
