+++ cmake/HPX_AddConfigTest.cmake
@@ -446,6 +446,14 @@ function(hpx_check_for_cxx20_coroutines)
 endfunction()
 
 # ##############################################################################
+function(hpx_check_for_cxx20_disable_sized_sentinel_for)
+  add_hpx_config_test(
+    HPX_WITH_CXX20_DISABLE_SIZED_SENTINEL_FOR SOURCE cmake/tests/cxx20_disable_sized_sentinel_for.cpp FILE
+                                     ${ARGN}
+  )
+endfunction()
+
+# ##############################################################################
 function(hpx_check_for_builtin_integer_pack)
   add_hpx_config_test(
     HPX_WITH_BUILTIN_INTEGER_PACK SOURCE cmake/tests/builtin_integer_pack.cpp
