### 2020-06-17 #2335 @danielskatz
👋 @G-071, @davidbeckingsale - Would one of of you be willing to review this submission for JOSS?  We openly review both the short paper and the software in a GitHub issue, based on a [checklist](https://joss.readthedocs.io/en/latest/review_criteria.html#review-items)
