### 2020-06-17 #2352 @whedon
[REVIEW]: HPX - The C++ Standard Library for Parallelism and Concurrency
**Submitting author:** @diehlpk (<a href="http://orcid.org/0000-0003-3922-8419">Patrick Diehl</a>)
**Repository:** <a href="https://github.com/STEllAR-GROUP/hpx" target ="_blank">https://github.com/STEllAR-GROUP/hpx</a>
**Version:** 1.5.0
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.12907034" target="_blank">10.6084/m9.figshare.12907034</a>
