### 2020-06-17 #2352 @danielskatz
👋 @bhatele & @davidbeckingsale - thanks for agreeing to review this submission. Please be sure to read the comments above, and let me know if you have any questions. Basically, your job is to check the [article proof](https://github.com/openjournals/joss-papers/blob/joss.02352/joss.02352/10.21105.joss.02352.pdf) and [repository](https://github.com/STEllAR-GROUP/hpx) and check items off your checklist above. (And note that you will have to accept the invitation before you will be able to do so.)

If you see small problems that need to be discussed, feel free to discuss them here. But if you can, create a new issue in the [target repository]](https://github.com/STEllAR-GROUP/hpx) and link to [this review thread](https://github.com/openjournals/joss-reviews/issues/2352) in that issue to create a corresponding breadcrumb trail here.

I look forward to seeing how this review goes
