### 2020-06-17 STEllAR-GROUP/hpx#4763 @davidbeckingsale
License file name
@diehlpk - is there a specific reason the license filename is `LICENSE_0_1.txt`, rather than the standard `LICENSE`?

cc https://github.com/openjournals/joss-reviews/issues/2352
