### 2020-06-17 STEllAR-GROUP/hpx#4763 @hkaiser
@davidbeckingsale it's the standard way to reference the Boost Software License (see https://www.boost.org/users/license.html)
