### 2020-06-17 #2352 @davidbeckingsale
@diehlpk - can you offer a quick bit of rationale behind author selection? A couple of contributors are not on the author list, but have more contributions (in terms of commits/loc) than some of the other authors.
