### 2020-06-17 #2352 @diehlpk
@davidbeckingsale We sent some an email and one reminder to our mailing lists and invited people to contribute to the paper before the deadline. The current list of authors followed the invitation and contributed to the paper. However, we could not contact over 100 contributors in person.
