### 2020-06-17 #2352 @davidbeckingsale
> @davidbeckingsale We sent some an email and one reminder to our mailing lists and invited people to contribute to the paper before the deadline. The current list of authors followed the invitation and contributed to the paper. However, we could not contact over 100 contributors in person.

Perfect. Sorry for being pedantic, I am just trying to make sure that the rationale for each of the checklist points is document somewhere 😄
