### 2020-06-18 f4f0057
Ignore compatibility headers when checking for circular dependencies
