+++ .circleci/config.yml
@@ -107,7 +107,12 @@ jobs:
       - run:
           name: Check circular dependencies with cpp-dependencies
           command: |
-              cpp-dependencies --dir /hpx/source/libs --graph-cycles /tmp/circular_deps.dot
+              # Ignore compatibility headers (the last cut -c3- is to remove a
+              # leading ./ in path names)
+              cpp-dependencies \
+                  --dir /hpx/source/libs \
+                  --ignore $(find /hpx/source/libs -type d -wholename '*/include_compatibility' | cut -d'/' -f5-) \
+                  --graph-cycles /tmp/circular_deps.dot
               dot /tmp/circular_deps.dot -Tsvg -o /tmp/circular_deps.svg
               if [[ $(wc -l /tmp/circular_deps.dot | awk '{print $1}') -gt 2 ]]; then exit 1; fi
       - store_artifacts:
