### 2020-06-18 68f9f11
Make distance dispatch according to sized_sentinel_for and test
