### 2020-06-18 97fff87
Remove all force linking leftovers
