+++ libs/async_cuda/CMakeLists.txt
@@ -41,7 +41,7 @@ include(HPX_AddModule)
 add_hpx_module(
   async_cuda
   COMPATIBILITY_HEADERS ON
-  DEPRECATION_WARNINGS FORCE_LINKING_GEN
+  DEPRECATION_WARNINGS
   GLOBAL_HEADER_GEN ON
   SOURCES ${async_cuda_sources}
   HEADERS ${async_cuda_headers}
