+++ libs/config/cmake/templates/config_version.hpp.in
@@ -1,10 +1,12 @@
-//  Copyright (c) 2007-2017 Hartmut Kaiser
+//  Copyright (c) 2007-2020 Hartmut Kaiser
 //  Copyright (c)      2011 Bryce Lelbach
 //  Copyright (c)      2013 Adrian Serio
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+// Do not edit this file! It has been generated by the cmake configuration step.
 
 // Make HPX inspect tool happy: hpxinspect:nounnamed
 
