+++ examples/distributed_stencil/1d_stencil.cpp
@@ -1,0 +1,169 @@
+// Copyright (c) 2020 Nikunj Gupta
+//
+// Distributed under the Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include "communicator.hpp"
+#include "stencil.hpp"
+
+#include <hpx/hpx_init.hpp>
+#include <hpx/include/compute.hpp>
+#include <hpx/include/lcos.hpp>
+#include <hpx/include/async.hpp>
+#include <hpx/include/components.hpp>
+#include <hpx/include/parallel_algorithm.hpp>
+#include <hpx/include/util.hpp>
+#include <hpx/program_options/options_description.hpp>
+
+#include <array>
+#include <algorithm>
+#include <vector>
+#include <iostream>
+#include <string>
+
+using communication_type = double;
+
+HPX_REGISTER_CHANNEL_DECLARATION(communication_type);
+HPX_REGISTER_CHANNEL(communication_type, stencil_communication);
+
+int hpx_main(boost::program_options::variables_map& vm)
+{
+    std::size_t Nx_global = vm["Nx"].as<std::size_t>();
+    std::size_t steps = vm["steps"].as<std::size_t>();
+    std::size_t nlp = vm["Nlp"].as<std::size_t>();
+
+    typedef hpx::compute::host::block_allocator<double> allocator_type;
+    typedef hpx::compute::host::block_executor<> executor_type;
+    typedef hpx::compute::vector<double, allocator_type> data_type;
+    
+    std::array<data_type, 2> U;
+
+    auto numa_domains = hpx::compute::host::numa_domains();
+    allocator_type alloc(numa_domains);
+
+    std::size_t num_localities = hpx::get_num_localities(hpx::launch::sync);
+    std::size_t num_worker_threads = hpx::get_num_worker_threads();
+    std::size_t rank = hpx::get_locality_id();
+
+    hpx::util::high_resolution_timer t_main;
+
+    // Keep only partial data
+    std::size_t Nx = Nx_global / num_localities;
+    std::size_t local_nx = Nx / nlp;
+
+    U[0] = data_type(Nx, 0.0, alloc);
+    U[1] = data_type(Nx, 0.0, alloc);
+
+    init(U, Nx, rank, num_localities);
+
+    // setup communicator
+    using communicator_type = communicator<double>;
+    communicator_type comm(rank, num_localities);
+
+    if (rank == 0)
+    {
+        std::cout << "Starting benchmark with " << num_localities << 
+                     " nodes and " << num_worker_threads <<
+                     " threads.\n";
+    }
+
+    if (comm.has_neighbor(communicator_type::left))
+    {
+        // Send initial value to the left neighbor
+        comm.set(communicator_type::left, U[0][0], 0);
+    }
+    if (comm.has_neighbor(communicator_type::right))
+    {
+        // Send initial value to the right neighbor
+        comm.set(communicator_type::right, U[0][Nx-1], 0);
+    }
+    
+    auto range = boost::irange(static_cast<std::size_t>(0), nlp);
+
+    executor_type executor(numa_domains);
+    auto policy = hpx::parallel::execution::par.on(executor);
+
+    hpx::util::high_resolution_timer t;
+    for (std::size_t t = 0; t < steps; ++t)
+    {
+        data_type& curr = U[t % 2];
+        data_type& next = U[(t + 1) % 2];
+
+        hpx::parallel::for_each(
+            policy,
+            std::begin(range), std::end(range),
+            [&U, local_nx, nlp, t] (std::size_t i)
+            {
[…]
