+++ examples/distributed_stencil/CMakeLists.txt
@@ -1,0 +1,32 @@
+# Copyright (c) 2020 Nikunj Gupta
+#
+# SPDX-License-Identifier: BSL-1.0
+# Distributed under the Boost Software License, Version 1.0. (See accompanying
+# file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+if(NOT HPX_WITH_DISTRIBUTED_RUNTIME)
+  return()
+endif()
+
+set(example_programs 1d_stencil)
+
+foreach(example_program ${example_programs})
+  set(sources ${example_program}.cpp)
+
+  source_group("Source Files" FILES ${sources})
+
+  # add example executable
+  add_hpx_executable(
+    ${example_program} INTERNAL_FLAGS
+    SOURCES ${sources} ${${example_program}_FLAGS}
+    FOLDER "Examples/1dStencil"
+  )
+
+  add_hpx_example_target_dependencies("1d_stencil" ${example_program})
+
+  if(HPX_WITH_TESTS AND HPX_WITH_TESTS_EXAMPLES)
+    add_hpx_example_test(
+      "1d_stencil" ${example_program} ${${example_program}_PARAMETERS}
+    )
+  endif()
+endforeach()
