### 2020-06-18 e959e9c
Add folder to example cmake
