### 2020-06-19 8e87ce4
Fix compilation with HPX_WITH_NETWORKING=OFF
