### 2020-06-19 21ee2dd
Make internal algorithms functions const
Kokkos expects functions in parallel regions to be const. This satisfies
Kokkos' requirements when calling HPX's algorithm functions in Kokkos parallel
regions.
