### 2020-06-19 9c580de
Reduce running time of some tests
