### 2020-06-19 de47e01
Make most collectives tests run with only one locality as well
