### 2020-06-19 e81dbec
Add Jenkins configuration for Piz Daint
