### 2020-06-19 #2352 @diehlpk
@danielskatz We fixed all issues with the DOI reported by whedon.
