+++ libs/coroutines/include/hpx/coroutines/detail/coroutine_stackful_self.hpp
@@ -81,6 +81,19 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
             return pimpl_->set_thread_data(data);
         }
 
+#ifdef HPX_HAVE_LIBCDS
+        std::size_t get_libcds_data() const override
+        {
+            HPX_ASSERT(pimpl_);
+            return pimpl_->get_libcds_data();
+        }
+        std::size_t set_libcds_data(std::size_t data) override
+        {
+            HPX_ASSERT(pimpl_);
+            return pimpl_->set_libcds_data(data);
+        }
+#endif
+
         tss_storage* get_thread_tss_data() override
         {
 #if defined(HPX_HAVE_THREAD_LOCAL_STORAGE)
