+++ libs/coroutines/include/hpx/coroutines/detail/context_base.hpp
@@ -85,6 +85,9 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
           , m_thread_data(nullptr)
 #else
           , m_thread_data(0)
+#   ifdef HPX_HAVE_LIBCDS
+          , libcds_thread_data(0)
+#   endif
 #endif
           , m_type_info()
           , m_thread_id(id)
@@ -98,6 +101,9 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
             delete_tss_storage(m_thread_data);
 #else
             m_thread_data = 0;
+#   ifdef HPX_HAVE_LIBCDS
+            libcds_thread_data = 0;
+#   endif
 #endif
         }
 
@@ -204,6 +210,9 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
             delete_tss_storage(m_thread_data);
 #else
             m_thread_data = 0;
+#   ifdef HPX_HAVE_LIBCDS
+            libcds_thread_data = 0;
+#   endif
 #endif
         }
 
@@ -228,6 +237,20 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
             return olddata;
 #endif
         }
+
+#ifdef HPX_HAVE_LIBCDS
+        std::size_t get_libcds_data() const
+        {
+            return libcds_thread_data;
+        }
+
+        std::size_t set_libcds_data(std::size_t data)
+        {
+            std::size_t olddata = libcds_thread_data;
+            libcds_thread_data = data;
+            return olddata;
+        }
+#endif
 
 #if defined(HPX_HAVE_THREAD_LOCAL_STORAGE)
         tss_storage* get_thread_tss_data(bool create_if_needed) const
@@ -284,6 +307,9 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
             HPX_ASSERT(m_thread_data == nullptr);
 #else
             HPX_ASSERT(m_thread_data == 0);
+#endif
+#ifdef HPX_HAVE_LIBCDS
+
 #endif
             // NOLINTNEXTLINE(bugprone-throw-keyword-missing)
             m_type_info = std::exception_ptr();
@@ -345,6 +371,9 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
         mutable detail::tss_storage* m_thread_data;
 #else
         mutable std::size_t m_thread_data;
+#   ifdef HPX_HAVE_LIBCDS
+        mutable std::size_t libcds_thread_data;
+#   endif
 #endif
 
         // This is used to generate a meaningful exception trace.
