+++ libs/threading_base/include/hpx/threading_base/thread_helpers.hpp
@@ -392,6 +392,14 @@ namespace hpx { namespace threads {
     HPX_EXPORT std::size_t set_thread_data(
         thread_id_type const& id, std::size_t data, error_code& ec = throws);
 
+#ifdef HPX_HAVE_LIBCDS
+    HPX_EXPORT std::size_t get_libcds_data(
+        thread_id_type const& id, error_code& ec = throws);
+
+    HPX_EXPORT std::size_t set_libcds_data(
+        thread_id_type const& id, std::size_t data, error_code& ec = throws);
+#endif
+
     HPX_EXPORT std::size_t& get_continuation_recursion_count();
     HPX_EXPORT void reset_continuation_recursion_count();
     /// \endcond
