+++ libs/threading/include/hpx/threading/thread.hpp
@@ -120,6 +120,11 @@ namespace hpx {
         std::size_t get_thread_data() const;
         std::size_t set_thread_data(std::size_t);
 
+#ifdef HPX_HAVE_LIBCDS
+        std::size_t get_libcds_data() const;
+        std::size_t set_libcds_data(std::size_t);
+#endif
+
     private:
         bool joinable_locked() const noexcept
         {
@@ -255,6 +260,11 @@ namespace hpx {
 
         HPX_EXPORT std::size_t get_thread_data();
         HPX_EXPORT std::size_t set_thread_data(std::size_t);
+
+#ifdef HPX_HAVE_LIBCDS
+        HPX_EXPORT std::size_t get_libcds_data();
+        HPX_EXPORT std::size_t set_libcds_data(std::size_t);
+#endif
 
         class HPX_EXPORT disable_interruption
         {

[…]
