### 2020-06-19 7f3b4ec
Minor cleanup of checkpoint functionality
- adding more access helper functions
