### 2020-06-20 40c25f0
Adding hpx::util::prepare_checkpoint API
- More cleanup of checkpointing functionality
- Adding testing for new API
