+++ libs/checkpoint/include/hpx/checkpoint/checkpoint.hpp
@@ -21,6 +21,7 @@
 #include <hpx/runtime/components/new.hpp>
 #include <hpx/runtime/get_ptr.hpp>
 #include <hpx/runtime/naming_fwd.hpp>
+#include <hpx/runtime/serialization/detail/preprocess_container.hpp>
 #include <hpx/serialization/serialize.hpp>
 #include <hpx/serialization/vector.hpp>
 #include <hpx/traits/is_client.hpp>
@@ -58,6 +59,7 @@ namespace hpx { namespace util {
     /// serialized including components.
     class checkpoint
     {
+    private:
         std::vector<char> data_;
 
         friend std::ostream& operator<<(
@@ -79,6 +81,9 @@ namespace hpx { namespace util {
 
         template <typename T, typename... Ts>
         friend void restore_checkpoint(checkpoint const& c, T& t, Ts&... ts);
+
+        template <typename T, typename... Ts>
+        friend checkpoint prepare_checkpoint(T const& t, Ts const&... ts);
 
     public:
         checkpoint() = default;
@@ -245,7 +250,7 @@ namespace hpx { namespace util {
 
                 // Trick to expand the variable pack, takes advantage of the
                 // comma operator.
-                int const sequencer[] = {0, (ar << ts, 0)...};
+                int const sequencer[] = {0, (ar << prep(ts), 0)...};
                 (void) sequencer;    // Suppress unused param. warnings
 
                 return std::move(c);
@@ -295,8 +300,7 @@ namespace hpx { namespace util {
     hpx::future<checkpoint> save_checkpoint(T&& t, Ts&&... ts)
     {
         return hpx::dataflow(detail::save_funct_obj{}, checkpoint{},
-            detail::prep(std::forward<T>(t)),
-            detail::prep(std::forward<Ts>(ts))...);
+            std::forward<T>(t), std::forward<Ts>(ts)...);
     }
 
     ///////////////////////////////////////////////////////////////////////////
@@ -337,8 +341,7 @@ namespace hpx { namespace util {
     hpx::future<checkpoint> save_checkpoint(checkpoint&& c, T&& t, Ts&&... ts)
     {
         return hpx::dataflow(detail::save_funct_obj{}, std::move(c),
-            detail::prep(std::forward<T>(t)),
-            detail::prep(std::forward<Ts>(ts))...);
+            std::forward<T>(t), std::forward<Ts>(ts)...);
     }
 
     ///////////////////////////////////////////////////////////////////////////
@@ -380,8 +383,7 @@ namespace hpx { namespace util {
     hpx::future<checkpoint> save_checkpoint(hpx::launch p, T&& t, Ts&&... ts)
     {
         return hpx::dataflow(p, detail::save_funct_obj{}, checkpoint{},
-            detail::prep(std::forward<T>(t)),
-            detail::prep(std::forward<Ts>(ts))...);
+            std::forward<T>(t), std::forward<Ts>(ts)...);
     }
 
     ///////////////////////////////////////////////////////////////////////////
@@ -427,8 +429,7 @@ namespace hpx { namespace util {
         hpx::launch p, checkpoint&& c, T&& t, Ts&&... ts)
     {
         return hpx::dataflow(p, detail::save_funct_obj{}, std::move(c),
-            detail::prep(std::forward<T>(t)),
-            detail::prep(std::forward<Ts>(ts))...);
+            std::forward<T>(t), std::forward<Ts>(ts)...);
     }
 
     ///////////////////////////////////////////////////////////////////////////
@@ -475,8 +476,7 @@ namespace hpx { namespace util {
     {
         hpx::future<checkpoint> f_chk =
[…]
