+++ libs/threading_base/include/hpx/threading_base/thread_data_stackful.hpp
@@ -93,12 +93,12 @@ namespace hpx { namespace threads {
         }
 
 #ifdef HPX_HAVE_LIBCDS
-        std::size_t get_libcds_data() const override
+        std::vector<std::size_t> get_libcds_data() const override
         {
             return coroutine_.get_libcds_data();
         }
 
-        std::size_t set_libcds_data(std::size_t data) override
+        std::vector<std::size_t> set_libcds_data(std::vector<std::size_t> data) override
         {
             return coroutine_.set_libcds_data(data);
         }
