+++ libs/coroutines/include/hpx/coroutines/detail/coroutine_self.hpp
@@ -130,8 +130,8 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
         virtual std::size_t set_thread_data(std::size_t data) = 0;
 
 #ifdef HPX_HAVE_LIBCDS
-        virtual std::size_t get_libcds_data() const = 0;
-        virtual std::size_t set_libcds_data(std::size_t data) = 0;
+        virtual std::vector<std::size_t> get_libcds_data() const = 0;
+        virtual std::vector<std::size_t> set_libcds_data(std::vector<std::size_t> data) = 0;
 #endif
 
         virtual tss_storage* get_thread_tss_data() = 0;
