### 2020-06-20 75d5be7
Adding nullary versions of checkpointing APIs
