+++ libs/checkpoint/include/hpx/checkpoint/checkpoint.hpp
@@ -45,7 +45,10 @@ namespace hpx { namespace util {
 
     namespace detail {
         struct save_funct_obj;
-    }
+
+        template <typename T, typename... Ts>
+        checkpoint prepare(checkpoint&& c, T const& t, Ts const&... ts);
+    }    // namespace detail
 
     ///////////////////////////////////////////////////////////////////////////
     /// Checkpoint Object
@@ -83,7 +86,8 @@ namespace hpx { namespace util {
         friend void restore_checkpoint(checkpoint const& c, T& t, Ts&... ts);
 
         template <typename T, typename... Ts>
-        friend checkpoint prepare_checkpoint(T const& t, Ts const&... ts);
+        friend checkpoint detail::prepare(
+            checkpoint&& c, T const& t, Ts const&... ts);
 
     public:
         checkpoint() = default;
@@ -303,6 +307,14 @@ namespace hpx { namespace util {
             std::forward<T>(t), std::forward<Ts>(ts)...);
     }
 
+    /// \cond NOINTERNAL
+    // Same as above, just nullary
+    inline hpx::future<checkpoint> save_checkpoint()
+    {
+        return hpx::make_ready_future(checkpoint{});
+    }
+    /// \endcond
+
     ///////////////////////////////////////////////////////////////////////////
     /// Save_checkpoint - Take a pre-initialized checkpoint
     ///
@@ -336,13 +348,21 @@ namespace hpx { namespace util {
     ///          exception: if you pass hpx::launch::sync as the first
     ///          argument. In this case save_checkpoint will simply return
     ///          a checkpoint.
-
+    ///
     template <typename T, typename... Ts>
     hpx::future<checkpoint> save_checkpoint(checkpoint&& c, T&& t, Ts&&... ts)
     {
         return hpx::dataflow(detail::save_funct_obj{}, std::move(c),
             std::forward<T>(t), std::forward<Ts>(ts)...);
     }
+
+    /// \cond NOINTERNAL
+    // Same as above, just nullary
+    inline hpx::future<checkpoint> save_checkpoint(checkpoint&& c)
+    {
+        return hpx::make_ready_future(std::move(c));
+    }
+    /// \endcond
 
     ///////////////////////////////////////////////////////////////////////////
     /// Save_checkpoint - Policy overload
@@ -378,13 +398,21 @@ namespace hpx { namespace util {
     ///          exception: if you pass hpx::launch::sync as the first
     ///          argument. In this case save_checkpoint will simply return
     ///          a checkpoint.
-
+    ///
     template <typename T, typename... Ts>
     hpx::future<checkpoint> save_checkpoint(hpx::launch p, T&& t, Ts&&... ts)
     {
         return hpx::dataflow(p, detail::save_funct_obj{}, checkpoint{},
             std::forward<T>(t), std::forward<Ts>(ts)...);
     }
+
+    /// \cond NOINTERNAL
+    // Same as above, just nullary
+    inline hpx::future<checkpoint> save_checkpoint(hpx::launch p)
+    {
+        return hpx::make_ready_future(checkpoint{});
+    }
+    /// \endcond
 
     ///////////////////////////////////////////////////////////////////////////
     /// Save_checkpoint - Policy overload & pre-initialized checkpoint
@@ -423,7 +451,7 @@ namespace hpx { namespace util {
     ///          exception: if you pass hpx::launch::sync as the first
     ///          argument. In this case save_checkpoint will simply return
     ///          a checkpoint.
-
+    ///
     template <typename T, typename... Ts>
     hpx::future<checkpoint> save_checkpoint(
[…]
