### 2020-06-20 58b10cf
Move preprocess_container to serialization module
- more cleaning up for checkpointing facilities
