+++ libs/checkpoint/include/hpx/checkpoint/checkpoint.hpp
@@ -21,7 +21,7 @@
 #include <hpx/runtime/components/new.hpp>
 #include <hpx/runtime/get_ptr.hpp>
 #include <hpx/runtime/naming_fwd.hpp>
-#include <hpx/runtime/serialization/detail/preprocess_container.hpp>
+#include <hpx/serialization/detail/preprocess_container.hpp>
 #include <hpx/serialization/serialize.hpp>
 #include <hpx/serialization/vector.hpp>
 #include <hpx/traits/is_client.hpp>
@@ -45,9 +45,7 @@ namespace hpx { namespace util {
 
     namespace detail {
         struct save_funct_obj;
-
-        template <typename T, typename... Ts>
-        checkpoint prepare(checkpoint&& c, T const& t, Ts const&... ts);
+        struct prepare_checkpoint;
     }    // namespace detail
 
     ///////////////////////////////////////////////////////////////////////////
@@ -81,13 +79,10 @@ namespace hpx { namespace util {
         }
 
         friend struct detail::save_funct_obj;
+        friend struct detail::prepare_checkpoint;
 
         template <typename T, typename... Ts>
         friend void restore_checkpoint(checkpoint const& c, T& t, Ts&... ts);
-
-        template <typename T, typename... Ts>
-        friend checkpoint detail::prepare(
-            checkpoint&& c, T const& t, Ts const&... ts);
 
     public:
         checkpoint() = default;
@@ -221,7 +216,7 @@ namespace hpx { namespace util {
         template <typename T,
             typename U = typename std::enable_if<!hpx::traits::is_client<
                 typename std::decay<T>::type>::value>::type>
-        T&& prep(T&& t) noexcept
+        T&& prepare_client(T&& t) noexcept
         {
             return std::forward<T>(t);
         }
@@ -230,7 +225,7 @@ namespace hpx { namespace util {
         template <typename Client, typename Server>
         hpx::future<std::shared_ptr<typename hpx::components::client_base<
             Client, Server>::server_component_type>>
-        prep(hpx::components::client_base<Client, Server> const& c)
+        prepare_client(hpx::components::client_base<Client, Server> const& c)
         {
             // Use shared pointer to serialize server
             using server_type = typename hpx::components::client_base<Client,
@@ -254,7 +249,7 @@ namespace hpx { namespace util {
 
                 // Trick to expand the variable pack, takes advantage of the
                 // comma operator.
-                int const sequencer[] = {0, (ar << prep(ts), 0)...};
+                int const sequencer[] = {0, (ar << ts, 0)...};
                 (void) sequencer;    // Suppress unused param. warnings
 
                 return std::move(c);
@@ -304,7 +299,8 @@ namespace hpx { namespace util {
     hpx::future<checkpoint> save_checkpoint(T&& t, Ts&&... ts)
     {
         return hpx::dataflow(detail::save_funct_obj{}, checkpoint{},
-            std::forward<T>(t), std::forward<Ts>(ts)...);
+            detail::prepare_client(std::forward<T>(t)),
+            detail::prepare_client(std::forward<Ts>(ts))...);
     }
 
     /// \cond NOINTERNAL
@@ -353,7 +349,8 @@ namespace hpx { namespace util {
     hpx::future<checkpoint> save_checkpoint(checkpoint&& c, T&& t, Ts&&... ts)
     {
         return hpx::dataflow(detail::save_funct_obj{}, std::move(c),
-            std::forward<T>(t), std::forward<Ts>(ts)...);
+            detail::prepare_client(std::forward<T>(t)),
+            detail::prepare_client(std::forward<Ts>(ts))...);
     }
 
     /// \cond NOINTERNAL
@@ -399,11 +396,14 @@ namespace hpx { namespace util {
[…]
