### 2020-06-21 8fb80d6
Complete functional version of is_sentinel_for
