### 2020-06-21 f021ebc
Add missing includes in distance header and test and fix clang formatting
