### 2020-06-21 07957a7
Splitting checkpoint module into checkpoint_base and checkpoint
- the checkpoint_base module exposes the basic checkpointing functionality
- the checkpoint module is built on top of checkpoint_base and additionally
  supports asynchrony, futures, and component-client
