### 2020-06-21 95b6b21
Add is_sized_sentinel_for test
