### 2020-06-22 34c1705
Add debug::scope entry/exit, fix const correctness, ip-addr print
