+++ libs/debugging/include/hpx/debugging/print.hpp
@@ -11,8 +11,10 @@
 #include <array>
 #include <bitset>
 #include <chrono>
+#include <cmath>
 #include <cstddef>
 #include <cstdint>
+#include <cstring>
 #include <iomanip>
 #include <iostream>
 #include <iterator>
@@ -23,8 +25,9 @@
 #include <vector>
 
 #if defined(__linux) || defined(linux) || defined(__linux__)
-#include <linux/unistd.h>
+#include <unistd.h>
 #include <sys/mman.h>
+extern char **environ;
 #define DEBUGGING_PRINT_LINUX
 #endif
 
@@ -223,15 +226,21 @@ namespace hpx { namespace debug {
     {
         ipaddr(const void* a)
           : data_(reinterpret_cast<const uint8_t*>(a))
+          , ipdata_(0)
+        {
+        }
+        ipaddr(const uint32_t a)
+          : data_(reinterpret_cast<const uint8_t*>(&ipdata_))
+          , ipdata_(a)
         {
         }
         const uint8_t* data_;
+        const uint32_t ipdata_;
+
         friend std::ostream& operator<<(std::ostream& os, const ipaddr& p)
         {
-            os << std::dec << int((reinterpret_cast<const uint8_t*>(&p))[0])
-               << "." << int((reinterpret_cast<const uint8_t*>(&p))[1]) << "."
-               << int((reinterpret_cast<const uint8_t*>(&p))[2]) << "."
-               << int((reinterpret_cast<const uint8_t*>(&p))[3]);
+            os << std::dec << int(p.data_[0]) << "." << int(p.data_[1]) << "."
+               << int(p.data_[2]) << "." << int(p.data_[3]);
             return os;
         }
     };
@@ -239,7 +248,7 @@ namespace hpx { namespace debug {
     // ------------------------------------------------------------------
     // helper fuction for printing CRC32
     // ------------------------------------------------------------------
-    inline uint32_t crc32(const void* address, size_t length)
+    inline uint32_t crc32(const void* /*address*/, size_t /*length*/)
     {
         //        boost::crc_32_type result;
         //        result.process_bytes(address, length);
@@ -249,7 +258,7 @@ namespace hpx { namespace debug {
 
     // ------------------------------------------------------------------
     // helper fuction for printing short memory dump and crc32
-    // useful for debugging corruptions in buffers during parcelport
+    // useful for debugging corruptions in buffers during
     // rma or other transfers
     // ------------------------------------------------------------------
     struct mem_crc32
@@ -268,9 +277,12 @@ namespace hpx { namespace debug {
             const uint64_t* uintBuf = static_cast<const uint64_t*>(p.addr_);
             os << "Memory:";
             os << " address " << hpx::debug::ptr(p.addr_) << " length "
-               << hpx::debug::hex<8>(p.len_) << " CRC32:\n"
-               << hpx::debug::hex<8>(crc32(p.addr_, p.len_)) << " ";
-            for (size_t i = 0; i < (std::min)(p.len_ / 8, size_t(128)); i++)
+               << hpx::debug::hex<6>(p.len_)
+               << " CRC32:" << hpx::debug::hex<8>(crc32(p.addr_, p.len_))
+               << "\n";
+            for (size_t i = 0;
+                 i < (std::min)(size_t(std::ceil(p.len_ / 8.0)), size_t(128));
+                 i++)
             {
                 os << hpx::debug::hex<16>(*uintBuf++) << " ";
             }
@@ -281,7 +293,7 @@ namespace hpx { namespace debug {
 
     namespace detail {
 
-#ifdef HPX_HAVE_CXX17_FOLD_EXPRESSIONS
[…]
