### 2020-06-22 b4a0542
Add debug hostname_print_helper
