### 2020-06-22 e9ca43f
Add functions and classes to the public API listing in documentation
