### 2020-06-22 43e66b2
clang-format and inspect violations
