### 2020-06-22 9fa36a3
Remove compatibility headers in /detail
