+++ libs/execution/CMakeLists.txt
@@ -110,7 +110,6 @@ add_hpx_module(
     hpx_threadmanager
     hpx_thread_support
     hpx_timing
-    hpx_topology
     hpx_type_support
     hpx_pack_traversal
   CMAKE_SUBDIRS examples tests
