### 2020-06-23 be52269
Fix RUN_SERIAL handling for verbs parcelport
