+++ cmake/HPX_AddTest.cmake
@@ -113,10 +113,7 @@ function(add_hpx_test category name)
       endif()
       if(_add_test)
         set(_full_name "${category}.distributed.verbs.${name}")
-        add_test(NAME "${_full_name}"
-                 COMMAND ${cmd} "-p" "verbs" ${args} set_tests_properties
-                         ("${_full_name}" PROPERTIES RUN_SERIAL TRUE)
-        )
+        add_test(NAME "${_full_name}" COMMAND ${cmd} "-p" "verbs" ${args})
         set_tests_properties("${_full_name}" PROPERTIES RUN_SERIAL TRUE)
       endif()
     endif()
