### 2020-06-23 7ed8a5a
Clean up CircleCI config
