+++ .circleci/config.yml
@@ -361,7 +361,11 @@ jobs:
       - attach_workspace:
           at: /hpx
       - run:
-          name: Running Unit Tests
+          name: Building Examples
+          command: |
+              ninja -j2 -k 0 examples
+      - run:
+          name: Running Example Tests
           when: always
           command: |
               ulimit -c unlimited
@@ -381,389 +385,27 @@ jobs:
           path: tests.examples
       - store_artifacts:
           path: tests.examples
+      - persist_to_workspace:
+          root: /hpx
+          paths:
+            - ./build
 
-  tests.unit.actions:
+  tests.unit:
     <<: *defaults
     steps:
       - attach_workspace:
           at: /hpx
       - run:
           name: Building Unit Tests
-          command: |
-              ninja -j2 -k 0 tests.unit.actions
-      - run:
-          name: Running Unit Tests
-          when: always
-          command: |
-              ulimit -c unlimited
-              ctest --timeout 60 -T test --no-compress-output --output-on-failure -R tests.unit.actions
-      - run:
-          <<: *convert_xml
-      - run:
-          <<: *move_core_dump
-      - run:
-          <<: *move_debug_log
-      - store_test_results:
-          path: tests.unit.actions
-      - store_artifacts:
-          path: tests.unit.actions
-
-  tests.unit.agas:
-    <<: *defaults
-    steps:
-      - attach_workspace:
-          at: /hpx
-      - run:
-          name: Building Unit Tests
-          command: |
-              ninja -j2 -k 0 tests.unit.agas
-      - run:
-          name: Running Unit Tests
-          when: always
-          command: |
-              ulimit -c unlimited
-              ctest --timeout 60 -T test --no-compress-output --output-on-failure -R tests.unit.agas
-      - run:
-          <<: *convert_xml
-      - run:
-          <<: *move_core_dump
-      - run:
-          <<: *move_debug_log
-      - store_test_results:
-          path: tests.unit.agas
-      - store_artifacts:
-          path: tests.unit.agas
-
-  tests.unit.build:
-    <<: *defaults
-    steps:
-      - attach_workspace:
-          at: /hpx
-      - run:
-          name: Installing HPX for build unit tests
-          command: |
-              # Disable documentation just for this step
-              cmake \
-              -DHPX_WITH_DOCUMENTATION=Off .
-              ninja -j1 install
-      - run:
-          name: Building Unit Tests
-          command: |
-              ninja -j2 -k 0 tests.unit.build
-      - run:
-          name: Running Unit Tests
-          when: always
-          command: |
-              ulimit -c unlimited
-              ctest --timeout 60 -T test --no-compress-output --output-on-failure -R tests.unit.build
-      - run:
-          <<: *convert_xml
-      - run:
-          <<: *move_core_dump
-      - run:
-          <<: *move_debug_log
-      - store_test_results:
-          path: tests.unit.build
-      - store_artifacts:
-          path: tests.unit.build
-
-  tests.unit.component:
-    <<: *defaults
-    steps:
-      - attach_workspace:
-          at: /hpx
-      - run:
-          name: Building Unit Tests
-          command: |
-              ninja -j2 -k 0 tests.unit.component tests.unit.components
-      - run:
-          name: Running Unit Tests
-          when: always
-          command: |
-              ulimit -c unlimited
-              ctest --timeout 60 -T test --no-compress-output --output-on-failure -R tests.unit.component
-      - run:
-          <<: *convert_xml
-      - run:
-          <<: *move_core_dump
-      - run:
-          <<: *move_debug_log
-      - store_test_results:
-          path: tests.unit.component
-      - store_artifacts:
-          path: tests.unit.component
-
-  tests.unit.lcos:
-    <<: *defaults
-    steps:
-      - attach_workspace:
-          at: /hpx
-      - run:
[…]
