### 2020-06-23 #2352 @davidbeckingsale
@diehlpk - I don't feel that the paper addresses the following review point: "State of the field: Do the authors describe how this software compares to other commonly-used packages?"

Could you please add a brief overview of related work and compare hpx to these other packages?
