### 2020-06-24 0625cef
Clean up cxx14 version of variadic printing
