### 2020-06-24 5c1f7ed
Convert West const to const East format
To convert files changed relative to stellar/master branch
git diff --name-only stellar/master | xargs -n 1 \
    sed -i -E 's/const ([a-zA-Z0-9_<>,:]+)&/\1 const\&/g'

NB. the character set includes alphanumeric as well as "_<>,:"
to allow const std::vector<std::vector, allocator> expressions
to be correctly detected. Other might be needed for more complex
parameter lists.
