+++ libs/debugging/include/hpx/debugging/print.hpp
@@ -115,7 +115,7 @@ namespace hpx { namespace debug {
         {
         }
         void const* data_;
-        friend std::ostream& operator<<(std::ostream& os, const ptr& d)
+        friend std::ostream& operator<<(std::ostream& os, ptr const& d)
         {
             os << d.data_;
             return os;
@@ -134,11 +134,11 @@ namespace hpx { namespace debug {
         struct hex<N, T,
             typename std::enable_if<!std::is_pointer<T>::value>::type>
         {
-            hex(const T& v)
+            hex(T const& v)
               : data_(v)
             {
             }
-            const T& data_;
+            T const& data_;
             friend std::ostream& operator<<(
                 std::ostream& os, const hex<N, T>& d)
             {
@@ -152,11 +152,11 @@ namespace hpx { namespace debug {
         struct hex<N, T,
             typename std::enable_if<std::is_pointer<T>::value>::type>
         {
-            hex(const T& v)
+            hex(T const& v)
               : data_(v)
             {
             }
-            const T& data_;
+            T const& data_;
             friend std::ostream& operator<<(
                 std::ostream& os, const hex<N, T>& d)
             {
@@ -181,11 +181,11 @@ namespace hpx { namespace debug {
         template <int N = 8, typename T = int>
         struct bin
         {
-            bin(const T& v)
+            bin(T const& v)
               : data_(v)
             {
             }
-            const T& data_;
+            T const& data_;
             friend std::ostream& operator<<(
                 std::ostream& os, const bin<N, T>& d)
             {
@@ -212,7 +212,7 @@ namespace hpx { namespace debug {
         {
         }
         const char* data_;
-        friend std::ostream& operator<<(std::ostream& os, const str<N>& d)
+        friend std::ostream& operator<<(std::ostream& os, str<N> const& d)
         {
             os << std::left << std::setfill(' ') << std::setw(N) << d.data_;
             return os;
@@ -237,7 +237,7 @@ namespace hpx { namespace debug {
         const uint8_t* data_;
         const uint32_t ipdata_;
 
-        friend std::ostream& operator<<(std::ostream& os, const ipaddr& p)
+        friend std::ostream& operator<<(std::ostream& os, ipaddr const& p)
         {
             os << std::dec << int(p.data_[0]) << "." << int(p.data_[1]) << "."
                << int(p.data_[2]) << "." << int(p.data_[3]);
@@ -272,7 +272,7 @@ namespace hpx { namespace debug {
         const uint64_t* addr_;
         const std::size_t len_;
         const char* txt_;
-        friend std::ostream& operator<<(std::ostream& os, const mem_crc32& p)
+        friend std::ostream& operator<<(std::ostream& os, mem_crc32 const& p)
         {
             const uint64_t* uintBuf = static_cast<const uint64_t*>(p.addr_);
             os << "Memory:";
@@ -296,7 +296,7 @@ namespace hpx { namespace debug {
 #if defined(HPX_HAVE_CXX17_FOLD_EXPRESSIONS)
         template <typename TupleType, std::size_t... I>
         void tuple_print(
-            std::ostream& os, const TupleType& t, std::index_sequence<I...>)
+            std::ostream& os, TupleType const& t, std::index_sequence<I...>)
         {
             (..., (os << (I == 0 ? "" : " ") << std::get<I>(t)));
         }
@@ -312,7 +312,7 @@ namespace hpx { namespace debug {
         // helper function to print a tuple of any size
         template <typename TupleType, std::size_t... I>
         void tuple_print(
-            std::ostream& os, const TupleType& t, std::index_sequence<I...>)
+            std::ostream& os, TupleType const& t, std::index_sequence<I...>)
         {
             using expander = int[];
[…]
