### 2020-06-24 c0fb1d2
Fix typo in init_ini_data.cpp
