### 2020-06-24 ba53078
Add default value for getting exception_verbosity config entry
