### 2020-06-24 b7add99
Add coveralls badge
