### 2020-06-24 7f30a32
Fixing deleter argument bindings in serialize_buffer
- flyby: touching up on 1d_stencil_7 example
