### 2020-06-25 c017899
use std::array to maintain libcds data instead of vector
