### 2020-06-25 15b5df2
Fixing inclusive transform scan algorithm to properly handle initial value
