### 2020-06-26 5f43cc0
Get correct runtime pointer (i.e. distributed) in interface.cpp
The operations after the check for the runtime pointer depend on the
distributed runtime pointer being available, not just the local one. This
particular failure was triggered in the runtime_distributed constructor when
the local runtime TSS pointer has been set, but the distributed one has not
been set.
