### 2020-06-26 3e2f8ae
Ignore internal deprecation warnings from result_of etc.
