### 2020-06-26 f694ad2
Fixing argument sequence for secondary sweep in transform_scan
