+++ libs/algorithms/include/hpx/parallel/algorithms/transform_exclusive_scan.hpp
@@ -133,7 +133,7 @@ namespace hpx { namespace parallel { inline namespace v1 {
 
                     util::loop_n<ExPolicy>(
                         dst, part_size - 1, [&op, &val](FwdIter2 it) -> void {
-                            *it = hpx::util::invoke(op, *it, val);
+                            *it = hpx::util::invoke(op, val, *it);
                         });
                 };
 
