### 2020-06-26 9dbb6c8
Apply small formatting and language changes to hpxcxx documentation
