+++ docs/sphinx/manual/creating_hpx_projects.rst
@@ -299,26 +299,31 @@ variable to point to the installation of |hpx|.
 
 .. _hpxcxx_documentation:
 
-Building |hpx| components and applications
-------------------------------------------
+Using the |hpx| compiler wrapper ``hpxcxx``
+-------------------------------------------
 
-The ``hpxcxx`` compiler wrapper helps to build a |hpx| component, an application or generate a ``<file-name>.o`` file, based on the arguments passed to it.
+The ``hpxcxx`` compiler wrapper helps to compile a |hpx| component, application,
+or object file, based on the arguments passed to it.
 
-.. code-block :: bash
+.. code-block:: bash
 
    hpxcxx [--exe=<APPLICATION_NAME> | --comp=<COMPONENT_NAME> | -c] FLAGS FILES
 
-The ``hpxcxx`` command **requires** that either an application or a component is built or ``-c`` flag is specified. If the build is against a debug build, the ``-g`` is to be specified while building.
+The ``hpxcxx`` command **requires** that either an application or a component is
+built or ``-c`` flag is specified. If the build is against a debug build, the
+``-g`` is to be specified while building.
 
 Optional ``FLAGS``
 ..................
 
 * ``-l <LIBRARY> | -l<LIBRARY>``: Links ``<LIBRARY>`` to the build
-* ``-g``: Specifies that the application or component build is against a debug build
+* ``-g``: Specifies that the application or component build is against a debug
+  build
 * ``-rd``: Sets ``release-with-debug-info`` option
 * ``-mr``: Sets ``minsize-release`` option
 
-All other flags (like ``-o OUTPUT_FILE``) are directly passed to the underlying C++ compiler.
+All other flags (like ``-o OUTPUT_FILE``) are directly passed to the underlying
+C++ compiler.
  
 .. _cmake_integrate_hpx:
 
