### 2020-06-26 6cd4ef6
Create codeql-analysis.yml
