### 2020-06-26 65f584b
Replace enable_if with HPX_CONCEPT_REQUIRES_ and add is_sentinel_for constraint
