### 2020-06-26 baaaa76
Add lazy eval function for function return arguments / expressions
When function returns are passed to debug print statements
it may be necessary to prevent functions being called if their
return values are not used/needed

This may be more concise that using blocks of the kind
if (debug_thing.is_enabled() {
    ...
}
