### 2020-06-26 66c7dc0
Add unit test for debug::print
