### 2020-06-26 162b17e
Split up main library into smaller libraries
