+++ libs/parallelism/futures/src/future_data.cpp
@@ -1,0 +1,315 @@
+//  Copyright (c) 2015 Hartmut Kaiser
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include <hpx/futures/detail/future_data.hpp>
+#include <hpx/futures/future.hpp>
+
+#include <hpx/config.hpp>
+#include <hpx/assert.hpp>
+#include <hpx/async_base/launch_policy.hpp>
+#include <hpx/execution_base/this_thread.hpp>
+#include <hpx/functional/deferred_call.hpp>
+#include <hpx/functional/unique_function.hpp>
+#include <hpx/futures/futures_factory.hpp>
+#include <hpx/modules/errors.hpp>
+#include <hpx/modules/memory.hpp>
+#include <hpx/threading_base/annotated_function.hpp>
+
+#include <cstddef>
+#include <exception>
+#include <functional>
+#include <mutex>
+#include <utility>
+
+namespace hpx { namespace lcos { namespace detail {
+    static run_on_completed_error_handler_type run_on_completed_error_handler;
+
+    void set_run_on_completed_error_handler(
+        run_on_completed_error_handler_type f)
+    {
+        run_on_completed_error_handler = f;
+    }
+
+    future_data_refcnt_base::~future_data_refcnt_base() = default;
+
+    ///////////////////////////////////////////////////////////////////////////
+    struct handle_continuation_recursion_count
+    {
+        handle_continuation_recursion_count()
+          : count_(threads::get_continuation_recursion_count())
+        {
+            ++count_;
+        }
+        ~handle_continuation_recursion_count()
+        {
+            --count_;
+        }
+
+        std::size_t& count_;
+    };
+
+    ///////////////////////////////////////////////////////////////////////////
+    template <typename Callback>
+    static void run_on_completed_on_new_thread(Callback&& f)
+    {
+        lcos::local::futures_factory<void()> p(std::forward<Callback>(f));
+
+        bool is_hpx_thread = nullptr != hpx::threads::get_self_ptr();
+        hpx::launch policy = launch::fork;
+        if (!is_hpx_thread)
+            policy = launch::async;
+
+        // launch a new thread executing the given function
+        threads::thread_id_type tid = p.apply("run_on_completed_on_new_thread",
+            policy, threads::thread_priority_boost,
+            threads::thread_stacksize_current, threads::thread_schedule_hint());
+
+        // wait for the task to run
+        if (is_hpx_thread)
+        {
+            // make sure this thread is executed last
+            this_thread::suspend(threads::pending, std::move(tid));
+            return p.get_future().get();
+        }
+        // If we are not on a HPX thread, we need to return immediately, to
+        // allow the newly spawned thread to execute.
+    }
+
+    ///////////////////////////////////////////////////////////////////////////
+    future_data_base<traits::detail::future_data_void>::~future_data_base() {}
+
+    static util::unused_type unused_;
+
+    util::unused_type*
+    future_data_base<traits::detail::future_data_void>::get_result_void(
+        void const* storage, error_code& ec)
+    {
+        // yields control if needed
+        state s = wait(ec);
+        if (ec)
+            return nullptr;
+
+        // No locking is required. Once a future has been made ready, which
+        // is a postcondition of wait, either:
+        //
+        // - there is only one writer (future), or
+        // - there are multiple readers only (shared_future, lock hurts
+        //   concurrency)
+
+        // Avoid retrieving state twice. If wait() returns 'empty' then this
+        // thread was suspended, in this case we need to load it again.
+        if (s == empty)
[…]
