### 2020-06-26 fa62c6e
Further optimizations
