### 2020-06-26 #2352 @diehlpk
@davidbeckingsale

We highlighted the review paper

> Thoman, P., Dichev, K., Heller, T. et al. A taxonomy of task-based parallel programming technologies for high-performance > computing. J Supercomput 74, 1422–1434 (2018). https://doi.org/10.1007/s11227-018-2238-4

which is a very nice overview of AMTS. In addition, we added a paragraph to paper and highlighted that the main features of HPX is the distributed computing capability with some other notable AMTS. Among this, the future-proof C++ standard conforming API is the major difference.
