### 2020-06-26 55526a3
Fixing inclusive_scan if no initial value was supplied
- flyby: remove predicate from local inclusive_scan/exclusive_scan
  implementation
- adding feature test for std::inclusive_scan/std::exclusive_scan
- adding tests comparing with std algorithms (if available)
