+++ libs/algorithms/include/hpx/parallel/algorithms/transform_exclusive_scan.hpp
@@ -168,7 +168,7 @@ namespace hpx { namespace parallel { inline namespace v1 {
             typename T, typename Op, typename Conv>
         typename util::detail::algorithm_result<ExPolicy, FwdIter2>::type
         transform_exclusive_scan_(ExPolicy&& policy, FwdIter1 first,
-            FwdIter1 last, FwdIter2 dest, Conv&& conv, T init, Op&& op,
+            FwdIter1 last, FwdIter2 dest, Conv&& conv, T&& init, Op&& op,
             std::false_type)
         {
             static_assert((hpx::traits::is_forward_iterator<FwdIter1>::value),
@@ -180,7 +180,7 @@ namespace hpx { namespace parallel { inline namespace v1 {
 
             return detail::transform_exclusive_scan<FwdIter2>().call(
                 std::forward<ExPolicy>(policy), is_seq(), first, last, dest,
-                std::forward<Conv>(conv), std::move(init),
+                std::forward<Conv>(conv), std::forward<T>(init),
                 std::forward<Op>(op));
         }
 
@@ -188,7 +188,7 @@ namespace hpx { namespace parallel { inline namespace v1 {
             typename T, typename Op, typename Conv>
         typename util::detail::algorithm_result<ExPolicy, FwdIter2>::type
         transform_exclusive_scan_(ExPolicy&& policy, FwdIter1 first,
-            FwdIter1 last, FwdIter2 dest, Conv&& conv, T init, Op&& op,
+            FwdIter1 last, FwdIter2 dest, Conv&& conv, T&& init, Op&& op,
             std::true_type);
         /// \endcond
     }    // namespace detail
