### 2020-06-26 add00e9
Replace obsolete debug::init with declare_variable/set
Variables that are not needed when debug::print is <false>
are replaced with constexpr bool and should be almost
zero cost
