### 2020-06-26 f23b851
Fix bad CMake after module name change
