### 2020-06-26 3d334a9
Fix a bad #include when MPI is not built
