### 2020-06-26 582175c
Rename macro to HPX_DP_LAZY
