### 2020-06-26 73b51c0
Add support for mac/windows to environment checks
