### 2020-06-27 10e123e
Updating README
Fixing documentation links, adding Coveralls to the text
