### 2020-06-27 1176314
Try to fix compilation errors on mac/windows and change set signature
