### 2020-06-27 7874865
clang-format,inspect and #include violations
