+++ libs/debugging/tests/unit/print.cpp
@@ -9,21 +9,24 @@
 // are completely elided
 
 #include <hpx/debugging/print.hpp>
-#include <hpx/threading_base/print.hpp>
-#include <hpx/util/high_resolution_timer.hpp>
 #include <hpx/modules/testing.hpp>
+#include <hpx/threading_base/print.hpp>
+#include <hpx/chrono.hpp>
 
-#include <string>
+#include <atomic>
 #include <chrono>
+#include <iostream>
+#include <string>
 
 namespace hpx {
     // this is an enabled debug object that should output messages
-    static hpx::debug::enable_print<true>  p_enabled("PRINT  ");
+    static hpx::debug::enable_print<true> p_enabled("PRINT  ");
     // this is disabled and we want it to have zero footprint
     static hpx::debug::enable_print<false> p_disabled("PRINT  ");
-}   // namespace hpx
+}    // namespace hpx
 
-int increment(std::atomic<int> &counter) {
+int increment(std::atomic<int>& counter)
+{
     return ++counter;
 }
 
@@ -44,44 +47,49 @@ int main()
     HPX_TEST_EQ(enabled_counter, 1);
 
     // we expect the counter to increment as LAZY will be evaluated
-    p_enabled.debug("Increment", HPX_DP_LAZY(increment(enabled_counter), p_enabled));
+    p_enabled.debug(
+        "Increment", HPX_DP_LAZY(increment(enabled_counter), p_enabled));
     HPX_TEST_EQ(enabled_counter, 2);
 
     // we do not expect the counter to increment
-    if (p_disabled.is_enabled()) {
+    if (p_disabled.is_enabled())
+    {
         p_disabled.debug("Increment", increment(disabled_counter));
     }
     HPX_TEST_EQ(disabled_counter, 0);
 
     // we do not expect the counter to increment: HPX_DP_LAZY will not be evaluated
-    p_disabled.debug("Increment", HPX_DP_LAZY(increment(disabled_counter), p_disabled));
+    p_disabled.debug(
+        "Increment", HPX_DP_LAZY(increment(disabled_counter), p_disabled));
     HPX_TEST_EQ(disabled_counter, 0);
-
 
     // ---------------------------------------------------------
     // Test that scoped log messages behave as expected
     {
-        auto s_enabled  = p_enabled.scope("scoped block", HPX_DP_LAZY(increment(enabled_counter), p_enabled));
-        auto s_disabled = p_disabled.scope("scoped block", HPX_DP_LAZY(increment(disabled_counter), p_disabled));
-        (void)s_disabled; // silence warning about unused var
+        auto s_enabled = p_enabled.scope(
+            "scoped block", HPX_DP_LAZY(increment(enabled_counter), p_enabled));
+        auto s_disabled = p_disabled.scope("scoped block",
+            HPX_DP_LAZY(increment(disabled_counter), p_disabled));
+        (void) s_disabled;    // silence warning about unused var
     }
     HPX_TEST_EQ(enabled_counter, 3);
     HPX_TEST_EQ(disabled_counter, 0);
 
-
     // ---------------------------------------------------------
     // Test that debug only variables behave as expected
     // create high resolution timers to see if they count
-    auto var1 = p_enabled.declare_variable<int>(HPX_DP_LAZY(enabled_counter+4, p_enabled));
-    (void)var1; // silenced unused var when optimized out
+    auto var1 = p_enabled.declare_variable<int>(
[…]
