### 2020-06-29 cb4cbda
Add CUDA flag to async_cuda CMakeLists
