### 2020-06-29 cc4bcb8
Removing 128bit integer emulation
