### 2020-06-30 bfc0f17
Make hpx default target without wrap_main
Add the hpx_wrap_main target where it is needed. This change
addresses the linking problems with googletest main.
