+++ src/CMakeLists.txt
@@ -536,54 +536,51 @@ set(_library_types
 set(_is_executable "$<STREQUAL:$<TARGET_PROPERTY:TYPE>,EXECUTABLE>")
 set(_is_library "$<IN_LIST:$<TARGET_PROPERTY:TYPE>,${_library_types}>")
 
-add_library(hpx_no_wrap_main INTERFACE)
 add_library(hpx INTERFACE)
+add_library(wrap_main INTERFACE)
 
-target_link_libraries(hpx_no_wrap_main INTERFACE hpx_core)
 target_link_libraries(hpx INTERFACE hpx_core)
 
-# hpx_interface and hpx_interface_no_wrap_main contain additional interface
-# options to be passed to dependent targets. We create these as separate targets
-# to easily filter out the generator expressions that can't be handled by the
+# hpx_interface and hpx_interface_wrap_main contain additional interface options
+# to be passed to dependent targets. We create these as separate targets to
+# easily filter out the generator expressions that can't be handled by the
 # pkgconfig file generation.
-add_library(hpx_interface_no_wrap_main INTERFACE)
+add_library(hpx_interface INTERFACE)
 target_link_libraries(
-  hpx_interface_no_wrap_main
+  hpx_interface
   INTERFACE $<${_is_executable}:$<TARGET_NAME_IF_EXISTS:hpx_init>>
 )
 target_link_libraries(
-  hpx_interface_no_wrap_main
+  hpx_interface
   INTERFACE $<${_is_executable}:$<TARGET_NAME_IF_EXISTS:HPXInternal::hpx_init>>
 )
 target_compile_definitions(
-  hpx_interface_no_wrap_main
+  hpx_interface
   INTERFACE
     "$<${_is_executable}:HPX_APPLICATION_NAME_DEFAULT=$<TARGET_PROPERTY:NAME>>"
 )
 target_compile_definitions(
-  hpx_interface_no_wrap_main
+  hpx_interface
   INTERFACE "$<${_is_executable}:HPX_PREFIX_DEFAULT=\"${HPX_PREFIX}\">"
 )
 target_compile_definitions(
-  hpx_interface_no_wrap_main
-  INTERFACE "$<${_is_executable}:HPX_APPLICATION_EXPORTS>"
+  hpx_interface INTERFACE "$<${_is_executable}:HPX_APPLICATION_EXPORTS>"
 )
 target_compile_definitions(
-  hpx_interface_no_wrap_main INTERFACE "$<${_is_library}:HPX_LIBRARY_EXPORTS>"
+  hpx_interface INTERFACE "$<${_is_library}:HPX_LIBRARY_EXPORTS>"
 )
 
-add_library(hpx_interface INTERFACE)
-target_link_libraries(hpx_interface INTERFACE hpx_interface_no_wrap_main)
+add_library(hpx_interface_wrap_main INTERFACE)
 target_link_libraries(
-  hpx_interface
+  hpx_interface_wrap_main
   INTERFACE $<${_is_executable}:$<TARGET_NAME_IF_EXISTS:hpx_wrap>>
 )
 target_link_libraries(
-  hpx_interface
+  hpx_interface_wrap_main
   INTERFACE $<${_is_executable}:$<TARGET_NAME_IF_EXISTS:HPXInternal::hpx_wrap>>
 )
 
-target_link_libraries(hpx_no_wrap_main INTERFACE hpx_interface_no_wrap_main)
+target_link_libraries(wrap_main INTERFACE hpx_interface_wrap_main)
 target_link_libraries(hpx INTERFACE hpx_interface)
 
 # HPX::component is to be linked privately to all HPX components NOTE: The
[…]
