+++ libs/futures/include_compatibility/hpx/lcos/local/packaged_continuation.hpp
@@ -8,15 +8,15 @@
 
 #include <hpx/config.hpp>
 #include <hpx/futures/config/defines.hpp>
-#include <hpx/future.hpp>
+#include <hpx/futures/packaged_continuation.hpp>
 
 #if HPX_FUTURES_HAVE_DEPRECATION_WARNINGS
 #if defined(HPX_MSVC)
 #pragma message(                                                               \
     "The header hpx/lcos/local/packaged_continuation.hpp is deprecated, \
-    please include hpx/future.hpp instead")
+    please include hpx/futures/packaged_continuation.hpp instead")
 #else
 #warning "The header hpx/lcos/local/packaged_continuation.hpp is deprecated, \
-    please include hpx/future.hpp instead"
+    please include hpx/futures/packaged_continuation.hpp instead"
 #endif
 #endif
