+++ libs/async_combinators/include_compatibility/hpx/lcos/wait_all.hpp
@@ -8,14 +8,14 @@
 
 #include <hpx/config.hpp>
 #include <hpx/async_combinators/config/defines.hpp>
-#include <hpx/future.hpp>
+#include <hpx/async_combinators/wait_all.hpp>
 
 #if HPX_ASYNC_COMBINATORS_HAVE_DEPRECATION_WARNINGS
 #if defined(HPX_MSVC)
 #pragma message("The header hpx/lcos/wait_all.hpp is deprecated, \
-    please include hpx/future.hpp instead")
+    please include hpx/async_combinators/wait_all.hpp instead")
 #else
 #warning "The header hpx/lcos/wait_all.hpp is deprecated, \
-    please include hpx/future.hpp instead"
+    please include hpx/async_combinators/wait_all.hpp instead"
 #endif
 #endif
