+++ libs/futures/include_compatibility/hpx/traits/get_remote_result.hpp
@@ -8,14 +8,14 @@
 
 #include <hpx/config.hpp>
 #include <hpx/futures/config/defines.hpp>
-#include <hpx/future.hpp>
+#include <hpx/futures/traits/get_remote_result.hpp>
 
 #if HPX_FUTURES_HAVE_DEPRECATION_WARNINGS
 #if defined(HPX_MSVC)
 #pragma message("The header hpx/traits/get_remote_result.hpp is deprecated, \
-    please include hpx/future.hpp instead")
+    please include hpx/futures/traits/get_remote_result.hpp instead")
 #else
 #warning "The header hpx/traits/get_remote_result.hpp is deprecated, \
-    please include hpx/future.hpp instead"
+    please include hpx/futures/traits/get_remote_result.hpp instead"
 #endif
 #endif
