+++ libs/futures/include_compatibility/hpx/traits/future_then_result.hpp
@@ -8,14 +8,18 @@
 
 #include <hpx/config.hpp>
 #include <hpx/futures/config/defines.hpp>
-#include <hpx/future.hpp>
+#include <hpx/futures/traits/future_then_result.hpp>
 
+// different versions of clang-format do different things
+// clang-format off
 #if HPX_FUTURES_HAVE_DEPRECATION_WARNINGS
 #if defined(HPX_MSVC)
-#pragma message("The header hpx/traits/future_then_result.hpp is deprecated, \
-    please include hpx/future.hpp instead")
+#pragma message(                                                               \
+    "The header hpx/traits/future_then_result.hpp is deprecated, \
+    please include hpx/futures/traits/future_then_result.hpp instead")
 #else
 #warning "The header hpx/traits/future_then_result.hpp is deprecated, \
-    please include hpx/future.hpp instead"
+    please include hpx/futures/traits/future_then_result.hpp instead"
 #endif
 #endif
+// clang-format on
