+++ libs/futures/include_compatibility/hpx/traits/promise_local_result.hpp
@@ -8,15 +8,15 @@
 
 #include <hpx/config.hpp>
 #include <hpx/futures/config/defines.hpp>
-#include <hpx/future.hpp>
+#include <hpx/futures/traits/promise_local_result.hpp>
 
 #if HPX_FUTURES_HAVE_DEPRECATION_WARNINGS
 #if defined(HPX_MSVC)
 #pragma message(                                                               \
     "The header hpx/traits/promise_local_result.hpp is deprecated, \
-    please include hpx/future.hpp instead")
+    please include hpx/futures/traits/promise_local_result.hpp instead")
 #else
 #warning "The header hpx/traits/promise_local_result.hpp is deprecated, \
-    please include hpx/future.hpp instead"
+    please include hpx/futures/traits/promise_local_result.hpp instead"
 #endif
 #endif
