+++ libs/actions/include/hpx/actions/transfer_base_action.hpp
@@ -1,0 +1,359 @@
+//  Copyright (c) 2007-2017 Hartmut Kaiser
+//  Copyright (c)      2011 Bryce Lelbach
+//  Copyright (c) 2011-2016 Thomas Heller
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+/// \file transfer_action.hpp
+
+#pragma once
+
+#include <hpx/config/defines.hpp>
+
+#if defined(HPX_HAVE_NETWORKING)
+#include <hpx/actions/action_support.hpp>
+#include <hpx/actions/actions_fwd.hpp>
+#include <hpx/actions/base_action.hpp>
+#include <hpx/actions_base/traits/action_priority.hpp>
+#include <hpx/actions_base/traits/action_stacksize.hpp>
+#include <hpx/assert.hpp>
+#include <hpx/components_base/pinned_ptr.hpp>
+#include <hpx/datastructures/tuple.hpp>
+#include <hpx/runtime/actions/detail/invocation_count_registry.hpp>
+#include <hpx/serialization/input_archive.hpp>
+#include <hpx/serialization/output_archive.hpp>
+#include <hpx/serialization/unique_ptr.hpp>
+#include <hpx/traits/action_does_termination_detection.hpp>
+#include <hpx/traits/action_message_handler.hpp>
+#include <hpx/traits/action_schedule_thread.hpp>
+#include <hpx/traits/action_serialization_filter.hpp>
+#include <hpx/traits/action_was_object_migrated.hpp>
+#include <hpx/util/get_and_reset_value.hpp>
+#include <hpx/util/serialize_exception.hpp>
+#if HPX_HAVE_ITTNOTIFY != 0 && !defined(HPX_HAVE_APEX)
+#include <hpx/modules/itt_notify.hpp>
+#endif
+
+#include <atomic>
+#include <cstddef>
+#include <cstdint>
+#include <memory>
+#include <type_traits>
+#include <utility>
+
+namespace hpx { namespace actions {
+    ///////////////////////////////////////////////////////////////////////////
+    // If one or more arguments of the action are non-default-constructible,
+    // the transfer_action does not store the argument tuple directly but a
+    // unique_ptr to the tuple instead.
+    namespace detail {
+        template <typename Args>
+        struct argument_holder
+        {
+            argument_holder() = default;
+
+            explicit argument_holder(Args&& args)
+              : data_(new Args(std::move(args)))
+            {
+            }
+
+            template <typename... Ts>
+            argument_holder(Ts&&... ts)
+              : data_(new Args(std::forward<Ts>(ts)...))
+            {
+            }
+
+            template <typename Archive>
+            void serialize(Archive& ar, unsigned int const)
+            {
+                // clang-format off
+                ar & data_;
+                // clang-format on
+            }
+
+            HPX_HOST_DEVICE HPX_FORCEINLINE Args& data()
+            {
+                HPX_ASSERT(!!data_);
+                return *data_;
+            }
+
+#if defined(HPX_DISABLE_ASSERTS) || defined(BOOST_DISABLE_ASSERTS) ||          \
+    defined(NDEBUG)
+            constexpr HPX_HOST_DEVICE HPX_FORCEINLINE Args const& data() const
+            {
+                return *data_;
+            }
+#else
+            HPX_HOST_DEVICE HPX_FORCEINLINE Args const& data() const
+            {
+                HPX_ASSERT(!!data_);
+                return *data_;
+            }
+#endif
+
+        private:
[…]

[…]
