### 2020-07-02 443fca4
Adapt other headers to include/ and modules/ headers
Add hpx/modules/string_util.hpp in hpx/include/util.hpp
