+++ examples/quickstart/fibonacci_dataflow.cpp
@@ -15,8 +15,6 @@
 #include <hpx/include/lcos.hpp>
 #include <hpx/include/util.hpp>
 
-#include <hpx/pack_traversal/unwrap.hpp>
-
 #include <cstdint>
 #include <iostream>
 #include <utility>
