+++ examples/quickstart/fibonacci_futures_distributed.cpp
@@ -9,7 +9,6 @@
 #include <hpx/include/lcos.hpp>
 #include <hpx/include/runtime.hpp>
 #include <hpx/include/util.hpp>
-#include <hpx/pack_traversal/unwrap.hpp>
 
 #include <atomic>
 #include <cstddef>
