+++ examples/future_reduce/rnd_future_reduce.cpp
@@ -9,7 +9,9 @@
 #include <hpx/future.hpp>
 #include <hpx/include/runtime.hpp>
 #include <hpx/distributed/iostream.hpp>
+#include <hpx/include/runtime.hpp>
 #include <hpx/modules/timing.hpp>
+#include <hpx/include/util.hpp>
 //
 #include <random>
 #include <utility>
