+++ examples/accumulators/accumulator_client.cpp
@@ -7,10 +7,7 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/hpx_init.hpp>
-#include <hpx/string_util/split.hpp>
-#include <hpx/string_util/trim.hpp>
-#include <hpx/string_util/classification.hpp>
-#include <hpx/util/from_string.hpp>
+#include <hpx/include/util.hpp>
 
 #include "accumulator.hpp"
 
