+++ examples/quickstart/executor_with_thread_hooks.cpp
@@ -15,7 +15,6 @@
 #include <hpx/modules/algorithms.hpp>
 #include <hpx/assert.hpp>
 #include <hpx/execution.hpp>
-#include <hpx/modules/functional.hpp>
 
 #include <algorithm>
 #include <atomic>
