### 2020-07-02 2dee708
Replace boost::shared_array<T> with std::shared_ptr<T[]> if supported
