+++ libs/serialization/include/hpx/serialization/serialize_buffer.hpp
@@ -15,7 +15,11 @@
 #include <hpx/serialization/serialization_fwd.hpp>
 #include <hpx/serialization/serialize.hpp>
 
+#if !defined(HPX_HAVE_CXX17_SHARED_PTR_ARRAY)
 #include <boost/shared_array.hpp>
+#else
+#include <memory>
+#endif
 
 #include <algorithm>
 #include <cstddef>
@@ -29,6 +33,11 @@ namespace hpx { namespace serialization {
     {
     private:
         using allocator_type = Allocator;
+#if defined(HPX_HAVE_CXX17_SHARED_PTR_ARRAY)
+        using buffer_type = std::shared_ptr<T[]>;
+#else
+        using buffer_type = boost::shared_array<T>;
+#endif
 
         static void no_deleter(T*) {}
 
@@ -89,13 +98,12 @@ namespace hpx { namespace serialization {
             }
             else if (mode == reference)
             {
-                data_ =
-                    boost::shared_array<T>(data, &serialize_buffer::no_deleter);
+                data_ = buffer_type(data, &serialize_buffer::no_deleter);
             }
             else
             {
                 // take ownership
-                data_ = boost::shared_array<T>(
+                data_ = buffer_type(
                     data, [alloc = this->alloc_, size = this->size_](T* p) {
                         serialize_buffer::deleter<allocator_type>(
                             p, alloc, size);
@@ -111,7 +119,7 @@ namespace hpx { namespace serialization {
           , alloc_(alloc)
         {
             // if 2 allocators are specified we assume mode 'take'
-            data_ = boost::shared_array<T>(data, [this, dealloc](T* p) {
+            data_ = buffer_type(data, [this, dealloc](T* p) {
                 serialize_buffer::deleter<Deallocator>(p, dealloc, size_);
             });
         }
@@ -133,7 +141,7 @@ namespace hpx { namespace serialization {
             else
             {
                 // reference or take ownership, behavior is defined by deleter
-                data_ = boost::shared_array<T>(data, deleter);
+                data_ = buffer_type(data, deleter);
             }
         }
 
@@ -154,7 +162,7 @@ namespace hpx { namespace serialization {
             }
             else if (mode == reference)
             {
-                data_ = boost::shared_array<T>(const_cast<T*>(data), deleter);
+                data_ = buffer_type(const_cast<T*>(data), deleter);
             }
             else
             {
@@ -174,7 +182,7 @@ namespace hpx { namespace serialization {
           , alloc_(alloc)
         {
             // if 2 allocators are specified we assume mode 'take'
-            data_ = boost::shared_array<T>(data, deleter);
+            data_ = buffer_type(data, deleter);
         }
 
         // same set of constructors, but taking const data
@@ -225,7 +233,7 @@ namespace hpx { namespace serialization {
             }
             else if (mode == reference)
             {
-                data_ = boost::shared_array<T>(
+                data_ = buffer_type(
                     const_cast<T*>(data), &serialize_buffer::no_deleter);
             }
             else
@@ -265,7 +273,7 @@ namespace hpx { namespace serialization {
             return data_[idx];
         }
 
-        boost::shared_array<T> data_array() const
+        buffer_type data_array() const
         {
             return data_;
         }
@@ -320,7 +328,7 @@ namespace hpx { namespace serialization {
         }
 
     private:
-        boost::shared_array<T> data_;
+        buffer_type data_;
         std::size_t size_;
         Allocator alloc_;
     };
