+++ examples/1d_stencil/1d_stencil_8.cpp
@@ -12,7 +12,7 @@
 #include <hpx/hpx_init.hpp>
 #include <hpx/hpx.hpp>
 #include <hpx/modules/collectives.hpp>
-#include <hpx/serialization/serialize.hpp>
+#include <hpx/serialization.hpp>
 #include <hpx/type_support/unused.hpp>
 
 #include <boost/shared_array.hpp>
