### 2020-07-02 d425040
Fixing inspect problem that slipped through
