+++ libs/algorithms/include/hpx/parallel/datapar/iterator_helpers.hpp
@@ -9,9 +9,9 @@
 #include <hpx/config.hpp>
 
 #if defined(HPX_HAVE_DATAPAR)
-#include <hpx/execution/parallel/traits/vector_pack_alignment_size.hpp>
-#include <hpx/execution/parallel/traits/vector_pack_load_store.hpp>
-#include <hpx/execution/parallel/traits/vector_pack_type.hpp>
+#include <hpx/execution/traits/vector_pack_alignment_size.hpp>
+#include <hpx/execution/traits/vector_pack_load_store.hpp>
+#include <hpx/execution/traits/vector_pack_type.hpp>
 #include <hpx/functional/invoke_result.hpp>
 
 #include <cstddef>
