### 2020-07-03 136361c
Change inspect tool for intrusive_ptr
+ silence inspect tool for from_string
