+++ examples/accumulators/accumulator_client.cpp
@@ -6,6 +6,8 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
+// hpxinspect:noinclude:hpx::util::from_string
+
 #include <hpx/hpx_init.hpp>
 #include <hpx/include/util.hpp>
 
