+++ libs/collectives/src/detail/barrier_node.cpp
@@ -10,9 +10,9 @@
 #include <hpx/collectives/detail/barrier_node.hpp>
 #include <hpx/futures/future.hpp>
 #include <hpx/lcos_local/promise.hpp>
-#include <hpx/memory/intrusive_ptr.hpp>
 #include <hpx/modules/async_distributed.hpp>
 #include <hpx/modules/execution.hpp>
+#include <hpx/modules/memory.hpp>
 #include <hpx/pack_traversal/unwrap.hpp>
 #include <hpx/runtime/components/component_type.hpp>
 #include <hpx/util/from_string.hpp>
