+++ hpx/runtime/naming/id_type.hpp
@@ -7,7 +7,7 @@
 #pragma once
 
 #include <hpx/config.hpp>
-#include <hpx/memory/intrusive_ptr.hpp>
+#include <hpx/modules/memory.hpp>
 #include <hpx/runtime/naming_fwd.hpp>
 #include <hpx/serialization/serialization_fwd.hpp>
 
