### 2020-07-03 69c19d9
Add a test emulating gtest to avoid linking to googletest
