### 2020-07-04 023791a
Adding Francisco Jose Tapia's implementation of stable_sort
