+++ libs/algorithms/tests/unit/algorithms/util/test_merge_vector.cpp
@@ -1,0 +1,286 @@
+//  Copyright (c) 2015-2017 Francisco Jose Tapia
+//  Copyright (c) 2020 Hartmut Kaiser
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include <hpx/hpx_main.hpp>
+#include <hpx/modules/testing.hpp>
+#include <hpx/parallel/util/merge_vector.hpp>
+
+#include <algorithm>
+#include <cstdint>
+#include <type_traits>
+#include <utility>
+#include <vector>
+
+using namespace hpx::parallel::util;
+
+// template <typename Iter, typename Sent>
+// std::ostream& operator<<(std::ostream& out, range<Iter, Sent> R)
+// {
+//     out << "[ " << (R.last - R.first) << "] ";
+//     if (not R.valid())
+//         return out;
+//     while (R.first != R.last)
+//         out << (*(R.first++)) << " ";
+//     out << std::endl;
+//     return out;
+// }
+
+struct xk
+{
+    unsigned tail : 3;
+    unsigned num : 24;
+
+    xk(unsigned N = 0, unsigned T = 0)
+      : tail(T)
+      , num(N)
+    {
+    }
+
+    bool operator<(xk A) const
+    {
+        return (unsigned) num < (unsigned) A.num;
+    }
+};
+
+// std::ostream& operator<<(std::ostream& out, xk x)
+// {
+//     out << "[" << x.num << "-" << x.tail << "] ";
+//     return out;
+// }
+
+// TEST OF  MERGE_LEVEL4
+void test1()
+{
+    std::uint64_t X[10][10], Y[100];
+    range<std::uint64_t*> RY(&Y[0], &Y[100]);
+
+    for (std::uint32_t i = 0; i < 4; ++i)
+    {
+        for (std::uint32_t k = 0; k < 10; ++k)
+            X[i][k] = i + k * 4;
+    }
+    for (std::uint32_t i = 0; i < 3; ++i)
+    {
+        for (std::uint32_t k = 0; k < 10; ++k)
+            X[i + 4][k] = i + 40 + k * 3;
+    }
+    for (std::uint32_t i = 0; i < 3; ++i)
+    {
+        for (std::uint32_t k = 0; k < 10; ++k)
+            X[i + 7][k] = i + 70 + k * 3;
+    }
+
+    for (std::uint32_t i = 0; i < 100; ++i)
+        Y[i] = 1000;
+    std::vector<range<std::uint64_t*>> V, Z;
+    for (std::uint32_t i = 0; i < 10; ++i)
+        V.emplace_back(&X[i][0], &X[i][10]);
+    merge_level4(RY, V, Z, std::less<std::uint64_t>());
+
+    for (std::uint32_t i = 0; i < 100; ++i)
+        HPX_TEST(Y[i] == i);
+    HPX_TEST(Z.size() == 3);
+}
+
+void test2()
+{
+    typedef typename std::vector<xk>::iterator iter_t;
+    typedef range<iter_t> rng;
+    typedef std::less<xk> compare;
+
+    std::vector<xk> VA, VB;
+    VB.resize(90);
+    rng RB(VB.begin(), VB.end());
+
+    for (std::uint32_t i = 0; i < 30; ++i)
+    {
+        std::uint32_t k = i % 10;
+        VA.emplace_back(k, i / 10);
+    }
[…]
