+++ libs/algorithms/include/hpx/parallel/util/range.hpp
@@ -1,0 +1,324 @@
+//  Copyright (c) 2015-2017 Francisco Jose Tapia
+//  Copyright (c) 2020 Hartmut Kaiser
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#pragma once
+
+#include <hpx/modules/iterator_support.hpp>
+#include <hpx/parallel/algorithms/detail/distance.hpp>
+#include <hpx/parallel/util/low_level.hpp>
+
+#include <functional>
+#include <iterator>
+#include <memory>
+#include <type_traits>
+#include <utility>
+#include <vector>
+
+namespace hpx { namespace parallel { namespace util {
+
+    /// \struct range
+    /// \brief this represent a range between two iterators
+    /// \tparam Iter type of parameters of the range
+    /// \remarks
+    template <typename Iterator, typename Sentinel = Iterator>
+    using range = hpx::util::iterator_range<Iterator, Sentinel>;
+
+    /// \brief concatenate two contiguous ranges
+    /// \param [in] it1 : first range
+    /// \param [in] it2 : second range
+    /// \return  range resulting of the concatenation
+    template <typename Iter, typename Sent>
+    range<Iter, Sent> concat(
+        range<Iter, Sent> const& it1, range<Iter, Sent> const& it2)
+    {
+        return range<Iter, Sent>(it1.begin(), it2.end());
+    }
+
+    /// \brief Move objects from the range src to dest
+    /// \param [in] dest : range where move the objects
+    /// \param [in] src : range from where move the objects
+    /// \return range with the objects moved and the size adjusted
+    template <typename Iter1, typename Sent1, typename Iter2, typename Sent2>
+    inline range<Iter2, Iter2> init_move(
+        range<Iter2, Sent2> const& dest, range<Iter1, Sent1> const& src)
+    {
+        using type1 = typename std::iterator_traits<Iter1>::value_type;
+        using type2 = typename std::iterator_traits<Iter2>::value_type;
+
+        static_assert(
+            std::is_same<type1, type2>::value, "Incompatible iterators\n");
+
+        if (src.size() == 0)
+        {
+            return range<Iter2, Iter2>(dest.begin(), dest.begin());
+        }
+
+        init_move(dest.begin(), src.begin(), src.end());
+        return range<Iter2, Iter2>(
+            dest.begin(), std::next(dest.begin(), src.size()));
+    }
+
+    //-----------------------------------------------------------------------------
+    //  function : uninit_move
+    /// \brief Move objects from the range src creating them in  dest
+    /// \param [in] dest : range where move and create the objects
+    /// \param [in] src : range from where move the objects
+    /// \return range with the objects moved and the size adjusted
+    //-----------------------------------------------------------------------------
+    template <typename Iter1, typename Sent1, typename Iter2, typename Sent2>
+    inline range<Iter2, Sent2> uninit_move(
+        range<Iter2, Sent2> const& dest, range<Iter1, Sent1> const& src)
+    {
+        using type1 = typename std::iterator_traits<Iter1>::value_type;
+        using type2 = typename std::iterator_traits<Iter2>::value_type;
+
+        static_assert(
+            std::is_same<type1, type2>::value, "Incompatible iterators\n");
+
+        if (src.size() == 0)
+        {
+            return range<Iter2, Iter2>(dest.begin(), dest.begin());
+        }
+
+        uninit_move(dest.begin(), src.begin(), src.end());
+        return range<Iter2, Iter2>(
+            dest.begin(), std::next(dest.begin(), src.size()));
+    }
+
+    //  function : destroy
[…]

[…]
