### 2020-07-04 43f0e22
for_each adaptation
