+++ libs/algorithms/include/hpx/parallel/algorithms/for_each.hpp
@@ -26,6 +26,7 @@
 #include <hpx/execution/algorithms/detail/is_negative.hpp>
 #include <hpx/executors/execution_policy.hpp>
 #include <hpx/parallel/algorithms/detail/dispatch.hpp>
+#include <hpx/parallel/algorithms/detail/distance.hpp>
 #include <hpx/parallel/util/detail/algorithm_result.hpp>
 #include <hpx/parallel/util/foreach_partitioner.hpp>
 #include <hpx/parallel/util/loop.hpp>
@@ -359,25 +360,25 @@ namespace hpx { namespace parallel { inline namespace v1 {
             {
             }
 
-            template <typename ExPolicy, typename InIter, typename F,
+            template <typename ExPolicy, typename InIterB,typename InIterE, typename F,
                 typename Proj>
             static typename std::enable_if<
-                hpx::traits::is_random_access_iterator<InIter>::value,
-                InIter>::type
-            sequential(ExPolicy&& policy, InIter first, InIter last, F&& f,
+                hpx::traits::is_random_access_iterator<InIterB>::value,
+                InIterB>::type
+            sequential(ExPolicy&& policy, InIterB first, InIterE last, F&& f,
                 Proj&& proj)
             {
                 return util::loop_n<typename std::decay<ExPolicy>::type>(first,
-                    static_cast<std::size_t>(std::distance(first, last)),
+                    static_cast<std::size_t>(detail::distance(first, last)),
                     invoke_projected<F, Proj>{f, proj});
             }
 
-            template <typename ExPolicy, typename InIter, typename F,
+            template <typename ExPolicy, typename InIterB, typename InIterE, typename F,
                 typename Proj>
             static typename std::enable_if<
-                !hpx::traits::is_random_access_iterator<InIter>::value,
-                InIter>::type
-            sequential(ExPolicy&& policy, InIter first, InIter last, F&& f,
+                !hpx::traits::is_random_access_iterator<InIterB>::value,
+                InIterB>::type
+            sequential(ExPolicy&& policy, InIterB first, InIterE last, F&& f,
                 Proj&& proj)
             {
                 return util::loop(std::forward<ExPolicy>(policy), first, last,
@@ -398,7 +399,7 @@ namespace hpx { namespace parallel { inline namespace v1 {
 
                     return util::foreach_partitioner<ExPolicy>::call(
                         std::forward<ExPolicy>(policy), first,
-                        std::distance(first, last), std::move(f1),
+                        detail::distance(first, last), std::move(f1),
                         util::projection_identity());
                 }
 
@@ -409,10 +410,10 @@ namespace hpx { namespace parallel { inline namespace v1 {
 
         ///////////////////////////////////////////////////////////////////////
         // non-segmented implementation
-        template <typename ExPolicy, typename FwdIter, typename F,
+        template <typename ExPolicy, typename FwdIterB, typename FwdIterE, typename F,
             typename Proj>
-        inline typename util::detail::algorithm_result<ExPolicy, FwdIter>::type
-        for_each_(ExPolicy&& policy, FwdIter first, FwdIter last, F&& f,
+        inline typename util::detail::algorithm_result<ExPolicy, FwdIterB>::type
+        for_each_(ExPolicy&& policy, FwdIterB first, FwdIterE last, F&& f,
             Proj&& proj, std::false_type)
         {
             typedef parallel::execution::is_sequenced_execution_policy<ExPolicy>
@@ -420,12 +421,12 @@ namespace hpx { namespace parallel { inline namespace v1 {
 
             if (first == last)
             {
[…]
