### 2020-07-06 299aa8c
Use longer timeouts in certain tests
