### 2020-07-06 e326586
Fix clang format error
