### 2020-07-06 ee31af5
Fix missing files and dependencies in modules CMakeLists.txt
