+++ libs/thread_pools/CMakeLists.txt
@@ -30,6 +30,7 @@ add_hpx_module(
   SOURCES ${thread_pools_sources}
   HEADERS ${thread_pools_headers}
   COMPAT_HEADERS ${thread_pools_compat_headers}
-  DEPENDENCIES hpx_assertion hpx_config hpx_errors hpx_logging hpx_schedulers
+  DEPENDENCIES hpx_assertion hpx_config hpx_errors hpx_itt_notify hpx_logging
+               hpx_schedulers
   CMAKE_SUBDIRS examples tests
 )
