+++ libs/serialization/CMakeLists.txt
@@ -158,6 +158,7 @@ add_hpx_module(
     hpx_debugging
     hpx_errors
     hpx_format
+    hpx_functional
     hpx_hashing
     hpx_preprocessor
     hpx_type_support
