+++ libs/algorithms/include/hpx/parallel/algorithms/sort.hpp
@@ -1,6 +1,6 @@
 //  Copyright (c) 2015 John Biddiscombe
-//  Copyright (c) 2015-2017 Hartmut Kaiser
-//  Copyright (c) 2015 Francisco Jose Tapia
+//  Copyright (c) 2015-2020 Hartmut Kaiser
+//  Copyright (c) 2015-2019 Francisco Jose Tapia
 //  Copyright (c) 2018 Taeguk Kwon
 //
 //  SPDX-License-Identifier: BSL-1.0
@@ -40,6 +40,7 @@
 #include <utility>
 
 namespace hpx { namespace parallel { inline namespace v1 {
+
     ///////////////////////////////////////////////////////////////////////////
     // sort
     namespace detail {
@@ -64,19 +65,76 @@ namespace hpx { namespace parallel { inline namespace v1 {
             return sorted;
         }
 
-        //------------------------------------------------------------------------
-        //  function : sort_thread
+        /// Return the iterator to the mid value of the three values
+        /// passed as parameters
+        ///
+        /// \param iter_1 : iterator to the first value
+        /// \param iter_2 : iterator to the second value
+        /// \param iter_3 : iterator to the third value
+        /// \param comp : object for to compare two values
+        /// \return iterator to mid value
+        template <typename Iter, typename Compare>
+        inline Iter mid3(Iter iter_1, Iter iter_2, Iter iter_3, Compare comp)
+        {
+            return comp(*iter_1, *iter_2) ?
+                (comp(*iter_2, *iter_3) ?
+                        iter_2 :
+                        (comp(*iter_1, *iter_3) ? iter_3 : iter_1)) :
+                (comp(*iter_3, *iter_2) ?
+                        iter_2 :
+                        (comp(*iter_3, *iter_1) ? iter_3 : iter_1));
+        }
+
+        /// Return the iterator to the mid value of the nine values
+        /// passed as parameters
+        //
+        /// \param iter_1   iterator to the first value
+        /// \param iter_2   iterator to the second value
+        /// \param iter_3   iterator to the third value
+        /// \param iter_4   iterator to the fourth value
+        /// \param iter_5   iterator to the fifth value
+        /// \param iter_6   iterator to the sixth value
+        /// \param iter_7   iterator to the seventh value
+        /// \param iter_8   iterator to the eighth value
+        /// \param iter_9   iterator to the ninth value
+        /// \return iterator to the mid value
+        template <typename Iter, typename Compare>
+        inline Iter mid9(Iter iter_1, Iter iter_2, Iter iter_3, Iter iter_4,
+            Iter iter_5, Iter iter_6, Iter iter_7, Iter iter_8, Iter iter_9,
+            Compare const& comp)
+        {
+            return mid3(mid3(iter_1, iter_2, iter_3, comp),
+                mid3(iter_4, iter_5, iter_6, comp),
+                mid3(iter_7, iter_8, iter_9, comp), comp);
+        }
+
+        /// Receive a range between first and last, obtain 9 values
+        /// between the elements  including the first and the previous
+        /// to the last. Obtain the iterator to the mid value and swap
+        /// with the first position
+        //
+        /// \param first    iterator to the first element
+        /// \param last     iterator to the last element
+        /// \param comp     object for to compare two elements
+        template <typename Iter, typename Compare>
+        inline void pivot9(Iter first, Iter last, Compare const& comp)
+        {
+            std::size_t chunk = (last - first) >> 3;
+            Iter itaux = mid9(first + 1, first + chunk, first + 2 * chunk,
+                first + 3 * chunk, first + 4 * chunk, first + 5 * chunk,
+                first + 6 * chunk, first + 7 * chunk, last - 1, comp);
+            std::iter_swap(first, itaux);
[…]
