### 2020-07-06 #2352 @davidbeckingsale
@danielskatz - my checklist shows the tasks I have completed. I had some other priorities last week but will make more review progress this week.
