### 2020-07-07 18578f4
Update documentation links to use stellar-group.org domain
