+++ libs/compute/README.rst
@@ -13,4 +13,4 @@ compute
 This library is part of HPX.
 
 Documentation can be found `here
-<https://stellar-group.github.io/hpx-docs/latest/html/libs/compute/docs/index.html>`__.
+<https://hpx-docs.stellar-group.org/latest/html/libs/compute/docs/index.html>`__.
