+++ libs/lcos_local/README.rst
@@ -13,4 +13,4 @@ lcos_local
 This library is part of HPX.
 
 Documentation can be found `here
-<https://stellar-group.github.io/hpx-docs/latest/html/libs/lcos_local/docs/index.html>`__.
+<https://hpx-docs.stellar-group.org/latest/html/libs/lcos_local/docs/index.html>`__.
