### 2020-07-07 1e40d87
Add for_each iterator adaptation
