+++ libs/algorithms/tests/unit/container_algorithms/foreach_adapt.cpp
@@ -12,25 +12,12 @@ void myfunction (int i) {
   std::cout << ' ' << i;
 }
 
-struct set_42
-{
-    template <typename T>
-    void operator()(T& val)
-    {
-        val = T(42);
-    }
-};
-
 int main()
 {
      hpx::parallel::for_each(hpx::parallel::execution::seq,
-         Iterator<std::int64_t>{0}, Sentinel<int64_t>{100}, myfunction);
+         Iterator<std::int64_t>{0}, Sentinel<int64_t>{100}, &myfunction);
 
     //HPX_TEST_EQ(result, std::int64_t(4950));
-
-    //hpx::parallel::for_each(hpx::parallel::execution::seq,
-
-    //)
 
     return hpx::util::report_errors();
 }
