### 2020-07-07 5bc4717
Add license and clang-format code
