+++ libs/algorithms/tests/unit/container_algorithms/foreach_adapt.cpp
@@ -21,10 +21,11 @@ void myfunction(int i)
 
 int main()
 {
-    hpx::parallel::for_each(hpx::parallel::execution::seq,
-        Iterator<std::int64_t>{0}, Sentinel<int64_t>{100}, &myfunction);
+    Iterator<std::int64_t> iter =
+        hpx::parallel::for_each(hpx::parallel::execution::seq,
+            Iterator<std::int64_t>{0}, Sentinel<int64_t>{100}, &myfunction);
 
-    //HPX_TEST_EQ(result, std::int64_t(4950));
+    HPX_TEST_EQ(*iter, std::int64_t(100));
 
     return hpx::util::report_errors();
 }
