### 2020-07-07 2d5e9a1
Adapt for parallel execution policy
