### 2020-07-08 e704cd7
Adapt segmented algorithm version
