### 2020-07-08 030f25c
Add is_sentinel_for and fix testing format
