### 2020-07-08 0aee109
Replace hpx::function_nonser with std::function in error module
