+++ libs/functional/tests/regressions/function_argument.cpp
@@ -6,8 +6,8 @@
 
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
-#include <hpx/distributed/iostream.hpp>
 
+#include <hpx/distributed/iostream.hpp>
 #include <hpx/functional/function.hpp>
 #include <hpx/modules/testing.hpp>
 #include <hpx/modules/timing.hpp>
