+++ libs/collectives/tests/performance/osu/osu_scatter.cpp
@@ -8,6 +8,7 @@
 
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
+
 #include <hpx/distributed/iostream.hpp>
 #include <hpx/serialization/serialize_buffer.hpp>
 
