+++ libs/functional/tests/regressions/function_serialization_728.cpp
@@ -7,6 +7,8 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #include <hpx/hpx_init.hpp>
+
+#include <hpx/distributed/iostream.hpp>
 #include <hpx/include/actions.hpp>
 #include <hpx/include/async.hpp>
 #include <hpx/include/lcos.hpp>
@@ -15,7 +17,6 @@
 #include <hpx/include/compression_zlib.hpp>
 #endif
 #include <hpx/include/parcel_coalescing.hpp>
-#include <hpx/distributed/iostream.hpp>
 #include <hpx/modules/testing.hpp>
 
 #include <cstddef>
