+++ libs/async_distributed/tests/regressions/dataflow_future_swap2.cpp
@@ -7,9 +7,10 @@
 // This test case demonstrates the issue described in #778: swapping futures
 // segfaults.
 
-#include <hpx/async_local/dataflow.hpp>
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_main.hpp>
+
+#include <hpx/async_local/dataflow.hpp>
 #include <hpx/distributed/iostream.hpp>
 #include <hpx/modules/format.hpp>
 #include <hpx/pack_traversal/unwrap.hpp>
