+++ libs/async_combinators/tests/regressions/wait_all_hang_1946.cpp
@@ -9,8 +9,9 @@
 
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
-#include <hpx/include/actions.hpp>
+
 #include <hpx/distributed/iostream.hpp>
+#include <hpx/include/actions.hpp>
 #include <hpx/serialization/serialize.hpp>
 
 #include <cstddef>
