+++ libs/algorithms/tests/unit/block/task_block_par.cpp
@@ -6,8 +6,9 @@
 
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
-#include <hpx/include/parallel_task_block.hpp>
+
 #include <hpx/distributed/iostream.hpp>
+#include <hpx/include/parallel_task_block.hpp>
 #include <hpx/modules/testing.hpp>
 #include <hpx/runtime_local/custom_exception_info.hpp>
 
