### 2020-07-09 881861e
Tweaks to CMake/cxx to fix MPI + NETWORKING ON/OFF build issues
