+++ libs/command_line_handling/CMakeLists.txt
@@ -21,6 +21,11 @@ set(command_line_handling_sources command_line_handling.cpp
                                   parse_command_line.cpp
 )
 
+if (HPX_WITH_NETWORKING OR HPX_MPI_WITH_FUTURES)
+    set(MPI_DEPS hpx_mpi_base)
+endif()
+
+
 include(HPX_AddModule)
 add_hpx_module(
   command_line_handling
@@ -47,7 +52,7 @@ add_hpx_module(
     hpx_functional
     hpx_format
     hpx_memory
-    hpx_mpi_base
+    ${MPI_DEPS}
     hpx_prefix
     hpx_program_options
     hpx_runtime_configuration
