+++ libs/resource_partitioner/CMakeLists.txt
@@ -22,6 +22,10 @@ set(resource_partitioner_compat_headers
 
 set(resource_partitioner_sources detail_partitioner.cpp partitioner.cpp)
 
+if (HPX_WITH_NETWORKING)
+    set(MPI_DEPS hpx_mpi_base)
+endif()
+
 include(HPX_AddModule)
 add_hpx_module(
   resource_partitioner
@@ -46,7 +50,7 @@ add_hpx_module(
     hpx_functional
     hpx_hardware
     hpx_memory
-    hpx_mpi_base
+    ${MPI_DEPS}
     hpx_pack_traversal
     hpx_plugin
     hpx_program_options
