### 2020-07-09 bd53780
Adding range version of stable sort
