+++ libs/algorithms/include/hpx/parallel/container_algorithms/stable_sort.hpp
@@ -1,10 +1,10 @@
-//  Copyright (c) 2015-2020 Hartmut Kaiser
+//  Copyright (c) 2020 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-/// \file parallel/container_algorithms/sort.hpp
+/// \file parallel/container_algorithms/stable_sort.hpp
 
 #pragma once
 
@@ -14,20 +14,21 @@
 #include <hpx/iterator_support/traits/is_range.hpp>
 
 #include <hpx/algorithms/traits/projected_range.hpp>
-#include <hpx/parallel/algorithms/sort.hpp>
+#include <hpx/parallel/algorithms/stable_sort.hpp>
 #include <hpx/parallel/util/projection_identity.hpp>
 
 #include <type_traits>
 #include <utility>
 
 namespace hpx { namespace parallel { inline namespace rangev1 {
-    /// Sorts the elements in the range \a rng  in ascending order. The
-    /// order of equal elements is not guaranteed to be preserved. The function
+
+    /// Sorts the elements in the range [first, last) in ascending order. The
+    /// relative order of equal elements is preserved. The function
     /// uses the given comparison function object comp (defaults to using
     /// operator<()).
     ///
-    /// \note   Complexity: O(Nlog(N)),
-    ///             where N = std::distance(begin(rng), end(rng)) comparisons.
+    /// \note   Complexity: O(Nlog(N)), where N = std::distance(first, last)
+    ///                     comparisons.
     ///
     /// A sequence is sorted with respect to a comparator \a comp and a
     /// projection \a proj if for every iterator i pointing to the sequence and
@@ -76,21 +77,23 @@ namespace hpx { namespace parallel { inline namespace rangev1 {
     /// permitted to execute in an unordered fashion in unspecified
     /// threads, and indeterminately sequenced within each thread.
     ///
-    /// \returns  The \a sort algorithm returns a
-    ///           \a hpx::future<Iter> if the execution policy is of
+    /// \returns  The \a stable_sort algorithm returns a
+    ///           \a hpx::future<RandomIt> if the execution policy is of
     ///           type
     ///           \a sequenced_task_policy or
-    ///           \a parallel_task_policy and returns \a Iter
+    ///           \a parallel_task_policy and returns \a RandomIt
     ///           otherwise.
-    ///           It returns \a last.
+    ///           The algorithm returns an iterator pointing to the first
+    ///           element after the last element in the input sequence.
+    //-----------------------------------------------------------------------------
     // clang-format off
     template <typename ExPolicy, typename Rng,
         typename Compare = v1::detail::less,
         typename Proj = util::projection_identity,
         HPX_CONCEPT_REQUIRES_(
             execution::is_execution_policy<ExPolicy>::value &&
-            hpx::traits::is_range<Rng>::value &&
-            traits::is_projected_range<Proj, Rng>::value &&
+            hpx::traits::is_range<Rng>::value&&
+            traits::is_projected_range<Proj, Rng>::value&&
             traits::is_indirect_callable<ExPolicy, Compare,
                 traits::projected_range<Proj, Rng>,
                 traits::projected_range<Proj, Rng>
@@ -99,11 +102,11 @@ namespace hpx { namespace parallel { inline namespace rangev1 {
     // clang-format on
     typename util::detail::algorithm_result<ExPolicy,
[…]
