### 2020-07-10 f5e09fd
Add missing dependency to algorithms module
