### 2020-07-10 ddd2295
Add missing module dependencies
