### 2020-07-10 8fd96a4
Adapt segmented algorithm version
