### 2020-07-10 6e6a06c
Fix build errors caused by missing mpi_base module
