### 2020-07-11 f17912a
Implement customization point for hpx::copy and hpx::ranges::copy
- This creates two customization points for the copy algorithm. This
  also makes hpx::copy and hpx::ranges::copy fully conforming to C++20
  by adding support for sentinels and algorithms that do not expect
  an execution policy as their first argument
- Replace all uses of hpx::parallel::copy with the appropriate new
  versions of the algorithms
- flyby: add HPX_NO_UNIQUE_ADDRESS expanding to [[no_unique_address]]
  on systems that support it
- flyby: fixing hpx::functional::tag_invoke for MSVC and adding
  hpx::functional::ta
[…]
