+++ libs/algorithms/tests/performance/benchmark_inplace_merge.cpp
@@ -60,8 +60,7 @@ double run_inplace_merge_benchmark_std(int test_count, OrgIter org_first,
     for (int i = 0; i < test_count; ++i)
     {
         // Restore [first, last) with original data.
-        hpx::parallel::copy(
-            hpx::parallel::execution::par, org_first, org_last, first);
+        hpx::copy(hpx::parallel::execution::par, org_first, org_last, first);
 
         std::uint64_t elapsed = hpx::util::high_resolution_clock::now();
         std::inplace_merge(first, middle, last);
@@ -82,8 +81,7 @@ double run_inplace_merge_benchmark_hpx(int test_count, ExPolicy policy,
     for (int i = 0; i < test_count; ++i)
     {
         // Restore [first, last) with original data.
-        hpx::parallel::copy(
-            hpx::parallel::execution::par, org_first, org_last, first);
+        hpx::copy(hpx::parallel::execution::par, org_first, org_last, first);
 
         std::uint64_t elapsed = hpx::util::high_resolution_clock::now();
         hpx::parallel::inplace_merge(policy, first, middle, last);
