+++ libs/segmented_algorithms/tests/unit/partitioned_vector_copy.cpp
@@ -78,8 +78,8 @@ void copy_algo_tests_with_policy(std::size_t size, std::size_t localities,
     fill_vector(v1, T(43));
 
     hpx::partitioned_vector<T> v2(size, policy);
-    auto p = hpx::parallel::copy(copy_policy, v1.begin(), v1.end(), v2.begin());
-    HPX_TEST(p.out() == v2.end());
+    auto p = hpx::ranges::copy(copy_policy, v1.begin(), v1.end(), v2.begin());
+    HPX_TEST(p.out == v2.end());
     compare_vectors(v1, v2);
 }
 
@@ -93,10 +93,10 @@ void copy_algo_tests_with_policy_async(std::size_t size, std::size_t localities,
     using hpx::parallel::execution::task;
 
     hpx::partitioned_vector<T> v2(size, policy);
-    auto f = hpx::parallel::copy(
-        copy_policy(task), v1.begin(), v1.end(), v2.begin());
+    auto f =
+        hpx::ranges::copy(copy_policy(task), v1.begin(), v1.end(), v2.begin());
 
-    HPX_TEST(f.get().out() == v2.end());
+    HPX_TEST(f.get().out == v2.end());
     compare_vectors(v1, v2);
 }
 
