+++ tests/performance/local/stream.cpp
@@ -94,11 +94,11 @@ void check_results(std::size_t iterations, Vector const& a_res,
     std::vector<STREAM_TYPE> b(b_res.size());
     std::vector<STREAM_TYPE> c(c_res.size());
 
-    hpx::parallel::copy(
+    hpx::copy(
         hpx::parallel::execution::par, a_res.begin(), a_res.end(), a.begin());
-    hpx::parallel::copy(
+    hpx::copy(
         hpx::parallel::execution::par, b_res.begin(), b_res.end(), b.begin());
-    hpx::parallel::copy(
+    hpx::copy(
         hpx::parallel::execution::par, c_res.begin(), c_res.end(), c.begin());
 
     STREAM_TYPE aj, bj, cj, scalar;
@@ -362,7 +362,7 @@ std::vector<std::vector<double>> run_benchmark(std::size_t iterations,
     {
         // Copy
         timing[0][iteration] = mysecond();
-        hpx::parallel::copy(policy, a.begin(), a.end(), c.begin());
+        hpx::copy(policy, a.begin(), a.end(), c.begin());
         timing[0][iteration] = mysecond() - timing[0][iteration];
 
         // Scale
