### 2020-07-11 2ffa250
Fix format violations clang/cmake
