+++ libs/command_line_handling/src/command_line_handling.cpp
@@ -12,7 +12,7 @@
 #include <hpx/modules/command_line_handling.hpp>
 #include <hpx/modules/debugging.hpp>
 #include <hpx/modules/format.hpp>
-#if defined (HPX_HAVE_LIB_MPI_BASE)
+#if defined(HPX_HAVE_LIB_MPI_BASE)
 #include <hpx/modules/mpi_base.hpp>
 #endif
 #include <hpx/modules/program_options.hpp>
