+++ src/runtime/components/server/runtime_support_server.cpp
@@ -7,24 +7,24 @@
 
 #include <hpx/config.hpp>
 #include <hpx/modules/async_distributed.hpp>
-#include <hpx/runtime_local/runtime_local.hpp>
-#include <hpx/runtime_distributed.hpp>
-#include <hpx/modules/errors.hpp>
 #include <hpx/modules/errors.hpp>
 #include <hpx/modules/filesystem.hpp>
+#include <hpx/runtime_distributed.hpp>
+#include <hpx/runtime_local/runtime_local.hpp>
 #ifdef HPX_HAVE_LIB_MPI_BASE
 #include <hpx/modules/mpi_base.hpp>
 #endif
 #include <hpx/modules/logging.hpp>
-#include <hpx/runtime_local/runtime_local.hpp>
-#include <hpx/thread_support/unlock_guard.hpp>
 #include <hpx/modules/timing.hpp>
-#include <hpx/util/from_string.hpp>
 #include <hpx/prefix/find_prefix.hpp>
 #include <hpx/runtime_configuration/ini.hpp>
+#include <hpx/runtime_local/runtime_local.hpp>
 #include <hpx/string_util/case_conv.hpp>
+#include <hpx/thread_support/unlock_guard.hpp>
+#include <hpx/util/from_string.hpp>
 
 #include <hpx/async_combinators/wait_all.hpp>
+#include <hpx/modules/threadmanager.hpp>
 #include <hpx/performance_counters/counters.hpp>
 #include <hpx/runtime/actions/continuation.hpp>
 #include <hpx/runtime/actions/plain_action.hpp>
@@ -34,29 +34,27 @@
 #include <hpx/runtime/components/server/component_database.hpp>
 #include <hpx/runtime/components/server/create_component.hpp>
 #include <hpx/runtime/components/server/runtime_support.hpp>
-#include <hpx/runtime_configuration/static_factory_data.hpp>
 #include <hpx/runtime/components/stubs/runtime_support.hpp>
 #include <hpx/runtime/find_localities.hpp>
 #include <hpx/runtime/naming/resolver_client.hpp>
 #include <hpx/runtime/naming/unmanaged.hpp>
 #include <hpx/runtime/runtime_fwd.hpp>
-#include <hpx/serialization/serialize.hpp>
-#include <hpx/serialization/vector.hpp>
+#include <hpx/runtime_configuration/static_factory_data.hpp>
 #include <hpx/runtime_local/shutdown_function.hpp>
 #include <hpx/runtime_local/startup_function.hpp>
-#include <hpx/modules/threadmanager.hpp>
+#include <hpx/serialization/serialize.hpp>
+#include <hpx/serialization/vector.hpp>
 
-#include <hpx/modules/collectives.hpp>
 #include <hpx/lcos_local/packaged_task.hpp>
+#include <hpx/modules/collectives.hpp>
 
 #include <hpx/assert.hpp>
-#include <hpx/execution_base/this_thread.hpp>
 #include <hpx/command_line_handling/command_line_handling.hpp>
 #include <hpx/command_line_handling/parse_command_line.hpp>
+#include <hpx/execution_base/this_thread.hpp>
 
-#include <hpx/plugins/message_handler_factory_base.hpp>
 #include <hpx/plugins/binary_filter_factory_base.hpp>
-
+#include <hpx/plugins/message_handler_factory_base.hpp>
 
 #include <boost/tokenizer.hpp>
 
@@ -80,8 +78,7 @@
 // Serialization support for the runtime_support actions
 HPX_REGISTER_ACTION_ID(
     hpx::components::server::runtime_support::load_components_action,
-    load_components_action,
-    hpx::actions::load_components_action_id)
[…]
