+++ libs/mpi_base/include/hpx/mpi_base/mpi_environment.hpp
@@ -25,8 +25,8 @@ namespace hpx { namespace util {
     {
         static bool check_mpi_environment(runtime_configuration const& cfg);
 
-        static int  init(int* argc, char*** argv, const int required,
-                         const int minimal, int &provided);
+        static int init(int* argc, char*** argv, const int required,
+            const int minimal, int& provided);
         static void init(int* argc, char*** argv, runtime_configuration& cfg);
         static void finalize();
 
