+++ libs/command_line_handling/CMakeLists.txt
@@ -22,9 +22,8 @@ set(command_line_handling_sources command_line_handling.cpp
 )
 
 if((HPX_WITH_NETWORKING AND HPX_WITH_PARCELPORT_MPI) OR HPX_WITH_ASYNC_MPI)
-    set(MPI_DEPS hpx_mpi_base)
+  set(MPI_DEPS hpx_mpi_base)
 endif()
-
 
 include(HPX_AddModule)
 add_hpx_module(
