### 2020-07-11 8c4ef28
Add doxygen comments for begin/end iterators and fix CONCEPT_REQUIRES format
