### 2020-07-11 f709563
Fix #4810 - Remove decay used by invoke_projected in parallel::for_each
