+++ libs/algorithms/include/hpx/parallel/algorithms/copy.hpp
@@ -159,7 +159,7 @@ namespace hpx {
     ///
     template <typename ExPolicy, typename FwdIter1, typename Size,
         typename FwdIter2>
-    typename util::detail::algorithm_result<ExPolicy,
+    typename hpx::parallel::util::detail::algorithm_result<ExPolicy,
         hpx::util::tagged_pair<tag::in(FwdIter1), tag::out(FwdIter2)>>::type
     copy_n(ExPolicy&& policy, FwdIter1 first, Size count, FwdIter2 dest);
 
@@ -183,12 +183,10 @@ namespace hpx {
     ///                     destination range (deduced).
     ///                     This iterator type must meet the requirements of an
     ///                     forward iterator.
-    /// \tparam F           The type of the function/function object to use
+    /// \tparam Pred        The type of the function/function object to use
     ///                     (deduced). Unlike its sequential form, the parallel
     ///                     overload of \a copy_if requires \a F to meet the
     ///                     requirements of \a CopyConstructible.
-    /// \tparam Proj        The type of an optional projection function. This
-    ///                     defaults to \a util::projection_identity
     ///
     /// \param policy       The execution policy to use for the scheduling of
     ///                     the iterations.
@@ -197,7 +195,7 @@ namespace hpx {
     /// \param last         Refers to the end of the sequence of elements the
     ///                     algorithm will be applied to.
     /// \param dest         Refers to the beginning of the destination range.
-    /// \param f            Specifies the function (or function object) which
+    /// \param pred         Specifies the function (or function object) which
     ///                     will be invoked for each of the elements in the
     ///                     sequence specified by [first, last).This is an
     ///                     unary predicate which returns \a true for the
@@ -211,10 +209,6 @@ namespace hpx {
     ///                     it. The type \a Type must be such that an object of
     ///                     type \a FwdIter1 can be dereferenced and then
     ///                     implicitly converted to Type.
-    /// \param proj         Specifies the function (or function object) which
-    ///                     will be invoked for each of the elements as a
-    ///                     projection operation before the actual predicate
-    ///                     \a is invoked.
     ///
     /// The assignments in the parallel \a copy_if algorithm invoked with
     /// an execution policy object of type \a sequenced_policy
@@ -227,12 +221,11 @@ namespace hpx {
     /// within each thread.
     ///
     /// \returns  The \a copy_if algorithm returns a
-    ///           \a hpx::future<tagged_pair<tag::in(FwdIter1), tag::out(FwdIter2)> >
+    ///           \a hpx::future<FwdIter2> >
     ///           if the execution policy is of type
     ///           \a sequenced_task_policy or
     ///           \a parallel_task_policy and
-    ///           returns \a tagged_pair<tag::in(FwdIter1), tag::out(FwdIter2)>
-    ///           otherwise.
+    ///           returns \a FwdIter2 otherwise.
     ///           The \a copy algorithm returns the pair of the input iterator
     ///           forwarded to the first element after the last in the input
     ///           sequence and the output iterator to the
@@ -240,11 +233,9 @@ namespace hpx {
     ///           copied.
     ///
     template <typename ExPolicy, typename FwdIter1, typename FwdIter2,
-        typename F, typename Proj = util::projection_identity>
-        typename util::detail::algorithm_result<ExPolicy,
[…]
