### 2020-07-12 c9e2a2c
Simplify creating generic CPOs by deriving them from a CRTP base hpx::functional::tag
