+++ libs/algorithms/include/hpx/parallel/container_algorithms/copy.hpp
@@ -21,6 +21,7 @@
 #include <hpx/parallel/algorithms/copy.hpp>
 #include <hpx/parallel/util/result_types.hpp>
 
+#include <cstddef>
 #include <type_traits>
 #include <utility>
 
@@ -131,6 +132,64 @@ namespace hpx { namespace ranges {
             FwdIter>
         >::type
     copy(ExPolicy&& policy, Rng&& rng, FwdIter dest);
+
+    /// Copies the elements in the range [first, first + count), starting from
+    /// first and proceeding to first + count - 1., to another range beginning
+    /// at dest.
+    ///
+    /// \note   Complexity: Performs exactly \a count assignments, if
+    ///         count > 0, no assignments otherwise.
+    ///
+    /// \tparam ExPolicy    The type of the execution policy to use (deduced).
+    ///                     It describes the manner in which the execution
+    ///                     of the algorithm may be parallelized and the manner
+    ///                     in which it executes the assignments.
+    /// \tparam FwdIter1    The type of the source iterators used (deduced).
+    ///                     This iterator type must meet the requirements of an
+    ///                     forward iterator.
+    /// \tparam Size        The type of the argument specifying the number of
+    ///                     elements to apply \a f to.
+    /// \tparam FwdIter2    The type of the iterator representing the
+    ///                     destination range (deduced).
+    ///                     This iterator type must meet the requirements of an
+    ///                     forward iterator.
+    ///
+    /// \param policy       The execution policy to use for the scheduling of
+    ///                     the iterations.
+    /// \param first        Refers to the beginning of the sequence of elements
+    ///                     the algorithm will be applied to.
+    /// \param count        Refers to the number of elements starting at
+    ///                     \a first the algorithm will be applied to.
+    /// \param dest         Refers to the beginning of the destination range.
+    ///
+    /// The assignments in the parallel \a copy_n algorithm invoked with
+    /// an execution policy object of type \a sequenced_policy
+    /// execute in sequential order in the calling thread.
+    ///
+    /// The assignments in the parallel \a copy_n algorithm invoked with
+    /// an execution policy object of type \a parallel_policy or
+    /// \a parallel_task_policy are permitted to execute in an unordered
+    /// fashion in unspecified threads, and indeterminately sequenced
+    /// within each thread.
+    ///
+    /// \returns  The \a copy_n algorithm returns a
+    ///           \a hpx::future<ranges::copy_n_result<FwdIter1, FwdIter2> >
+    ///           if the execution policy is of type
+    ///           \a sequenced_task_policy or
+    ///           \a parallel_task_policy and
+    ///           returns \a ranges::copy_n_result<FwdIter1, FwdIter2>
+    ///           otherwise.
+    ///           The \a copy algorithm returns the pair of the input iterator
+    ///           forwarded to the first element after the last in the input
+    ///           sequence and the output iterator to the
+    ///           element in the destination range, one past the last element
+    ///           copied.
+    ///
+    template <typename ExPolicy, typename FwdIter1, typename Size,
+        typename FwdIter2>
+    typename hpx::parallel::util::detail::algorithm_result<ExPolicy,
+        hpx::ranges::copy_n_result<FwdIter1, FwdIter2>>::type
+    copy_n(ExPolicy&& policy, FwdIter1 first, Size count, FwdIter2 dest);
 
     /// Copies the elements in the range, defined by [first, last) to another
     /// range beginning at \a dest. Copies only the elements for which the
[…]
