+++ libs/algorithms/include/hpx/parallel/algorithms/copy.hpp
@@ -145,11 +145,11 @@ namespace hpx {
     /// within each thread.
     ///
     /// \returns  The \a copy_n algorithm returns a
-    ///           \a hpx::future<tagged_pair<tag::in(FwdIter1), tag::out(FwdIter2)> >
+    ///           \a hpx::future<FwdIter2>
     ///           if the execution policy is of type
     ///           \a sequenced_task_policy or
     ///           \a parallel_task_policy and
-    ///           returns \a tagged_pair<tag::in(FwdIter1), tag::out(FwdIter2)>
+    ///           returns \a FwdIter2
     ///           otherwise.
     ///           The \a copy algorithm returns the pair of the input iterator
     ///           forwarded to the first element after the last in the input
@@ -157,10 +157,8 @@ namespace hpx {
     ///           element in the destination range, one past the last element
     ///           copied.
     ///
-    template <typename ExPolicy, typename FwdIter1, typename Size,
-        typename FwdIter2>
-    typename hpx::parallel::util::detail::algorithm_result<ExPolicy,
-        hpx::util::tagged_pair<tag::in(FwdIter1), tag::out(FwdIter2)>>::type
+    template <typename ExPolicy, typename FwdIter1, typename Size, typename FwdIter2>
+    typename hpx::parallel::util::detail::algorithm_result<ExPolicy, FwdIter2>::type
     copy_n(ExPolicy&& policy, FwdIter1 first, Size count, FwdIter2 dest);
 
     /// Copies the elements in the range, defined by [first, last), to another
@@ -243,9 +241,9 @@ namespace hpx {
 #else    // DOXYGEN
 
 namespace hpx { namespace parallel { inline namespace v1 {
+
     ///////////////////////////////////////////////////////////////////////////
     // copy
-
     namespace detail {
 
         struct copy_iteration
@@ -289,7 +287,8 @@ namespace hpx { namespace parallel { inline namespace v1 {
 #if defined(HPX_COMPUTE_DEVICE_CODE)
                 HPX_ASSERT(false);
                 typename util::detail::algorithm_result<ExPolicy,
-                    std::pair<FwdIter1, FwdIter2>>::type* dummy = nullptr;
+                    util::in_out_result<FwdIter1, FwdIter2>>::type* dummy =
+                    nullptr;
                 return std::move(*dummy);
 #else
                 typedef hpx::util::zip_iterator<FwdIter1, FwdIter2>
@@ -313,7 +312,7 @@ namespace hpx { namespace parallel { inline namespace v1 {
 
 #if defined(HPX_COMPUTE_DEVICE_CODE)
         template <typename FwdIter1, typename FwdIter2, typename Enable = void>
-        struct copy_iter : public copy<std::pair<FwdIter1, FwdIter2>>
+        struct copy_iter : public copy<util::in_out_result<FwdIter1, FwdIter2>>
         {
         };
 #else
@@ -434,9 +433,11 @@ namespace hpx { namespace parallel { inline namespace v1 {
             hpx::traits::is_iterator<FwdIter1>::value &&
             hpx::traits::is_iterator<FwdIter2>::value)>
     // clang-format on
-    typename util::detail::algorithm_result<ExPolicy,
-        util::in_out_result<FwdIter1, FwdIter2>>::type
-    copy_n(ExPolicy&& policy, FwdIter1 first, Size count, FwdIter2 dest)
+    HPX_DEPRECATED(
+        "hpx::parallel::copy_n is deprecated, use hpx::copy_n instead")
+        typename util::detail::algorithm_result<ExPolicy,
+            util::in_out_result<FwdIter1, FwdIter2>>::type
[…]

[…]
