### 2020-07-12 20309cf
add Hazard pointer overhead test
