### 2020-07-13 008617d
Add support for building non-PRs in Jenkins configuration
