+++ .jenkins/cscs/ctest.cmake
@@ -12,9 +12,14 @@ set(CTEST_BUILD_PARALLELISM 20)
 set(CTEST_TEST_PARALLELISM 4)
 set(CTEST_CMAKE_GENERATOR Ninja)
 set(CTEST_SITE "cscs(daint)")
-set(CTEST_BUILD_NAME "$ENV{ghprbPullId}-${CTEST_BUILD_CONFIGURATION_NAME}")
 set(CTEST_UPDATE_COMMAND "git")
 set(CTEST_UPDATE_VERSION_ONLY "ON")
+
+if(NOT "$ENV{ghprbPullId}" STREQUAL "")
+  set(CTEST_BUILD_NAME "$ENV{ghprbPullId}-${CTEST_BUILD_CONFIGURATION_NAME}")
+else()
+  set(CTEST_BUILD_NAME "$ENV{GIT_LOCAL_BRANCH}-${CTEST_BUILD_CONFIGURATION_NAME}")
+endif()
 
 message("CTEST_BUILD_NAME=${CTEST_BUILD_NAME}")
 
