### 2020-07-13 e5da88b
Format Jenkins ctest script
