+++ .jenkins/cscs/ctest.cmake
@@ -18,29 +18,37 @@ set(CTEST_UPDATE_VERSION_ONLY "ON")
 if(NOT "$ENV{ghprbPullId}" STREQUAL "")
   set(CTEST_BUILD_NAME "$ENV{ghprbPullId}-${CTEST_BUILD_CONFIGURATION_NAME}")
 else()
-  set(CTEST_BUILD_NAME "$ENV{GIT_LOCAL_BRANCH}-${CTEST_BUILD_CONFIGURATION_NAME}")
+  set(CTEST_BUILD_NAME
+      "$ENV{GIT_LOCAL_BRANCH}-${CTEST_BUILD_CONFIGURATION_NAME}"
+  )
 endif()
 
 message("CTEST_BUILD_NAME=${CTEST_BUILD_NAME}")
 
 set(CTEST_CONFIGURE_COMMAND "${CMAKE_COMMAND} ${CTEST_SOURCE_DIRECTORY}")
-set(CTEST_CONFIGURE_COMMAND "${CTEST_CONFIGURE_COMMAND} -G${CTEST_CMAKE_GENERATOR}")
-set(CTEST_CONFIGURE_COMMAND "${CTEST_CONFIGURE_COMMAND} -B${CTEST_BINARY_DIRECTORY}")
-set(CTEST_CONFIGURE_COMMAND "${CTEST_CONFIGURE_COMMAND} -DHPX_WITH_PARALLEL_TESTS_BIND_NONE=ON")
-set(CTEST_CONFIGURE_COMMAND "${CTEST_CONFIGURE_COMMAND} ${CTEST_CONFIGURE_EXTRA_OPTIONS}")
+set(CTEST_CONFIGURE_COMMAND
+    "${CTEST_CONFIGURE_COMMAND} -G${CTEST_CMAKE_GENERATOR}"
+)
+set(CTEST_CONFIGURE_COMMAND
+    "${CTEST_CONFIGURE_COMMAND} -B${CTEST_BINARY_DIRECTORY}"
+)
+set(CTEST_CONFIGURE_COMMAND
+    "${CTEST_CONFIGURE_COMMAND} -DHPX_WITH_PARALLEL_TESTS_BIND_NONE=ON"
+)
+set(CTEST_CONFIGURE_COMMAND
+    "${CTEST_CONFIGURE_COMMAND} ${CTEST_CONFIGURE_EXTRA_OPTIONS}"
+)
 
 ctest_start(Experimental TRACK "Pull_Requests")
 ctest_update()
 ctest_submit(PARTS Update BUILD_ID CTEST_BUILD_ID)
-file(WRITE "jenkins-hpx-$ENV{configuration_name}-cdash-build-id.txt" "${CTEST_BUILD_ID}")
+file(WRITE "jenkins-hpx-$ENV{configuration_name}-cdash-build-id.txt"
+     "${CTEST_BUILD_ID}"
+)
 ctest_configure()
 ctest_submit(PARTS Configure)
-ctest_build(
-  TARGET all
-  FLAGS "-k0 -j ${CTEST_BUILD_PARALLELISM}")
-ctest_build(
-  TARGET tests
-  FLAGS "-k0 -j ${CTEST_BUILD_PARALLELISM}")
+ctest_build(TARGET all FLAGS "-k0 -j ${CTEST_BUILD_PARALLELISM}")
+ctest_build(TARGET tests FLAGS "-k0 -j ${CTEST_BUILD_PARALLELISM}")
 ctest_submit(PARTS Build)
 ctest_test(PARALLEL_LEVEL "${CTEST_TEST_PARALLELISM}")
 ctest_submit(PARTS Test)
