### 2020-07-13 012cff7
fix PR comments related to cmake and headers order
