### 2020-07-13 3d95f1a
typedef std::array for libcds
