+++ libs/coroutines/include/hpx/coroutines/detail/coroutine_stackful_self.hpp
@@ -83,13 +83,12 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
         }
 
 #ifdef HPX_HAVE_LIBCDS
-        std::array<std::size_t, 3> get_libcds_data() const override
+        cds_type get_libcds_data() const override
         {
             HPX_ASSERT(pimpl_);
             return pimpl_->get_libcds_data();
         }
-        std::array<std::size_t, 3> set_libcds_data(
-            std::array<std::size_t, 3> data) override
+        cds_type set_libcds_data(cds_type data) override
         {
             HPX_ASSERT(pimpl_);
             return pimpl_->set_libcds_data(data);
