+++ libs/coroutines/include/hpx/coroutines/coroutine.hpp
@@ -98,13 +98,13 @@ namespace hpx { namespace threads { namespace coroutines {
         }
 
 #ifdef HPX_HAVE_LIBCDS
-        std::array<std::size_t, 3> get_libcds_data() const
+        typedef std::array<std::size_t, 3> cds_type;
+        cds_type get_libcds_data() const
         {
             return impl_.get_libcds_data();
         }
 
-        std::array<std::size_t, 3> set_libcds_data(
-            std::array<std::size_t, 3> data)
+        cds_type set_libcds_data(cds_type data)
         {
             return impl_.set_libcds_data(data);
         }
