+++ libs/threading_base/include/hpx/threading_base/thread_data.hpp
@@ -554,9 +554,9 @@ namespace hpx { namespace threads {
         virtual std::size_t set_thread_data(std::size_t data) = 0;
 
 #ifdef HPX_HAVE_LIBCDS
-        virtual std::array<std::size_t, 3> get_libcds_data() const = 0;
-        virtual std::array<std::size_t, 3> set_libcds_data(
-            std::array<std::size_t, 3> data) = 0;
+        typedef std::array<std::size_t, 3> cds_type;
+        virtual cds_type get_libcds_data() const = 0;
+        virtual cds_type set_libcds_data(cds_type data) = 0;
 #endif
 
         virtual void rebind(thread_init_data& init_data) = 0;
