+++ libs/threading_base/src/thread_helpers.cpp
@@ -206,27 +206,26 @@ namespace hpx { namespace threads {
     }
 
 #ifdef HPX_HAVE_LIBCDS
-    std::array<std::size_t, 3> get_libcds_data(
-        thread_id_type const& id, error_code& ec)
+    cds_type get_libcds_data(thread_id_type const& id, error_code& ec)
     {
         if (HPX_UNLIKELY(!id))
         {
             HPX_THROWS_IF(ec, null_thread_id, "hpx::threads::get_libcds_data",
                 "null thread id encountered");
-            return std::array<size_t, 3>{{0, 0, 0}};
+            return cds_type{{0, 0, 0}};
         }
 
         return get_thread_id_data(id)->get_libcds_data();
     }
 
-    std::array<std::size_t, 3> set_libcds_data(thread_id_type const& id,
-        std::array<std::size_t, 3> data, error_code& ec)
+    cds_type set_libcds_data(
+        thread_id_type const& id, cds_type data, error_code& ec)
     {
         if (HPX_UNLIKELY(!id))
         {
             HPX_THROWS_IF(ec, null_thread_id, "hpx::threads::set_libcds_data",
                 "null thread id encountered");
-            return std::array<size_t, 3>{{0, 0, 0}};
+            return cds_type{{0, 0, 0}};
         }
 
         return get_thread_id_data(id)->set_libcds_data(data);
