+++ libs/coroutines/include/hpx/coroutines/detail/coroutine_self.hpp
@@ -131,9 +131,9 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
         virtual std::size_t set_thread_data(std::size_t data) = 0;
 
 #ifdef HPX_HAVE_LIBCDS
-        virtual std::array<std::size_t, 3> get_libcds_data() const = 0;
-        virtual std::array<std::size_t, 3> set_libcds_data(
-            std::array<std::size_t, 3> data) = 0;
+        typedef std::array<std::size_t, 3> cds_type;
+        virtual cds_type get_libcds_data() const = 0;
+        virtual cds_type set_libcds_data(cds_type data) = 0;
 #endif
 
         virtual tss_storage* get_thread_tss_data() = 0;
