### 2020-07-13 f4fc288
update CXX standard for libcds due to its master changes
