### 2020-07-14 71105d7
Move getting the CDash build id to the test submission phase
