### 2020-07-14 5eb4b72
Ignore unset pull request id for master branch builds in Jenkins config
