### 2020-07-14 3ad8e1a
Add Jenkins configuration for rostam
