+++ .jenkins/lsu/slurm-configuration-gcc-10-cuda-11.sh
@@ -1,0 +1,8 @@
+# Copyright (c) 2020 ETH Zurich
+#
+# SPDX-License-Identifier: BSL-1.0
+# Distributed under the Boost Software License, Version 1.0. (See accompanying
+# file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+configuration_slurm_partition="cuda"
+configuration_slurm_num_nodes="1"
