+++ .jenkins/lsu/env-clang-8.sh
@@ -4,24 +4,18 @@
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
 # file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-source $SPACK_ROOT/share/spack/setup-env.sh
+module purge
+module load cmake
+module load clang/8
+module load boost/1.69.0-${build_type,,}
+module load hwloc/2
+module load mpich
 
-export CRAYPE_LINK_TYPE=dynamic
+export HPXRUN_RUNWRAPPER=srun
 
-module load daint-mc
-module switch PrgEnv-cray PrgEnv-icc
-module load Boost
-module load hwloc/.2.0.3
-spack load cmake
-spack load ninja
-
-export CXX=`which CC`
-export CC=`which cc`
-
-configure_extra_options="-DCMAKE_BUILD_TYPE=Debug"
-configure_extra_options+=" -DHPX_WITH_MAX_CPU_COUNT=128"
+configure_extra_options="-DCMAKE_BUILD_TYPE=${build_type}"
 configure_extra_options+=" -DHPX_WITH_MALLOC=system"
 configure_extra_options+=" -DHPX_WITH_DEPRECATION_WARNINGS=OFF"
 configure_extra_options+=" -DHPX_WITH_COMPILER_WARNINGS=ON"
 configure_extra_options+=" -DHPX_WITH_COMPILER_WARNINGS_AS_ERRORS=ON"
-configure_extra_options+=" -DHPX_WITH_SPINLOCK_DEADLOCK_DETECTION=ON"
+configure_extra_options+=" -DHPX_WITH_PARCELPORT_MPI=ON"
