### 2020-07-14 2a6ff81
Make sure get_action_name is defined with networking off
