+++ hpx/lcos/base_lco.hpp
@@ -7,6 +7,7 @@
 #pragma once
 
 #include <hpx/config.hpp>
+#include <hpx/actions/action_support.hpp>
 #include <hpx/actions/base_action.hpp>
 #include <hpx/actions_base/component_action.hpp>
 #include <hpx/components_base/get_lva.hpp>
