### 2020-07-14 0f67708
Fix warnings reported by clang 10
