### 2020-07-15 7cad821
Update documentation on using HPX::wrap_main
