### 2020-07-15 61a3a45
Fixing assertion in split_gids
