### 2020-07-15 28e7e78
Fix deprecated and duplicated headers
