+++ tests/unit/component/migrate_polymorphic_component.cpp
@@ -9,7 +9,6 @@
 #include <hpx/distributed/iostream.hpp>
 #include <hpx/include/actions.hpp>
 #include <hpx/include/components.hpp>
-#include <hpx/include/iostreams.hpp>
 #include <hpx/include/lcos.hpp>
 #include <hpx/include/serialization.hpp>
 #include <hpx/modules/testing.hpp>
