### 2020-07-16 43ebffb
Remove workaround for compiler problem (it's not needed anymore)
