### 2020-07-16 1efa06e
Change the warning message if HPX is not installe
