### 2020-07-16 038462a
Add more assertions to coroutines
