### 2020-07-16 e123fbe
Rearrange coroutine reset so that yielding is allowed in thread function destructor
