### 2020-07-16 8e910a6
Mark captured variable as unused in release build
