### 2020-07-16 947918d
Remove `using util::result_of` from namespace hpx
