### 2020-07-16 e895b80
separate std::array into individual elements same as threading functions
