+++ libs/threading_base/include/hpx/threading_base/thread_helpers.hpp
@@ -394,12 +394,23 @@ namespace hpx { namespace threads {
         thread_id_type const& id, std::size_t data, error_code& ec = throws);
 
 #ifdef HPX_HAVE_LIBCDS
-    typedef std::array<std::size_t, 3> cds_type;
-    HPX_EXPORT cds_type get_libcds_data(
+    HPX_EXPORT std::size_t& get_libcds_data(
         thread_id_type const& id, error_code& ec = throws);
 
-    HPX_EXPORT cds_type set_libcds_data(
-        thread_id_type const& id, cds_type, error_code& ec = throws);
+    HPX_EXPORT std::size_t& set_libcds_data(
+        thread_id_type const& id, std::size_t& data, error_code& ec = throws);
+
+    HPX_EXPORT std::size_t& get_libcds_hp_data(
+        thread_id_type const& id, error_code& ec = throws);
+
+    HPX_EXPORT std::size_t& set_libcds_hp_data(
+        thread_id_type const& id, std::size_t& data, error_code& ec = throws);
+
+    HPX_EXPORT std::size_t& get_libcds_dhp_data(
+        thread_id_type const& id, error_code& ec = throws);
+
+    HPX_EXPORT std::size_t& set_libcds_dhp_data(
+        thread_id_type const& id, std::size_t& data, error_code& ec = throws);
 #endif
 
     HPX_EXPORT std::size_t& get_continuation_recursion_count();
