### 2020-07-16 da362a5
Fix async_cuda build problems when distributed runtime is disabled
