+++ libs/async_cuda/src/get_targets.cpp
@@ -8,7 +8,6 @@
 
 #include <hpx/async_cuda/target.hpp>
 #include <hpx/futures/future.hpp>
-#include <hpx/modules/async_distributed.hpp>
 #include <hpx/modules/errors.hpp>
 #include <hpx/runtime/threads/thread_helpers.hpp>
 #include <hpx/runtime_fwd.hpp>
@@ -17,6 +16,7 @@
 #include <hpx/threading_base/thread_helpers.hpp>
 
 #if defined(HPX_HAVE_DISTRIBUTED_RUNTIME)
+#include <hpx/modules/async_distributed.hpp>
 #include <hpx/actions_base/plain_action.hpp>
 #include <hpx/runtime/find_here.hpp>
 #include <hpx/runtime/naming/id_type.hpp>
