+++ libs/algorithms/include/hpx/parallel/algorithms/copy.hpp
@@ -360,9 +360,10 @@ namespace hpx { namespace parallel { inline namespace v1 {
     // clang-format off
     template <typename ExPolicy, typename FwdIter1, typename FwdIter2,
         HPX_CONCEPT_REQUIRES_(
-            execution::is_execution_policy<ExPolicy>::value&&
-            hpx::traits::is_iterator<FwdIter1>::value&&
-            hpx::traits::is_iterator<FwdIter2>::value)>
+            execution::is_execution_policy<ExPolicy>::value &&
+            hpx::traits::is_iterator<FwdIter1>::value &&
+            hpx::traits::is_iterator<FwdIter2>::value
+        )>
     // clang-format on
     HPX_DEPRECATED("hpx::parallel::copy is deprecated, use hpx::copy instead")
         typename util::detail::algorithm_result<ExPolicy, FwdIter2>::type
