### 2020-07-16 94a5527
Adding CPOs for hpx::reduce
- also add hpx::ranges::reduce implementations
- deprecate hpx::parallel::reduce
