+++ libs/algorithms/include/hpx/parallel/container_algorithms/reduce.hpp
@@ -1,16 +1,16 @@
-//  Copyright (c) 2007-2017 Hartmut Kaiser
+//  Copyright (c) 2007-2020 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-/// \file parallel/algorithms/reduce.hpp
+/// \file parallel/container_algorithms/reduce.hpp
 
 #pragma once
 
 #if defined(DOXYGEN)
 
-namespace hpx {
+namespace hpx { namespace ranges {
 
     // clang-format off
 
@@ -23,8 +23,10 @@ namespace hpx {
     ///                     It describes the manner in which the execution
     ///                     of the algorithm may be parallelized and the manner
     ///                     in which it executes the assignments.
-    /// \tparam FwdIter     The type of the source begin and end iterators used
-    ///                     (deduced).
+    /// \tparam FwdIter     The type of the source begin iterator used (deduced).
+    ///                     This iterator type must meet the requirements of an
+    ///                     forward iterator.
+    /// \tparam Sent        The type of the source sentinel used (deduced).
     ///                     This iterator type must meet the requirements of an
     ///                     forward iterator.
     /// \tparam F           The type of the function/function object to use
@@ -50,7 +52,7 @@ namespace hpx {
     ///                     \endcode \n
     ///                     The signature does not need to have const&.
     ///                     The types \a Type1 \a Ret must be
-    ///                     such that an object of type \a FwdIter can be
+    ///                     such that an object of type \a FwdIterB can be
     ///                     dereferenced and then implicitly converted to any
     ///                     of those types.
     /// \param init         The initial value for the generalized sum.
@@ -85,9 +87,10 @@ namespace hpx {
     /// that the behavior of reduce may be non-deterministic for
     /// non-associative or non-commutative binary predicate.
     ///
-    template <typename ExPolicy, typename FwdIter, typename T, typename F>
-    typename util::detail::algorithm_result<ExPolicy, T>::type
-    reduce(ExPolicy&& policy, FwdIter first, FwdIter last, T init, F&& f);
+    template <typename ExPolicy, typename FwdIter, typename Sent, typename T,
+        typename F>
+    typename util::detail::algorithm_result<ExPolicy, T>::type reduce(
+        ExPolicy&& policy, FwdIter first, Sent last, T init, F&& f);
 
     /// Returns GENERALIZED_SUM(+, init, *first, ..., *(first + (last - first) - 1)).
     ///
@@ -98,8 +101,10 @@ namespace hpx {
     ///                     It describes the manner in which the execution
     ///                     of the algorithm may be parallelized and the manner
     ///                     in which it executes the assignments.
-    /// \tparam FwdIter     The type of the source begin and end iterators used
-    ///                     (deduced).
+    /// \tparam FwdIter     The type of the source begin iterator used (deduced).
+    ///                     This iterator type must meet the requirements of an
+    ///                     forward iterator.
+    /// \tparam Sent        The type of the source sentinel used (deduced).
     ///                     This iterator type must meet the requirements of an
     ///                     forward iterator.
     /// \tparam T           The type of the value to be used as initial (and
@@ -143,9 +148,9 @@ namespace hpx {
[…]
