+++ libs/segmented_algorithms/include/hpx/parallel/segmented_algorithms/reduce.hpp
@@ -191,9 +191,7 @@ namespace hpx { namespace parallel { inline namespace v1 {
         // segmented implementation
         template <typename ExPolicy, typename InIterB, typename InIterE,
             typename T, typename F>
-        inline typename std::enable_if<
-            execution::is_execution_policy<ExPolicy>::value,
-            typename util::detail::algorithm_result<ExPolicy, T>::type>::type
+        typename util::detail::algorithm_result<ExPolicy, T>::type
         reduce_(ExPolicy&& policy, InIterB first, InIterE last, T init, F&& f,
             std::true_type)
         {
@@ -215,9 +213,7 @@ namespace hpx { namespace parallel { inline namespace v1 {
         // forward declare the non-segmented version of this algorithm
         template <typename ExPolicy, typename InIterB, typename InIterE,
             typename T, typename F>
-        inline typename std::enable_if<
-            execution::is_execution_policy<ExPolicy>::value,
-            typename util::detail::algorithm_result<ExPolicy, T>::type>::type
+        typename util::detail::algorithm_result<ExPolicy, T>::type
         reduce_(ExPolicy&& policy, InIterB first, InIterE last, T init, F&& f,
             std::false_type);
 

[…]
