### 2020-07-16 31b8e5f
Restore compatibility for old (now deprecated) copy algorithms
