### 2020-07-16 23ffd62
use value semantics and descriptive function names
