+++ libs/threading_base/include/hpx/threading_base/thread_data_stackless.hpp
@@ -88,34 +88,35 @@ namespace hpx { namespace threads {
         }
 
 #ifdef HPX_HAVE_LIBCDS
-        std::size_t& get_libcds_data() const override
+        std::size_t get_libcds_data() const override
         {
             return coroutine_.get_libcds_data();
         }
 
-        std::size_t& set_libcds_data(std::size_t& data) override
+        std::size_t set_libcds_data(std::size_t data) override
         {
             return coroutine_.set_libcds_data(data);
         }
 
-        std::size_t& get_libcds_hp_data() const override
+        std::size_t get_libcds_hazard_pointer_data() const override
         {
-            return coroutine_.get_libcds_hp_data();
+            return coroutine_.get_libcds_hazard_pointer_data();
         }
 
-        std::size_t& set_libcds_hp_data(std::size_t& data) override
+        std::size_t set_libcds_hazard_pointer_data(std::size_t data) override
         {
-            return coroutine_.set_libcds_hp_data(data);
+            return coroutine_.set_libcds_hazard_pointer_data(data);
         }
 
-        std::size_t& get_libcds_dhp_data() const override
+        std::size_t get_libcds_dynamic_hazard_pointer_data() const override
         {
-            return coroutine_.get_libcds_dhp_data();
+            return coroutine_.get_libcds_dynamic_hazard_pointer_data();
         }
 
-        std::size_t& set_libcds_dhp_data(std::size_t& data) override
+        std::size_t set_libcds_dynamic_hazard_pointer_data(
+            std::size_t data) override
         {
-            return coroutine_.set_libcds_dhp_data(data);
+            return coroutine_.set_libcds_dynamic_hazard_pointer_data(data);
         }
 #endif
 
