### 2020-07-16 a7f4a50
add error handling if libcds is not fetched successfully
