+++ CMakeLists.txt
@@ -934,8 +934,13 @@ endif()
 # LibCDS option
 hpx_option(HPX_WITH_LIBCDS BOOL "Enable LibCDS support." OFF)
 if(HPX_WITH_LIBCDS)
-  hpx_add_config_define(HPX_HAVE_LIBCDS) # tell HPX that we use LibCDS
   include(HPX_SetupLibCDS)
+  if(NOT libcds_POPULATED)
+    hpx_error(
+      "HPX_WITH_LIBCDS was set to ON, but HPX failed to fetch LibCDS"
+    )
+  endif()
+  hpx_add_config_define(HPX_HAVE_LIBCDS) # tell HPX that we use LibCDS
 endif()
 
 hpx_option(
