### 2020-07-17 52f9197
remove unecessary static data, headers, move preprocessor directives, rename test file
