### 2020-07-17 22639cc
Fix includes in transform_reduce_scaling test
