### 2020-07-17 c981cf9
Fix typo in reduce
