### 2020-07-17 206a7c7
Adding docs for range based algorithms
