### 2020-07-17 48c7c70
Small tweaks to Jenkins configuration
