### 2020-07-17 6cbbe6b
Fix build name for master/release builds on Jenkins
