### 2020-07-17 73819fe
adjust #ifdef to #if define(X) style
