+++ libs/threading/include/hpx/threading/thread.hpp
@@ -120,7 +120,7 @@ namespace hpx {
         std::size_t get_thread_data() const;
         std::size_t set_thread_data(std::size_t);
 
-#ifdef HPX_HAVE_LIBCDS
+#if defined(HPX_HAVE_LIBCDS)
         std::size_t get_libcds_data() const;
         std::size_t set_libcds_data(std::size_t);
         std::size_t get_libcds_hazard_pointer_data() const;
@@ -265,7 +265,7 @@ namespace hpx {
         HPX_EXPORT std::size_t get_thread_data();
         HPX_EXPORT std::size_t set_thread_data(std::size_t);
 
-#ifdef HPX_HAVE_LIBCDS
+#if defined(HPX_HAVE_LIBCDS)
         HPX_EXPORT std::size_t get_libcds_data();
         HPX_EXPORT std::size_t set_libcds_data(std::size_t);
         HPX_EXPORT std::size_t get_libcds_hazard_pointer_data();
