+++ libs/coroutines/include/hpx/coroutines/coroutine.hpp
@@ -96,7 +96,7 @@ namespace hpx { namespace threads { namespace coroutines {
             return impl_.set_thread_data(data);
         }
 
-#ifdef HPX_HAVE_LIBCDS
+#if defined(HPX_HAVE_LIBCDS)
         std::size_t get_libcds_data() const
         {
             return impl_.get_libcds_data();
