+++ libs/core/execution_base/tests/unit/basic_receiver.cpp
@@ -4,8 +4,8 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#include <hpx/basic_execution/receiver.hpp>
-#include <hpx/testing.hpp>
+#include <hpx/execution_base/receiver.hpp>
+#include <hpx/modules/testing.hpp>
 
 #include <exception>
 #include <string>
@@ -181,120 +181,127 @@ namespace mylib {
 int main()
 {
     static_assert(
-        hpx::basic_execution::traits::is_receiver<mylib::receiver_1>::value,
+        hpx::execution_base::traits::is_receiver<mylib::receiver_1>::value,
         "mylib::receiver_1 should be a receiver");
-    static_assert(
-        hpx::basic_execution::traits::is_receiver_of<mylib::receiver_1,
-            int>::value,
+    static_assert(hpx::execution_base::traits::is_receiver_of<mylib::receiver_1,
+                      int>::value,
         "mylib::receiver_1 should be a receiver of an int");
     static_assert(
-        !hpx::basic_execution::traits::is_nothrow_receiver_of<mylib::receiver_1,
+        !hpx::execution_base::traits::is_nothrow_receiver_of<mylib::receiver_1,
             int>::value,
         "mylib::receiver_1 should not be a nothrow receiver of an int");
     static_assert(
-        !hpx::basic_execution::traits::is_receiver_of<mylib::receiver_1,
+        !hpx::execution_base::traits::is_receiver_of<mylib::receiver_1,
             std::string>::value,
         "mylib::receiver_1 should not be a receiver of a std::string");
 
     static_assert(
-        !hpx::basic_execution::traits::is_receiver<mylib::receiver_2>::value,
+        !hpx::execution_base::traits::is_receiver<mylib::receiver_2>::value,
         "mylib::receiver_2 should not be a receiver of std::exception_ptr");
-    static_assert(hpx::basic_execution::traits::is_receiver<mylib::receiver_2,
-                      int>::value,
+    static_assert(
+        hpx::execution_base::traits::is_receiver<mylib::receiver_2, int>::value,
         "mylib::receiver_2 should be a receiver");
     static_assert(
-        !hpx::basic_execution::traits::is_receiver_of<mylib::receiver_2,
+        !hpx::execution_base::traits::is_receiver_of<mylib::receiver_2,
             int>::value,
         "mylib::receiver_2 should not be a receiver of int");
     static_assert(
-        !hpx::basic_execution::traits::is_nothrow_receiver_of<mylib::receiver_2,
+        !hpx::execution_base::traits::is_nothrow_receiver_of<mylib::receiver_2,
             int>::value,
         "mylib::receiver_2 should not be a nothrow receiver of int");
 
     static_assert(
-        hpx::basic_execution::traits::is_receiver<mylib::receiver_1>::value,
+        hpx::execution_base::traits::is_receiver<mylib::receiver_1>::value,
         "mylib::receiver_1 should be a receiver");
-    static_assert(
-        hpx::basic_execution::traits::is_receiver_of<mylib::receiver_3,
-            int>::value,
+    static_assert(hpx::execution_base::traits::is_receiver_of<mylib::receiver_3,
+                      int>::value,
         "mylib::receiver_3 should be a receiver of an int");
     static_assert(
-        hpx::basic_execution::traits::is_nothrow_receiver_of<mylib::receiver_3,
+        hpx::execution_base::traits::is_nothrow_receiver_of<mylib::receiver_3,
             int>::value,
         "mylib::receiver_3 should be a nothrow receiver of an int");
     static_assert(
[…]
