+++ libs/core/functional/CMakeLists.txt
@@ -32,6 +32,7 @@ set(functional_headers
     hpx/functional/mem_fn.hpp
     hpx/functional/one_shot.hpp
     hpx/functional/protect.hpp
+    hpx/functional/tag_fallback_invoke.hpp
     hpx/functional/tag_invoke.hpp
     hpx/functional/unique_function.hpp
     hpx/functional/serialization/detail/serializable_basic_function.hpp
