### 2020-07-17 a16c14a
Simplify implementation of receiver related CPOs
- this is based on new tag_member and tag_member_noexcept base classes
  that take care of the special dispatching that falls back to invoking
  a tag_member_invoke function on the CPO.
- updating file locations and namespaces for recent module renamings
