+++ libs/core/execution_base/include/hpx/execution_base/receiver.hpp
@@ -1,0 +1,255 @@
+//  Copyright (c) 2020 Thomas Heller
+//  Copyright (c) 2020 Hartmut Kaiser
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+// hpxinspect:nodeprecatedname:hpx::traits::is_callable
+
+#pragma once
+
+#include <hpx/config/constexpr.hpp>
+#include <hpx/functional/tag_fallback_invoke.hpp>
+#include <hpx/functional/tag_invoke.hpp>
+
+#include <exception>
+#include <type_traits>
+#include <utility>
+
+namespace hpx { namespace execution_base {
+
+#if defined(DOXYGEN)
+    /// set_value is a customization point object. The expression
+    /// `hpx::basic_execution::set_value(r, as...)` is equivalent to:
+    ///     * `r.set_value(as...)`, if that expression is valid. If the function selected
+    ///       does not send to value(s) `as...` to the Receiver `r`'s value channel,
+    ///       the program is ill-formed (no diagnostic required).
+    ///     * Otherwise, `set_value(r, as...), if that expression is valid, with
+    ///       overload resolution performed in a context that include the declaration
+    ///       `void set_value();`
+    ///     * Otherwise, the expression is ill-formed.
+    ///
+    /// The customization is implemented in terms of `hpx::function::tag_invoke`
+    template <typename R, typename... As>
+    void set_value(R&& r, As&&... as);
+
+    /// set_done is a customization point object. The expression
+    /// `hpx::basic_execution::set_done(r)` is equivalent to:
+    ///     * `r.set_done()`, if that expression is valid. If the function selected
+    ///       does not signal the Receiver `r`'s done channel,
+    ///       the program is ill-formed (no diagnostic required).
+    ///     * Otherwise, `set_done(r), if that expression is valid, with
+    ///       overload resolution performed in a context that include the declaration
+    ///       `void set_done();`
+    ///     * Otherwise, the expression is ill-formed.
+    ///
+    /// The customization is implemented in terms of `hpx::function::tag_invoke`
+    template <typename R>
+    void set_done(R&& r);
+
+    /// set_error is a customization point object. The expression
+    /// `hpx::basic_execution::set_error(r, e)` is equivalent to:
+    ///     * `r.set_done(e)`, if that expression is valid. If the function selected
+    ///       does not send the error e the Receiver `r`'s error channel,
+    ///       the program is ill-formed (no diagnostic required).
+    ///     * Otherwise, `set_error(r, e), if that expression is valid, with
+    ///       overload resolution performed in a context that include the declaration
+    ///       `void set_error();`
+    ///     * Otherwise, the expression is ill-formed.
+    ///
+    /// The customization is implemented in terms of `hpx::function::tag_invoke`
+    template <typename R, typename E>
+    void set_error(R&& r, E&& e);
+#endif
+
+    namespace traits {
+
+        /// Receiving values from asynchronous computations is handled by the `Receiver`
+        /// concept. A `Receiver` needs to be able to receive an error or be marked as
+        /// being canceled. As such, the Receiver concept is defined by having the
+        /// following two customization points defined, which form the completion-signal
+        /// operations:
+        ///     * `hpx::basic_execution::set_done`
+        ///     * `hpx::basic_execution::set_error`
+        ///
+        /// Those two functions denote the completion-signal operations. The Receiver
+        /// contract is as follows:
+        ///     * None of a Receiver's completion-signal operation shall be invoked
+        ///       before `hpx::basic_execution::start` has been called on the operation
[…]
