+++ libs/core/functional/include/hpx/functional/tag_fallback_invoke.hpp
@@ -1,0 +1,280 @@
+//  Copyright (c) 2020 Thomas Heller
+//  Copyright (c) 2020 Hartmut Kaiser
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#pragma once
+
+#if defined(DOXYGEN)
+namespace hpx { namespace functional {
+    inline namespace unspecified {
+        /// The `hpx::functional::tag_fallback_invoke` name defines a constexpr object
+        /// that is invocable with one or more arguments. The first argument
+        /// is a 'tag' (typically a CPO). It is only invocable if an overload
+        /// of tag_invoke() that accepts the same arguments could be found via
+        /// ADL.
+        ///
+        /// The evaluation of the expression
+        /// `hpx::functional::tag_fallback_invoke(tag, args...)` is
+        /// equivalent to evaluating the unqualified call to
+        /// `tag_fallback_invoke(decay-copy(tag), std::forward<Args>(args)...)`.
+        ///
+        /// `hpx::functional::tag_fallback_invoke` is implemented against P1895.
+        ///
+        /// Example:
+        /// Defining a new customization point `foo`:
+        /// ```
+        /// namespace mylib {
+        ///     inline constexpr
+        ///         struct foo_fn final : hpx::functional::tag_fallback<foo_fn>
+        ///         {
+        ///         } foo{};
+        /// }
+        /// ```
+        ///
+        /// Defining an object `bar` which customizes `foo`:
+        /// ```
+        /// struct bar
+        /// {
+        ///     int x = 42;
+        ///
+        ///     friend constexpr int tag_fallback_invoke(mylib::foo_fn, bar const& x)
+        ///     {
+        ///         return b.x;
+        ///     }
+        /// };
+        /// ```
+        ///
+        /// Using the customization point:
+        /// ```
+        /// static_assert(42 == mylib::foo(bar{}), "The answer is 42");
+        /// ```
+        inline constexpr unspecified tag_fallback_invoke = unspecified;
+    }    // namespace unspecified
+
+    /// `hpx::functional::is_tag_fallback_invocable<Tag, Args...>` is std::true_type if
+    /// an overload of `tag_fallback_invoke(tag, args...)` can be found via ADL.
+    template <typename Tag, typename... Args>
+    struct is_tag_fallback_invocable;
+
+    /// `hpx::functional::is_tag_fallback_invocable_v<Tag, Args...>` evaluates to
+    /// `hpx::functional::is_tag_fallback_invocable<Tag, Args...>::value`
+    template <typename Tag, typename... Args>
+    constexpr bool is_tag_fallback_invocable_v =
+        is_tag_fallback_invocable<Tag, Args...>::value;
+
+    /// `hpx::functional::is_nothrow_tag_fallback_invocable<Tag, Args...>` is
+    /// std::true_type if an overload of `tag_fallback_invoke(tag, args...)` can be
+    /// found via ADL and is noexcept.
+    template <typename Tag, typename... Args>
+    struct is_nothrow_tag_fallback_invocable;
+
+    /// `hpx::functional::is_tag_fallback_invocable_v<Tag, Args...>` evaluates to
+    /// `hpx::functional::is_tag_fallback_invocable<Tag, Args...>::value`
+    template <typename Tag, typename... Args>
+    constexpr bool is_nothrow_tag_fallback_invocable_v =
+        is_nothrow_tag_fallback_invocable<Tag, Args...>::value;
+
+    /// `hpx::functional::tag_fallback_invoke_result<Tag, Args...>` is the trait
[…]
