### 2020-07-18 32cb053
Simplify implementation of operation_state CPO
