+++ libs/execution_base/tests/unit/basic_receiver.cpp
@@ -1,288 +1,0 @@
-//  Copyright (c) 2020 Thomas Heller
-//
-//  SPDX-License-Identifier: BSL-1.0
-//  Distributed under the Boost Software License, Version 1.0. (See accompanying
-//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-#include <hpx/basic_execution/receiver.hpp>
-#include <hpx/testing.hpp>
-
-#include <exception>
-#include <string>
-#include <type_traits>
-#include <utility>
-
-bool done_called = false;
-bool error_called = false;
-bool value_called = false;
-
-namespace mylib {
-    struct receiver_1
-    {
-        void set_done() noexcept
-        {
-            done_called = true;
-        }
-
-        void set_error(std::exception_ptr) noexcept
-        {
-            error_called = true;
-        }
-
-        void set_value(int)
-        {
-            value_called = true;
-        }
-    };
-
-    struct receiver_2
-    {
-        void set_done() noexcept
-        {
-            done_called = true;
-        }
-
-        void set_error(int) noexcept
-        {
-            error_called = true;
-        }
-    };
-
-    struct receiver_3
-    {
-        void set_done() noexcept
-        {
-            done_called = true;
-        }
-
-        void set_error(std::exception_ptr) noexcept
-        {
-            error_called = true;
-        }
-
-        void set_value(int) noexcept
-        {
-            value_called = true;
-        }
-    };
-
-    struct non_receiver_1
-    {
-        void set_done() & noexcept
-        {
-            done_called = true;
-        }
-
-        void set_error(std::exception_ptr) noexcept
-        {
-            error_called = true;
-        }
-
-        void set_value(int)
-        {
-            value_called = true;
-        }
-    };
-
-    struct non_receiver_2
-    {
-        void set_done() noexcept
-        {
-            done_called = true;
-        }
-
-        void set_error(std::exception_ptr) & noexcept
-        {
-            error_called = true;
-        }
-
-        void set_value(int)
-        {
-            value_called = true;
-        }
-    };
-
-    struct non_receiver_3
-    {
-        void set_done() & noexcept
-        {
-            done_called = true;
-        }
-
-        void set_error(std::exception_ptr) & noexcept
-        {
-            error_called = true;
-        }
-
-        void set_value(int)
-        {
-            value_called = true;
-        }
-    };
-
-    struct non_receiver_4
-    {
-        void set_done() noexcept
-        {
-            done_called = true;
-        }
-
-        void set_error(std::exception_ptr) noexcept
-        {
-            error_called = true;
-        }
-
-        void set_value(int) &
-        {
-            value_called = true;
-        }
-    };
-
-    struct non_receiver_5
-    {
-        void set_done()
-        {
-            done_called = true;
-        }
-
-        void set_error(std::exception_ptr) noexcept
-        {
-            error_called = true;
-        }
-    };
-
-    struct non_receiver_6
-    {
-        void set_done() noexcept
-        {
-            done_called = true;
-        }
-
-        void set_error(std::exception_ptr)
-        {
-            error_called = true;
-        }
-    };
-
-    struct non_receiver_7
-    {
-        void set_done()
-        {
-            done_called = true;
-        }
-
-        void set_error(std::exception_ptr)
-        {
-            error_called = true;
-        }
-    };
-}    // namespace mylib
-
-int main()
-{
-    static_assert(hpx::basic_execution::is_receiver<mylib::receiver_1>::value,
[…]
