+++ libs/basic_execution/include/hpx/basic_execution/receiver.hpp
@@ -1,369 +1,0 @@
-//  Copyright (c) 2020 Thomas Heller
-//
-//  SPDX-License-Identifier: BSL-1.0
-//  Distributed under the Boost Software License, Version 1.0. (See accompanying
-//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-// hpxinspect:nodeprecatedname:hpx::traits::is_callable
-
-#pragma once
-
-#include <hpx/config/constexpr.hpp>
-#include <hpx/functional/tag_invoke.hpp>
-
-#include <exception>
-#include <type_traits>
-#include <utility>
-
-namespace hpx { namespace basic_execution {
-#if defined(DOXYGEN)
-    /// set_value is a customization point object. The expression
-    /// `hpx::basic_execution::set_value(r, as...)` is equivalent to:
-    ///     * `r.set_value(as...)`, if that expression is valid. If the function selected
-    ///       does not send to value(s) `as...` to the receiver `r`'s value channel,
-    ///       the program is ill-formed (no diagnostic required).
-    ///     * Otherwise, `set_value(r, as...), if that expression is valid, with
-    ///       overload resolution performed in a context that include the declaration
-    ///       `void set_value();`
-    ///     * Otherwise, the expression is ill-formed.
-    ///
-    /// The customization is implemented in terms of `hpx::function::tag_invoke`
-    template <typename R, typename... As>
-    void set_value(R&& r, As&&... as);
-
-    /// set_done is a customization point object. The expression
-    /// `hpx::basic_execution::set_done(r)` is equivalent to:
-    ///     * `r.set_done()`, if that expression is valid. If the function selected
-    ///       does not signal the receiver `r`'s done channel,
-    ///       the program is ill-formed (no diagnostic required).
-    ///     * Otherwise, `set_done(r), if that expression is valid, with
-    ///       overload resolution performed in a context that include the declaration
-    ///       `void set_done();`
-    ///     * Otherwise, the expression is ill-formed.
-    ///
-    /// The customization is implemented in terms of `hpx::function::tag_invoke`
-    template <typename R>
-    void set_done(R&& r);
-
-    /// set_error is a customization point object. The expression
-    /// `hpx::basic_execution::set_error(r, e)` is equivalent to:
-    ///     * `r.set_done(e)`, if that expression is valid. If the function selected
-    ///       does not send the error e the receiver `r`'s error channel,
-    ///       the program is ill-formed (no diagnostic required).
-    ///     * Otherwise, `set_error(r, e), if that expression is valid, with
-    ///       overload resolution performed in a context that include the declaration
-    ///       `void set_error();`
-    ///     * Otherwise, the expression is ill-formed.
-    ///
-    /// The customization is implemented in terms of `hpx::function::tag_invoke`
-    template <typename R, typename E>
-    void set_error(R&& r, E&& e);
-#endif
-
-    namespace traits {
-    /// Receiving values from asynchronous computations is handled by the `receiver`
-    /// concept. A `receiver` needs to be able to receive an error or be marked as
-    /// being cancelled. As such, the Receiver concept is defined by having the
-    /// following two customization points defined, which form the completion-signal
[…]
