### 2020-07-18 e3abd82
Simplify implementation of operation_state CPO
