### 2020-07-18 19d0ee2
Simplify implementation of operation_state CPO
