### 2020-07-18 0d0ee62
Simplify implementation of sender CPO
