+++ libs/core/execution_base/include/hpx/execution_base/operation_state.hpp
@@ -60,7 +60,7 @@ namespace hpx { namespace execution_base {
         {
             return std::forward<OperationState>(o).start();
         }
-    } start;
+    } start{};
 
     namespace traits {
         namespace detail {
