+++ libs/core/execution_base/CMakeLists.txt
@@ -16,6 +16,7 @@ set(execution_base_headers
     hpx/execution_base/receiver.hpp
     hpx/execution_base/register_locks.hpp
     hpx/execution_base/resource_base.hpp
+    hpx/execution_base/sender.hpp
     hpx/execution_base/this_thread.hpp
 )
 

[…]
