### 2020-07-19 ce9644c
Fixing TSS problems during thread termination
