+++ libs/coroutines/src/detail/coroutine_impl.cpp
@@ -74,9 +74,6 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
                     result_last = m_fun(*this->args());
                     HPX_ASSERT(
                         result_last.first == thread_state_enum::terminated);
-
-                    // Reset early as the destructors may still yield.
-                    this->reset();
                 }
                 catch (...)
                 {
@@ -84,7 +81,9 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
                     tinfo = std::current_exception();
                 }
 
+                // Reset early as the destructors may still yield.
                 this->reset_tss();
+                this->reset();
 
                 // return value to other side of the fence
                 this->bind_result(result_last);
