### 2020-07-19 fe1a1b5
Fixing clang-format issues
