+++ libs/include/include/hpx/include/parallel_reduce.hpp
@@ -9,8 +9,8 @@
 
 #include <hpx/config.hpp>
 #include <hpx/parallel/algorithms/reduce.hpp>
-#include <hpx/parallel/container_algorithms/reduce.hpp>
 #include <hpx/parallel/algorithms/reduce_by_key.hpp>
+#include <hpx/parallel/container_algorithms/reduce.hpp>
 
 #if defined(HPX_HAVE_DISTRIBUTED_RUNTIME)
 #include <hpx/parallel/segmented_algorithms/reduce.hpp>
