+++ libs/segmented_algorithms/include/hpx/parallel/segmented_algorithms/reduce.hpp
@@ -191,8 +191,8 @@ namespace hpx { namespace parallel { inline namespace v1 {
         // segmented implementation
         template <typename ExPolicy, typename InIterB, typename InIterE,
             typename T, typename F>
-        typename util::detail::algorithm_result<ExPolicy, T>::type
-        reduce_(ExPolicy&& policy, InIterB first, InIterE last, T init, F&& f,
+        typename util::detail::algorithm_result<ExPolicy, T>::type reduce_(
+            ExPolicy&& policy, InIterB first, InIterE last, T init, F&& f,
             std::true_type)
         {
             typedef parallel::execution::is_sequenced_execution_policy<ExPolicy>
@@ -213,8 +213,8 @@ namespace hpx { namespace parallel { inline namespace v1 {
         // forward declare the non-segmented version of this algorithm
         template <typename ExPolicy, typename InIterB, typename InIterE,
             typename T, typename F>
-        typename util::detail::algorithm_result<ExPolicy, T>::type
-        reduce_(ExPolicy&& policy, InIterB first, InIterE last, T init, F&& f,
+        typename util::detail::algorithm_result<ExPolicy, T>::type reduce_(
+            ExPolicy&& policy, InIterB first, InIterE last, T init, F&& f,
             std::false_type);
 
         /// \endcond
