### 2020-07-20 9e18313
cleanup #ifdefs around code needing distributed runtime
