+++ libs/async_cuda/src/get_targets.cpp
@@ -11,15 +11,15 @@
 #include <hpx/modules/errors.hpp>
 #include <hpx/runtime/threads/thread_helpers.hpp>
 #include <hpx/runtime_fwd.hpp>
-#include <hpx/serialization/serialize.hpp>
-#include <hpx/serialization/vector.hpp>
 #include <hpx/threading_base/thread_helpers.hpp>
 
 #if defined(HPX_HAVE_DISTRIBUTED_RUNTIME)
-#include <hpx/modules/async_distributed.hpp>
 #include <hpx/actions_base/plain_action.hpp>
+#include <hpx/modules/async_distributed.hpp>
 #include <hpx/runtime/find_here.hpp>
 #include <hpx/runtime/naming/id_type.hpp>
+#include <hpx/serialization/serialize.hpp>
+#include <hpx/serialization/vector.hpp>
 #endif
 
 #include <hpx/async_cuda/target.hpp>
