### 2020-07-20 0fe6211
Simplify typedefs slightly with using statement
