### 2020-07-20 1109820
Rearrange public API headers
