### 2020-07-20 d55d311
Remove dependency on Boost.Assign
