+++ libs/synchronization/tests/unit/shared_mutex/shared_mutex2.cpp
@@ -288,34 +288,6 @@ void test_if_other_thread_has_shared_lock_try_lock_upgrade_returns_true()
     writer.join();
 }
 
-void test_if_other_thread_has_upgrade_lock_try_lock_upgrade_returns_false()
-{
-    typedef hpx::lcos::local::shared_mutex shared_mutex_type;
-    typedef hpx::lcos::local::mutex mutex_type;
-
-    shared_mutex_type rw_mutex;
-    mutex_type finish_mutex;
-    mutex_type unblocked_mutex;
-    unsigned unblocked_count = 0;
-    std::unique_lock<mutex_type> finish_lock(finish_mutex);
-    hpx::thread writer(test::simple_upgrade_thread(
-        rw_mutex, finish_mutex, unblocked_mutex, unblocked_count));
-
-    hpx::this_thread::sleep_for(std::chrono::seconds(1));
-
-    CHECK_LOCKED_VALUE_EQUAL(unblocked_mutex, unblocked_count, 1u);
-
-    bool const try_succeeded = rw_mutex.try_lock_upgrade();
-    HPX_TEST(!try_succeeded);
-    if (try_succeeded)
-    {
-        rw_mutex.unlock_upgrade();
-    }
-
-    finish_lock.unlock();
-    writer.join();
-}
-
 ///////////////////////////////////////////////////////////////////////////////
 int hpx_main()
 {
@@ -328,7 +300,6 @@ int hpx_main()
     test_if_no_thread_has_lock_try_lock_upgrade_returns_true();
     test_if_other_thread_has_shared_lock_try_lock_shared_returns_true();
     test_if_other_thread_has_shared_lock_try_lock_upgrade_returns_true();
-    test_if_other_thread_has_upgrade_lock_try_lock_upgrade_returns_false();
 
     return hpx::finalize();
 }
