+++ libs/synchronization/tests/unit/shared_mutex/shared_mutex1.cpp
@@ -13,10 +13,9 @@
 
 #include <hpx/modules/testing.hpp>
 
-#include <boost/thread/locks.hpp>
-
 #include <chrono>
 #include <mutex>
+#include <shared_mutex>
 #include <string>
 #include <vector>
 
@@ -52,7 +51,7 @@ void test_multiple_readers()
         for (unsigned i = 0; i != number_of_threads; ++i)
         {
             pool.create_thread(
-                test::locking_thread<boost::shared_lock<shared_mutex_type>>(
+                test::locking_thread<std::shared_lock<shared_mutex_type>>(
                     rw_mutex, unblocked_count, unblocked_count_mutex,
                     unblocked_condition, finish_mutex,
                     simultaneous_running_count, max_simultaneous_running));
