### 2020-07-21 f46e311
Update release procedure notes
- Remove the outdated steps
- Add an easier way to update the release branch
- Replace hpxMP (deprecated) with libCDS
- Explain more in details certain steps
