### 2020-07-21 3c7173a
Use Linux syscall gettid for PAPI counters
PAPI_attach_event only works with gettid on Linux. Add checks to CMake
configuration to only enable HPX_WITH_PAPI on Linux.
