### 2020-07-21 f785c13
Ignore lock in PAPI performance counters in case of error
