### 2020-07-21 f5b3158
Add general notes for the 1.5.0 release
