+++ docs/sphinx/releases/whats_new_1_5_0.rst
@@ -1,5 +1,6 @@
 ..
     Copyright (C) 2007-2020 Hartmut Kaiser
+    Copyright (C)      2020 ETH Zurich
 
     SPDX-License-Identifier: BSL-1.0
     Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -8,17 +9,570 @@
 .. _hpx_1_5_0:
 
 ===========================
-|hpx| V1.5.0 (TBD)
+|hpx| V1.5.0 (?)
 ===========================
 
 General changes
 ===============
 
+The main focus of this release is on APIs and C++20 conformance. We have added
+many new C++20 features and adapted multiple algorithms to be fully C++20
+conformant. As part of the modularization we have begun specifying the public
+API of |hpx| in terms of headers and functionality, and aligning it more closely
+to the C++ standard. All non-distributed modules are now in place, along with an
+experimental option to completely disable distributed features in |hpx|. We have
+also added an experimental asynchronous MPI executor. Lastly this release
+contains official |cmake| targets, performance improvements, and many bug fixes.
+
+* We have added the C++20 features ``hpx::jthread`` and ``hpx::stop_token``. We
+  have also added ``hpx::stable_sort`` based on Francisco Tapia's
+  implementation.
+* We have adapted existing sychronization primitives to be fully conformant
+  C++20: ``barrier``, ``latch``, ``counting_semaphore``.
+* We have started using customization point objects to make the corresponding
+  algorithms fully conformant to C++20 as well as to make algorithm extension
+  easier for the user. ``all_of``/``any_of``/``none_of``, ``copy``, ``count``,
+  ``destroy``, ``equal``, ``fill``, ``mismatch``, ``move``, ``reduce`` are using
+  those CPOs.  We also have adapted their corresponding ``hpx::ranges`` versions
+  in this release.
+* We have adapted support for ``co_await`` to C++20, it now supports
+  ``shared_future`` and have added allocator support for futures returned by
+  ``co_return``. It is no longer in the ``experimental`` namespace.
+* We added serialization support for ``std::variant`` and ``std::tuple``.
+* ``result_of`` and ``is_callable`` are now deprecated and replaced by
+  ``invoke_result`` and ``is_invocable`` to conform to C++20.
+* We continued with the modularization, making it easier for us to add the new
+  experimental ``HPX_WITH_DISTRIBUTED_RUNTIME`` |cmake| option (see below) . An
+  important amount of headers have been deprecated. We adapted the namespaces
+  and headers we could to be closer to the standard ones (:ref:`api
+  <_public_api>`).
+* It is now possible to have a basic CUDA support including a helper function to
+  get a future from a CUDA stream and target handling. They are available under
+  the ``hpx::cuda::experimental`` namespace and they can be enabled with the
+  ``HPX_WITH_ASYNC_CUDA=ON`` cmake option.
+* We added a new ``hpx::mpi::experimental`` namespace for getting futures from
+  an asynchronous MPI call and a new minimal MPI executor (add the name).
+* A polymorphic executor has been implemented to reduce compile times as a
+  function accepting executor can potentially be instantiated only once instead
+  of multiple times with different executors. It accepts the function signature
+  as a template argument. It needs to be constructed from any other executor.
[…]
