### 2020-07-21 3396dcb
Use CMAKE_INSTALL_LIBDIR for component_path_suffixes
