+++ CMakeLists.txt
@@ -67,8 +67,8 @@ endif()
 set(HPX_VERSION_MAJOR 1)
 set(HPX_VERSION_MINOR 5)
 set(HPX_VERSION_SUBMINOR 0)
-set(HPX_VERSION_DATE 20200212)
-set(HPX_VERSION_TAG "-trunk")
+set(HPX_VERSION_DATE 20200723)
+set(HPX_VERSION_TAG "-rc1")
 
 set(HPX_VERSION
     "${HPX_VERSION_MAJOR}.${HPX_VERSION_MINOR}.${HPX_VERSION_SUBMINOR}"
