### 2020-07-21 7325686
Remove endian utilities
