+++ hpx/util/integer/cover_operators.hpp
@@ -1,101 +1,0 @@
-//  boost/integer/cover_operators.hpp ----------------------------------------//
-
-//  (C) Copyright Darin Adler 2000
-
-//  SPDX-License-Identifier: BSL-1.0
-//  Distributed under the Boost Software License, Version 1.0. (See accompanying
-//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-//----------------------------------------------------------------------------//
-
-#pragma once
-
-#include <iosfwd>
-
-namespace hpx { namespace util
-{
-  namespace integer
-  {
-
-  // A class that adds integer operators to an integer cover class
-
-    template <typename T, typename IntegerType>
-    class cover_operators
-    {
-      // The other operations take advantage of the type conversion that's
-      // built into unary +.
-
-      // Unary operations.
-      friend IntegerType operator+(const T& x) { return x; }
-      friend IntegerType operator-(const T& x) { return -+x; }
-      friend IntegerType operator~(const T& x) { return ~+x; }
-      friend IntegerType operator!(const T& x) { return !+x; }
-
-      // The basic ordering operations.
-      friend bool operator==(const T& x, IntegerType y) { return +x == y; }
-      friend bool operator!=(const T& x, IntegerType y) { return +x != y; }
-      friend bool operator<(const T& x, IntegerType y) { return +x < y; }
-      friend bool operator>(const T& x, IntegerType y) { return +x > y; }
-      friend bool operator<=(const T& x, IntegerType y) { return +x <= y; }
-      friend bool operator>=(const T& x, IntegerType y) { return +x >= y; }
-
-      // The basic arithmetic operations.
-      friend T& operator+=(T& x, IntegerType y) { return x = +x + y; }
-      friend T& operator-=(T& x, IntegerType y) { return x = +x - y; }
-      friend T& operator*=(T& x, IntegerType y) { return x = +x * y; }
-      friend T& operator/=(T& x, IntegerType y) { return x = +x / y; }
-      friend T& operator%=(T& x, IntegerType y) { return x = +x % y; }
-      friend T& operator&=(T& x, IntegerType y) { return x = +x & y; }
-      friend T& operator|=(T& x, IntegerType y) { return x = +x | y; }
-      friend T& operator^=(T& x, IntegerType y) { return x = +x ^ y; }
-      friend T& operator<<=(T& x, IntegerType y) { return x = +x << y; }
-      friend T& operator>>=(T& x, IntegerType y) { return x = +x >> y; }
-
-      friend IntegerType operator+(T const& x, IntegerType y) { return +x + y; }
-      friend IntegerType operator-(T const& x, IntegerType y) { return +x - y; }
-      friend IntegerType operator*(T const& x, IntegerType y) { return +x * y; }
-      friend IntegerType operator/(T const& x, IntegerType y) { return +x / y; }
-      friend IntegerType operator%(T const& x, IntegerType y) { return +x % y; }
-      friend IntegerType operator&(T const& x, IntegerType y) { return +x & y; }
-      friend IntegerType operator|(T const& x, IntegerType y) { return +x | y; }
-      friend IntegerType operator^(T const& x, IntegerType y) { return +x ^ y; }
-
-      // A few binary arithmetic operations not covered by operators base class.
-      friend IntegerType operator<<(const T& x, IntegerType y) { return +x << y; }
-      friend IntegerType operator>>(const T& x, IntegerType y) { return +x >> y; }
-
-      // Auto-increment and auto-decrement can be defined in terms of the
-      // arithmetic operations.
-      friend T& operator++(T& x) { return x += 1; }
-      friend T operator++(T& x, int)
-      {
-          T tmp(x);
[…]
