### 2020-07-21 8add246
Updating Codacy badge
