### 2020-07-21 7bb0b15
More Codacy badge updates
