### 2020-07-21 9945b68
Fixing 'error prone' constructs as reported by Codacy
