### 2020-07-21 62a6ec3
Fixing performance' issues flagged by codacy
