### 2020-07-21 ac016de
Fixing some Python problems as reported by codacy
