### 2020-07-21 553dcea
Fix CUDA compilation with distributed runtime off
