### 2020-07-21 1353c5d
Add API headers for barrier, channel, and latch
