+++ libs/collectives/include/hpx/distributed/barrier.hpp
@@ -1,0 +1,15 @@
+//  Copyright (c) 2020 ETH Zurich
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+/// \file hpx/collectives/barrier.hpp
+
+#pragma once
+
+#include <hpx/collectives/barrier.hpp>
+
+namespace hpx { namespace distributed {
+    using hpx::lcos::barrier;
+}}    // namespace hpx::distributed
