### 2020-07-21 23d5569
Remove obsolete polling mode flag and set/get calls
