+++ libs/async_cuda/src/cuda_future.cpp
@@ -244,23 +244,16 @@ namespace hpx { namespace cuda { namespace experimental { namespace detail {
     // -------------------------------------------------------------
     void register_polling(hpx::threads::thread_pool_base& pool)
     {
+        cud_debug.debug(debug::str<>("enable polling"));
         auto* sched = pool.get_scheduler();
-
-        cud_debug.debug(debug::str<>("Setting mode"), "enable_user_polling");
-
-        // always set polling function before enabling polling
-        sched->set_cuda_polling_function(
-            &hpx::cuda::experimental::detail::poll);
-        sched->add_remove_scheduler_mode(threads::policies::enable_user_polling,
-            threads::policies::do_background_work);
+        sched->set_cuda_polling_function(&hpx::cuda::experimental::detail::poll);
     }
 
     // -------------------------------------------------------------
     void unregister_polling(hpx::threads::thread_pool_base& pool)
     {
-        cud_debug.debug(debug::str<>("Setting mode"), "disable_user_polling");
+        cud_debug.debug(debug::str<>("disable polling"));
         auto* sched = pool.get_scheduler();
-        sched->remove_scheduler_mode(threads::policies::enable_user_polling);
         sched->clear_cuda_polling_function();
     }
 
