+++ libs/threading_base/include/hpx/threading_base/scheduler_mode.hpp
@@ -58,9 +58,6 @@ namespace hpx { namespace threads { namespace policies {
         /// This option allows for certain schedulers to explicitly disable
         /// exponential idle-back off
         enable_idle_backoff = 0x0800,
-        /// this option tells the scheduler to call a user supplied function
-        /// on each iteration (after execution of each task on each worker thread)
-        enable_user_polling = 0x1000,
 
         // clang-format off
         /// This option represents the default mode.
@@ -86,8 +83,7 @@ namespace hpx { namespace threads { namespace policies {
             assign_work_thread_parent |
             steal_high_priority_first |
             steal_after_local |
-            enable_idle_backoff |
-            enable_user_polling
+            enable_idle_backoff
         // clang-format on
     };
 }}}    // namespace hpx::threads::policies
