+++ libs/async_mpi/src/mpi_future.cpp
@@ -307,28 +307,19 @@ namespace hpx { namespace mpi { namespace experimental {
     }
 
     namespace detail {
-
         // -------------------------------------------------------------
         void register_polling(hpx::threads::thread_pool_base& pool)
         {
+            mpi_debug.debug(debug::str<>("enable polling"));
             auto* sched = pool.get_scheduler();
-
-            mpi_debug.debug(debug::str<>("Setting mode"),
-                detail::get_mpi_info(), "enable_user_polling");
-
-            // always set polling function before enabling polling
             sched->set_mpi_polling_function(&hpx::mpi::experimental::poll);
-            sched->add_remove_scheduler_mode(
-                threads::policies::enable_user_polling,
-                threads::policies::do_background_work);
         }
 
         // -------------------------------------------------------------
         void unregister_polling(hpx::threads::thread_pool_base& pool)
         {
-            mpi_debug.debug(debug::str<>("Setting mode"), "disable_user_polling");
+            mpi_debug.debug(debug::str<>("disable polling"));
             auto* sched = pool.get_scheduler();
-            sched->remove_scheduler_mode(threads::policies::enable_user_polling);
             sched->clear_mpi_polling_function();
         }
     }    // namespace detail
