### 2020-07-21 ea21ef3
Fixing unused variable warnings as reported by Codacy
