+++ tests/unit/agas/refcnted_symbol_to_remote_object.cpp
@@ -97,6 +97,7 @@ void hpx_test_main(
             HPX_TEST_EQ(false, monitor.is_ready(milliseconds(delay)));
 
             gid = id.get_gid();
+            (void) gid;
 
             // let id go out of scope
         }
