+++ tests/unit/lcos/await.cpp
@@ -37,6 +37,7 @@ hpx::future<int> test1()
 hpx::future<int> test2()
 {
     auto result = co_await hpx::make_ready_future(42);
+    (void) result;
     co_return result;
 }
 
@@ -44,6 +45,7 @@ hpx::future<int> test3()
 {
     int local_variable[128] = {42};      // large local variable
     auto result = co_await hpx::make_ready_future(local_variable[0]);
+    (void) result;
     co_return result;
 }
 
@@ -55,6 +57,7 @@ hpx::future<int> async_test1()
 hpx::future<int> async_test2()
 {
     auto result = co_await hpx::async(just_wait, 42);
+    (void) result;
     co_return result;
 }
 
@@ -62,6 +65,7 @@ hpx::future<int> async_test3()
 {
     int local_variable[128] = {42};      // large local variable
     auto result = co_await hpx::async(just_wait, local_variable[0]);
+    (void) result;
     co_return result;
 }
 
@@ -81,6 +85,7 @@ hpx::future<int> fib1(int n)
 {
     if (n >= 2)
         n = co_await fib1(n - 1) + co_await fib1(n - 2);
+    (void) n;
     co_return n;
 }
 
@@ -88,6 +93,7 @@ hpx::future<int> fib2(int n)
 {
     if (n >= 2)
         n = co_await hpx::async(&fib2, n - 1) + co_await fib2(n - 2);
+    (void) n;
     co_return n;
 }
 
@@ -106,6 +112,7 @@ hpx::shared_future<int> shared_test1()
 hpx::shared_future<int> shared_test2()
 {
     auto result = co_await hpx::make_ready_future(42);
+    (void) result;
     co_return result;
 }
 
@@ -113,6 +120,7 @@ hpx::shared_future<int> shared_test3()
 {
     int local_variable[128] = {42};      // large local variable
     auto result = co_await hpx::make_ready_future(local_variable[0]);
+    (void) result;
     co_return result;
 }
 
@@ -124,6 +132,7 @@ hpx::shared_future<int> async_shared_test1()
 hpx::shared_future<int> async_shared_test2()
 {
     auto result = co_await hpx::async(just_wait, 42);
+    (void) result;
     co_return result;
 }
 
@@ -131,6 +140,7 @@ hpx::shared_future<int> async_shared_test3()
 {
     int local_variable[128] = {42};      // large local variable
     auto result = co_await hpx::async(just_wait, local_variable[0]);
+    (void) result;
     co_return result;
 }
 
@@ -150,6 +160,7 @@ hpx::shared_future<int> shared_fib1(int n)
 {
     if (n >= 2)
         n = co_await shared_fib1(n - 1) + co_await shared_fib1(n - 2);
+    (void) n;
     co_return n;
 }
 
@@ -157,6 +168,7 @@ hpx::shared_future<int> shared_fib2(int n)
 {
     if (n >= 2)
         n = co_await hpx::async(&fib2, n - 1) + co_await shared_fib2(n - 2);
+    (void) n;
     co_return n;
 }
 
