+++ tests/regressions/component/client_base_registration.cpp
@@ -77,6 +77,8 @@ void test_client_registration()
 
         HPX_TEST_EQ(42, c.call());
     }
+
+    (void) c_outer;
 }
 
 ///////////////////////////////////////////////////////////////////////////////
