### 2020-07-21 02c6313
Fix snake_case type and clang-format
