### 2020-07-22 769cfee
Put back atomic load/store of polling functions
