### 2020-07-22 7225b01
Add lib/hpx directory to component path suffixes for build directory
