### 2020-07-22 4e6a8d1
Add missing header for std::terminate
