+++ plugins/parcelport/verbs/rdma/verbs_shared_receive_queue.hpp
@@ -12,6 +12,7 @@
 #include <rdma/rdma_cma.h>
 #include <infiniband/verbs.h>
 //
+#include <exception>
 #include <iostream>
 #include <memory>
 
