### 2020-07-22 76b4779
Cloning upgrade_lock from Boost
