### 2020-07-22 2e1787d
Restoring broken MRU cache test
