+++ libs/cache/include/hpx/cache/entries/entry.hpp
@@ -29,6 +29,25 @@ namespace hpx { namespace util { namespace cache { namespace entries {
         {
             typedef entry<Value> type;
         };
+
+        template <typename Derived>
+        struct less_than_comparable
+        {
+            friend bool operator>(Derived const& lhs, Derived const& rhs)
+            {
+                return rhs < lhs;
+            }
+
+            friend bool operator<=(Derived const& lhs, Derived const& rhs)
+            {
+                return !(rhs < lhs);
+            }
+
+            friend bool operator>=(Derived const& lhs, Derived const& rhs)
+            {
+                return !(lhs < rhs);
+            }
+        };
     }    // namespace detail
 
     ///////////////////////////////////////////////////////////////////////////
@@ -42,13 +61,15 @@ namespace hpx { namespace util { namespace cache { namespace entries {
     ///
     template <typename Value, typename Derived>
     class entry
+      : detail::less_than_comparable<
+            typename detail::derived<Value, Derived>::type>
     {
     public:
         typedef Value value_type;
 
     public:
         /// \brief Any cache entry has to be default constructible
-        entry() {}
+        entry() = default;
 
         /// \brief Construct a new instance of a cache entry holding the given
         ///        value.
@@ -117,31 +138,6 @@ namespace hpx { namespace util { namespace cache { namespace entries {
         friend bool operator<(entry const& lhs, entry const& rhs)
         {
             return lhs.value_ < rhs.value_;
-        }
-
-        friend bool operator>(entry const& lhs, entry const& rhs)
-        {
-            return lhs.value_ > rhs.value_;
-        }
-
-        friend bool operator<=(entry const& lhs, entry const& rhs)
-        {
-            return lhs.value_ <= rhs.value_;
-        }
-
-        friend bool operator>=(entry const& lhs, entry const& rhs)
-        {
-            return lhs.value_ >= rhs.value_;
-        }
-
-        friend bool operator==(entry const& lhs, entry const& rhs)
-        {
-            return lhs.value_ == rhs.value_;
-        }
-
-        friend bool operator!=(entry const& lhs, entry const& rhs)
-        {
-            return lhs.value_ != rhs.value_;
         }
 
         /// \brief Get a reference to the stored data value
