+++ libs/resiliency/tests/performance/replicate/1d_stencil_replicate_checksum.cpp
@@ -236,7 +236,7 @@ struct stepper
         std::uint64_t nd, std::uint64_t n_value, double error,
         hpx::lcos::local::sliding_semaphore& sem)
     {
-        using hpx::resiliency::dataflow_replicate_validate;
+        using hpx::experimental::dataflow_replicate_validate;
         using hpx::util::unwrapping;
 
         // U[t][i] is the state of position i at time t.
