+++ libs/resiliency/include/hpx/resiliency/dataflow_replay.hpp
@@ -1,6 +1,6 @@
 //  Copyright (c) 2019 National Technology & Engineering Solutions of Sandia,
 //                     LLC (NTESS).
-//  Copyright (c) 2018-2019 Hartmut Kaiser
+//  Copyright (c) 2018-2020 Hartmut Kaiser
 //  Copyright (c) 2018-2019 Adrian Serio
 //
 //  SPDX-License-Identifier: BSL-1.0
@@ -11,6 +11,7 @@
 
 #include <hpx/resiliency/config.hpp>
 #include <hpx/resiliency/async_replay.hpp>
+#include <hpx/resiliency/resiliency_cpos.hpp>
 
 #include <hpx/futures/future.hpp>
 #include <hpx/modules/async_local.hpp>
@@ -20,16 +21,17 @@
 
 namespace hpx { namespace resiliency {
 
-    /// Asynchronously launch given function \a f. Verify the result of
-    /// those invocations using the given predicate \a pred. Repeat launching
-    /// on error exactly \a n times.
-    ///
-    /// Delay the invocation of \a f if any of the arguments to \a f are
-    /// futures.
+    // Asynchronously launch given function \a f. Verify the result of
+    // those invocations using the given predicate \a pred. Repeat launching
+    // on error exactly \a n times.
+    //
+    // Delay the invocation of \a f if any of the arguments to \a f are
+    // futures.
     template <typename Pred, typename F, typename... Ts>
     hpx::future<
         typename hpx::util::detail::invoke_deferred_result<F, Ts...>::type>
-    dataflow_replay_validate(std::size_t n, Pred&& pred, F&& f, Ts&&... ts)
+    tag_invoke(dataflow_replay_validate_t, std::size_t n, Pred&& pred, F&& f,
+        Ts&&... ts)
     {
         return hpx::dataflow(
             hpx::resiliency::functional::async_replay_validate{}, n,
@@ -37,15 +39,15 @@ namespace hpx { namespace resiliency {
             std::forward<Ts>(ts)...);
     }
 
-    /// Asynchronously launch given function \a f. Repeat launching on error
-    /// exactly \a n times.
-    ///
-    /// Delay the invocation of \a f if any of the arguments to \a f are
-    /// futures.
+    // Asynchronously launch given function \a f. Repeat launching on error
+    // exactly \a n times.
+    //
+    // Delay the invocation of \a f if any of the arguments to \a f are
+    // futures.
     template <typename F, typename... Ts>
     hpx::future<
         typename hpx::util::detail::invoke_deferred_result<F, Ts...>::type>
-    dataflow_replay(std::size_t n, F&& f, Ts&&... ts)
+    tag_invoke(dataflow_replay_t, std::size_t n, F&& f, Ts&&... ts)
     {
         return hpx::dataflow(hpx::resiliency::functional::async_replay{}, n,
             std::forward<F>(f), std::forward<Ts>(ts)...);
