+++ libs/resiliency/tests/performance/replay/1d_stencil_checksum.cpp
@@ -237,7 +237,7 @@ struct stepper
         std::uint64_t nd, std::uint64_t n_value, double error,
         hpx::lcos::local::sliding_semaphore& sem)
     {
-        using hpx::resiliency::dataflow_replay_validate;
+        using hpx::experimental::dataflow_replay_validate;
         using hpx::util::unwrapping;
 
         // U[t][i] is the state of position i at time t.
