+++ libs/resiliency/tests/performance/replicate/async_replicate_vote.cpp
@@ -98,7 +98,7 @@ int hpx_main(hpx::program_options::variables_map& vm)
 
         for (int i = 0; i < num_iterations; ++i)
         {
-            hpx::future<int> f = hpx::resiliency::async_replicate_validate(
+            hpx::future<int> f = hpx::experimental::async_replicate_validate(
                 n, &validate, &universal_ans, delay * 1000, error);
             vect.push_back(std::move(f));
         }
