+++ libs/resiliency/tests/performance/replicate/async_replicate_vote_validate.cpp
@@ -95,7 +95,7 @@ int hpx_main(hpx::program_options::variables_map& vm)
 
         for (int i = 0; i < num_iterations; ++i)
         {
-            hpx::future<int> f = hpx::resiliency::async_replicate_vote(
+            hpx::future<int> f = hpx::experimental::async_replicate_vote(
                 n, &vote, &universal_ans, delay * 1000, error);
             vect.push_back(std::move(f));
         }
