### 2020-07-22 8302b9e
Fixing inspect errors
