+++ libs/synchronization/tests/unit/shared_mutex/shared_mutex1.cpp
@@ -150,9 +150,9 @@ void test_reader_blocks_writer()
     try
     {
         pool.create_thread(
-            test::locking_thread<std::shared_lock<shared_mutex_type>>(
-                rw_mutex, unblocked_count, unblocked_count_mutex,
-                unblocked_condition, finish_mutex, simultaneous_running_count,
+            test::locking_thread<std::shared_lock<shared_mutex_type>>(rw_mutex,
+                unblocked_count, unblocked_count_mutex, unblocked_condition,
+                finish_mutex, simultaneous_running_count,
                 max_simultaneous_running));
 
         {
