### 2020-07-22 158bfff
Remove functional helpers as those are not needed anymore
