+++ libs/resiliency/include/hpx/resiliency/async_replicate.hpp
@@ -199,65 +199,4 @@ namespace hpx { namespace resiliency {
             detail::replicate_voter{}, detail::replicate_validator{},
             std::forward<F>(f), std::forward<Ts>(ts)...);
     }
-
-    ///////////////////////////////////////////////////////////////////////////
-    /// Functional version of \a hpx::experimental::async_replicate_validate and
-    /// \a hpx::experimental::async_replicate
-    namespace functional {
-
-        struct async_replicate_vote_validate
-        {
-            template <typename Vote, typename Pred, typename F, typename... Ts>
-            auto operator()(std::size_t n, Vote&& vote, Pred&& pred, F&& f,
-                Ts&&... ts) const
-                -> decltype(hpx::resiliency::async_replicate_vote_validate(n,
-                    std::forward<Vote>(vote), std::forward<Pred>(pred),
-                    std::forward<F>(f), std::forward<Ts>(ts)...))
-            {
-                return hpx::resiliency::async_replicate_vote_validate(n,
-                    std::forward<Vote>(vote), std::forward<Pred>(pred),
-                    std::forward<F>(f), std::forward<Ts>(ts)...);
-            }
-        };
-
-        struct async_replicate_vote
-        {
-            template <typename Vote, typename F, typename... Ts>
-            auto operator()(std::size_t n, Vote&& vote, F&& f, Ts&&... ts) const
-                -> decltype(hpx::resiliency::async_replicate_vote(n,
-                    std::forward<Vote>(vote), std::forward<F>(f),
-                    std::forward<Ts>(ts)...))
-            {
-                return hpx::resiliency::async_replicate_vote(n,
-                    std::forward<Vote>(vote), std::forward<F>(f),
-                    std::forward<Ts>(ts)...);
-            }
-        };
-
-        struct async_replicate_validate
-        {
-            template <typename Pred, typename F, typename... Ts>
-            auto operator()(std::size_t n, Pred&& pred, F&& f, Ts&&... ts) const
-                -> decltype(hpx::resiliency::async_replicate_validate(n,
-                    std::forward<Pred>(pred), std::forward<F>(f),
-                    std::forward<Ts>(ts)...))
-            {
-                return hpx::resiliency::async_replicate_validate(n,
-                    std::forward<Pred>(pred), std::forward<F>(f),
-                    std::forward<Ts>(ts)...);
-            }
-        };
-
-        struct async_replicate
-        {
-            template <typename F, typename... Ts>
-            auto operator()(std::size_t n, F&& f, Ts&&... ts) const
-                -> decltype(hpx::resiliency::async_replicate(
-                    n, std::forward<F>(f), std::forward<Ts>(ts)...))
-            {
-                return hpx::resiliency::async_replicate(
-                    n, std::forward<F>(f), std::forward<Ts>(ts)...);
-            }
-        };
-    }    // namespace functional
-}}       // namespace hpx::resiliency
+}}    // namespace hpx::resiliency
