### 2020-07-22 fbfd291
Adapt transform begin-end iterator
