+++ libs/algorithms/include/hpx/parallel/algorithms/transform.hpp
@@ -174,14 +174,15 @@ namespace hpx { namespace parallel { inline namespace v1 {
         };
 
         /// non_segmented version
-        template <typename ExPolicy, typename FwdIter1, typename FwdIter2,
-            typename F, typename Proj>
+        template <typename ExPolicy, typename FwdIter1B, typename FwdIter1E,
+            typename FwdIter2, typename F, typename Proj>
         typename util::detail::algorithm_result<ExPolicy,
-            hpx::util::tagged_pair<tag::in(FwdIter1), tag::out(FwdIter2)>>::type
-        transform_(ExPolicy&& policy, FwdIter1 first, FwdIter1 last,
+            hpx::util::tagged_pair<tag::in(FwdIter1B),
+                tag::out(FwdIter2)>>::type
+        transform_(ExPolicy&& policy, FwdIter1B first, FwdIter1E last,
             FwdIter2 dest, F&& f, Proj&& proj, std::false_type)
         {
-            static_assert((hpx::traits::is_forward_iterator<FwdIter1>::value),
+            static_assert((hpx::traits::is_forward_iterator<FwdIter1B>::value),
                 "Requires at least forward iterator.");
             static_assert((hpx::traits::is_forward_iterator<FwdIter2>::value),
                 "Requires at least forward iterator.");
@@ -189,17 +190,18 @@ namespace hpx { namespace parallel { inline namespace v1 {
             typedef execution::is_sequenced_execution_policy<ExPolicy> is_seq;
 
             return hpx::util::make_tagged_pair<tag::in, tag::out>(
-                detail::transform<std::pair<FwdIter1, FwdIter2>>().call(
+                detail::transform<std::pair<FwdIter1B, FwdIter2>>().call(
                     std::forward<ExPolicy>(policy), is_seq(), first, last, dest,
                     std::forward<F>(f), std::forward<Proj>(proj)));
         }
 
         /// forward declare the segmented version
-        template <typename ExPolicy, typename FwdIter1, typename FwdIter2,
-            typename F, typename Proj>
+        template <typename ExPolicy, typename FwdIter1B, typename FwdIter1E,
+            typename FwdIter2, typename F, typename Proj>
         typename util::detail::algorithm_result<ExPolicy,
-            hpx::util::tagged_pair<tag::in(FwdIter1), tag::out(FwdIter2)>>::type
-        transform_(ExPolicy&& policy, FwdIter1 first, FwdIter1 last,
+            hpx::util::tagged_pair<tag::in(FwdIter1B),
+                tag::out(FwdIter2)>>::type
+        transform_(ExPolicy&& policy, FwdIter1B first, FwdIter1E last,
             FwdIter2 dest, F&& f, Proj&& proj, std::true_type);
         /// \endcond
     }    // namespace detail
@@ -276,24 +278,24 @@ namespace hpx { namespace parallel { inline namespace v1 {
     ///           copied.
     ///
     //clang-format off
-    template <typename ExPolicy, typename FwdIter1, typename FwdIter2,
-        typename F, typename Proj = util::projection_identity,
-        HPX_CONCEPT_REQUIRES_(
-            execution::is_execution_policy<ExPolicy>::value &&
-            hpx::traits::is_iterator<FwdIter1>::value &&
-            hpx::traits::is_iterator<FwdIter2>::value &&
-            traits::is_projected<Proj, FwdIter1>::value &&
-                traits::is_indirect_callable<ExPolicy, F,
-                    traits::projected<Proj, FwdIter1>>::value
-        )>
[…]
