### 2020-07-22 3a6db82
Use only one node in rostam Jenkins configurations
