### 2020-07-22 1e53c02
Adding async_replay and async_replicate for executors
- adding tests
- flyby: enable single integral value as Shape parameter for
  bulk_async_execute (and friends)
