### 2020-07-23 4fb13ab
Add missing SPDX license tags to channel 1d stencil example
