+++ examples/1d_stencil/1d_stencil_channel.cpp
@@ -1,5 +1,6 @@
 // Copyright (c) 2020 Nikunj Gupta
 //
+// SPDX-License-Identifier: BSL-1.0
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
