### 2020-07-23 8074316
Use medusa partition for clang builds on rostam
