+++ .jenkins/lsu/slurm-configuration-clang-9.sh
@@ -4,5 +4,5 @@
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
 # file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-configuration_slurm_partition="marvin"
+configuration_slurm_partition="medusa"
 configuration_slurm_num_nodes="1"
