### 2020-07-23 4d80c53
minior modification to fc pq example
