### 2020-07-23 b8800fc
adapt hpx header changes
