+++ examples/libcds/libcds_flat_combining_priority_queue.cpp
@@ -6,9 +6,9 @@
 //  http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/assert.hpp>
-#include <hpx/distributed/iostream.hpp>
 #include <hpx/hpx_init.hpp>
 #include <hpx/include/threads.hpp>
+#include <hpx/iostream.hpp>
 
 #include <cds/container/fcpriority_queue.h>
 #include <cds/gc/hp.h>    // for cds::HP (Hazard Pointer) SMR
