+++ libs/resiliency/tests/performance/replay/1d_stencil_replay.cpp
@@ -21,8 +21,8 @@
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
 #include <hpx/include/parallel_algorithm.hpp>
+#include <hpx/modules/resiliency.hpp>
 #include <hpx/modules/synchronization.hpp>
-#include <hpx/resiliency/resiliency.hpp>
 #include <boost/range/irange.hpp>
 
 #include <cstddef>
@@ -194,7 +194,7 @@ struct stepper
         std::uint64_t nd, std::uint64_t n_value, double error,
         hpx::lcos::local::sliding_semaphore& sem)
     {
-        using hpx::experimental::dataflow_replay;
+        using hpx::resiliency::experimental::dataflow_replay;
         using hpx::util::unwrapping;
 
         // U[t][i] is the state of position i at time t.
