+++ libs/resiliency/tests/unit/dataflow_replicate_executor.cpp
@@ -1,6 +1,6 @@
 //  Copyright (c) 2019 National Technology & Engineering Solutions of Sandia,
 //                     LLC (NTESS).
-//  Copyright (c) 2018-2020 Hartmut Kaiser
+//  Copyright (c) 2018 Hartmut Kaiser
 //  Copyright (c) 2019 Adrian Serio
 //  Copyright (c) 2019 Nikunj Gupta
 //
@@ -55,17 +55,17 @@ int hpx_main()
         hpx::parallel::execution::parallel_executor exec;
 
         // successful replicate
-        hpx::future<int> f = hpx::resiliency::experimental::async_replicate(
+        hpx::future<int> f = hpx::resiliency::experimental::dataflow_replicate(
             exec, 10, &deep_thought);
         HPX_TEST(f.get() == 42);
 
         // successful replicate_validate
-        f = hpx::resiliency::experimental::async_replicate_validate(
+        f = hpx::resiliency::experimental::dataflow_replicate_validate(
             exec, 10, &validate, &universal_answer);
         HPX_TEST(f.get() == 42);
 
         // unsuccessful replicate
-        f = hpx::resiliency::experimental::async_replicate(
+        f = hpx::resiliency::experimental::dataflow_replicate(
             exec, 6, &deep_thought);
 
         bool exception_caught = false;
@@ -84,7 +84,7 @@ int hpx_main()
         HPX_TEST(exception_caught);
 
         // unsuccessful replicate_validate
-        f = hpx::resiliency::experimental::async_replicate_validate(
+        f = hpx::resiliency::experimental::dataflow_replicate_validate(
             exec, 6, &validate, &universal_answer);
 
         exception_caught = false;
@@ -103,7 +103,8 @@ int hpx_main()
         HPX_TEST(exception_caught);
 
         // aborted replicate
-        f = hpx::resiliency::experimental::async_replicate(exec, 1, &no_answer);
+        f = hpx::resiliency::experimental::dataflow_replicate(
+            exec, 1, &no_answer);
 
         exception_caught = false;
         try
@@ -121,7 +122,7 @@ int hpx_main()
         HPX_TEST(exception_caught);
 
         // aborted replicate validate
-        f = hpx::resiliency::experimental::async_replicate_validate(
+        f = hpx::resiliency::experimental::dataflow_replicate_validate(
             exec, 1, &validate, &no_answer);
 
         exception_caught = false;
