### 2020-07-23 82a4883
Fixing clang-tidy problems: 0 --> nullptr
