### 2020-07-23 697d1d8
Fix mpi_ring test in distributed mode by ensuring all ranks run hpx_main
