### 2020-07-23 410dd55
Fixing linker problem and formatting errors
