+++ libs/resiliency/tests/performance/replay/1d_stencil_replay.cpp
@@ -221,9 +221,11 @@ struct stepper
 
             for (std::size_t i = 0; i != subdomains; ++i)
             {
-                next[i] = dataflow_replay(n_value, Op, sti, error,
-                    current[(i - 1 + subdomains) % subdomains], current[i],
-                    current[(i + 1) % subdomains]);
+                // explicitly unwrap future
+                hpx::future<partition_data> f = dataflow_replay(n_value, Op,
+                    sti, error, current[(i - 1 + subdomains) % subdomains],
+                    current[i], current[(i + 1) % subdomains]);
+                next[i] = std::move(f);
             }
 
             // every nd time steps, attach additional continuation which will
