### 2020-07-23 #2352 @danielskatz
👋 @bhatele & @davidbeckingsale - how are things going with your reviews?
