### 2020-07-23 6a7d2ed
Removing obsolete code
