+++ hpx/runtime/components/runtime_support.hpp
@@ -139,14 +139,6 @@ namespace hpx { namespace components
         }
 
         ///////////////////////////////////////////////////////////////////////
-#if defined(HPX_HAVE_COMPONENT_GET_GID_COMPATIBILITY)
-        HPX_DEPRECATED(HPX_DEPRECATED_MSG)
-        naming::id_type const& get_gid() const
-        {
-            return gid_;
-        }
-#endif
-
         naming::id_type const& get_id() const
         {
             return gid_;
