### 2020-07-24 28d8b67
Little updates to the docs
- Change main to main() to be consistent with the rest of the docs.
- Fix display of the compilers table
- Update versions and old bash examples
- Change macports to brew for macos (brew is the package manager
used in github actions)
