+++ docs/sphinx/manual/building_hpx.rst
@@ -56,7 +56,7 @@ are case-sensitive. Example:
 
 .. code-block:: bash
 
-   cmake -G "Visual Studio 9 2008" path/to/hpx
+   cmake -G "Visual Studio 16 2019" path/to/hpx
 
 For a given development platform there can be more than one adequate generator.
 If you use Visual Studio ``"NMake Makefiles"`` is a generator you can use for
@@ -85,7 +85,7 @@ Here, you will use the command-line, non-interactive CMake interface.
       mkdir mybuilddir
       cd mybuilddir
 
-#. Execute this command on the shell replacing ``path/to/hpx/`` with the path to
+#. Execute this command on the shell replacing ``path/to/hpx`` with the path to
    the root of your |hpx| source tree:
 
    .. code-block:: bash
@@ -171,7 +171,7 @@ on your target machine. |hpx| currently requires at least Boost V1.61.0 to work
 properly. It may build and run with older versions, but we do not test |hpx|
 with those versions, so please be warned.
 
-The installation of Boost is described in detail in Boost's `Getting Started <https://www.boost.org/doc/libs/1_71_0/more/getting_started/index.html>`_
+The installation of Boost is described in detail in Boost's `Getting Started <https://www.boost.org/doc/libs/1_73_0/more/getting_started/index.html>`_
 document. However, if you've never used the Boost
 libraries (or even if you have), here's a quick primer:
 :ref:`boost_installation`.
@@ -201,7 +201,6 @@ favorite compiler with |hpx| visit |hpx_buildbot|_.
    * * **Compilers**
      *
      *
-     *
    * * |gcc|_
      * 7.0
      *
@@ -209,10 +208,9 @@ favorite compiler with |hpx| visit |hpx_buildbot|_.
      * 2014
      *
    * * |clang|_
-     * 4.0
+     * 5.0
      *
    * * **Build System**
-     *
      *
      *
    * * |cmake|_
@@ -221,12 +219,11 @@ favorite compiler with |hpx| visit |hpx_buildbot|_.
    * * **Required Libraries**
      *
      *
-     *
    * * |boost_libraries|_
-     * 1.61.0
+     * 1.64.0
      *
    * * |hwloc|_
-     * 1.2 (Xeon Phi: 1.6)
+     * 1.5 (Xeon Phi: 1.6)
      *
 
 .. note::
@@ -254,9 +251,8 @@ favorite compiler with |hpx| visit |hpx_buildbot|_.
    * * **Required Libraries**
      *
      *
-     *
    * * |boost|_
-     * 1.61.0
+     * 1.64.0
      *
    * * |hwloc|_
      * 1.5
@@ -353,7 +349,7 @@ Documentation
 To build the |hpx| documentation, you need recent versions of the following
 packages:
 
-- ``python`` (2 or 3)
+- ``python3``
 - ``sphinx`` (Python package)
 - ``sphinx_rtd_theme`` (Python package)
 - ``breathe`` (Python package)
@@ -661,7 +657,7 @@ How to install |hpx| on Unix variants
 
   .. code-block:: bash
 
-     cmake -DBOOST_ROOT=~/packages/boost -DHWLOC_ROOT=/packages/hwloc -DCMAKE_INSTALL_PREFIX=~/packages/hpx ~/downloads/hpx_0.9.10
+     cmake -DBOOST_ROOT=~/packages/boost -DHWLOC_ROOT=/packages/hwloc -DCMAKE_INSTALL_PREFIX=~/packages/hpx ~/downloads/hpx_1.5.0
 
 * Invoke GNU make. If you are on a machine with multiple cores, add the -jN flag
   to your make invocation, where N is the number of parallel processes |hpx|
@@ -752,7 +748,7 @@ remember to replace ``/usr/local/include/boost-1_53`` with whatever prefix
[…]
