### 2020-07-24 02a8dad
Add projection_identity specialization on transform_projected
