### 2020-07-24 a41fa3f
Adapt transform sequential/ parallel versions in detail namespace
