### 2020-07-24 382d521
Adapt transform segmented end-begin iterators
