### 2020-07-24 709c254
Cleaning up libCDS benchmark
