+++ tests/performance/local/libcds_hazard_pointer_overhead.cpp
@@ -10,15 +10,13 @@
 #include <hpx/hpx_init.hpp>
 #include <hpx/include/apply.hpp>
 #include <hpx/include/async.hpp>
-#include <hpx/include/iostreams.hpp>
 #include <hpx/include/parallel_executors.hpp>
 #include <hpx/include/parallel_for_loop.hpp>
 #include <hpx/include/threads.hpp>
+#include <hpx/iostream.hpp>
 #include <hpx/modules/format.hpp>
 #include <hpx/modules/testing.hpp>
 #include <hpx/modules/timing.hpp>
-#include <hpx/runtime/actions/continuation.hpp>
-#include <hpx/runtime/actions/plain_action.hpp>
 #include <hpx/threading_base/annotated_function.hpp>
 
 #include <hpx/include/parallel_execution.hpp>
@@ -115,10 +113,27 @@ double global_scratch = 0;
 std::uint64_t num_iterations = 0;
 
 ///////////////////////////////////////////////////////////////////////////////
+struct libcds_thread_manager_wrapper
+{
+    libcds_thread_manager_wrapper(bool uselibcds)
+      : uselibcds_(uselibcds)
+    {
+        if (uselibcds_)
+            cds::threading::Manager::attachThread();
+    }
+    ~libcds_thread_manager_wrapper()
+    {
+        if (uselibcds_)
+            cds::threading::Manager::detachThread();
+    }
+
+    bool uselibcds_;
+};
+
 double null_function(bool uselibcds) noexcept
 {
-    if (uselibcds)
-        cds::threading::Manager::attachThread();
+    libcds_thread_manager_wrapper wrap(uselibcds);
+
     if (num_iterations > 0)
     {
         const int array_size = 4096;
@@ -132,8 +147,6 @@ double null_function(bool uselibcds) noexcept
         }
         return dummy[0];
     }
-    if (uselibcds)
-        cds::threading::Manager::detachThread();
     return 0.0;
 }
 
@@ -210,10 +223,10 @@ void measure_function_futures_create_thread_hierarchical_placement(
     };
     auto const thread_func =
         hpx::threads::detail::thread_function_nullary<decltype(func)>{func};
-    auto const desc = hpx::util::thread_description();
+    auto desc = hpx::util::thread_description();
     auto prio = hpx::threads::thread_priority_normal;
-    auto const stack_size = hpx::threads::thread_stacksize_small;
-    auto const num_threads = hpx::get_num_worker_threads();
+    auto stack_size = hpx::threads::thread_stacksize_small;
+    auto num_threads = hpx::get_num_worker_threads();
     hpx::error_code ec;
 
     // start the clock
@@ -223,7 +236,7 @@ void measure_function_futures_create_thread_hierarchical_placement(
         auto const hint =
             hpx::threads::thread_schedule_hint(static_cast<std::int16_t>(t));
         auto spawn_func = [&thread_func, sched, hint, t, count, num_threads,
-                              desc, prio]() {
+                              desc, prio, stack_size]() {
             std::uint64_t const count_start = t * count / num_threads;
             std::uint64_t const count_end = (t + 1) * count / num_threads;
             hpx::error_code ec;
@@ -253,10 +266,25 @@ void measure_function_futures_create_thread_hierarchical_placement(
 }
 
 ///////////////////////////////////////////////////////////////////////////////
+struct libcds_wrapper
+{
+    libcds_wrapper()
+    {
+        // Initialize libcds
+        cds::Initialize();
+    }
+
+    ~libcds_wrapper()
+    {
+        // Terminate libcds
+        cds::Terminate();
+    }
+};
+
 int hpx_main(variables_map& vm)
 {
[…]
