### 2020-07-24 ae33624
Fix codacy warning
