### 2020-07-24 e9a9584
Removing nostack tests
