+++ tests/performance/local/libcds_hazard_pointer_overhead.cpp
@@ -315,9 +315,6 @@ int hpx_main(variables_map& vm)
         hpx::parallel::execution::parallel_executor par;
         hpx::parallel::execution::parallel_executor_aggregated par_agg;
         hpx::parallel::execution::thread_pool_executor tpe;
-        hpx::parallel::execution::thread_pool_executor tpe_nostack(
-            hpx::threads::thread_priority_default,
-            hpx::threads::thread_stacksize_nostack);
 
         for (int i = 0; i < repetitions; i++)
         {
@@ -331,8 +328,6 @@ int hpx_main(variables_map& vm)
                     count, csv, par_agg, bool(cds));
                 measure_function_futures_thread_count(
                     count, csv, tpe, bool(cds));
-                measure_function_futures_thread_count(
-                    count, csv, tpe_nostack, bool(cds));
             }
         }
     }
