### 2020-07-24 28e08dd
explict use hazard pointer threading and constructor
