### 2020-07-24 0c975b4
remove parameters in hp constructor, use default
