### 2020-07-25 0500516
force unwrapping of the inner future on return of dataflow functions
