+++ libs/execution_base/CMakeLists.txt
@@ -44,7 +44,13 @@ add_hpx_module(
   SOURCES ${execution_base_sources}
   HEADERS ${execution_base_headers}
   COMPAT_HEADERS ${execution_base_compat_headers}
-  DEPENDENCIES hpx_assertion hpx_config hpx_errors hpx_format hpx_functional
-               hpx_iterator_support hpx_timing
+  DEPENDENCIES
+    hpx_assertion
+    hpx_config
+    hpx_errors
+    hpx_format
+    hpx_functional
+    hpx_iterator_support
+    hpx_timing
   CMAKE_SUBDIRS examples tests
 )
