### 2020-07-26 e662145
Adapting all_of, any_of, and none_of to C++20
- this introduces CPOs for the algorithms
- flyby: make detail::accumulate constexpr
