+++ libs/algorithms/tests/unit/algorithms/all_of.cpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2014-2017 Hartmut Kaiser
+//  Copyright (c) 2014-2020 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -19,9 +19,8 @@
 #include "test_utils.hpp"
 
 ///////////////////////////////////////////////////////////////////////////////
-template <typename ExPolicy, typename IteratorTag,
-    typename Proj = hpx::parallel::util::projection_identity>
-void test_all_of(ExPolicy policy, IteratorTag, Proj proj = Proj())
+template <typename ExPolicy, typename IteratorTag>
+void test_all_of(ExPolicy policy, IteratorTag)
 {
     static_assert(
         hpx::parallel::execution::is_execution_policy<ExPolicy>::value,
@@ -36,7 +35,35 @@ void test_all_of(ExPolicy policy, IteratorTag, Proj proj = Proj())
         std::vector<std::size_t> c =
             test::fill_all_any_none(10007, i);    //-V106
 
-        bool result = hpx::parallel::all_of(
+        bool result = hpx::all_of(policy, iterator(std::begin(c)),
+            iterator(std::end(c)), [](std::size_t v) { return v != 0; });
+
+        // verify values
+        bool expected = std::all_of(
+            std::begin(c), std::end(c), [](std::size_t v) { return v != 0; });
+
+        HPX_TEST_EQ(result, expected);
+    }
+}
+
+template <typename ExPolicy, typename IteratorTag,
+    typename Proj = hpx::parallel::util::projection_identity>
+void test_all_of_ranges(ExPolicy policy, IteratorTag, Proj proj = Proj())
+{
+    static_assert(
+        hpx::parallel::execution::is_execution_policy<ExPolicy>::value,
+        "hpx::parallel::execution::is_execution_policy<ExPolicy>::value");
+
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::size_t iseq[] = {0, 23, 10007};
+    for (std::size_t i : iseq)
+    {
+        std::vector<std::size_t> c =
+            test::fill_all_any_none(10007, i);    //-V106
+
+        bool result = hpx::ranges::all_of(
             policy, iterator(std::begin(c)), iterator(std::end(c)),
             [](std::size_t v) { return v != 0; }, proj);
 
@@ -48,9 +75,8 @@ void test_all_of(ExPolicy policy, IteratorTag, Proj proj = Proj())
     }
 }
 
-template <typename ExPolicy, typename IteratorTag,
-    typename Proj = hpx::parallel::util::projection_identity>
-void test_all_of_async(ExPolicy p, IteratorTag, Proj proj = Proj())
+template <typename ExPolicy, typename IteratorTag>
+void test_all_of_async(ExPolicy p, IteratorTag)
 {
     typedef std::vector<std::size_t>::iterator base_iterator;
     typedef test::test_iterator<base_iterator, IteratorTag> iterator;
@@ -61,7 +87,32 @@ void test_all_of_async(ExPolicy p, IteratorTag, Proj proj = Proj())
         std::vector<std::size_t> c =
             test::fill_all_any_none(10007, i);    //-V106
 
-        hpx::future<bool> f = hpx::parallel::all_of(
+        hpx::future<bool> f = hpx::all_of(p, iterator(std::begin(c)),
+            iterator(std::end(c)), [](std::size_t v) { return v != 0; });
+        f.wait();
+
+        // verify values
[…]
