### 2020-07-26 7524dd9
Making fill and fill_n compatible with C++20
- converting algorithms to use CPOs
