+++ components/containers/partitioned_vector/tests/unit/serialization_partitioned_vector.cpp
@@ -31,14 +31,14 @@ void test(T minval, T maxval)
     {
         std::vector<char> buffer;
 
-        hpx::serialization::output_archive oarchive(buffer,
-            hpx::serialization::disable_data_chunking);
+        hpx::serialization::output_archive oarchive(
+            buffer, hpx::serialization::disable_data_chunking);
 
-        std::size_t sz = static_cast<std::size_t>(maxval-minval);
+        std::size_t sz = static_cast<std::size_t>(maxval - minval);
 
         hpx::partitioned_vector<T> os(sz);
         os.register_as("test_vector");
-        hpx::parallel::fill(
+        hpx::fill(
             hpx::parallel::execution::par, std::begin(os), std::end(os), 42);
 
         oarchive << os;
@@ -46,7 +46,7 @@ void test(T minval, T maxval)
         hpx::serialization::input_archive iarchive(buffer);
 
         hpx::partitioned_vector<T> is(os.size());
-        hpx::parallel::fill(
+        hpx::fill(
             hpx::parallel::execution::par, std::begin(is), std::end(is), 0);
 
         iarchive >> is;
@@ -85,4 +85,3 @@ int main()
 
     return hpx::util::report_errors();
 }
-
