+++ libs/algorithms/include/hpx/parallel/container_algorithms/fill.hpp
@@ -1,4 +1,5 @@
 //  Copyright (c) 2018 Christopher Ogle
+//  Copyright (c) 2020 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -8,17 +9,10 @@
 
 #pragma once
 
-#include <hpx/config.hpp>
-#include <hpx/execution/traits/is_execution_policy.hpp>
-#include <hpx/iterator_support/range.hpp>
-#include <hpx/iterator_support/traits/is_range.hpp>
+#if defined(DOXYGEN)
+namespace hpx {
+    // clang-format off
 
-#include <hpx/parallel/algorithms/fill.hpp>
-
-#include <type_traits>
-#include <utility>
-
-namespace hpx { namespace parallel { inline namespace v1 {
     /// Assigns the given value to the elements in the range [first, last).
     ///
     /// \note   Complexity: Performs exactly \a last - \a first assignments.
@@ -55,15 +49,9 @@ namespace hpx { namespace parallel { inline namespace v1 {
     ///           returns \a difference_type otherwise (where \a difference_type
     ///           is defined by \a void.
     ///
-    template <typename ExPolicy, typename Rng, typename T,
-        HPX_CONCEPT_REQUIRES_(execution::is_execution_policy<ExPolicy>::value&&
-                hpx::traits::is_range<Rng>::value)>
-    typename util::detail::algorithm_result<ExPolicy>::type fill(
-        ExPolicy&& policy, Rng&& rng, T value)
-    {
-        return fill(std::forward<ExPolicy>(policy), hpx::util::begin(rng),
-            hpx::util::end(rng), value);
-    }
+    template <typename ExPolicy, typename Rng, typename T>
+    typename util::detail::algorithm_result<ExPolicy>::type
+    fill(ExPolicy&& policy, Rng&& rng, T const& value);
 
     /// Assigns the given value value to the first count elements in the range
     /// beginning at first if count > 0. Does nothing otherwise.
@@ -75,17 +63,17 @@ namespace hpx { namespace parallel { inline namespace v1 {
     ///                     It describes the manner in which the execution
     ///                     of the algorithm may be parallelized and the manner
     ///                     in which it executes the assignments.
-    /// \tparam Rng         The type of the source range used (deduced).
+    /// \tparam Iterator    The type of the source range used (deduced).
     ///                     The iterators extracted from this range type must
-    ///                     meet the requirements of an input iterator.
+    ///                     meet the requirements of an forward iterator.
     /// \tparam Size        The type of the argument specifying the number of
     ///                     elements to apply \a f to.
     /// \tparam T           The type of the value to be assigned (deduced).
     ///
     /// \param policy       The execution policy to use for the scheduling of
     ///                     the iterations.
-    /// \param rng          Refers to the sequence of elements the algorithm
-    ///                     will be applied to.
+    /// \param first        Refers to the beginning of the sequence of elements
+    ///                     the algorithm will be applied to.
     /// \param count        Refers to the number of elements starting at
     ///                     \a first the algorithm will be applied to.
     /// \param value        The value to be assigned.
@@ -107,15 +95,286 @@ namespace hpx { namespace parallel { inline namespace v1 {
     ///           returns \a difference_type otherwise (where \a difference_type
     ///           is defined by \a void.
     ///
+    template <typename ExPolicy, typename Iterator, typename Size, typename T>
[…]
