+++ CMakeLists.txt
@@ -67,7 +67,7 @@ endif()
 set(HPX_VERSION_MAJOR 1)
 set(HPX_VERSION_MINOR 5)
 set(HPX_VERSION_SUBMINOR 0)
-set(HPX_VERSION_DATE 20200723)
+set(HPX_VERSION_DATE 20200727)
 set(HPX_VERSION_TAG "-rc1")
 
 set(HPX_VERSION
