### 2020-07-27 742b240
Allow limiting executor to use external counting mechanism
The limiting executor does not work with some special executors
like the MPI one since it returns immediately but leave a task
that is 'detached' from the runtime (the counters count up and down
almost immediately, but the task completes later).

To solve this we add a specialization inside the limiting executor
counting mechanism that allows the executor being wrapped to return
a counter that can be used to throttle tasks.

Any executor that implements the "in_flight_estimate" member can
override the counting mechanism to control thro
[…]
