+++ libs/executors/include/hpx/executors/limiting_executor.hpp
@@ -7,14 +7,12 @@
 #pragma once
 
 #include <hpx/config.hpp>
+#include <hpx/concepts/has_member_xxx.hpp>
 #include <hpx/execution/traits/executor_traits.hpp>
 #include <hpx/execution/traits/is_executor.hpp>
 #include <hpx/execution_base/execution.hpp>
 #include <hpx/execution_base/this_thread.hpp>
 #include <hpx/functional/invoke.hpp>
-#include <hpx/futures/futures_factory.hpp>
-#include <hpx/runtime/threads/thread_pool_base.hpp>
-#include <hpx/type_support/void_guard.hpp>
 
 #include <atomic>
 #include <cstddef>
@@ -30,15 +28,23 @@
 //
 // --------------------------------------------------------------------
 namespace hpx { namespace parallel { namespace execution {
+
+    // by convention the title is 7 chars (for alignment)
+    using print_on = debug::enable_print<false>;
+    static constexpr print_on lim_debug("LIMEXEC");
+
+    ///////////////////////////////////////////////////////////////////////////
+    namespace detail {
+        HPX_HAS_MEMBER_XXX_TRAIT_DEF(in_flight_estimate)
+    }    // namespace detail
+
     template <typename BaseExecutor>
     struct limiting_executor
     {
-      private:
         // --------------------------------------------------------------------
         // RAII wrapper for counting task completions (count_down)
         // count_up is done in the executor when the task is first scheduled
-        // This object is contructed when the task actually _runs_
-        // and destructed when it completes
+        // This object is destructed when the task has completed
         // --------------------------------------------------------------------
         struct on_exit
         {
@@ -48,33 +54,141 @@ namespace hpx { namespace parallel { namespace execution {
             }
             ~on_exit()
             {
+                lim_debug.debug(hpx::debug::str<>("Count Down"));
                 executor_.count_down();
             }
             limiting_executor const& executor_;
         };
 
-        template <typename F>
-        struct executor_wrapper
+        // --------------------------------------------------------------------
+        // this is the default wrapper struct that invokes count up / down
+        // and uses the limiting executor counter to control throttling
+        // --------------------------------------------------------------------
+        template <typename F, typename Enable = void>
+        struct throttling_wrapper
         {
+            throttling_wrapper(
+                limiting_executor const& lim, BaseExecutor const& base, F&& f)
+              : limiting_(lim)
+              , f_(std::forward<F>(f))
+            {
+                limiting_.count_up();
+                if (exceeds_upper())
+                {
+                    lim_debug.debug(hpx::debug::str<>("Exceeds_upper"));
+                    hpx::util::yield_while([&]() { return exceeds_lower(); });
+                    lim_debug.debug(hpx::debug::str<>("Below_lower"));
+                }
+            }
+
+            // when task completes, on_exit destructor calls count_down
             template <typename... Ts>
             decltype(auto) operator()(Ts&&... ts)
             {
-                on_exit _{executor_};
+                on_exit _{limiting_};
                 return hpx::util::invoke(f_, std::forward<Ts>(ts)...);
             }
 
-            limiting_executor const& executor_;
+            // returns true if too many tasks would be in flight
+            // NB. we use ">" because we count up right before testing
+            bool exceeds_upper()
+            {
+                return (limiting_.count_ > limiting_.upper_threshold_);
+            }
+            // returns true if we have not yet reached the lower threshold
+            bool exceeds_lower()
+            {
+                return (limiting_.count_ > limiting_.lower_threshold_);
+            }
+
[…]
