### 2020-07-27 811e871
Adding version specific deprecation options
