### 2020-07-27 9714485
Adapting count and count_if to be conforming to C++20
- both algorithms are now exposed through CPOs
