+++ libs/algorithms/tests/unit/algorithms/countif_tests.hpp
@@ -43,8 +43,25 @@ struct always_true
 };
 
 ///////////////////////////////////////////////////////////////////////////////
+template <typename IteratorTag>
+void test_count_if(IteratorTag)
+{
+    typedef std::vector<int>::iterator base_iterator;
+    typedef std::vector<int>::difference_type diff_type;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<int> c(100007);
+    std::iota(std::begin(c), std::begin(c) + 50, 0);
+    std::iota(std::begin(c) + 50, std::end(c), dis(gen) + 50);
+
+    diff_type num_items = hpx::count_if(
+        iterator(std::begin(c)), iterator(std::end(c)), smaller_than_50());
+
+    HPX_TEST_EQ(num_items, 50u);
+}
+
 template <typename ExPolicy, typename IteratorTag>
-void test_count_if(ExPolicy policy, IteratorTag)
+void test_count_if(ExPolicy&& policy, IteratorTag)
 {
     static_assert(
         hpx::parallel::execution::is_execution_policy<ExPolicy>::value,
@@ -58,14 +75,14 @@ void test_count_if(ExPolicy policy, IteratorTag)
     std::iota(std::begin(c), std::begin(c) + 50, 0);
     std::iota(std::begin(c) + 50, std::end(c), dis(gen) + 50);
 
-    diff_type num_items = hpx::parallel::count_if(policy,
-        iterator(std::begin(c)), iterator(std::end(c)), smaller_than_50());
+    diff_type num_items = hpx::count_if(policy, iterator(std::begin(c)),
+        iterator(std::end(c)), smaller_than_50());
 
     HPX_TEST_EQ(num_items, 50u);
 }
 
 template <typename ExPolicy, typename IteratorTag>
-void test_count_if_async(ExPolicy p, IteratorTag)
+void test_count_if_async(ExPolicy&& p, IteratorTag)
 {
     typedef std::vector<int>::iterator base_iterator;
     typedef std::vector<int>::difference_type diff_type;
@@ -75,15 +92,48 @@ void test_count_if_async(ExPolicy p, IteratorTag)
     std::iota(std::begin(c), std::begin(c) + 50, 0);
     std::iota(std::begin(c) + 50, std::end(c), dis(gen) + 50);
 
-    hpx::future<diff_type> f = hpx::parallel::count_if(
+    hpx::future<diff_type> f = hpx::count_if(
         p, iterator(std::begin(c)), iterator(std::end(c)), smaller_than_50());
 
     HPX_TEST_EQ(f.get(), 50);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
+template <typename IteratorTag>
+void test_count_if_exception(IteratorTag)
+{
+    typedef std::vector<int>::iterator base_iterator;
+    typedef test::decorated_iterator<base_iterator, IteratorTag>
+        decorated_iterator;
+
+    std::vector<int> c(10007);
+    std::iota(std::begin(c), std::end(c), dis(gen));
+
+    bool caught_exception = false;
+    try
+    {
+        // pred should never proc, so simple 'returns true'
+        hpx::count_if(decorated_iterator(std::begin(c),
+                          []() { throw std::runtime_error("test"); }),
+            decorated_iterator(std::end(c)), always_true());
+        HPX_TEST(false);
+    }
+    catch (hpx::exception_list const& e)
+    {
+        caught_exception = true;
+        test::test_num_exceptions<hpx::parallel::execution::sequenced_policy,
+            IteratorTag>::call(hpx::parallel::execution::seq, e);
+    }
+    catch (...)
+    {
+        HPX_TEST(false);
+    }
+
+    HPX_TEST(caught_exception);
+}
[…]
