+++ libs/algorithms/include/hpx/parallel/algorithms/count.hpp
@@ -1,4 +1,5 @@
 //  Copyright (c) 2014 Grant Mercer
+//  Copyright (c) 2020 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -8,8 +9,153 @@
 
 #pragma once
 
+#if defined(DOXYGEN)
+namespace hpx {
+    // clang-format off
+
+    /// Returns the number of elements in the range [first, last) satisfying
+    /// a specific criteria. This version counts the elements that are equal to
+    /// the given \a value.
+    ///
+    /// \note   Complexity: Performs exactly \a last - \a first comparisons.
+    ///
+    /// \tparam ExPolicy    The type of the execution policy to use (deduced).
+    ///                     It describes the manner in which the execution
+    ///                     of the algorithm may be parallelized and the manner
+    ///                     in which it executes the comparisons.
+    /// \tparam FwdIterB    The type of the source begin iterator used (deduced).
+    ///                     This iterator type must meet the requirements of an
+    ///                     forward iterator.
+    /// \tparam FwdIterE    The type of the source end iterator used (deduced).
+    ///                     This iterator type must meet the requirements of an
+    ///                     forward iterator.
+    /// \tparam T           The type of the value to search for (deduced).
+    /// \tparam Proj        The type of an optional projection function. This
+    ///                     defaults to \a util::projection_identity
+    ///
+    /// \param policy       The execution policy to use for the scheduling of
+    ///                     the iterations.
+    /// \param first        Refers to the beginning of the sequence of elements
+    ///                     the algorithm will be applied to.
+    /// \param last         Refers to the end of the sequence of elements the
+    ///                     algorithm will be applied to.
+    /// \param value        The value to search for.
+    /// \param proj         Specifies the function (or function object) which
+    ///                     will be invoked for each of the elements as a
+    ///                     projection operation before the actual predicate
+    ///                     \a is invoked.
+    ///
+    /// The comparisons in the parallel \a count algorithm invoked with
+    /// an execution policy object of type \a sequenced_policy
+    /// execute in sequential order in the calling thread.
+    ///
+    /// \note The comparisons in the parallel \a count algorithm invoked with
+    ///       an execution policy object of type \a parallel_policy or
+    ///       \a parallel_task_policy are permitted to execute in an unordered
+    ///       fashion in unspecified threads, and indeterminately sequenced
+    ///       within each thread.
+    ///
+    /// \returns  The \a count algorithm returns a
+    ///           \a hpx::future<difference_type> if the execution policy is of
+    ///           type
+    ///           \a sequenced_task_policy or
+    ///           \a parallel_task_policy and
+    ///           returns \a difference_type otherwise (where \a difference_type
+    ///           is defined by \a std::iterator_traits<FwdIterB>::difference_type.
+    ///           The \a count algorithm returns the number of elements
+    ///           satisfying the given criteria.
+    ///
+    template <typename ExPolicy, typename FwdIterB, typename FwdIterE,
+        typename T, typename Proj = util::projection_identity>
+    typename util::detail::algorithm_result<ExPolicy,
+        typename std::iterator_traits<FwdIterB>::difference_type>::type
+    count(ExPolicy&& policy, FwdIterB first, FwdIterE last, T const& value,
+            Proj&& proj = Proj());
+
+    /// Returns the number of elements in the range [first, last) satisfying
+    /// a specific criteria. This version counts elements for which predicate
+    /// \a f returns true.
+    ///
+    /// \note   Complexity: Performs exactly \a last - \a first applications of
+    ///         the predicate.
+    ///
+    /// \tparam ExPolicy    The type of the execution policy to use (deduced).
+    ///                     It describes the manner in which the execution
[…]
