+++ libs/segmented_algorithms/tests/unit/test_transform_binary2.hpp
@@ -80,10 +80,9 @@ void verify_values_count(
     ExPolicy&& policy, hpx::partitioned_vector<T> const& v, U const& val)
 {
     HPX_TEST_EQ(
-        std::size_t(hpx::parallel::count(policy, v.begin(), v.end(), val)),
-        v.size());
-    HPX_TEST_EQ(std::size_t(hpx::parallel::count_if(
-                    policy, v.begin(), v.end(), cmp<T>(val))),
+        std::size_t(hpx::count(policy, v.begin(), v.end(), val)), v.size());
+    HPX_TEST_EQ(
+        std::size_t(hpx::count_if(policy, v.begin(), v.end(), cmp<T>(val))),
         v.size());
 }
 
@@ -91,13 +90,11 @@ template <typename ExPolicy, typename T, typename U = T>
 void verify_values_count_async(
     ExPolicy&& policy, hpx::partitioned_vector<T> const& v, U const& val)
 {
+    HPX_TEST_EQ(std::size_t(hpx::count(policy, v.begin(), v.end(), val).get()),
+        v.size());
     HPX_TEST_EQ(
         std::size_t(
-            hpx::parallel::count(policy, v.begin(), v.end(), val).get()),
-        v.size());
-    HPX_TEST_EQ(std::size_t(hpx::parallel::count_if(
-                    policy, v.begin(), v.end(), cmp<T>(val))
-                                .get()),
+            hpx::count_if(policy, v.begin(), v.end(), cmp<T>(val)).get()),
         v.size());
 }
 
