### 2020-07-27 4bd383d
Adding HPX_DEPRECATED_V(major, minor, "message")
- also removing the CMake options
