### 2020-07-27 3805821
Applying versioned deprecation macro to more spots
