+++ hpx/lcos_fwd.hpp
@@ -65,10 +65,9 @@ namespace hpx
     template <typename Result,
         typename RemoteResult =
             typename traits::promise_remote_result<Result>::type>
-    using promise HPX_DEPRECATED(
-        "The alias hpx::lcos::promise to "
-        "hpx::lcos::promise is deprecated. Please "
-        "use hpx::lcos::promise directly instead. "
+    using promise HPX_DEPRECATED_V(1, 5,
+        "The alias hpx::lcos::promise to hpx::lcos::promise is deprecated. "
+        "Please use hpx::lcos::promise directly instead. "
         "hpx::promise will refer to the local-only promise in the future.") =
         lcos::promise<Result, RemoteResult>;
 #endif
