+++ libs/algorithms/tests/unit/container_algorithms/foreach_range.cpp
@@ -19,6 +19,8 @@ void test_for_each()
 {
     using namespace hpx::parallel;
 
+    test_for_each_seq(IteratorTag());
+
     test_for_each(execution::seq, IteratorTag());
     test_for_each(execution::par, IteratorTag());
     test_for_each(execution::par_unseq, IteratorTag());
@@ -42,6 +44,8 @@ void test_for_each_exception()
     // If the execution policy object is of type vector_execution_policy,
     // std::terminate shall be called. therefore we do not test exceptions
     // with a vector execution policy
+    test_for_each_exception_seq(IteratorTag());
+
     test_for_each_exception(execution::seq, IteratorTag());
     test_for_each_exception(execution::par, IteratorTag());
 
@@ -66,6 +70,8 @@ void test_for_each_bad_alloc()
     // If the execution policy object is of type vector_execution_policy,
     // std::terminate shall be called. therefore we do not test exceptions
     // with a vector execution policy
+    test_for_each_bad_alloc_seq(IteratorTag());
+
     test_for_each_bad_alloc(execution::seq, IteratorTag());
     test_for_each_bad_alloc(execution::par, IteratorTag());
 
