+++ libs/algorithms/include/hpx/parallel/util/result_types.hpp
@@ -86,40 +86,83 @@ namespace hpx { namespace parallel { namespace util {
     };
 
     ///////////////////////////////////////////////////////////////////////////
-    template <typename ZipIter>
-    in_out_result<typename hpx::util::tuple_element<0,
-                      typename ZipIter::iterator_tuple_type>::type,
-        typename hpx::util::tuple_element<1,
-            typename ZipIter::iterator_tuple_type>::type>
-    get_in_out_result(ZipIter&& zipiter)
+    template <typename I, typename F>
+    struct in_fun_result
     {
-        using iterator_tuple_type = typename ZipIter::iterator_tuple_type;
+        HPX_NO_UNIQUE_ADDRESS I in;
+        HPX_NO_UNIQUE_ADDRESS F fun;
 
-        using result_type = in_out_result<
-            typename hpx::util::tuple_element<0, iterator_tuple_type>::type,
-            typename hpx::util::tuple_element<1, iterator_tuple_type>::type>;
+        template <typename I2, typename F2,
+            typename Enable = typename std::enable_if<
+                std::is_convertible<I const&, I2&>::value &&
+                std::is_convertible<F const&, F2&>::value>::type>
+        constexpr operator in_fun_result<I2, F2>() const&
+        {
+            return {in, fun};
+        }
 
-        iterator_tuple_type t = zipiter.get_iterator_tuple();
-        return result_type{hpx::util::get<0>(t), hpx::util::get<1>(t)};
-    }
+        template <typename I2, typename F2,
+            typename Enable =
+                typename std::enable_if<std::is_convertible<I, I2>::value &&
+                    std::is_convertible<F, F2>::value>::type>
+        constexpr operator in_fun_result<I2, F2>() &&
+        {
+            return {std::move(in), std::move(fun)};
+        }
 
-    template <typename ZipIter>
-    hpx::future<in_out_result<typename hpx::util::tuple_element<0,
-                                  typename ZipIter::iterator_tuple_type>::type,
-        typename hpx::util::tuple_element<1,
-            typename ZipIter::iterator_tuple_type>::type>>
-    get_in_out_result(hpx::future<ZipIter>&& zipiter)
-    {
-        using iterator_tuple_type = typename ZipIter::iterator_tuple_type;
+        template <typename Archive>
+        void serialize(Archive& ar, unsigned)
+        {
+            // clang-format off
+            ar & in & fun;
+            // clang-format on
+        }
+    };
 
-        using result_type = in_out_result<
-            typename hpx::util::tuple_element<0, iterator_tuple_type>::type,
-            typename hpx::util::tuple_element<1, iterator_tuple_type>::type>;
+    ///////////////////////////////////////////////////////////////////////////
+    namespace detail {
+        template <typename ZipIter>
+        in_out_result<typename hpx::util::tuple_element<0,
+                          typename ZipIter::iterator_tuple_type>::type,
+            typename hpx::util::tuple_element<1,
+                typename ZipIter::iterator_tuple_type>::type>
+        get_in_out_result(ZipIter&& zipiter)
+        {
+            using iterator_tuple_type = typename ZipIter::iterator_tuple_type;
 
-        return lcos::make_future<result_type>(
-            std::move(zipiter), [](ZipIter zipiter) {
-                return get_in_out_result(std::move(zipiter));
-            });
-    }
+            using result_type = in_out_result<
+                typename hpx::util::tuple_element<0, iterator_tuple_type>::type,
+                typename hpx::util::tuple_element<1,
+                    iterator_tuple_type>::type>;
 
-}}}    // namespace hpx::parallel::util
+            iterator_tuple_type t = zipiter.get_iterator_tuple();
[…]
