+++ libs/execution/tests/regressions/parallel_executor_1781.cpp
@@ -7,8 +7,8 @@
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
 
-#include <hpx/include/parallel_algorithm.hpp>
-#include <hpx/include/parallel_executors.hpp>
+#include <hpx/algorithm.hpp>
+#include <hpx/execution.hpp>
 
 #include <vector>
 
@@ -20,9 +20,9 @@ int hpx_main(int argc, char* argv[])
     {
         hpx::parallel::execution::static_chunk_size block(1);
         hpx::parallel::execution::parallel_executor exec;
-        hpx::parallel::for_each(
-            hpx::parallel::execution::par.on(exec).with(block), v.begin(),
-            v.end(), [](int i) {});
+        hpx::ranges::for_each(
+            hpx::parallel::execution::par.on(exec).with(block), v,
+            [](int i) {});
     }
 
     return hpx::finalize();
