+++ libs/executors/include/hpx/executors/restricted_thread_pool_executor.hpp
@@ -12,7 +12,7 @@
 #include <hpx/config.hpp>
 #include <hpx/assert.hpp>
 #include <hpx/execution/executors/execution_parameters.hpp>
-#include <hpx/execution/executors/thread_pool_executor.hpp>
+#include <hpx/executors/thread_pool_executor.hpp>
 
 #include <atomic>
 #include <cstddef>
