### 2020-07-28 4e080c8
Fix warning guards in compatibility headers
