+++ libs/actions/include_compatibility/hpx/runtime/actions/action_support.hpp
@@ -10,7 +10,7 @@
 #include <hpx/actions/config/defines.hpp>
 #include <hpx/actions/action_support.hpp>
 
-#if defined(HPX_ACTIONS_HAVE_DEPRECATION_WARNINGS)
+#if HPX_ACTIONS_HAVE_DEPRECATION_WARNINGS
 #if defined(HPX_MSVC)
 #pragma message(                                                               \
     "The header hpx/runtime/actions/action_support.hpp is deprecated, \
