+++ libs/actions_base/include_compatibility/hpx/traits/extract_action.hpp
@@ -10,7 +10,7 @@
 #include <hpx/actions_base/config/defines.hpp>
 #include <hpx/actions_base/traits/extract_action.hpp>
 
-#if defined(HPX_ACTION_BASE_HAVE_DEPRECATION_WARNINGS)
+#if HPX_ACTION_BASE_HAVE_DEPRECATION_WARNINGS
 #if defined(HPX_MSVC)
 #pragma message("The header hpx/traits/extract_action.hpp is deprecated, \
     please include hpx/actions_base/traits/extract_action.hpp instead")
