+++ libs/lcos_local/include_compatibility/hpx/local_lcos/trigger.hpp
@@ -10,7 +10,7 @@
 #include <hpx/lcos_local/config/defines.hpp>
 #include <hpx/lcos_local/trigger.hpp>
 
-#if defined(HPX_LCOS_LOCAL_HAVE_DEPRECATION_WARNINGS)
+#if HPX_LCOS_LOCAL_HAVE_DEPRECATION_WARNINGS
 #if defined(HPX_MSVC)
 #pragma message("The header hpx/local_lcos/trigger.hpp is deprecated, \
     please include hpx/lcos_local/trigger.hpp instead")
