+++ libs/actions_base/include_compatibility/hpx/runtime/actions/action_priority.hpp
@@ -10,7 +10,7 @@
 #include <hpx/actions_base/config/defines.hpp>
 #include <hpx/actions_base/action_priority.hpp>
 
-#if defined(HPX_ACTIONS_BASE_HAVE_DEPRECATION_WARNINGS)
+#if HPX_ACTIONS_BASE_HAVE_DEPRECATION_WARNINGS
 #if defined(HPX_MSVC)
 #pragma message(                                                               \
     "The header hpx/runtime/actions/action_priority.hpp is deprecated, \
