### 2020-07-28 a4e9f5d
Making destroy and destroy_n conforming to C++20
- adding ranges::destroy
