### 2020-07-28 18153c9
Making move conforming with C++20
- using CPOs for move
