+++ libs/algorithms/include/hpx/parallel/container_algorithms/move.hpp
@@ -9,25 +9,9 @@
 
 #pragma once
 
-#include <hpx/config.hpp>
-#include <hpx/concepts/concepts.hpp>
-#include <hpx/iterator_support/range.hpp>
-#include <hpx/iterator_support/traits/is_iterator.hpp>
-#include <hpx/iterator_support/traits/is_range.hpp>
-#include <hpx/parallel/util/result_types.hpp>
-
-#include <hpx/parallel/algorithms/move.hpp>
-
-#include <type_traits>
-#include <utility>
-
-namespace hpx { namespace ranges {
-
-    template <typename I, typename O>
-    using move_result = parallel::util::in_out_result<I, O>;
-}}    // namespace hpx::ranges
-
-namespace hpx { namespace parallel { inline namespace v1 {
+#if defined(DOXYGEN)
+namespace hpx {
+    // clang-format off
 
     /// Moves the elements in the range \a rng to another range beginning
     /// at \a dest. After this operation the elements in the moved-from
@@ -80,23 +64,11 @@ namespace hpx { namespace parallel { inline namespace v1 {
     ///           \a last and the output iterator to the element in the
     ///           destination range, one past the last element moved.
     ///
-    // clang-format off
     template <typename ExPolicy, typename FwdIter1, typename Sent1,
-        typename FwdIter,
-        HPX_CONCEPT_REQUIRES_(
-            execution::is_execution_policy<ExPolicy>::value &&
-            hpx::traits::is_iterator<FwdIter1>::value&&
-            hpx::traits::is_sentinel_for<Sent1, FwdIter1>::value&&
-            hpx::traits::is_iterator<FwdIter>::value)>
-    // clang-format on
-    typename util::detail::algorithm_result<ExPolicy,
-        ranges::move_result<FwdIter1, FwdIter>>::type
-    move(ExPolicy&& policy, FwdIter1 iter, Sent1 sent, FwdIter dest)
-    {
-        using move_iter_t = detail::move<FwdIter1, FwdIter>;
-        return detail::transfer<move_iter_t>(
-            std::forward<ExPolicy>(policy), iter, sent, dest);
-    }
+        typename FwdIter>
+    typename util::detail::algorithm_result<
+        ExPolicy, ranges::move_result<FwdIter1, FwdIter>>::type
+    move(ExPolicy&& policy, FwdIter1 iter, Sent1 sent, FwdIter dest);
 
     /// Moves the elements in the range \a rng to another range beginning
     /// at \a dest. After this operation the elements in the moved-from
@@ -144,18 +116,157 @@ namespace hpx { namespace parallel { inline namespace v1 {
     ///           \a last and the output iterator to the element in the
     ///           destination range, one past the last element moved.
     ///
+    template <typename ExPolicy, typename Rng, typename FwdIter>
+    typename util::detail::algorithm_result<
+        ExPolicy, ranges::move_result<
+            typename hpx::traits::range_traits<Rng>::iterator_type, FwdIter>
+    >::type
+    move(ExPolicy&& policy, Rng&& rng, FwdIter dest);
+
+    // clang-format on
+}    // namespace hpx
+
+#else    // DOXYGEN
+
+#include <hpx/config.hpp>
+#include <hpx/concepts/concepts.hpp>
+#include <hpx/iterator_support/range.hpp>
+#include <hpx/iterator_support/traits/is_iterator.hpp>
+#include <hpx/iterator_support/traits/is_range.hpp>
+#include <hpx/parallel/util/result_types.hpp>
+
+#include <hpx/parallel/algorithms/move.hpp>
+
+#include <type_traits>
+#include <utility>
+
+namespace hpx { namespace ranges {
+
+    template <typename I, typename O>
[…]
