+++ libs/algorithms/tests/unit/algorithms/move.cpp
@@ -23,6 +23,30 @@
 int seed = std::random_device{}();
 std::mt19937 gen(seed);
 
+template <typename IteratorTag>
+void test_move(IteratorTag)
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c(10007);
+    std::vector<std::size_t> d(c.size());
+    std::iota(std::begin(c), std::end(c), gen());
+    hpx::move(iterator(std::begin(c)), iterator(std::end(c)), std::begin(d));
+
+    //copy contents of d back into c for testing
+    std::copy(std::begin(d), std::end(d), std::begin(d));
+
+    std::size_t count = 0;
+    HPX_TEST(std::equal(std::begin(c), std::end(c), std::begin(d),
+        [&count](std::size_t v1, std::size_t v2) -> bool {
+            HPX_TEST_EQ(v1, v2);
+            ++count;
+            return v1 == v2;
+        }));
+    HPX_TEST_EQ(count, d.size());
+}
+
 template <typename ExPolicy, typename IteratorTag>
 void test_move(ExPolicy policy, IteratorTag)
 {
@@ -36,7 +60,7 @@ void test_move(ExPolicy policy, IteratorTag)
     std::vector<std::size_t> c(10007);
     std::vector<std::size_t> d(c.size());
     std::iota(std::begin(c), std::end(c), gen());
-    hpx::parallel::move(
+    hpx::move(
         policy, iterator(std::begin(c)), iterator(std::end(c)), std::begin(d));
 
     //copy contents of d back into c for testing
@@ -62,7 +86,7 @@ void test_move_async(ExPolicy p, IteratorTag)
     std::vector<std::size_t> d(c.size());
     std::iota(std::begin(c), std::end(c), gen());
 
-    auto f = hpx::parallel::move(
+    auto f = hpx::move(
         p, iterator(std::begin(c)), iterator(std::end(c)), std::begin(d));
 
     hpx::future<void> g = f.then([&d, &c](hpx::future<void> f) {
@@ -85,6 +109,9 @@ template <typename IteratorTag>
 void test_move()
 {
     using namespace hpx::parallel;
+
+    test_move(IteratorTag());
+
     test_move(execution::seq, IteratorTag());
     test_move(execution::par, IteratorTag());
     test_move(execution::par_unseq, IteratorTag());
@@ -119,7 +146,7 @@ void test_move_exception(ExPolicy policy, IteratorTag)
     bool caught_exception = false;
     try
     {
-        hpx::parallel::move(policy,
+        hpx::move(policy,
             decorated_iterator(
                 std::begin(c), []() { throw std::runtime_error("test"); }),
             decorated_iterator(std::end(c)), std::begin(d));
@@ -154,7 +181,7 @@ void test_move_exception_async(ExPolicy p, IteratorTag)
     bool returned_from_algorithm = false;
     try
     {
-        auto f = hpx::parallel::move(p,
+        auto f = hpx::move(p,
             decorated_iterator(
                 std::begin(c), []() { throw std::runtime_error("test"); }),
             decorated_iterator(std::end(c)), std::begin(d));
@@ -218,7 +245,7 @@ void test_move_bad_alloc(ExPolicy policy, IteratorTag)
     bool caught_bad_alloc = false;
     try
     {
-        hpx::parallel::move(policy,
+        hpx::move(policy,
             decorated_iterator(std::begin(c), []() { throw std::bad_alloc(); }),
             decorated_iterator(std::end(c)), std::begin(d));
         HPX_TEST(false);
[…]
