+++ libs/include/include/hpx/algorithm.hpp
@@ -44,7 +44,6 @@ namespace hpx {
     using hpx::parallel::min_element;
     using hpx::parallel::minmax_element;
     using hpx::parallel::mismatch;
-    using hpx::parallel::move;
     using hpx::parallel::none_of;
     using hpx::parallel::partition;
     using hpx::parallel::partition_copy;
