### 2020-07-29 a51e00d
Fix Inspect and format-violations
