### 2020-07-29 6050a36
Set CUDA compute capability on rostam Jenkins builds
