### 2020-07-29 32a1af8
Replace std::distance with detail::distance in parallel:transform
