### 2020-07-29 f2e4774
Adding non-policy tests to all_of, any_of, and none_of
