+++ libs/algorithms/tests/unit/container_algorithms/any_of_range.cpp
@@ -20,9 +20,33 @@
 #include "test_utils.hpp"
 
 ///////////////////////////////////////////////////////////////////////////////
+template <typename IteratorTag,
+    typename Proj = hpx::parallel::util::projection_identity>
+void test_any_of_seq(IteratorTag, Proj proj = Proj())
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::size_t iseq[] = {0, 23, 10007};
+    for (std::size_t i : iseq)
+    {
+        std::vector<std::size_t> c =
+            test::fill_all_any_none(10007, i);    //-V106
+
+        bool result = hpx::ranges::any_of(
+            c, [](std::size_t v) { return v != 0; }, proj);
+
+        // verify values
+        bool expected = std::any_of(std::begin(c), std::end(c),
+            [proj](std::size_t v) { return proj(v) != 0; });
+
+        HPX_TEST_EQ(result, expected);
+    }
+}
+
 template <typename ExPolicy, typename IteratorTag,
     typename Proj = hpx::parallel::util::projection_identity>
-void test_any_of(ExPolicy policy, IteratorTag, Proj proj = Proj())
+void test_any_of(ExPolicy&& policy, IteratorTag, Proj proj = Proj())
 {
     static_assert(
         hpx::parallel::execution::is_execution_policy<ExPolicy>::value,
@@ -87,6 +111,9 @@ void test_any_of()
     };
     using namespace hpx::parallel;
 
+    test_any_of_seq(IteratorTag());
+    test_any_of_seq(IteratorTag(), proj());
+
     test_any_of(execution::seq, IteratorTag());
     test_any_of(execution::par, IteratorTag());
     test_any_of(execution::par_unseq, IteratorTag());
@@ -133,8 +160,45 @@ void any_of_test()
 }
 
 ///////////////////////////////////////////////////////////////////////////////
+template <typename IteratorTag>
+void test_any_of_exception(IteratorTag)
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::size_t iseq[] = {0, 23, 10007};
+    for (std::size_t i : iseq)
+    {
+        std::vector<std::size_t> c =
+            test::fill_all_any_none(10007, i);    //-V106
+
+        bool caught_exception = false;
+        try
+        {
+            hpx::ranges::any_of(c, [](std::size_t v) {
+                return throw std::runtime_error("test"), v != 0;
+            });
+
+            HPX_TEST(false);
+        }
+        catch (hpx::exception_list const& e)
+        {
+            caught_exception = true;
+            test::test_num_exceptions<
+                hpx::parallel::execution::sequenced_policy,
+                IteratorTag>::call(hpx::parallel::execution::seq, e);
+        }
+        catch (...)
+        {
+            HPX_TEST(false);
+        }
+
+        HPX_TEST(caught_exception);
+    }
+}
+
 template <typename ExPolicy, typename IteratorTag>
-void test_any_of_exception(ExPolicy policy, IteratorTag)
+void test_any_of_exception(ExPolicy&& policy, IteratorTag)
 {
     static_assert(
         hpx::parallel::execution::is_execution_policy<ExPolicy>::value,
@@ -216,6 +280,8 @@ void test_any_of_exception()
 {
     using namespace hpx::parallel;
 
+    test_any_of_exception(IteratorTag());
+
     // If the execution policy object is of type vector_execution_policy,
     // std::terminate shall be called. therefore we do not test exceptions
[…]
