### 2020-07-30 #2352 @bhatele
@danielskatz I wasn't sure what to do about this question:
> Has the submitting author (@diehlpk) made major contributions to the software?

As per the [contributor's list](https://github.com/STEllAR-GROUP/hpx/graphs/contributors), I would not consider the submitting author a major contributor in terms of coding. I can't speak to the design of the software because that's not reflected in git commits.
