### 2020-07-30 #2352 @bhatele
@danielskatz Regarding this question:
> Have the functional claims of the software been confirmed?

The software has a long list of functionality:
* Split-phase transaction model
* Message-driven
* Distributed shared memory (not cache coherent)
* Multi-threaded
* Futures synchronization
* Local Control Objects (LCOs)
* Synchronization for anonymous producer-consumer scenarios
* Percolation (pre-staging of task data)
* multi-threading, parcels, global name space support
* fine-grained control of power usage, enabling reductions in power consumption.

It's difficult to check all of this functionality without having an intimate knowledge of HPX unless there are readily available examples that demonstrate different functionality?
