### 2020-07-30 #2352 @bhatele
@diehlpk HPX makes these performance claims:

"By design, ParalleX overcomes limitations of current models of parallelism by alleviating contention, latency, overhead and starvation. With ParalleX, it is further possible to increase performance by at least an order of magnitude on challenging parallel algorithms, e.g., dynamic directed graph algorithms and adaptive mesh refinement methods for astrophysics."

Have these claims been demonstrated in performance studies and/or publications?
