### 2020-07-30 #2352 @diehlpk
> @danielskatz I wasn't sure what to do about this question:
>
> > Has the submitting author (@diehlpk) made major contributions to the software?
>
> As per the [contributor's list](https://github.com/STEllAR-GROUP/hpx/graphs/contributors), I would not consider the submitting author a major contributor in terms of coding. I can't speak to the design of the software because that's not reflected in git commits.

@bhatele Yes the amount of line of codes is not on the top of the list, however these changes results in following important items for the HPX community
* that HPX compiles on the ARM7 architecture, and we prepare an SC workshop paper with that functionality
* that HPX compiles on Power9 on Summit
* that HPX can be compiled and distributed as rpm packages
* That HPX was accepted in the official Fedora repository and one can install it with `sudo dnf install hpx  `
