### 2020-07-30 #2352 @diehlpk
> @danielskatz Regarding this question:
>
> > Have the functional claims of the software been confirmed?
>
> The software has a long list of functionality:
>
>     * Split-phase transaction model
>
>     * Message-driven
>
>     * Distributed shared memory (not cache coherent)
>
>     * Multi-threaded
>
>     * Futures synchronization
>
>     * Local Control Objects (LCOs)
>
>     * Synchronization for anonymous producer-consumer scenarios
>
>     * Percolation (pre-staging of task data)
>
>     * multi-threading, parcels, global name space support
>
>     * fine-grained control of power usage, enabling reductions in power consumption.
>
>
> It's difficult to check all of this functionality without having an intimate knowledge of HPX unless there are readily available examples that demonstrate different functionality?

@bhatele We have some examples to show the basic features of HPX

https://github.com/STEllAR-GROUP/hpx/tree/master/examples/quickstart

and some more advanced examples where we combine these basic features

https://github.com/STEllAR-GROUP/hpx/tree/master/examples/

In addition, we have eight steps to go from a serial 1D stencil-based to a fully asynchronous distributed one

https://github.com/STEllAR-GROUP/hpx/tree/master/examples/1d_stencil
