### 2020-07-30 #2352 @danielskatz
@diehlpk

1. In a comment above (https://github.com/openjournals/joss-reviews/issues/2352#issuecomment-666317854), it looks like you edited the reviewers comment to add an answer.  It would be much cleaner to just quote the question in a new comment and answer it there.  Please don't edit reviewer comments in the future, to keep the process as transparent as possible.
2. Your answer seems to imply you made changes to the paper in response to the question, but I don't see a new version of the paper or any recent commit in that branch - can you explain?
