### 2020-07-30 #2352 @diehlpk
> @diehlpk
>
>     1. In a comment above ([#2352 (comment)](https://github.com/openjournals/joss-reviews/issues/2352#issuecomment-666317854)), it looks like you edited the reviewer's comment to add an answer.  It would be much cleaner to just quote the question in a new comment and answer it there.  Please don't edit reviewers' comments in the future, to keep the process as transparent as possible.

I need to apologize and it was not my intention, since I quoted for the previous answers.

>
>     2. Your answer seems to imply you made changes to the paper in response to the question, but I don't see a new version of the paper or any recent commit in that branch - can you explain?

Yes, childcare interfered and I could not commit the change yet, however, I am in meetings now and will commit is after I finished the meetings.
