### 2020-07-30 #2352 @diehlpk
> Re:
>
> > Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> @diehlpk are all contributions to the software for non-members through patches? Can non-members submit PRs to the project. I didn't see any documentation regarding this. Also, the contribution guide does not mention any of this and that's in a separate document on governance here: http://hpx.stellar-group.org/governance/

@mstimberg and @hkaiser Should we refer to the governance document here?

>
> The Governance document says: "Anyone can become a contributor, and contributions can take many
> forms, as detailed in a separate document." -> there is no link to find this separate document.

@mstimberg Can you check why there is no link and may fix it?
