### 2020-07-30 cce66a7
Adding link to governance document to CONTRIBUTING.md
