+++ docs/sphinx/contributing/contributing.rst
@@ -1,0 +1,20 @@
+..
+    Copyright (C) 2020 Hartmut Kaiser
+
+    SPDX-License-Identifier: BSL-1.0
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+.. _contributing:
+
+===================
+Contributing to HPX
+===================
+
+The main source of information to understand the process of how to contribute
+to HPX can be found in
+`this document <https://github.com/STEllAR-GROUP/hpx/blob/master/.github/CONTRIBUTING.md>`_.
+This is a living document that is constantly updated with relevant information.
+
+
+
