+++ docs/sphinx/contributing/governance.rst
@@ -1,0 +1,22 @@
+..
+    Copyright (C) 2020 Hartmut Kaiser
+
+    SPDX-License-Identifier: BSL-1.0
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+.. _governance:
+
+====================
+HPX Governance model
+====================
+
+The HPX project is a meritocratic, consensus-based community project. Anyone
+with an interest in the project can join the community, contribute to the
+project design and participate in the decision making process.
+`This document <http://hpx.stellar-group.org/documents/governance/>`_ describes
+how that participation takes place and how to set about earning merit within
+the project community.
+
+
+
