### 2020-07-30 #2352 @diehlpk
@bhatele This pull request blow addresses your remarks to the documentation and governance document. Please let us know if these changes are sufficient.
