### 2020-07-30 #2352 @hkaiser
@bhatele I have created a PR to the HPX documentation that addresses the comments you made here: https://github.com/STEllAR-GROUP/hpx/pull/4875. I have also addressed the comment to add more information about PRs and added the missing link to the governance model document. Thanks!
