### 2020-07-30 3a0236f
Adding splittable executor
