+++ libs/executors/include/hpx/executors/detail/splittable_task.hpp
@@ -1,0 +1,272 @@
+//  Copyright (c) 2007-2020 Hartmut Kaiser
+//  Copyright (c) 2020 Shahrzad Shirzad
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#ifndef HPX_SPLITTABLE_TASK_HPP
+#define HPX_SPLITTABLE_TASK_HPP
+
+#include <hpx/execution/detail/post_policy_dispatch.hpp>
+#include <hpx/modules/datastructures.hpp>
+#include <hpx/modules/execution.hpp>
+#include <hpx/modules/runtime_local.hpp>
+#include <hpx/modules/synchronization.hpp>
+#include <hpx/threading_base/thread_description.hpp>
+
+#include <cstddef>
+#include <cstdint>
+#include <type_traits>
+#include <utility>
+
+namespace hpx { namespace parallel { namespace execution {
+
+    ///////////////////////////////////////////////////////////////////////////
+    enum class splittable_mode
+    {
+        all = 0,
+        idle = 1,
+        idle_mask = 2,
+        all_multiple_tasks = 3
+    };
+
+    inline char const* const get_splittable_mode_name(splittable_mode mode)
+    {
+        static constexpr char const* const splittable_mode_names[] = {
+            "all", "idle", "idle_mask", "all_multiple_tasks"};
+        return splittable_mode_names[static_cast<std::size_t>(mode)];
+    }
+
+    ///////////////////////////////////////////////////////////////////////////
+    template <typename Executor, typename F>
+    struct splittable_task
+    {
+        template <typename F_, typename Shape>
+        splittable_task(Executor& exec, F_&& f, Shape const& elem,
+            std::size_t num_free, splittable_mode split_type,
+            std::size_t min_task_size)
+          : start_(hpx::util::get<0>(elem))
+          , stop_(hpx::util::get<1>(elem))
+          , index_(hpx::util::get<2>(elem))
+          , num_free_(num_free)
+          , f_(std::forward<F_>(f))
+          , exec_(exec)
+          , split_type_(split_type)
+          , min_task_size_(min_task_size)
+        {
+        }
+
+        void operator()(hpx::latch* outer_latch = nullptr)
+        {
+            if (split_type_ == splittable_mode::idle_mask)
+            {
+                call_idle_mask();
+            }
+            else if (split_type_ == splittable_mode::all_multiple_tasks)
+            {
+                call_all_multiple_tasks();
+            }
+            else if (split_type_ == splittable_mode::idle)
+            {
+                call_idle();
+            }
+            else
+            {
+                call();
+            }
+
+            // notify outer waiting task
+            if (outer_latch != nullptr)
+            {
+                outer_latch->count_down(1);
+            }
+        }
+
+    private:
+        void call_idle()
+        {
+            auto mask = hpx::threads::get_idle_core_mask();
+            num_free_ = hpx::threads::count(mask);
+
+            hpx::latch l(2);
+
+            std::size_t task_size =
+                std::ceil((stop_ - start_) / (num_free_ + 1));
+            std::size_t remainder = stop_ - start_ - task_size;
+
+            hpx::util::thread_description desc(f_);
+
+            if ((num_free_ != 0 && task_size > min_task_size_) && remainder > 1)
+            {
+                // split the current task, create one for each idle core
+                for (std::size_t i = 0, j = 0; i != 1; ++j)
+                {
+                    if (hpx::threads::test(mask, j))
+                    {
+                        // pass schedule hint to place new task on empty core
[…]
