+++ libs/executors/include/hpx/executors/splittable_executor.hpp
@@ -1,0 +1,157 @@
+//  Copyright (c) 2007-2020 Hartmut Kaiser
+//  Copyright (c) 2020 Shahrzad Shirzad
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#ifndef HPX_SPLITTABLE_EXECUTOR_HPP
+#define HPX_SPLITTABLE_EXECUTOR_HPP
+
+#include <hpx/config.hpp>
+#include <hpx/execution/traits/is_executor.hpp>
+#include <hpx/executors/detail/splittable_task.hpp>
+#include <hpx/include/async.hpp>
+#include <hpx/modules/executors.hpp>
+#include <hpx/modules/iterator_support.hpp>
+#include <hpx/modules/serialization.hpp>
+#include <hpx/modules/timing.hpp>
+
+#include <algorithm>
+#include <cstddef>
+#include <cstdint>
+#include <type_traits>
+
+namespace hpx { namespace parallel { namespace execution {
+    ///////////////////////////////////////////////////////////////////////////
+    /// Loop iterations are divided into pieces and then assigned to threads.
+    /// The number of loop iterations combined is determined based on
+    /// measurements of how long the execution of 1% of the overall number of
+    /// iterations takes.
+    /// This executor parameters type makes sure that as many loop iterations
+    /// are combined as necessary to run for the amount of time specified.
+    ///
+    struct splittable_executor
+      : parallel_policy_executor<hpx::launch::async_policy>
+    {
+        using base_type = parallel_policy_executor<hpx::launch::async_policy>;
+
+    public:
+        /// Construct an \a splittable_executor executor parameters object
+        ///
+        /// \note Default constructed \a splittable_executor executor parameter
+        ///       types will use 80 microseconds as the minimal time for which
+        ///       any of the scheduled chunks should run.
+        ///
+        splittable_executor()
+          : split_type_(splittable_mode::all)
+          , min_task_size_(0)
+        {
+        }
+
+        /// Construct an \a splittable_executor executor parameters object
+        ///
+        /// \param rel_time     [in] The time duration to use as the minimum
+        ///                     to decide how many loop iterations should be
+        ///                     combined.
+
+        splittable_executor(splittable_mode split_type)
+          : split_type_(split_type)
+          , min_task_size_(0)
+        {
+            if (split_type != splittable_mode::all &&
+                split_type != splittable_mode::idle &&
+                split_type != splittable_mode::idle_mask &&
+                split_type != splittable_mode::all_multiple_tasks)
+            {
+                HPX_THROW_EXCEPTION(hpx::bad_parameter,
+                    "splittable_executor::splittable_executor",
+                    "unknown type, type should be either all, idle, "
+                    "idle_mask, or all_multiple_tasks");
+            }
+        }
+
+        splittable_executor(
+            splittable_mode split_type, std::size_t min_task_size)
+          : split_type_(split_type)
+          , min_task_size_(min_task_size)
+        {
+            if (split_type != splittable_mode::all &&
+                split_type != splittable_mode::idle &&
+                split_type != splittable_mode::idle_mask &&
+                split_type != splittable_mode::all_multiple_tasks)
+            {
+                HPX_THROW_EXCEPTION(hpx::bad_parameter,
+                    "splittable_executor::splittable_executor",
+                    "unknown type, type should be either all, idle, "
+                    "idle_mask, all_multiple_tasks");
+            }
+        }
+
+        /// \cond NOINTERNAL
[…]
