### 2020-07-30 44293c2
Adding grain size benchmark
