+++ tests/performance/local/grain_size.cpp
@@ -1,0 +1,278 @@
+//  Copyright (c) 2020 Shahrzad Shirzad
+//  Copyright (c) 2018-2019 Mikael Simberg
+//  Copyright (c) 2018-2019 John Biddiscombe
+//  Copyright (c) 2011 Bryce Adelstein-Lelbach
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include <hpx/hpx_init.hpp>
+#include <hpx/include/iostreams.hpp>
+#include <hpx/include/parallel_algorithm.hpp>
+#include <hpx/include/parallel_execution.hpp>
+#include <hpx/include/parallel_executor_parameters.hpp>
+#include <hpx/program_options.hpp>
+
+#include "worker_timed.hpp"
+
+#include <cstddef>
+#include <cstdint>
+#include <memory>
+#include <stdexcept>
+#include <string>
+
+using hpx::program_options::options_description;
+using hpx::program_options::value;
+using hpx::program_options::variables_map;
+
+using hpx::finalize;
+using hpx::init;
+
+using hpx::find_here;
+using hpx::naming::id_type;
+
+using hpx::apply;
+using hpx::async;
+using hpx::future;
+using hpx::lcos::wait_each;
+
+using hpx::util::high_resolution_timer;
+
+using hpx::cout;
+using hpx::flush;
+
+// global vars we stick here to make printouts easy for plotting
+static std::string queuing = "default";
+static std::size_t numa_sensitive = 0;
+static std::uint64_t num_threads = 1;
+static std::string info_string = "";
+
+///////////////////////////////////////////////////////////////////////////////
+void print_stats(const char* title, const char* wait, const char* exec,
+    std::int64_t count, double duration, bool csv)
+{
+    std::ostringstream temp;
+    double us = 1e6 * duration;
+    if (csv)
+    {
+        hpx::util::format_to(temp,
+            "{1}, {:27}, {:15}, {:18}, {:8}, {:8}, {:20}, {:4}, {:4}, "
+            "{:20}\n",
+            count, title, wait, exec, duration, us, queuing, numa_sensitive,
+            num_threads, info_string);
+    }
+    else
+    {
+        hpx::util::format_to(temp,
+            "num_iterations {:1}, {:27} {:15} {:18} in {:8} microseconds "
+            ", queue {:20}, numa {:4}, threads {:4}\n",
+            count, title, wait, exec, us, queuing, numa_sensitive, num_threads);
+    }
+    std::cout << temp.str() << std::endl;
+}
+
+const char* ExecName(const hpx::parallel::execution::parallel_executor& exec)
+{
+    return "parallel_executor";
+}
+
+///////////////////////////////////////////////////////////////////////////////
+// we use globals here to prevent the delay from being optimized away
+double global_scratch = 0;
+std::uint64_t num_num_iterations = 0;
+
+///////////////////////////////////////////////////////////////////////////////
+void measure_function_futures_for_loop(std::uint64_t count, bool csv,
+    std::uint64_t chunk_size, std::uint64_t iter_length)
+{
+    // start the clock
+    high_resolution_timer walltime;
+    hpx::parallel::for_loop(
+        hpx::parallel::execution::par.with(
+            hpx::parallel::execution::dynamic_chunk_size(chunk_size)),
+        0, count, [&](std::uint64_t) { worker_timed(iter_length * 1000); });
+
+    // stop the clock
+    const double duration = walltime.elapsed();
[…]
