### 2020-07-30 bfa3757
Adapt parallel::transform result type
