### 2020-07-30 7d3f37c
Add in_in_out_result and adapt corresponding transform result types
