+++ libs/algorithms/include/hpx/parallel/util/transform_loop.hpp
@@ -53,7 +53,7 @@ namespace hpx { namespace parallel { namespace util {
         {
             template <typename InIter1, typename InIter2, typename OutIter,
                 typename F>
-            HPX_HOST_DEVICE HPX_FORCEINLINE static hpx::util::tuple<InIter1,
+            HPX_HOST_DEVICE HPX_FORCEINLINE static util::in_in_out_result<InIter1,
                 InIter2, OutIter>
             call(InIter1 first1, InIter1 last1, InIter2 first2, OutIter dest,
                 F&& f)
@@ -63,13 +63,13 @@ namespace hpx { namespace parallel { namespace util {
                     *dest =
                         hpx::util::invoke(std::forward<F>(f), first1, first2);
                 }
-                return hpx::util::make_tuple(
+                return util::in_in_out_result<InIter1,InIter2, OutIter>(
                     std::move(first1), std::move(first2), std::move(dest));
             }
 
             template <typename InIter1, typename InIter2, typename OutIter,
                 typename F>
-            HPX_HOST_DEVICE HPX_FORCEINLINE static hpx::util::tuple<InIter1,
+            HPX_HOST_DEVICE HPX_FORCEINLINE static util::in_in_out_result<InIter1,
                 InIter2, OutIter>
             call(InIter1 first1, InIter1 last1, InIter2 first2, InIter2 last2,
                 OutIter dest, F&& f)
@@ -80,7 +80,8 @@ namespace hpx { namespace parallel { namespace util {
                     *dest =
                         hpx::util::invoke(std::forward<F>(f), first1, first2);
                 }
-                return hpx::util::make_tuple(first1, first2, dest);
+                return util::in_in_out_result<InIter1,InIter2, OutIter>
+                    (first1, first2, dest);
             }
         };
     }    // namespace detail
@@ -89,7 +90,7 @@ namespace hpx { namespace parallel { namespace util {
         typename OutIter, typename F>
     HPX_HOST_DEVICE HPX_FORCEINLINE typename std::enable_if<
         !execution::is_vectorpack_execution_policy<ExPolicy>::value,
-        hpx::util::tuple<InIter1, InIter2, OutIter>>::type
+        util::in_in_out_result<InIter1, InIter2, OutIter>>::type
     transform_binary_loop(
         InIter1 first1, InIter1 last1, InIter2 first2, OutIter dest, F&& f)
     {
@@ -101,7 +102,7 @@ namespace hpx { namespace parallel { namespace util {
         typename OutIter, typename F>
     HPX_HOST_DEVICE HPX_FORCEINLINE typename std::enable_if<
         !execution::is_vectorpack_execution_policy<ExPolicy>::value,
-        hpx::util::tuple<InIter1, InIter2, OutIter>>::type
+        util::in_in_out_result<InIter1, InIter2, OutIter>>::type
     transform_binary_loop(InIter1 first1, InIter1 last1, InIter2 first2,
         InIter2 last2, OutIter dest, F&& f)
     {
