### 2020-07-31 993acc5
Adapt transform result types completely (and clang-format)
