+++ libs/algorithms/include/hpx/parallel/algorithms/transform.hpp
@@ -31,7 +31,7 @@
 #include <hpx/parallel/util/detail/algorithm_result.hpp>
 #include <hpx/parallel/util/foreach_partitioner.hpp>
 #include <hpx/parallel/util/projection_identity.hpp>
-#include <hpx/parallel/util/result_types.hpp> 
+#include <hpx/parallel/util/result_types.hpp>
 #include <hpx/parallel/util/transform_loop.hpp>
 #include <hpx/parallel/util/zip_iterator.hpp>
 
@@ -159,9 +159,9 @@ namespace hpx { namespace parallel { inline namespace v1 {
 
             template <typename ExPolicy, typename InIterB, typename InIterE,
                 typename OutIter, typename F, typename Proj>
-            HPX_HOST_DEVICE static util::in_out_result<InIterB, OutIter> sequential(
-                ExPolicy&& policy, InIterB first, InIterE last, OutIter dest,
-                F&& f, Proj&& proj)
+            HPX_HOST_DEVICE static util::in_out_result<InIterB, OutIter>
+            sequential(ExPolicy&& policy, InIterB first, InIterE last,
+                OutIter dest, F&& f, Proj&& proj)
             {
                 return util::transform_loop(std::forward<ExPolicy>(policy),
                     first, last, dest, transform_projected<F, Proj>{f, proj});
@@ -178,7 +178,7 @@ namespace hpx { namespace parallel { inline namespace v1 {
                 {
                     auto f1 = transform_iteration<ExPolicy, F, Proj>(
                         std::forward<F>(f), std::forward<Proj>(proj));
-    
+
                     return util::get_in_out_result(
                         util::foreach_partitioner<ExPolicy>::call(
                             std::forward<ExPolicy>(policy),
@@ -190,8 +190,8 @@ namespace hpx { namespace parallel { inline namespace v1 {
                 using result_type = util::in_out_result<FwdIter1B, FwdIter2>;
 
                 return util::detail::algorithm_result<ExPolicy,
-                    result_type>::
-                    get(result_type{std::move(first), std::move(dest)});
+                    result_type>::get(result_type{
+                    std::move(first), std::move(dest)});
             }
         };
 
@@ -210,10 +210,9 @@ namespace hpx { namespace parallel { inline namespace v1 {
 
             typedef execution::is_sequenced_execution_policy<ExPolicy> is_seq;
 
-            return detail::transform<util::in_out_result<FwdIter1B, FwdIter2>>().call(
-                std::forward<ExPolicy>(policy), is_seq(), first, last, dest,
-                std::forward<F>(f), std::forward<Proj>(proj)); 
-            
+            return detail::transform<util::in_out_result<FwdIter1B, FwdIter2>>()
+                .call(std::forward<ExPolicy>(policy), is_seq(), first, last,
+                    dest, std::forward<F>(f), std::forward<Proj>(proj));
         }
 
         /// forward declare the segmented version
@@ -315,10 +314,10 @@ namespace hpx { namespace parallel { inline namespace v1 {
     // clang-format on
     // HPX_DEPRECATED(
     //     "hpx::parallel::transform is deprecated, use hpx::transform instead")
-        typename util::detail::algorithm_result<ExPolicy,
-            util::in_out_result<FwdIter1B, FwdIter2>>::
-            type transform(ExPolicy&& policy, FwdIter1B first, FwdIter1E last,
-                FwdIter2 dest, F&& f, Proj&& proj = Proj())
+    typename util::detail::algorithm_result<ExPolicy,
+        util::in_out_result<FwdIter1B, FwdIter2>>::type
[…]
