### 2020-07-31 e24deb0
Add projection_identity specialization on trasform_binary_projected and fix decay issue
