### 2020-07-31 496e0a3
Fix result types on transform tests
