### 2020-07-31 5f680da
Adding make_heap from ticket 1914
