+++ libs/parallelism/algorithms/tests/unit/algorithms/make_heap.cpp
@@ -1,0 +1,448 @@
+//  Copyright (c) 2015 Grant Mercer
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include <hpx/hpx.hpp>
+#include <hpx/hpx_init.hpp>
+#include <hpx/include/parallel_make_heap.hpp>
+#include <hpx/modules/testing.hpp>
+
+#include <algorithm>
+#include <cstddef>
+#include <iostream>
+#include <iterator>
+#include <numeric>
+#include <random>
+#include <string>
+#include <vector>
+
+#include "test_utils.hpp"
+
+///////////////////////////////////////////////////////////////////////////
+int seed = std::random_device{}();
+std::mt19937 gen(seed);
+
+///////////////////////////////////////////////////////////////////////////
+template <typename IteratorTag>
+void test_make_heap1(IteratorTag)
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c(10007);
+    std::iota(hpx::util::begin(c), hpx::util::end(c), 0);
+
+    hpx::make_heap(iterator(hpx::util::begin(c)), iterator(hpx::util::end(c)));
+
+    HPX_TEST_EQ(std::is_heap(hpx::util::begin(c), hpx::util::end(c)), true);
+}
+
+template <typename ExPolicy, typename IteratorTag>
+void test_make_heap1(ExPolicy&& policy, IteratorTag)
+{
+    static_assert(
+        hpx::parallel::execution::is_execution_policy<ExPolicy>::value,
+        "hpx::parallel::execution::is_execution_policy<ExPolicy>::value");
+
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c(10007);
+    std::iota(hpx::util::begin(c), hpx::util::end(c), 0);
+
+    hpx::make_heap(
+        policy, iterator(hpx::util::begin(c)), iterator(hpx::util::end(c)));
+
+    HPX_TEST_EQ(std::is_heap(hpx::util::begin(c), hpx::util::end(c)), true);
+}
+
+template <typename ExPolicy, typename IteratorTag>
+void test_make_heap_async1(ExPolicy&& p, IteratorTag)
+{
+    typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+
+    std::vector<std::size_t> c(10007);
+    std::iota(hpx::util::begin(c), hpx::util::end(c), 0);
+
+    hpx::future<void> test = hpx::make_heap(
+        p, iterator(hpx::util::begin(c)), iterator(hpx::util::end(c)));
+
+    test.wait();
+    HPX_TEST_EQ(std::is_heap(hpx::util::begin(c), hpx::util::end(c)), true);
+}
+
+template <typename IteratorTag>
+void test_make_heap1()
+{
+    using namespace hpx::parallel::execution;
+
+    test_make_heap1(IteratorTag());
+
+    test_make_heap1(seq, IteratorTag());
+    test_make_heap1(par, IteratorTag());
+    test_make_heap1(par_unseq, IteratorTag());
+
+    test_make_heap_async1(seq(task), IteratorTag());
+    test_make_heap_async1(par(task), IteratorTag());
+}
+
+void make_heap_test1()
+{
[…]
