### 2020-08-02 f651b39
Adds different runtime exception when registering thread without linking to libhpx_wrap
