+++ libs/threading_base/src/register_thread.cpp
@@ -16,6 +16,15 @@
 #include <string>
 #include <utility>
 
+namespace hpx_start {
+    // Redifining weak variables defined in hpx_main.hpp to facilitate error
+    // checking and make sure correct errors are thrown. It is added again
+    // to make sure that these variables are defined correctly in cases
+    // where hpx_main functionalities are not used.
+    HPX_SYMBOL_EXPORT bool is_linked __attribute__((weak)) = false;
+    HPX_SYMBOL_EXPORT bool include_libhpx_wrap __attribute__((weak)) = false;
+}
+
 namespace hpx { namespace threads { namespace detail {
     static get_default_pool_type get_default_pool;
 
@@ -39,6 +48,14 @@ namespace hpx { namespace threads { namespace detail {
         }
         else
         {
+            // hpx_main.hpp is included but not linked to libhpx_wrap
+            if (!hpx_start::is_linked && hpx_start::include_libhpx_wrap)
+                HPX_THROW_EXCEPTION(invalid_status,
+                    "hpx::threads::detail::get_self_or_default_pool",
+                    "Attempting to use hpx_main.hpp functionality without "
+                    "linking to libhpx_wrap. Did you link to HPX::wrap_main "
+                    "in your CMakeLists.txt?");
+
             HPX_THROW_EXCEPTION(invalid_status,
                 "hpx::threads::detail::get_self_or_default_pool",
                 "Attempting to register a thread outside the HPX runtime and "
