### 2020-08-02 a7fd362
Fix clang-format error
