+++ libs/threading_base/src/register_thread.cpp
@@ -1,5 +1,6 @@
 //  Copyright (c) 2007-2016 Hartmut Kaiser
 //  Copyright (c)      2011 Bryce Lelbach
+//  Copyright (c)      2020 Nikunj Gupta
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -23,7 +24,7 @@ namespace hpx_start {
     // where hpx_main functionalities are not used.
     HPX_SYMBOL_EXPORT bool is_linked __attribute__((weak)) = false;
     HPX_SYMBOL_EXPORT bool include_libhpx_wrap __attribute__((weak)) = false;
-}
+}   // namespace hpx_start
 
 namespace hpx { namespace threads { namespace detail {
     static get_default_pool_type get_default_pool;
