+++ wrap/src/hpx_wrap.cpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2018 Nikunj Gupta
+//  Copyright (c) 2018-2020 Nikunj Gupta
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
