### 2020-08-02 731d51b
Fix errors on windows
