### 2020-08-02 068768e
Add instruction for explicit linking
