### 2020-08-03 5b2d87d
fix clang-format errors
