+++ libs/threading_base/src/register_thread.cpp
@@ -18,9 +18,9 @@
 #include <utility>
 
 // The following implementation has been divided for Linux and Mac OSX
-#if defined(HPX_HAVE_DYNAMIC_HPX_MAIN) && \
-    (defined(__linux) || defined(__linux__) || defined(linux) || \
-    defined(__APPLE__))
+#if defined(HPX_HAVE_DYNAMIC_HPX_MAIN) &&                                      \
+    (defined(__linux) || defined(__linux__) || defined(linux) ||               \
+        defined(__APPLE__))
 
 namespace hpx_start {
     // Redefining weak variables defined in hpx_main.hpp to facilitate error
@@ -29,7 +29,7 @@ namespace hpx_start {
     // where hpx_main functionalities are not used.
     HPX_SYMBOL_EXPORT bool is_linked __attribute__((weak)) = false;
     HPX_SYMBOL_EXPORT bool include_libhpx_wrap __attribute__((weak)) = false;
-}   // namespace hpx_start
+}    // namespace hpx_start
 
 #endif
 
@@ -56,69 +56,78 @@ namespace hpx { namespace threads { namespace detail {
         }
         else
         {
-
 // The following implementation has been divided for Linux and Mac OSX
-#if defined(HPX_HAVE_DYNAMIC_HPX_MAIN) && \
-    (defined(__linux) || defined(__linux__) || defined(linux) || \
-    defined(__APPLE__))
+#if defined(HPX_HAVE_DYNAMIC_HPX_MAIN) &&                                      \
+    (defined(__linux) || defined(__linux__) || defined(linux) ||               \
+        defined(__APPLE__))
 
-            // hpx_main.hpp is included but not linked to libhpx_wrap
-            if (!hpx_start::is_linked && hpx_start::include_libhpx_wrap)
-                HPX_THROW_EXCEPTION(invalid_status,
-                    "hpx::threads::detail::get_self_or_default_pool",
-                    "Attempting to use hpx_main.hpp functionality without "
-                    "linking to libhpx_wrap. If you're using CMakeLists, make "
-                    "sure to add HPX::wrap_main to target_link_libraries. "
-                    "If you're using Makefile, make sure to link to "
-                    "libhpx_wrap when generating the executable. If you're "
-                    "linking explicitly, consult the HPX docs for library "
-                    "link order and other subtle nuances.");
+                    // hpx_main.hpp is included but not linked to libhpx_wrap
+                    if (!hpx_start::is_linked && hpx_start::include_libhpx_wrap)
+                        HPX_THROW_EXCEPTION(invalid_status,
+                            "hpx::threads::detail::get_self_or_default_pool",
+                            "Attempting to use hpx_main.hpp functionality "
+                            "without "
+                            "linking to libhpx_wrap. If you're using "
+                            "CMakeLists, make "
+                            "sure to add HPX::wrap_main to "
+                            "target_link_libraries. "
+                            "If you're using Makefile, make sure to link to "
+                            "libhpx_wrap when generating the executable. If "
+                            "you're "
+                            "linking explicitly, consult the HPX docs for "
+                            "library "
+                            "link order and other subtle nuances.");
 
 #endif
 
-            HPX_THROW_EXCEPTION(invalid_status,
-                "hpx::threads::detail::get_self_or_default_pool",
-                "Attempting to register a thread outside the HPX runtime and "
-                "no default pool handler is installed. Did you mean to run "
-                "this on an HPX thread?");
-        }
+                    HPX_THROW_EXCEPTION(invalid_status,
+                        "hpx::threads::detail::get_self_or_default_pool",
+                        "Attempting to register a thread outside the HPX "
+                        "runtime and "
+                        "no default pool handler is installed. Did you mean to "
[…]
