### 2020-08-03 151be04
Fixing exception handling for hpx::copy and adding missing tests
