### 2020-08-03 e36eb7c
Fixing split size and stop condition
