### 2020-08-03 f61c67e
disable thread stealing during the execution of splittable executor
