### 2020-08-03 1dc7022
Adapting hpx::equal to be conforming to C++20
- add hpx::ranges::equal
