+++ libs/algorithms/tests/unit/container_algorithms/equal_range.cpp
@@ -37,11 +37,11 @@ void test_equal1(IteratorTag)
     std::iota(std::begin(c2), std::end(c2), first_value);
 
     {
-        bool result = hpx::equal(
-            iterator(std::begin(c1)), iterator(std::end(c1)), std::begin(c2));
+        bool result = hpx::ranges::equal(iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2), std::end(c2));
 
-        bool expected =
-            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+        bool expected = std::equal(
+            std::begin(c1), std::end(c1), std::begin(c2), std::end(c2));
 
         // verify values
         HPX_TEST_EQ(result, expected);
@@ -50,11 +50,11 @@ void test_equal1(IteratorTag)
     {
         std::uniform_int_distribution<> dis(0, c1.size() - 1);
         ++c1[dis(gen)];    //-V104
-        bool result = hpx::equal(
-            iterator(std::begin(c1)), iterator(std::end(c1)), std::begin(c2));
+        bool result = hpx::ranges::equal(iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2), std::end(c2));
 
-        bool expected =
-            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+        bool expected = std::equal(
+            std::begin(c1), std::end(c1), std::begin(c2), std::end(c2));
 
         // verify values
         HPX_TEST_EQ(result, expected);
@@ -79,11 +79,11 @@ void test_equal1(ExPolicy&& policy, IteratorTag)
     std::iota(std::begin(c2), std::end(c2), first_value);
 
     {
-        bool result = hpx::equal(policy, iterator(std::begin(c1)),
-            iterator(std::end(c1)), std::begin(c2));
+        bool result = hpx::ranges::equal(policy, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2), std::end(c2));
 
-        bool expected =
-            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+        bool expected = std::equal(
+            std::begin(c1), std::end(c1), std::begin(c2), std::end(c2));
 
         // verify values
         HPX_TEST_EQ(result, expected);
@@ -92,11 +92,11 @@ void test_equal1(ExPolicy&& policy, IteratorTag)
     {
         std::uniform_int_distribution<> dis(0, c1.size() - 1);
         ++c1[dis(gen)];    //-V104
-        bool result = hpx::equal(policy, iterator(std::begin(c1)),
-            iterator(std::end(c1)), std::begin(c2));
+        bool result = hpx::ranges::equal(policy, iterator(std::begin(c1)),
+            iterator(std::end(c1)), std::begin(c2), std::end(c2));
 
-        bool expected =
-            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
+        bool expected = std::equal(
+            std::begin(c1), std::end(c1), std::begin(c2), std::end(c2));
 
         // verify values
         HPX_TEST_EQ(result, expected);
@@ -117,12 +117,13 @@ void test_equal1_async(ExPolicy&& p, IteratorTag)
     std::iota(std::begin(c2), std::end(c2), first_value);
 
     {
-        hpx::future<bool> result = hpx::equal(p, iterator(std::begin(c1)),
-            iterator(std::end(c1)), std::begin(c2));
+        hpx::future<bool> result =
+            hpx::ranges::equal(p, iterator(std::begin(c1)),
+                iterator(std::end(c1)), std::begin(c2), std::end(c2));
         result.wait();
 
-        bool expected =
-            std::equal(std::begin(c1), std::end(c1), std::begin(c2));
[…]
