### 2020-08-05 60521e0
Don't apply the deprecation process to the compatibility headers
.. as all the compatibility headers have changed for 1.5.0.

- Compatibility_headers to off for version and config
- Add a comment in the python script to specify added in 1.5.0.
- Change the comment "Added in 1.4.0" to 1.5.0 when the headers
have been changed just before 1.5.0 (api headers etc.)
