+++ libs/batch_environments/CMakeLists.txt
@@ -29,7 +29,7 @@ set(batch_environments_sources alps_environment.cpp batch_environment.cpp
 include(HPX_AddModule)
 add_hpx_module(
   batch_environments
-  COMPATIBILITY_HEADERS ON
+  COMPATIBILITY_HEADERS ON # Added in 1.5.0
   DEPRECATION_WARNINGS
   GLOBAL_HEADER_GEN ON
   SOURCES ${batch_environments_sources}
