+++ libs/statistics/CMakeLists.txt
@@ -26,7 +26,7 @@ set(statistics_sources)
 include(HPX_AddModule)
 add_hpx_module(
   statistics
-  COMPATIBILITY_HEADERS ON # Added in 1.4.0
+  COMPATIBILITY_HEADERS ON # Added in 1.5.0
   DEPRECATION_WARNINGS
   GLOBAL_HEADER_GEN ON
   SOURCES ${statistics_sources}
