+++ libs/static_reinit/CMakeLists.txt
@@ -23,7 +23,7 @@ set(static_reinit_sources static_reinit.cpp)
 include(HPX_AddModule)
 add_hpx_module(
   static_reinit
-  COMPATIBILITY_HEADERS ON
+  COMPATIBILITY_HEADERS ON # Added in 1.5.0
   DEPRECATION_WARNINGS
   GLOBAL_HEADER_GEN ON
   SOURCES ${static_reinit_sources}
