+++ libs/resource_partitioner/CMakeLists.txt
@@ -29,7 +29,7 @@ endif()
 include(HPX_AddModule)
 add_hpx_module(
   resource_partitioner
-  COMPATIBILITY_HEADERS ON # Added in 1.4.0
+  COMPATIBILITY_HEADERS ON # Added in 1.5.0
   DEPRECATION_WARNINGS
   GLOBAL_HEADER_GEN ON
   SOURCES ${resource_partitioner_sources}
