+++ libs/lcos_local/CMakeLists.txt
@@ -39,7 +39,7 @@ set(lcos_local_sources composable_guard.cpp)
 include(HPX_AddModule)
 add_hpx_module(
   lcos_local
-  COMPATIBILITY_HEADERS ON # Added in 1.4.0
+  COMPATIBILITY_HEADERS ON # Added in 1.5.0
   DEPRECATION_WARNINGS
   SOURCES ${lcos_local_sources}
   HEADERS ${lcos_local_headers}
