### 2020-08-05 7bf08ab
Add cmake deprecation warning for hpxmp
