### 2020-08-05 889e9b7
Adapting mismatch to C++20
- add overloads for mismatch taking ranges
- flyby: making projection_identity constexpr
