+++ libs/algorithms/tests/unit/container_algorithms/mismatch_binary_range.cpp
@@ -26,10 +26,12 @@ std::mt19937 gen(seed);
 std::uniform_int_distribution<> dis(0, 10006);
 
 template <typename IteratorTag>
-void test_mismatch1(IteratorTag)
+void test_mismatch_binary1(IteratorTag)
 {
     typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef hpx::util::iterator_range<base_iterator> base_range;
     typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+    typedef hpx::util::iterator_range<iterator> range;
 
     std::vector<std::size_t> c1(10007);
     std::vector<std::size_t> c2(c1.size());
@@ -42,12 +44,14 @@ void test_mismatch1(IteratorTag)
     iterator end1 = iterator(std::end(c1));
 
     {
-        auto result = hpx::mismatch(begin1, end1, std::begin(c2));
+        auto result =
+            hpx::ranges::mismatch(range(iterator(begin1), iterator(end1)),
+                base_range(std::begin(c2), std::end(c2)));
 
         // verify values
         HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), c1.size());
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
+            std::size_t(std::distance(begin1, result.in1)), c1.size());
+        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.in2)),
             c2.size());
     }
 
@@ -55,25 +59,29 @@ void test_mismatch1(IteratorTag)
         std::size_t changed_idx = dis(gen);    //-V104
         ++c1[changed_idx];
 
-        auto result = hpx::mismatch(begin1, end1, std::begin(c2));
+        auto result =
+            hpx::ranges::mismatch(range(iterator(begin1), iterator(end1)),
+                base_range(std::begin(c2), std::end(c2)));
 
         // verify values
         HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), changed_idx);
-        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.second)),
+            std::size_t(std::distance(begin1, result.in1)), changed_idx);
+        HPX_TEST_EQ(std::size_t(std::distance(std::begin(c2), result.in2)),
             changed_idx);
     }
 }
 
 template <typename ExPolicy, typename IteratorTag>
-void test_mismatch1(ExPolicy&& policy, IteratorTag)
+void test_mismatch_binary1(ExPolicy&& policy, IteratorTag)
 {
     static_assert(
         hpx::parallel::execution::is_execution_policy<ExPolicy>::value,
         "hpx::parallel::execution::is_execution_policy<ExPolicy>::value");
 
     typedef std::vector<std::size_t>::iterator base_iterator;
+    typedef hpx::util::iterator_range<base_iterator> base_range;
     typedef test::test_iterator<base_iterator, IteratorTag> iterator;
+    typedef hpx::util::iterator_range<iterator> range;
 
     std::vector<std::size_t> c1(10007);
     std::vector<std::size_t> c2(c1.size());
@@ -86,12 +94,14 @@ void test_mismatch1(ExPolicy&& policy, IteratorTag)
     iterator end1 = iterator(std::end(c1));
 
     {
-        auto result = hpx::mismatch(policy, begin1, end1, std::begin(c2));
+        auto result = hpx::ranges::mismatch(policy,
+            range(iterator(begin1), iterator(end1)),
+            base_range(std::begin(c2), std::end(c2)));
 
         // verify values
         HPX_TEST_EQ(
-            std::size_t(std::distance(begin1, result.first)), c1.size());
[…]
