+++ libs/algorithms/include/hpx/parallel/container_algorithms.hpp
@@ -20,6 +20,7 @@
 #include <hpx/parallel/container_algorithms/is_heap.hpp>
 #include <hpx/parallel/container_algorithms/merge.hpp>
 #include <hpx/parallel/container_algorithms/minmax.hpp>
+#include <hpx/parallel/container_algorithms/mismatch.hpp>
 #include <hpx/parallel/container_algorithms/move.hpp>
 #include <hpx/parallel/container_algorithms/partition.hpp>
 #include <hpx/parallel/container_algorithms/reduce.hpp>
