### 2020-08-05 c9a8680
Adapting hpx::find and friends to C++20
