+++ libs/algorithms/tests/unit/container_algorithms/find_first_of_range2.cpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2014 Grant Mercer
+//  copyright (c) 2014 Grant Mercer
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -22,29 +22,36 @@
 ////////////////////////////////////////////////////////////////////////////
 unsigned int seed = std::random_device{}();
 std::mt19937 gen(seed);
-std::uniform_int_distribution<> dis(2, 101);
+std::uniform_int_distribution<> dis(0, 10006);
+std::uniform_int_distribution<> dist(0, 2);
 
 template <typename IteratorTag>
-void test_find(IteratorTag)
+void test_find_first_of_proj(IteratorTag)
 {
     typedef std::vector<std::size_t>::iterator base_iterator;
     typedef test::test_iterator<base_iterator, IteratorTag> iterator;
 
+    int find_first_of_pos = dis(gen);
+    int random_sub_seq_pos = dist(gen);
+
     std::vector<std::size_t> c(10007);
-    //fill vector with random values about 1
-    std::fill(std::begin(c), std::end(c), dis(gen));
-    c.at(c.size() / 2) = 1;
+    std::iota(std::begin(c), std::end(c), (gen() % 32768) + 19);
+    std::size_t h[] = {1 + 65536, 7 + 65536, 18 + 65536, 3 + 65536};
+    c[find_first_of_pos] = h[random_sub_seq_pos];    //-V108
 
-    iterator index = hpx::find(
-        iterator(std::begin(c)), iterator(std::end(c)), std::size_t(1));
+    iterator index = hpx::ranges::find_first_of(
+        iterator(std::begin(c)), iterator(std::end(c)), std::begin(h),
+        std::end(h), std::equal_to<std::size_t>(),
+        [](std::size_t x) { return x % 65536; },
+        [](std::size_t x) { return x % 65536; });
 
-    base_iterator test_index = std::begin(c) + c.size() / 2;
+    base_iterator test_index = std::begin(c) + find_first_of_pos;
 
     HPX_TEST(index == iterator(test_index));
 }
 
 template <typename ExPolicy, typename IteratorTag>
-void test_find(ExPolicy&& policy, IteratorTag)
+void test_find_first_of_proj(ExPolicy&& policy, IteratorTag)
 {
     static_assert(
         hpx::parallel::execution::is_execution_policy<ExPolicy>::value,
@@ -53,78 +60,95 @@ void test_find(ExPolicy&& policy, IteratorTag)
     typedef std::vector<std::size_t>::iterator base_iterator;
     typedef test::test_iterator<base_iterator, IteratorTag> iterator;
 
+    int find_first_of_pos = dis(gen);
+    int random_sub_seq_pos = dist(gen);
+
     std::vector<std::size_t> c(10007);
-    //fill vector with random values about 1
-    std::fill(std::begin(c), std::end(c), dis(gen));
-    c.at(c.size() / 2) = 1;
+    std::iota(std::begin(c), std::end(c), (gen() % 32768) + 19);
+    std::size_t h[] = {1 + 65536, 7 + 65536, 18 + 65536, 3 + 65536};
+    c[find_first_of_pos] = h[random_sub_seq_pos];    //-V108
 
-    iterator index = hpx::find(
-        policy, iterator(std::begin(c)), iterator(std::end(c)), std::size_t(1));
+    iterator index = hpx::ranges::find_first_of(
+        policy, iterator(std::begin(c)), iterator(std::end(c)), std::begin(h),
+        std::end(h), std::equal_to<std::size_t>(),
[…]
