+++ libs/algorithms/include/hpx/parallel/util/loop.hpp
@@ -55,7 +55,7 @@ namespace hpx { namespace parallel { namespace util {
         {
             ///////////////////////////////////////////////////////////////////
             template <typename Begin, typename End, typename F>
-            HPX_HOST_DEVICE HPX_FORCEINLINE static Begin call(
+            HPX_HOST_DEVICE HPX_FORCEINLINE static constexpr Begin call(
                 Begin it, End end, F&& f)
             {
                 for (/**/; it != end; ++it)
@@ -66,7 +66,7 @@ namespace hpx { namespace parallel { namespace util {
 
             template <typename Begin, typename End, typename CancelToken,
                 typename F>
-            HPX_HOST_DEVICE HPX_FORCEINLINE static Begin call(
+            HPX_HOST_DEVICE HPX_FORCEINLINE static constexpr Begin call(
                 Begin it, End end, CancelToken& tok, F&& f)
             {
                 for (/**/; it != end; ++it)
@@ -81,7 +81,7 @@ namespace hpx { namespace parallel { namespace util {
     }    // namespace detail
 
     template <typename ExPolicy, typename Begin, typename End, typename F>
-    HPX_HOST_DEVICE HPX_FORCEINLINE Begin loop(
+    HPX_HOST_DEVICE HPX_FORCEINLINE constexpr Begin loop(
         ExPolicy&&, Begin begin, End end, F&& f)
     {
         return detail::loop<Begin>::call(begin, end, std::forward<F>(f));
@@ -89,7 +89,7 @@ namespace hpx { namespace parallel { namespace util {
 
     template <typename ExPolicy, typename Begin, typename End,
         typename CancelToken, typename F>
-    HPX_HOST_DEVICE HPX_FORCEINLINE Begin loop(
+    HPX_HOST_DEVICE HPX_FORCEINLINE constexpr Begin loop(
         ExPolicy&&, Begin begin, End end, CancelToken& tok, F&& f)
     {
         return detail::loop<Begin>::call(begin, end, tok, std::forward<F>(f));
@@ -106,8 +106,9 @@ namespace hpx { namespace parallel { namespace util {
             ///////////////////////////////////////////////////////////////////
             template <typename Begin1, typename End1, typename Begin2,
                 typename F>
-            HPX_HOST_DEVICE HPX_FORCEINLINE static std::pair<Begin1, Begin2>
-            call(Begin1 it1, End1 end1, Begin2 it2, F&& f)
+            HPX_HOST_DEVICE
+                HPX_FORCEINLINE static constexpr std::pair<Begin1, Begin2>
+                call(Begin1 it1, End1 end1, Begin2 it2, F&& f)
             {
                 for (/**/; it1 != end1; (void) ++it1, ++it2)
                     f(it1, it2);
@@ -119,7 +120,7 @@ namespace hpx { namespace parallel { namespace util {
 
     template <typename ExPolicy, typename VecOnly, typename Begin1,
         typename End1, typename Begin2, typename F>
-    HPX_HOST_DEVICE HPX_FORCEINLINE typename std::enable_if<
+    HPX_HOST_DEVICE HPX_FORCEINLINE constexpr typename std::enable_if<
         !execution::is_vectorpack_execution_policy<ExPolicy>::value,
         std::pair<Begin1, Begin2>>::type
     loop2(VecOnly, Begin1 begin1, End1 end1, Begin2 begin2, F&& f)
@@ -139,7 +140,7 @@ namespace hpx { namespace parallel { namespace util {
             ///////////////////////////////////////////////////////////////////
             // handle sequences of non-futures
             template <typename Iter, typename F>
-            HPX_HOST_DEVICE HPX_FORCEINLINE static Iter call(
[…]
