### 2020-08-06 8c1590c
Remove Xeon phi and Bluegene instruction (not tested)
The user can easily find the instructions back in old versions of
the docs.
