+++ docs/sphinx/manual/building_hpx.rst
@@ -223,7 +223,7 @@ favorite compiler with |hpx| visit |hpx_buildbot|_.
      * 1.64.0
      *
    * * |hwloc|_
-     * 1.5 (Xeon Phi: 1.6)
+     * 1.5
      *
 
 .. note::
@@ -1021,167 +1021,6 @@ How to build |hpx| under Windows 10 x64 with Visual Studio 2015
   .. figure:: ../_static/images/vs_build_output.png
 
      Visual Studio build output.
-
-.. _bgq_installation:
-
-How to Install |hpx| on BlueGene/Q
-----------------------------------
-
-So far we only support BGClang for compiling |hpx| on the BlueGene/Q.
-
-* Check if BGClang is available on your installation. If not, obtain and install
-  a copy from the `BGClang trac page
-  <https://trac.alcf.anl.gov/projects/llvm-bgq>`_.
-
-* Build (and install) a recent version of |hwloc_downloads|_. With the following
-  commands:
-
-  .. code-block:: bash
-
-     ./configure \
-       --host=powerpc64-bgq-linux \
-       --prefix=$HOME/install/hwloc \
-       --disable-shared \
-       --enable-static \
-       CPPFLAGS='-I/bgsys/drivers/ppcfloor -I/bgsys/drivers/ppcfloor/spi/include/kernel/cnk/'
-     make
-     make install
-
-* Build (and install) a recent version of Boost, using BGClang. To build Boost
-  with BGClang, you'll need to set up the following in your Boost
-  ``~/user-config.jam`` file:
-
-  .. code-block:: bash
-
-     # user-config.jam (put this file into your home directory)
-     using clang
-       :
-       : bgclang++11
-       :
-       ;
-
-  You can then use this as your build command:
-
-  .. code-block:: bash
-
-     ./bootstrap.sh
-     ./b2 --build-dir=/tmp/build-boost --layout=versioned toolset=clang -j12
-
-* Clone the master |hpx| git repository (or a stable tag):
-
-  .. code-block:: bash
-
-     git clone git://github.com/STEllAR-GROUP/hpx.git
-
-* Generate the |hpx| buildfiles using CMake:
-
-  .. code-block:: bash
-
-     cmake -DHPX_PLATFORM=BlueGeneQ \
-             -DCMAKE_TOOLCHAIN_FILE=/path/to/hpx/cmake/toolchains/BGQ.cmake \
-             -DCMAKE_CXX_COMPILER=bgclang++11 \
-             -DMPI_CXX_COMPILER=mpiclang++11 \
-             -DHWLOC_ROOT=/path/to/hwloc/installation \
-             -DBOOST_ROOT=/path/to/boost \
-             -DHPX_WITH_MALLOC=system \
-             /path/to/hpx
-
-* To complete the build and install |hpx|:
-
-  .. code-block:: bash
-
-     make -j24
-     make install
-
-  This will build and install the essential core components of |hpx| only. Use:
-
-  .. code-block:: bash
-
-     make -j24 examples
-     make -j24 install
-
-  to build and install the examples.
-
-.. _intel_mic_installation:
-
-How to Install |hpx| on the Xeon Phi
-------------------------------------
-
-Installation of the Boost Libraries
-...................................
-
-* Download |boost_downloads|_ for Linux and unpack the retrieved tarball.
-
-* Adapt your ``~/user-config.jam`` to contain the following lines:
-
-  .. code-block:: bash
-
-     ## Toolset to be used for compiling for the host
-     using intel
-         : host
-         :
-         : <cxxflags>"-std=c++0x"
-         ;
-
-     ## Toolset to be used for compiling for the Xeon Phi
-     using intel
-         : mic
-         :
-         : <cxxflags>"-std=c++0x -mmic"
-           <linkflags>"-std=c++0x -mmic"
[…]
