### 2020-08-06 b861ff2
Fix tag function object result type
