### 2020-08-06 ff19b7d
Add the resource partitioner examples in the docs
