### 2020-08-06 664ca17
Exit generate_issue_pr_list.sh if hub not installed
