### 2020-08-06 0da10a8
Move get_second_element from copy.hpp to result_types.hpp
