+++ libs/algorithms/include/hpx/parallel/algorithms/copy.hpp
@@ -220,7 +220,6 @@ namespace hpx {
 #include <hpx/parallel/util/projection_identity.hpp>
 #include <hpx/parallel/util/result_types.hpp>
 #include <hpx/parallel/util/scan_partitioner.hpp>
-#include <hpx/parallel/util/tagged_pair.hpp>
 #include <hpx/parallel/util/transfer.hpp>
 #include <hpx/parallel/util/zip_iterator.hpp>
 #include <hpx/type_support/unused.hpp>
@@ -338,21 +337,6 @@ namespace hpx { namespace parallel { inline namespace v1 {
         {
         };
 #endif
-
-        ///////////////////////////////////////////////////////////////////////
-        template <typename I, typename O>
-        O get_second_element(util::in_out_result<I, O>&& p)
-        {
-            return p.out;
-        }
-
-        template <typename I, typename O>
-        hpx::future<O> get_second_element(
-            hpx::future<util::in_out_result<I, O>>&& f)
-        {
-            return lcos::make_future<O>(std::move(f),
-                [](util::in_out_result<I, O>&& p) { return p.out; });
-        }
     }    // namespace detail
 
     // clang-format off
@@ -649,7 +633,7 @@ namespace hpx {
         tag_invoke(hpx::copy_t, ExPolicy&& policy, FwdIter1 first,
             FwdIter1 last, FwdIter2 dest)
         {
-            return parallel::v1::detail::get_second_element(
+            return parallel::util::get_second_element(
                 parallel::v1::detail::transfer<
                     parallel::v1::detail::copy_iter<FwdIter1, FwdIter2>>(
                     std::forward<ExPolicy>(policy), first, last, dest));
@@ -705,7 +689,7 @@ namespace hpx {
                 hpx::parallel::execution::is_sequenced_execution_policy<
                     ExPolicy>;
 
-            return hpx::parallel::v1::detail::get_second_element(
+            return hpx::parallel::util::get_second_element(
                 hpx::parallel::v1::detail::copy_n<
                     hpx::parallel::util::in_out_result<FwdIter1, FwdIter2>>()
                     .call(std::forward<ExPolicy>(policy), is_seq(), first,
@@ -758,7 +742,7 @@ namespace hpx {
                 hpx::parallel::execution::is_sequenced_execution_policy<
                     ExPolicy>;
 
-            return hpx::parallel::v1::detail::get_second_element(
+            return hpx::parallel::util::get_second_element(
                 hpx::parallel::v1::detail::copy_if<
                     hpx::parallel::util::in_out_result<FwdIter1, FwdIter2>>()
                     .call(std::forward<ExPolicy>(policy), is_seq(), first, last,
