+++ libs/algorithms/include/hpx/parallel/util/transform_loop.hpp
@@ -65,8 +65,8 @@ namespace hpx { namespace parallel { namespace util {
                     *dest =
                         hpx::util::invoke(std::forward<F>(f), first1, first2);
                 }
-                return util::in_in_out_result<InIter1, InIter2, OutIter>(
-                    std::move(first1), std::move(first2), std::move(dest));
+                return util::in_in_out_result<InIter1, InIter2, OutIter>{
+                    std::move(first1), std::move(first2), std::move(dest)};
             }
 
             template <typename InIter1, typename InIter2, typename OutIter,
@@ -82,8 +82,8 @@ namespace hpx { namespace parallel { namespace util {
                     *dest =
                         hpx::util::invoke(std::forward<F>(f), first1, first2);
                 }
-                return util::in_in_out_result<InIter1, InIter2, OutIter>(
-                    first1, first2, dest);
+                return util::in_in_out_result<InIter1, InIter2, OutIter>{
+                    first1, first2, dest};
             }
         };
     }    // namespace detail
@@ -117,14 +117,14 @@ namespace hpx { namespace parallel { namespace util {
         struct transform_loop_n
         {
             template <typename InIter, typename OutIter, typename F>
-            HPX_HOST_DEVICE HPX_FORCEINLINE static std::pair<InIter, OutIter>
+            HPX_HOST_DEVICE HPX_FORCEINLINE static util::in_out_result<InIter, OutIter>
             call(InIter first, std::size_t count, OutIter dest, F&& f)
             {
                 for (/* */; count != 0; (void) --count, ++first, ++dest)
                 {
                     *dest = hpx::util::invoke(f, first);
                 }
-                return std::make_pair(std::move(first), std::move(dest));
+                return util::in_out_result<InIter, OutIter>{std::move(first), std::move(dest)};
             }
         };
     }    // namespace detail
@@ -132,7 +132,7 @@ namespace hpx { namespace parallel { namespace util {
     template <typename ExPolicy, typename Iter, typename OutIter, typename F>
     HPX_HOST_DEVICE HPX_FORCEINLINE typename std::enable_if<
         !execution::is_vectorpack_execution_policy<ExPolicy>::value,
-        std::pair<Iter, OutIter>>::type
+        util::in_out_result<Iter, OutIter>>::type
     transform_loop_n(Iter it, std::size_t count, OutIter dest, F&& f)
     {
         return detail::transform_loop_n<Iter>::call(
@@ -146,7 +146,7 @@ namespace hpx { namespace parallel { namespace util {
         {
             template <typename InIter1, typename InIter2, typename OutIter,
                 typename F>
-            HPX_HOST_DEVICE HPX_FORCEINLINE static hpx::util::tuple<InIter1,
+            HPX_HOST_DEVICE HPX_FORCEINLINE static util::in_in_out_result<InIter1,
                 InIter2, OutIter>
             call(InIter1 first1, std::size_t count, InIter2 first2,
                 OutIter dest, F&& f)
@@ -156,8 +156,9 @@ namespace hpx { namespace parallel { namespace util {
                 {
                     *dest = hpx::util::invoke(f, first1, first2);
                 }
-                return hpx::util::make_tuple(
-                    std::move(first1), std::move(first2), std::move(dest));
+                return util::in_in_out_result<InIter1,
+                InIter2, OutIter>{
+                    std::move(first1), std::move(first2), std::move(dest)};
             }
         };
     }    // namespace detail
@@ -166,7 +167,7 @@ namespace hpx { namespace parallel { namespace util {
         typename OutIter, typename F>
[…]
