+++ libs/algorithms/tests/unit/container_algorithms/transform_range_binary.cpp
@@ -43,6 +43,10 @@ void test_transform_binary(ExPolicy policy, IteratorTag)
     auto result = hpx::parallel::transform(
         policy, c1, std::begin(c2), std::begin(d1), add);
 
+    // HPX_TEST(hpx::util::get<0>(result) == std::end(c1));
+    // HPX_TEST(hpx::util::get<1>(result) == std::end(c2));
+    // HPX_TEST(hpx::util::get<2>(result) == std::end(d1));
+
     HPX_TEST(result.in1() == std::end(c1));
     HPX_TEST(result.in2() == std::end(c2));
     HPX_TEST(result.out() == std::end(d1));
@@ -83,9 +87,12 @@ void test_transform_binary_async(ExPolicy p, IteratorTag)
     f.wait();
 
     auto result = f.get();
-    HPX_TEST(hpx::util::get<0>(result) == std::end(c1));
-    HPX_TEST(hpx::util::get<1>(result) == std::end(c2));
-    HPX_TEST(hpx::util::get<2>(result) == std::end(d1));
+    // HPX_TEST(hpx::util::get<0>(result) == std::end(c1));
+    // HPX_TEST(hpx::util::get<1>(result) == std::end(c2));
+    // HPX_TEST(hpx::util::get<2>(result) == std::end(d1));
+    HPX_TEST(result.in1 == std::end(c1));
+    HPX_TEST(result.in2 == std::end(c2));
+    HPX_TEST(result.out == std::end(d1));
 
     // verify values
     std::vector<std::size_t> d2(c1.size());
