### 2020-08-06 c80acf6
Unify predicates in namespace ranges.
