### 2020-08-09 8c8cd44
Change transform result type (segmented included) from pair to in_out_result
