+++ libs/parallelism/algorithms/include/hpx/parallel/algorithms/copy.hpp
@@ -633,7 +633,7 @@ namespace hpx {
         tag_invoke(hpx::copy_t, ExPolicy&& policy, FwdIter1 first,
             FwdIter1 last, FwdIter2 dest)
         {
-            return parallel::util::get_second_element(
+            return parallel::util::detail::get_second_element(
                 parallel::v1::detail::transfer<
                     parallel::v1::detail::copy_iter<FwdIter1, FwdIter2>>(
                     std::forward<ExPolicy>(policy), first, last, dest));
@@ -649,7 +649,7 @@ namespace hpx {
         friend FwdIter2 tag_invoke(
             hpx::copy_t, FwdIter1 first, FwdIter1 last, FwdIter2 dest)
         {
-            return parallel::util::get_second_element(
+            return parallel::util::detail::get_second_element(
                 parallel::v1::detail::transfer<
                     parallel::v1::detail::copy_iter<FwdIter1, FwdIter2>>(
                     hpx::execution::seq, first, last, dest));
@@ -692,7 +692,7 @@ namespace hpx {
                 hpx::parallel::execution::is_sequenced_execution_policy<
                     ExPolicy>;
 
-            return hpx::parallel::util::get_second_element(
+            return hpx::parallel::util::detail::get_second_element(
                 hpx::parallel::v1::detail::copy_n<
                     hpx::parallel::util::in_out_result<FwdIter1, FwdIter2>>()
                     .call(std::forward<ExPolicy>(policy), is_seq(), first,
@@ -722,7 +722,7 @@ namespace hpx {
                     FwdIter2>::get(std::move(dest));
             }
 
-            return hpx::parallel::util::get_second_element(
+            return hpx::parallel::util::detail::get_second_element(
                 hpx::parallel::v1::detail::copy_n<
                     hpx::parallel::util::in_out_result<FwdIter1, FwdIter2>>()
                     .call(hpx::execution::seq, std::true_type{}, first,
@@ -762,7 +762,7 @@ namespace hpx {
                 hpx::parallel::execution::is_sequenced_execution_policy<
                     ExPolicy>;
 
-            return hpx::parallel::util::get_second_element(
+            return hpx::parallel::util::detail::get_second_element(
                 hpx::parallel::v1::detail::copy_if<
                     hpx::parallel::util::in_out_result<FwdIter1, FwdIter2>>()
                     .call(std::forward<ExPolicy>(policy), is_seq(), first, last,
@@ -788,7 +788,7 @@ namespace hpx {
             static_assert((hpx::traits::is_forward_iterator<FwdIter2>::value),
                 "Requires at least forward iterator.");
 
-            return hpx::parallel::util::get_second_element(
+            return hpx::parallel::util::detail::get_second_element(
                 hpx::parallel::v1::detail::copy_if<
                     hpx::parallel::util::in_out_result<FwdIter1, FwdIter2>>()
                     .call(hpx::execution::seq, std::true_type{}, first, last,
