### 2020-08-10 #2352 @danielskatz
👋 @davidbeckingsale

> Fine, just need to finish reviewing the actual paper. I have verified the software itself 👍

How is this coming along?
