### 2020-08-10 d1dfd68
Update docs/sphinx/contributing/contributing.rst
Co-authored-by: Mikael Simberg <mikael.simberg@iki.fi>
