+++ docs/sphinx/contributing/contributing.rst
@@ -8,13 +8,12 @@
 .. _contributing:
 
 ===================
-Contributing to HPX
+Contributing to |hpx|
 ===================
 
 The main source of information to understand the process of how to contribute
 to HPX can be found in
 `this document <https://github.com/STEllAR-GROUP/hpx/blob/master/.github/CONTRIBUTING.md>`_.
 This is a living document that is constantly updated with relevant information.
-
 
 
