### 2020-08-10 ab48f74
Update docs/sphinx/contributing/governance.rst
Co-authored-by: Mikael Simberg <mikael.simberg@iki.fi>
