### 2020-08-10 8cb6866
Update docs/sphinx/contributing/governance.rst
Co-authored-by: Mikael Simberg <mikael.simberg@iki.fi>
