### 2020-08-10 d1c3bd1
Fix config includes to match module name
