+++ libs/distributed_executors/include_compatibility/hpx/parallel/executors/distribution_policy_executor.hpp
@@ -7,7 +7,7 @@
 #pragma once
 
 #include <hpx/config.hpp>
-#include <hpx/execution/config/defines.hpp>
+#include <hpx/distributed_executors/config/defines.hpp>
 #include <hpx/modules/distributed_executors.hpp>
 
 #if HPX_EXECUTION_HAVE_DEPRECATION_WARNINGS
