### 2020-08-10 66bf5a6
Move default main implementation header to hpx_init
