### 2020-08-10 c410844
Removing non-needed libCDS support functions (for now)
