+++ libs/coroutines/include/hpx/coroutines/detail/coroutine_self.hpp
@@ -130,15 +130,8 @@ namespace hpx { namespace threads { namespace coroutines { namespace detail {
         virtual std::size_t set_thread_data(std::size_t data) = 0;
 
 #if defined(HPX_HAVE_LIBCDS)
-        virtual std::size_t get_libcds_data() const = 0;
-        virtual std::size_t set_libcds_data(std::size_t data) = 0;
-
         virtual std::size_t get_libcds_hazard_pointer_data() const = 0;
         virtual std::size_t set_libcds_hazard_pointer_data(
-            std::size_t data) = 0;
-
-        virtual std::size_t get_libcds_dynamic_hazard_pointer_data() const = 0;
-        virtual std::size_t set_libcds_dynamic_hazard_pointer_data(
             std::size_t data) = 0;
 #endif
 
