+++ libs/threading_base/include/hpx/threading_base/thread_data_stackless.hpp
@@ -88,16 +88,6 @@ namespace hpx { namespace threads {
         }
 
 #if defined(HPX_HAVE_LIBCDS)
-        std::size_t get_libcds_data() const override
-        {
-            return coroutine_.get_libcds_data();
-        }
-
-        std::size_t set_libcds_data(std::size_t data) override
-        {
-            return coroutine_.set_libcds_data(data);
-        }
-
         std::size_t get_libcds_hazard_pointer_data() const override
         {
             return coroutine_.get_libcds_hazard_pointer_data();
@@ -106,17 +96,6 @@ namespace hpx { namespace threads {
         std::size_t set_libcds_hazard_pointer_data(std::size_t data) override
         {
             return coroutine_.set_libcds_hazard_pointer_data(data);
-        }
-
-        std::size_t get_libcds_dynamic_hazard_pointer_data() const override
-        {
-            return coroutine_.get_libcds_dynamic_hazard_pointer_data();
-        }
-
-        std::size_t set_libcds_dynamic_hazard_pointer_data(
-            std::size_t data) override
-        {
-            return coroutine_.set_libcds_dynamic_hazard_pointer_data(data);
         }
 #endif
 
