+++ libs/threading_base/include/hpx/threading_base/thread_helpers.hpp
@@ -393,22 +393,10 @@ namespace hpx { namespace threads {
         thread_id_type const& id, std::size_t data, error_code& ec = throws);
 
 #if defined(HPX_HAVE_LIBCDS)
-    HPX_EXPORT std::size_t get_libcds_data(
-        thread_id_type const& id, error_code& ec = throws);
-
-    HPX_EXPORT std::size_t set_libcds_data(
-        thread_id_type const& id, std::size_t data, error_code& ec = throws);
-
     HPX_EXPORT std::size_t get_libcds_hazard_pointer_data(
         thread_id_type const& id, error_code& ec = throws);
 
     HPX_EXPORT std::size_t set_libcds_hazard_pointer_data(
-        thread_id_type const& id, std::size_t data, error_code& ec = throws);
-
-    HPX_EXPORT std::size_t get_libcds_dynamic_hazard_pointer_data(
-        thread_id_type const& id, error_code& ec = throws);
-
-    HPX_EXPORT std::size_t set_libcds_dynamic_hazard_pointer_data(
         thread_id_type const& id, std::size_t data, error_code& ec = throws);
 #endif
 
