+++ libs/threading/src/thread.cpp
@@ -248,15 +248,6 @@ namespace hpx {
     }
 
 #if defined(HPX_HAVE_LIBCDS)
-    std::size_t thread::get_libcds_data() const
-    {
-        return threads::get_libcds_data(native_handle());
-    }
-    std::size_t thread::set_libcds_data(std::size_t data)
-    {
-        return threads::set_libcds_data(native_handle(), data);
-    }
-
     std::size_t thread::get_libcds_hazard_pointer_data() const
     {
         return threads::get_libcds_hazard_pointer_data(native_handle());
@@ -264,16 +255,6 @@ namespace hpx {
     std::size_t thread::set_libcds_hazard_pointer_data(std::size_t data)
     {
         return threads::set_libcds_hazard_pointer_data(native_handle(), data);
-    }
-
-    std::size_t thread::get_libcds_dynamic_hazard_pointer_data() const
-    {
-        return threads::get_libcds_dynamic_hazard_pointer_data(native_handle());
-    }
-    std::size_t thread::set_libcds_dynamic_hazard_pointer_data(std::size_t data)
-    {
-        return threads::set_libcds_dynamic_hazard_pointer_data(
-            native_handle(), data);
     }
 #endif
 
@@ -428,16 +409,6 @@ namespace hpx {
         }
 
 #if defined(HPX_HAVE_LIBCDS)
-        std::size_t get_libcds_data()
-        {
-            return threads::get_libcds_data(threads::get_self_id());
-        }
-
-        std::size_t set_libcds_data(std::size_t data)
-        {
-            return threads::set_libcds_data(threads::get_self_id(), data);
-        }
-
         std::size_t get_libcds_hazard_pointer_data()
         {
             return threads::get_libcds_hazard_pointer_data(
@@ -447,18 +418,6 @@ namespace hpx {
         std::size_t set_libcds_hazard_pointer_data(std::size_t data)
         {
             return threads::set_libcds_hazard_pointer_data(
-                threads::get_self_id(), data);
-        }
-
-        std::size_t get_libcds_dynamic_hazard_pointer_data()
-        {
-            return threads::get_libcds_dynamic_hazard_pointer_data(
-                threads::get_self_id());
-        }
-
-        std::size_t set_libcds_dynamic_hazard_pointer_data(std::size_t data)
-        {
-            return threads::set_libcds_dynamic_hazard_pointer_data(
                 threads::get_self_id(), data);
         }
 #endif

[…]
