### 2020-08-11 c2d9315
Adding replay_executor and replicate_executor
- this allows for instance for the parallel algorithms to be resilient
  to failing iterations
