+++ libs/resiliency/include/hpx/resiliency/async_replicate_executor.hpp
@@ -30,80 +30,150 @@ namespace hpx { namespace resiliency { namespace experimental {
     namespace detail {
 
         ///////////////////////////////////////////////////////////////////////
-        template <typename Executor, typename Vote, typename Pred, typename F,
-            typename... Ts>
-        typename hpx::traits::executor_future<Executor,
-            typename hpx::util::detail::invoke_deferred_result<F,
-                Ts...>::type>::type
-        async_replicate_vote_validate_executor(Executor&& exec, std::size_t n,
-            Vote&& vote, Pred&& pred, F&& f, Ts&&... ts)
+        template <typename Result>
+        struct async_replicate_vote_validate_executor
         {
-            using result_type =
-                typename hpx::util::detail::invoke_deferred_result<F,
-                    Ts...>::type;
+            template <typename Executor, typename Vote, typename Pred,
+                typename F, typename... Ts>
+            static typename hpx::traits::executor_future<Executor, Result>::type
+            call(Executor&& exec, std::size_t n, Vote&& vote, Pred&& pred,
+                F&& f, Ts&&... ts)
+            {
+                using result_type =
+                    typename hpx::util::detail::invoke_deferred_result<F,
+                        Ts...>::type;
 
-            using future_type = typename hpx::traits::executor_future<Executor,
-                result_type>::type;
+                using future_type =
+                    typename hpx::traits::executor_future<Executor,
+                        result_type>::type;
 
-            // launch given function n times
-            auto func = [f = std::forward<F>(f),
-                            t = hpx::util::make_tuple(std::forward<Ts>(ts)...)](
-                            std::size_t) -> result_type {
-                // ignore argument (invocation count of bulk_execute)
-                return hpx::util::invoke_fused(f, t);
-            };
+                // launch given function n times
+                auto func =
+                    [f = std::forward<F>(f),
+                        t = hpx::util::make_tuple(std::forward<Ts>(ts)...)](
+                        std::size_t) mutable -> result_type {
+                    // ignore argument (invocation count of bulk_execute)
+                    return hpx::util::invoke_fused(f, t);
+                };
 
-            std::vector<future_type> results =
-                hpx::parallel::execution::bulk_async_execute(
-                    std::forward<Executor>(exec), std::move(func), n);
+                std::vector<future_type> results =
+                    hpx::parallel::execution::bulk_async_execute(
+                        std::forward<Executor>(exec), std::move(func), n);
 
-            // wait for all threads to finish executing and return the first
-            // result that passes the predicate, properly handle exceptions
-            // do not schedule new thread for the lambda
-            return hpx::dataflow(
-                hpx::launch::sync,
-                [pred = std::forward<Pred>(pred),
-                    vote = std::forward<Vote>(vote), n](
-                    std::vector<future_type>&& results) mutable -> result_type {
-                    // Store all valid results
-                    std::vector<result_type> valid_results;
-                    valid_results.reserve(n);
+                // wait for all threads to finish executing and return the first
+                // result that passes the predicate, properly handle exceptions
+                // do not schedule new thread for the lambda
+                return hpx::dataflow(
+                    hpx::launch::sync,
+                    [pred = std::forward<Pred>(pred),
+                        vote = std::forward<Vote>(vote),
+                        n](std::vector<future_type>&& results) mutable
+                    -> result_type {
+                        // Store all valid results
+                        std::vector<result_type> valid_results;
+                        valid_results.reserve(n);
 
-                    std::exception_ptr ex;
+                        std::exception_ptr ex;
 
-                    for (auto&& f : std::move(results))
-                    {
-                        if (f.has_exception())
+                        for (auto&& f : std::move(results))
[…]
