### 2020-08-11 abd8ea0
Align heading underlines with heading length
