### 2020-08-11 d1ed41e
Add projection_identity specialization for transform and fix argument type in tag result type
