### 2020-08-12 a902f0a
Update the generate_issue_pr_list script to be faster
when just updating the last PRs/issues
