### 2020-08-12 e4cdd9d
Add a CMake option to handle pop_bulk parts
