### 2020-08-12 2d1bcf7
Fix exception handling for tasks
Setting the exception on a future may yield (a lock is required), and thus lead
to the task being resumed on another worker thread after setting the exception.
Before this change the exception for a future was set directly in the catch
block which could lead to the catch block being entered on one worker thread
and exited on another thread. This seems to not be allowed and leads to
segfaults at least with gcc and clang. This change only gets the current
exception in the catch block, but sets it after the catch block to avoid
potentially yielding in the catch bl
[…]
