### 2020-08-13 26a22fe
Update more catch blocks not to yield from setting exceptions
