+++ libs/lcos_local/include/hpx/lcos_local/packaged_task.hpp
@@ -135,28 +135,46 @@ namespace hpx { namespace lcos { namespace local {
         template <typename... Vs>
         void invoke_impl(/*is_void=*/std::false_type, Vs&&... vs)
         {
+            std::exception_ptr p;
+
             try
             {
                 promise_.set_value(function_(std::forward<Vs>(vs)...));
+                return;
             }
             catch (...)
             {
-                promise_.set_exception(std::current_exception());
+                p = std::current_exception();
             }
+
+            // The exception is set outside the catch block since
+            // set_exception may yield. Ending the catch block on a
+            // different worker thread than where it was started may lead
+            // to segfaults.
+            promise_.set_exception(std::move(p));
         }
 
         template <typename... Vs>
         void invoke_impl(/*is_void=*/std::true_type, Vs&&... vs)
         {
+            std::exception_ptr p;
+
             try
             {
                 function_(std::forward<Ts>(vs)...);
                 promise_.set_value();
+                return;
             }
             catch (...)
             {
-                promise_.set_exception(std::current_exception());
+                p = std::current_exception();
             }
+
+            // The exception is set outside the catch block since
+            // set_exception may yield. Ending the catch block on a
+            // different worker thread than where it was started may lead
+            // to segfaults.
+            promise_.set_exception(std::move(p));
         }
 
     private:
