### 2020-08-13 c1e1ad5
Making generate() and generate_n conforming to C++20
