+++ libs/algorithms/include/hpx/parallel/algorithms/generate.hpp
@@ -1,5 +1,5 @@
 //  Copyright (c) 2014 Grant Mercer
-//  Copyright (c) 2017 Hartmut Kaiser
+//  Copyright (c) 2017-2020 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -9,83 +9,9 @@
 
 #pragma once
 
-#include <hpx/config.hpp>
-#include <hpx/algorithms/traits/segmented_iterator_traits.hpp>
-#include <hpx/concepts/concepts.hpp>
-#include <hpx/iterator_support/traits/is_iterator.hpp>
-
-#include <hpx/execution/algorithms/detail/is_negative.hpp>
-#include <hpx/executors/execution_policy.hpp>
-#include <hpx/parallel/algorithms/detail/dispatch.hpp>
-#include <hpx/parallel/algorithms/for_each.hpp>
-#include <hpx/parallel/util/detail/algorithm_result.hpp>
-#include <hpx/parallel/util/projection_identity.hpp>
-
-#include <algorithm>
-#include <cstddef>
-#include <iterator>
-#include <type_traits>
-#include <utility>
-
-namespace hpx { namespace parallel { inline namespace v1 {
-    ///////////////////////////////////////////////////////////////////////////
-    // generate
-    namespace detail {
-        /// \cond NOINTERNAL
-        template <typename Iter>
-        struct generate : public detail::algorithm<generate<Iter>, Iter>
-        {
-            generate()
-              : generate::algorithm("generate")
-            {
-            }
-
-            template <typename ExPolicy, typename InIter, typename F>
-            static InIter sequential(ExPolicy, InIter first, InIter last, F&& f)
-            {
-                std::generate(first, last, std::forward<F>(f));
-                return last;
-            }
-
-            template <typename ExPolicy, typename FwdIter, typename F>
-            static
-                typename util::detail::algorithm_result<ExPolicy, FwdIter>::type
-                parallel(ExPolicy&& policy, FwdIter first, FwdIter last, F&& f)
-            {
-                typedef typename std::iterator_traits<FwdIter>::value_type type;
-
-                return for_each_n<FwdIter>().call(
-                    std::forward<ExPolicy>(policy), std::false_type(), first,
-                    std::distance(first, last),
-                    [f = std::forward<F>(f)](type& v) mutable { v = f(); },
-                    util::projection_identity());
-            }
-        };
-
-        ///////////////////////////////////////////////////////////////////////
-        // non-segmented implementation
-        template <typename ExPolicy, typename FwdIter, typename F>
-        inline typename util::detail::algorithm_result<ExPolicy, FwdIter>::type
-        generate_(ExPolicy&& policy, FwdIter first, FwdIter last, F&& f,
-            std::false_type)
-        {
-            typedef parallel::execution::is_sequenced_execution_policy<ExPolicy>
-                is_seq;
-
-            return detail::generate<FwdIter>().call(
-                std::forward<ExPolicy>(policy), is_seq(), first, last,
-                std::forward<F>(f));
-        }
-
-        ///////////////////////////////////////////////////////////////////////
-        // segmented implementation
-        template <typename ExPolicy, typename FwdIter, typename F>
-        inline typename util::detail::algorithm_result<ExPolicy, FwdIter>::type
-        generate_(ExPolicy&& policy, FwdIter first, FwdIter last, F&& f,
-            std::true_type);
-
-        /// \endcond
-    }    // namespace detail
+#if defined(DOXYGEN)
+namespace hpx {
+    // clang-format off
 
     /// Assign each element in range [first, last) a value generated by the
     /// given function object f
[…]
