+++ libs/include/include/hpx/algorithm.hpp
@@ -24,8 +24,6 @@ namespace hpx {
     using hpx::parallel::for_loop_n;
     using hpx::parallel::for_loop_n_strided;
     using hpx::parallel::for_loop_strided;
-    using hpx::parallel::generate;
-    using hpx::parallel::generate_n;
     using hpx::parallel::includes;
     using hpx::parallel::inplace_merge;
     using hpx::parallel::is_heap;
