### 2020-08-14 fb7edb3
Minor tweak to hpx::equal implementation
