### 2020-08-14 a082b2f
Suggested changes for HPX V1.5 release notes
