+++ docs/sphinx/releases/whats_new_1_5_0.rst
@@ -21,39 +21,48 @@ conformant. As part of the modularization we have begun specifying the public
 API of |hpx| in terms of headers and functionality, and aligning it more closely
 to the C++ standard. All non-distributed modules are now in place, along with an
 experimental option to completely disable distributed features in |hpx|. We have
-also added an experimental asynchronous MPI executor. Lastly this release
-contains official |cmake| targets, performance improvements, and many bug fixes.
+also added experimental asynchronous MPI and CUDA executors. Lastly this release
+introduces |cmake| targets for depending projects, performance improvements,
+and many bug fixes.
 
-* We have added the C++20 features ``hpx::jthread`` and ``hpx::stop_token``. We
-  have also added ``hpx::stable_sort`` based on Francisco Tapia's
+* We have added the C++20 features ``hpx::jthread`` and ``hpx::stop_token``.
+  ``hpx::condition_variable_any`` now exposes new functions supporting
+  ``hpx::stop_token``.
+* We have added ``hpx::stable_sort`` based on Francisco Tapia's
   implementation.
 * We have adapted existing sychronization primitives to be fully conformant
-  C++20: ``barrier``, ``latch``, ``counting_semaphore``.
-* We have started using customization point objects to make the corresponding
+  C++20: ``hpx::barrier``, ``hpx::latch``, ``hpx::counting_semaphore``, and
+  ``hpx::binary_semaphore``.
+* We have started using customization point objects (CPOs) to make the corresponding
   algorithms fully conformant to C++20 as well as to make algorithm extension
   easier for the user. ``all_of``/``any_of``/``none_of``, ``copy``, ``count``,
   ``destroy``, ``equal``, ``fill``, ``mismatch``, ``move``, ``reduce`` are using
-  those CPOs.  We also have adapted their corresponding ``hpx::ranges`` versions
-  in this release.
-* We have adapted support for ``co_await`` to C++20, it now supports
-  ``shared_future`` and have added allocator support for futures returned by
-  ``co_return``. It is no longer in the ``experimental`` namespace.
+  those CPOs (all in namespace ``hpx``).  We also have adapted their
+  corresponding ``hpx::ranges`` versions to be conforming to C++20 in this
+  release.
+* We have adapted support for ``co_await`` to C++20, in addition to
+  ``hpx::future`` it now also supports ``hpx::shared_future``. We have also
+  added allocator support for futures returned by ``co_return``. It is no longer
+  in the ``experimental`` namespace.
 * We added serialization support for ``std::variant`` and ``std::tuple``.
 * ``result_of`` and ``is_callable`` are now deprecated and replaced by
   ``invoke_result`` and ``is_invocable`` to conform to C++20.
 * We continued with the modularization, making it easier for us to add the new
   experimental ``HPX_WITH_DISTRIBUTED_RUNTIME`` |cmake| option (see below) . An
-  important amount of headers have been deprecated. We adapted the namespaces
+  significant amount of headers have been deprecated. We adapted the namespaces
   and headers we could to be closer to the standard ones (:ref:`api
-  <_public_api>`).
+  <_public_api>`). Depending code should stil compile, however warnings are now
+  generated instructing to change the ``#include``s accordingly.
 * It is now possible to have a basic CUDA support including a helper function to
   get a future from a CUDA stream and target handling. They are available under
   the ``hpx::cuda::experimental`` namespace and they can be enabled with the
-  ``HPX_WITH_ASYNC_CUDA=ON`` cmake option.
+  ``-DHPX_WITH_ASYNC_CUDA=ON`` |cmake| option.
 * We added a new ``hpx::mpi::experimental`` namespace for getting futures from
-  an asynchronous MPI call and a new minimal MPI executor (add the name).
+  an asynchronous MPI call and a new minimal MPI executor
[…]
