### 2020-08-14 a1568ab
Adding unity build option
