### 2020-08-14 834969e
Adding a note about required compilers and prerequisite versions to 'breaking changes'
