### 2020-08-16 6edcb5d
Changing new allocator to support jemalloc (on Windows, mostly) as well.
