### 2020-08-16 228f511
Changing new allocator to support jemalloc (on Windows, mostly) as well.
