### 2020-08-17 #2352 @bhatele
@danielskatz @diehlpk I read the software paper. I believe these two things from the list of requirements above are missing:
1. The paper does not have a clear statement of need -- what problems is this software designed to solve (note that the paper lists the benefits of using the software)
2. Even though the paper has list of AMT's that are similar, it does not provide how this software compares to them.
