### 2020-08-17 6d7dfcf
Cancel previous pull requests builds before starting a new one with Jenkins
