### 2020-08-17 9d580e7
Fixing unused variable warning
