+++ libs/coroutines/include/hpx/coroutines/stackless_coroutine.hpp
@@ -58,7 +58,7 @@ namespace hpx { namespace threads { namespace coroutines {
             util::unique_function_nonser<result_type(arg_type)>;
 
         stackless_coroutine(functor_type&& f, thread_id_type id,
-            std::ptrdiff_t stack_size = default_stack_size)
+            std::ptrdiff_t /*stack_size*/ = default_stack_size)
           : f_(std::move(f))
           , state_(ctx_ready)
           , id_(id)
