### 2020-08-17 96eea98
Adding support for aligned_alloc on systems that don't have it
- fixing formatting problems
- adding missing header
