### 2020-08-17 5c179b5
Adding support for aligned_alloc on systems that don't have it
- fixing formatting problems
- adding missing header
