+++ cmake/HPX_PerformCxxFeatureTests.cmake
@@ -27,6 +27,8 @@ function(hpx_perform_cxx_feature_tests)
     DEFINITIONS HPX_HAVE_CXX11_STD_SHARED_PTR_LWG3018
   )
 
+  hpx_check_for_c11_aligned_alloc(DEFINITIONS HPX_HAVE_C11_ALIGNED_ALLOC)
+
   hpx_check_for_cxx17_filesystem(DEFINITIONS HPX_HAVE_CXX17_FILESYSTEM)
 
   hpx_check_for_cxx17_fold_expressions(
