+++ cmake/HPX_AddConfigTest.cmake
@@ -317,6 +317,14 @@ function(hpx_check_for_cxx11_std_shared_ptr_lwg3018)
 endfunction()
 
 # ##############################################################################
+function(hpx_check_for_c11_aligned_alloc)
+  add_hpx_config_test(
+    HPX_WITH_C11_ALIGNED_ALLOC
+    SOURCE cmake/tests/c11_aligned_alloc.cpp FILE ${ARGN}
+  )
+endfunction()
+
+# ##############################################################################
 function(hpx_check_for_cxx11_std_quick_exit)
   add_hpx_config_test(
     HPX_WITH_CXX11_STD_QUICK_EXIT SOURCE cmake/tests/cxx11_std_quick_exit.cpp
