### 2020-08-17 #2352 @hkaiser
> 1. The paper does not have a clear statement of need -- what problems is this software designed to solve (note that the paper lists the benefits of using the software)

We thought that paragraph 4 (the one starting with
```
HPX’s main goal is to improve efficiency and scalability of parallel applications by increasing
resource utilization and reducing synchronization through providing an asynchronous API and
employing adaptive scheduling.
```
) would be sufficient. But I will try to sharpen this to emphasize the goal of HPX.
