### 2020-08-17 #2352 @danielskatz
@hkaiser - You might think of this the other way around: not what is the goal of HPX, but what is the goal of users of HPX that HPX satisfies.
