### 2020-08-17 0f23c50
Fixing allocators to throw on OOM conditions, fixing deque template parameters
