+++ libs/schedulers/include/hpx/schedulers/lockfree_queue_backends.hpp
@@ -37,7 +37,8 @@ namespace hpx { namespace threads { namespace policies {
     {
 #if defined(HPX_HAVE_CXX11_STD_ATOMIC_128BIT)
         using container_type =
-            boost::lockfree::deque<T, hpx::util::aligned_allocator<T>>;
+            boost::lockfree::deque<T, boost::lockfree::caching_freelist_t,
+                hpx::util::aligned_allocator<T>>;
 #else
         using container_type =
             boost::lockfree::queue<T, hpx::util::aligned_allocator<T>>;
@@ -142,162 +143,165 @@ namespace hpx { namespace threads { namespace policies {
         };
     };
 
-// LIFO
+    // LIFO
 #if defined(HPX_HAVE_CXX11_STD_ATOMIC_128BIT)
-            struct lockfree_lifo;
+    struct lockfree_lifo;
 
-            template <typename T>
-            struct lockfree_lifo_backend
-            {
-                using container_type =
-                    boost::lockfree::deque<T, hpx::util::aligned_allocator<T>>;
+    template <typename T>
+    struct lockfree_lifo_backend
+    {
+        using container_type =
+            boost::lockfree::deque<T, boost::lockfree::caching_freelist_t,
+                hpx::util::aligned_allocator<T>>;
 
-                using value_type = T;
-                using reference = T&;
-                using const_reference = T const&;
-                using size_type = std::uint64_t;
+        using value_type = T;
+        using reference = T&;
+        using const_reference = T const&;
+        using size_type = std::uint64_t;
 
-                lockfree_lifo_backend(size_type initial_size = 0,
-                    size_type num_thread = size_type(-1))
-                  : queue_(std::size_t(initial_size))
-                {
-                }
+        lockfree_lifo_backend(
+            size_type initial_size = 0, size_type num_thread = size_type(-1))
+          : queue_(std::size_t(initial_size))
+        {
+        }
 
-                bool push(const_reference val, bool other_end = false)
-                {
-                    if (other_end)
-                        return queue_.push_right(val);
-                    return queue_.push_left(val);
-                }
+        bool push(const_reference val, bool other_end = false)
+        {
+            if (other_end)
+                return queue_.push_right(val);
+            return queue_.push_left(val);
+        }
 
-                bool pop(reference val, bool steal = true)
-                {
-                    return queue_.pop_left(val);
-                }
+        bool pop(reference val, bool steal = true)
+        {
+            return queue_.pop_left(val);
+        }
 
-                bool empty()
-                {
-                    return queue_.empty();
-                }
+        bool empty()
+        {
+            return queue_.empty();
+        }
 
-            private:
-                container_type queue_;
-            };
+    private:
+        container_type queue_;
+    };
 
-            struct lockfree_lifo
-            {
-                template <typename T>
-                struct apply
-                {
-                    using type = lockfree_lifo_backend<T>;
-                };
-            };
+    struct lockfree_lifo
+    {
+        template <typename T>
+        struct apply
+        {
+            using type = lockfree_lifo_backend<T>;
+        };
+    };
 
 #if defined(HPX_HAVE_ABP_SCHEDULER)
-            ////////////////////////////////////////////////////////////////////////////
-            // FIFO + stealing at opposite end.
-            struct lockfree_abp_fifo;
-            struct lockfree_abp_lifo;
+    ////////////////////////////////////////////////////////////////////////////
+    // FIFO + stealing at opposite end.
+    struct lockfree_abp_fifo;
+    struct lockfree_abp_lifo;
 
-            template <typename T>
-            struct lockfree_abp_fifo_backend
-            {
-                using container_type =
-                    boost::lockfree::deque<T, hpx::util::aligned_allocator<T>>;
+    template <typename T>
+    struct lockfree_abp_fifo_backend
+    {
+        using container_type =
[…]
