### 2020-08-17 0984fd6
Fixing allocators to throw on OOM conditions, fixing deque template parameters
