+++ cmake/HPX_AddConfigTest.cmake
@@ -360,6 +360,13 @@ function(hpx_check_for_c11_aligned_alloc)
   )
 endfunction()
 
+function(hpx_check_for_cxx17_std_aligned_alloc)
+  add_hpx_config_test(
+    HPX_WITH_CXX17_STD_ALIGNED_ALLOC
+    SOURCE cmake/tests/cxx17_std_aligned_alloc.cpp FILE ${ARGN}
+  )
+endfunction()
+
 # ##############################################################################
 function(hpx_check_for_cxx11_std_quick_exit)
   add_hpx_config_test(
