### 2020-08-17 #2352 @bhatele
@diehlpk @hkaiser I thought the need should mention what's missing or challenging to do in the current models (whether bulk synchronous or AMT), and how HPX solves those challenges or "missing" features. I couldn't find this in the current paper, basically "Why HPX?" and not continue to use MPI or whatever model I like.
