### 2020-08-17 #2352 @bhatele
My understand of the comparison was that this would read something like the related work section of a paper, and have statements such as Uintah does X, and Chapel does Y but we do X better and/or provide Z. I'll leave this up to @danielskatz to decide what level of detail should go in this software paper.
