### 2020-08-18 #2352 @diehlpk
> My understand of the comparison was that this would read something like the related work section of a paper, and have statements such as Uintah does X, and Chapel does Y but we do X better and/or provide Z. I'll leave this up to @danielskatz to decide what level of detail should go in this software paper.

@bhatele We refer to this comparative review paper about AMTs

Thoman, P., Dichev, K., Heller, T. et al. A taxonomy of task-based parallel programming technologies for high-performance > computing. J Supercomput 74, 1422–1434 (2018). https://doi.org/10.1007/s11227-018-2238-4

and were thinking that we just mention the high-level differences. The interested reader could follow up in this reference for more technical details.
