### 2020-08-18 5032540
Add CPOs for for_loop algorithms
