+++ libs/algorithms/tests/unit/algorithms/for_loop_exception.cpp
@@ -4,11 +4,10 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
+#include <hpx/algorithm.hpp>
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
 #include <hpx/modules/testing.hpp>
-
-#include <hpx/include/parallel_for_loop.hpp>
 
 #include <algorithm>
 #include <atomic>
@@ -83,9 +82,8 @@ void test_for_loop_exception(ExPolicy&& policy, IteratorTag)
     bool caught_exception = false;
     try
     {
-        hpx::parallel::for_loop(std::forward<ExPolicy>(policy),
-            iterator(std::begin(c)), iterator(std::end(c)),
-            throw_always(dis(gen)));
+        hpx::for_loop(std::forward<ExPolicy>(policy), iterator(std::begin(c)),
+            iterator(std::end(c)), throw_always(dis(gen)));
 
         HPX_TEST(false);
     }
@@ -117,9 +115,9 @@ void test_for_loop_exception_async(ExPolicy&& p, IteratorTag)
     bool returned_from_algorithm = false;
     try
     {
-        auto f = hpx::parallel::for_loop(std::forward<ExPolicy>(p),
-            iterator(std::begin(c)), iterator(std::end(c)),
-            throw_always(dis(gen)));
+        auto f =
+            hpx::for_loop(std::forward<ExPolicy>(p), iterator(std::begin(c)),
+                iterator(std::end(c)), throw_always(dis(gen)));
         returned_from_algorithm = true;
         f.get();
 
@@ -158,9 +156,8 @@ void test_for_loop_bad_alloc(ExPolicy policy, IteratorTag)
     bool caught_exception = false;
     try
     {
-        hpx::parallel::for_loop(std::forward<ExPolicy>(policy),
-            iterator(std::begin(c)), iterator(std::end(c)),
-            throw_bad_alloc(dis(gen)));
+        hpx::for_loop(std::forward<ExPolicy>(policy), iterator(std::begin(c)),
+            iterator(std::end(c)), throw_bad_alloc(dis(gen)));
 
         HPX_TEST(false);
     }
@@ -191,9 +188,9 @@ void test_for_loop_bad_alloc_async(ExPolicy p, IteratorTag)
     bool returned_from_algorithm = false;
     try
     {
-        auto f = hpx::parallel::for_loop(std::forward<ExPolicy>(p),
-            iterator(std::begin(c)), iterator(std::end(c)),
-            throw_bad_alloc(dis(gen)));
+        auto f =
+            hpx::for_loop(std::forward<ExPolicy>(p), iterator(std::begin(c)),
+                iterator(std::end(c)), throw_bad_alloc(dis(gen)));
         returned_from_algorithm = true;
         f.get();
 
@@ -260,7 +257,7 @@ void test_for_loop_idx_exception(ExPolicy&& policy, IteratorTag)
     bool caught_exception = false;
     try
     {
-        hpx::parallel::for_loop(std::forward<ExPolicy>(policy), 0, c.size(),
+        hpx::for_loop(std::forward<ExPolicy>(policy), 0, c.size(),
             throw_always(dis(gen)));
 
         HPX_TEST(false);
@@ -292,7 +289,7 @@ void test_for_loop_idx_exception_async(ExPolicy&& p, IteratorTag)
     bool returned_from_algorithm = false;
     try
     {
-        auto f = hpx::parallel::for_loop(
+        auto f = hpx::for_loop(
[…]
