### 2020-08-18 a4e0f9d
Update docs with numeric.hpp header and which headers to include for different algorithms
