+++ docs/sphinx/manual/writing_single_node_hpx_applications.rst
@@ -349,7 +349,7 @@ Parallel algorithms
 
 |hpx| provides implementations of the following parallel algorithms:
 
-.. list-table:: Non-modifying parallel algorithms (in header: ``<hpx/include/parallel_algorithm.hpp>``)
+.. list-table:: Non-modifying parallel algorithms (in header: ``<hpx/algorithm.hpp>``)
 
    * * Name
      * Description
@@ -357,86 +357,78 @@ Parallel algorithms
      * Algorithm page at cppreference.com
    * * :cpp:func:`hpx::parallel::v1::adjacent_find`
      * Computes the differences between adjacent elements in a range.
-     * ``<hpx/include/parallel_adjacent_find.hpp>``
+     * ``<hpx/algorithm.hpp>``
      * :cppreference-algorithm:`adjacent_find`
-   * * :cpp:func:`hpx::parallel::v1::all_of`
+   * * :cpp:func:`hpx::all_of`
      * Checks if a predicate is ``true`` for all of the elements in a range.
-     * ``<hpx/include/parallel_all_any_none.hpp>``
+     * ``<hpx/algorithm.hpp>``
      * :cppreference-algorithm:`all_any_none_of`
-   * * :cpp:func:`hpx::parallel::v1::any_of`
+   * * :cpp:func:`hpx::any_of`
      * Checks if a predicate is ``true`` for any of the elements in a range.
-     * ``<hpx/include/parallel_all_any_none.hpp>``
+     * ``<hpx/algorithm.hpp>``
      * :cppreference-algorithm:`all_any_none_of`
-   * * :cpp:func:`hpx::parallel::v1::count`
+   * * :cpp:func:`hpx::count`
      * Returns the number of elements equal to a given value.
-     * ``<hpx/include/parallel_count.hpp>``
+     * ``<hpx/algorithm.hpp>``
      * :cppreference-algorithm:`count`
-   * * :cpp:func:`hpx::parallel::v1::count_if`
+   * * :cpp:func:`hpx::count_if`
      * Returns the number of elements satisfying a specific criteria.
-     * ``<hpx/include/parallel_count.hpp>``
+     * ``<hpx/algorithm.hpp>``
      * :cppreference-algorithm:`count_if`
-   * * :cpp:func:`hpx::parallel::v1::equal`
+   * * :cpp:func:`hpx::equal`
      * Determines if two sets of elements are the same.
-     * ``<hpx/include/parallel_equal.hpp>``
+     * ``<hpx/algorithm.hpp>``
      * :cppreference-algorithm:`equal`
-   * * :cpp:func:`hpx::parallel::v1::exclusive_scan`
-     * Does an exclusive parallel scan over a range of elements.
-     * ``<hpx/include/parallel_scan.hpp>``
-     * :cppreference-algorithm:`exclusive_scan`
-   * * :cpp:func:`hpx::parallel::v1::find`
+   * * :cpp:func:`hpx::find`
      * Finds the first element equal to a given value.
-     * ``<hpx/include/parallel_find.hpp>``
+     * ``<hpx/algorithm.hpp>``
      * :cppreference-algorithm:`find`
-   * * :cpp:func:`hpx::parallel::v1::find_end`
+   * * :cpp:func:`hpx::find_end`
      * Finds the last sequence of elements in a certain range.
-     * ``<hpx/include/parallel_find.hpp>``
+     * ``<hpx/algorithm.hpp>``
      * :cppreference-algorithm:`find_end`
-   * * :cpp:func:`hpx::parallel::v1::find_first_of`
+   * * :cpp:func:`hpx::find_first_of`
      * Searches for any one of a set of elements.
-     * ``<hpx/include/parallel_find.hpp>``
+     * ``<hpx/algorithm.hpp>``
[…]
