### 2020-08-18 #2352 @danielskatz
> were thinking that we just mention the high-level differences. The interested reader could follow up this reference for more technical details.

This seems reasonable to me.
