### 2020-08-18 d18986a
Explicitly specify hwloc root on Jenkins CSCS builds
