### 2020-08-18 e1e9559
Add tests for overloads without exec-policy argument
