+++ libs/parallelism/algorithms/tests/unit/algorithms/transform.cpp
@@ -19,6 +19,8 @@ void test_transform()
 {
     using namespace hpx::execution;
 
+    test_transform(IteratorTag());
+
     test_transform(seq, IteratorTag());
     test_transform(par, IteratorTag());
     test_transform(par_unseq, IteratorTag());
@@ -38,6 +40,7 @@ void test_transform_exception()
 {
     using namespace hpx::execution;
 
+    test_transform_exception(IteratorTag());
     // If the execution policy object is of type vector_execution_policy,
     // std::terminate shall be called. therefore we do not test exceptions
     // with a vector execution policy
