### 2020-08-18 808f4e5
Add in_in_out_result utility
