### 2020-08-18 ab78dbc
Change binary transform result type to in_in_out_result
