+++ libs/full/segmented_algorithms/include/hpx/parallel/segmented_algorithms/detail/dispatch.hpp
@@ -103,7 +103,36 @@ namespace hpx { namespace parallel { inline namespace v1 { namespace detail {
     };
 
     template <typename Iterator1, typename Iterator2, typename Iterator3>
-    struct algorithm_result_helper<hpx::tuple<Iterator1, Iterator2, Iterator3>,
+    struct algorithm_result_helper<
+        util::in_in_out_result<Iterator1, Iterator2, Iterator3>,
+        typename std::enable_if<
+            hpx::traits::is_segmented_local_iterator<Iterator1>::value ||
+            hpx::traits::is_segmented_local_iterator<Iterator2>::value ||
+            hpx::traits::is_segmented_local_iterator<Iterator3>::value>::type>
+    {
+        typedef hpx::traits::segmented_local_iterator_traits<Iterator1> traits1;
+        typedef hpx::traits::segmented_local_iterator_traits<Iterator2> traits2;
+        typedef hpx::traits::segmented_local_iterator_traits<Iterator3> traits3;
+
+        static HPX_FORCEINLINE util::in_in_out_result<
+            typename traits1::local_iterator, typename traits2::local_iterator,
+            typename traits3::local_iterator>
+        call(util::in_in_out_result<typename traits1::local_raw_iterator,
+            typename traits2::local_raw_iterator,
+            typename traits3::local_raw_iterator>&& p)
+        {
+            return util::in_in_out_result<typename traits1::local_iterator,
+                typename traits2::local_iterator,
+                typename traits3::local_iterator>{
+                traits1::remote(std::move(p.in1)),
+                traits2::remote(std::move(p.in2)),
+                traits3::remote(std::move(p.out))};
+        }
+    };
+
+    template <typename Iterator1, typename Iterator2, typename Iterator3>
+    struct algorithm_result_helper<
+        hpx::util::tuple<Iterator1, Iterator2, Iterator3>,
         typename std::enable_if<
             hpx::traits::is_segmented_local_iterator<Iterator1>::value ||
             hpx::traits::is_segmented_local_iterator<Iterator2>::value ||
@@ -208,7 +237,46 @@ namespace hpx { namespace parallel { inline namespace v1 { namespace detail {
 
     template <typename Iterator1, typename Iterator2, typename Iterator3>
     struct algorithm_result_helper<
-        future<hpx::tuple<Iterator1, Iterator2, Iterator3>>,
+        future<util::in_in_out_result<Iterator1, Iterator2, Iterator3>>,
+        typename std::enable_if<
+            hpx::traits::is_segmented_local_iterator<Iterator1>::value ||
+            hpx::traits::is_segmented_local_iterator<Iterator2>::value ||
+            hpx::traits::is_segmented_local_iterator<Iterator3>::value>::type>
+    {
+        typedef hpx::traits::segmented_local_iterator_traits<Iterator1> traits1;
+        typedef hpx::traits::segmented_local_iterator_traits<Iterator2> traits2;
+        typedef hpx::traits::segmented_local_iterator_traits<Iterator3> traits3;
+
+        typedef util::in_in_out_result<typename traits1::local_raw_iterator,
+            typename traits2::local_raw_iterator,
+            typename traits3::local_raw_iterator>
+            arg_type;
+
+        static HPX_FORCEINLINE future<util::in_in_out_result<
+            typename traits1::local_iterator, typename traits2::local_iterator,
+            typename traits3::local_iterator>>
+        call(future<arg_type>&& f)
+        {
+            // different versions of clang-format produce different results
+            // clang-format off
+            return f.then(hpx::launch::sync,
+                [](future<arg_type>&& f)
+                    -> util::in_in_out_result<typename traits1::local_iterator,
+                        typename traits2::local_iterator,
[…]
