+++ libs/parallelism/algorithms/tests/unit/container_algorithms/transform_range_binary2.cpp
@@ -41,9 +41,9 @@ void test_transform_binary(ExPolicy policy, IteratorTag)
 
     auto result = hpx::parallel::transform(policy, c1, c2, std::begin(d1), add);
 
-    HPX_TEST(hpx::get<0>(result) == std::end(c1));
-    HPX_TEST(hpx::get<1>(result) == std::end(c2));
-    HPX_TEST(hpx::get<2>(result) == std::end(d1));
+    HPX_TEST(result.in1 == std::end(c1));
+    HPX_TEST(result.in2 == std::end(c2));
+    HPX_TEST(result.out == std::end(d1));
 
     // verify values
     std::vector<std::size_t> d2(c1.size());
@@ -80,9 +80,9 @@ void test_transform_binary_async(ExPolicy p, IteratorTag)
     f.wait();
 
     auto result = f.get();
-    HPX_TEST(hpx::get<0>(result) == std::end(c1));
-    HPX_TEST(hpx::get<1>(result) == std::end(c2));
-    HPX_TEST(hpx::get<2>(result) == std::end(d1));
+    HPX_TEST(result.in1 == std::end(c1));
+    HPX_TEST(result.in2 == std::end(c2));
+    HPX_TEST(result.out == std::end(d1));
 
     // verify values
     std::vector<std::size_t> d2(c1.size());
