+++ libs/full/segmented_algorithms/include/hpx/parallel/segmented_algorithms/transform.hpp
@@ -264,7 +264,11 @@ namespace hpx { namespace parallel { inline namespace v1 {
             typename InIter2, typename OutIter, typename F, typename Proj1,
             typename Proj2>
         static typename util::detail::algorithm_result<ExPolicy,
+<<<<<<< HEAD
             hpx::tuple<InIter1, InIter2, OutIter>>::type
+=======
+            util::in_in_out_result<InIter1, InIter2, OutIter>>::type
+>>>>>>> b54a64cfee5... Change binary transform result type to in_in_out_result
         segmented_transform(Algo&& algo, ExPolicy&& policy, InIter1 first1,
             InIter1 last1, InIter2 first2, OutIter dest, F&& f, Proj1&& proj1,
             Proj2&& proj2, std::true_type)
@@ -280,7 +284,11 @@ namespace hpx { namespace parallel { inline namespace v1 {
             typedef typename traits3::local_iterator local_iterator_type3;
 
             typedef util::detail::algorithm_result<ExPolicy,
+<<<<<<< HEAD
                 hpx::tuple<InIter1, InIter2, OutIter>>
+=======
+                util::in_in_out_result<InIter1, InIter2, OutIter>>
+>>>>>>> b54a64cfee5... Change binary transform result type to in_in_out_result
                 result;
 
             auto last2 = first2;
@@ -300,6 +308,7 @@ namespace hpx { namespace parallel { inline namespace v1 {
                 local_iterator_type3 ldest = traits3::begin(sdest);
                 if (beg1 != end1)
                 {
+<<<<<<< HEAD
                     hpx::tuple<local_iterator_type1, local_iterator_type2,
                         local_iterator_type3>
                         out = dispatch(traits1::get_id(sit1), algo, policy,
@@ -308,6 +317,16 @@ namespace hpx { namespace parallel { inline namespace v1 {
                     last1 = traits1::compose(send1, hpx::get<0>(out));
                     last2 = traits2::compose(send2, hpx::get<1>(out));
                     dest = traits3::compose(sdest, hpx::get<2>(out));
+=======
+                    util::in_in_out_result<local_iterator_type1,
+                        local_iterator_type2, local_iterator_type3>
+                        out = dispatch(traits1::get_id(sit1), algo, policy,
+                            std::true_type(), beg1, end1, beg2, ldest, f, proj1,
+                            proj2);
+                    last1 = traits1::compose(send1, out.in1);
+                    last2 = traits2::compose(send2, out.in2);
+                    dest = traits3::compose(sdest, out.out);
+>>>>>>> b54a64cfee5... Change binary transform result type to in_in_out_result
                 }
             }
             else
@@ -317,8 +336,13 @@ namespace hpx { namespace parallel { inline namespace v1 {
                 local_iterator_type1 end1 = traits1::end(sit1);
                 local_iterator_type2 beg2 = traits2::local(first2);
                 local_iterator_type3 ldest = traits3::begin(sdest);
+<<<<<<< HEAD
                 hpx::tuple<local_iterator_type1, local_iterator_type2,
                     local_iterator_type3>
+=======
+                util::in_in_out_result<local_iterator_type1,
+                    local_iterator_type2, local_iterator_type3>
+>>>>>>> b54a64cfee5... Change binary transform result type to in_in_out_result
                     out;
                 if (beg1 != end1)
                 {
@@ -354,12 +378,22 @@ namespace hpx { namespace parallel { inline namespace v1 {
                         std::true_type(), beg1, end1, beg2, ldest, f, proj1,
                         proj2);
                 }
+<<<<<<< HEAD
                 last1 = traits1::compose(send1, hpx::get<0>(out));
[…]

[…]
