### 2020-08-19 #2352 @bhatele
> > @danielskatz @diehlpk I read the software paper. I believe these two things from the list of requirements above are missing:
> > ```
> > 2. Even though the paper has list of AMT's that are similar, it does not provide how this software compares to them.
> > ```
>
> @bhatele we have already following high-level statement
>
> ```
>  The major show piece of HPX compared to the mentioned distributed AMTs is its future-proof C++ standard conforming API.
> ```
>
> I would say that providing more details would go to deep in the implementation specifics of HPX and other distributed AMTs and might not appropriate for this paper. Can you please let us know how deep we should go for the comparison?
>
> We agree with 1. and are working on it.

@danielskatz The authors are working on 1. and that was my last suggestion. Other than that, my review is done. This looks good from my end.
