### 2020-08-19 6e74c48
Fix incomplete type compile error by adding a dummy template param
