+++ libs/execution/tests/unit/limiting_executor.cpp
@@ -4,143 +4,105 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
+#include <hpx/async.hpp>
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
 #include <hpx/include/parallel_executors.hpp>
 #include <hpx/modules/testing.hpp>
 #include <hpx/thread_executors/limiting_executor.hpp>
 
-#include <algorithm>
+#include <chrono>
 #include <cstdlib>
-#include <iterator>
-#include <numeric>
+#include <random>
 #include <string>
 #include <vector>
 
-///////////////////////////////////////////////////////////////////////////////
-hpx::thread::id thread_id_test(int passed_through)
-{
-    HPX_TEST_EQ(passed_through, 42);
-    return hpx::this_thread::get_id();
-}
+// this test launches many tasks continuously using a limiting executor
+// some of the tasks will suspend themselves randomly and so new tasks will be
+// spawned until the max tasks 'in flight' limit is reached. Once this happens
+// no new tasks should be created and the number 'in flight' should remain
+// below the limit. We track this using some counters and check that
+// at the end of the test, the max counter never exceeded the limit.
 
-void test2(int passed_through)
-{
-    HPX_TEST_EQ(passed_through, 42);
-}
-
-void test_sync()
-{
-    auto exec1 = hpx::parallel::execution::parallel_executor(
-        hpx::threads::thread_stacksize_small);
-    auto exec2 = hpx::parallel::execution::default_executor(
-        hpx::threads::thread_stacksize_large);
-
-    hpx::threads::executors::limiting_executor<decltype(exec1)> lexec1(
-        exec1, 100, 100);
-    hpx::threads::executors::limiting_executor<decltype(exec2)> lexec2(
-        exec2, 100, 100);
-
-    hpx::async(lexec1, &test2, 42);
-    hpx::async(lexec2, &test2, 42);
-
-    auto id1 =
-        hpx::parallel::execution::sync_execute(exec1, &thread_id_test, 42);
-    auto id2 = hpx::this_thread::get_id();
-    std::cout << "hpx::parallel::execution::sync_execute Ids " << id1 << " : "
-              << id2 << std::endl;
-
-    HPX_TEST(hpx::parallel::execution::sync_execute(
-                 exec1, &thread_id_test, 42) == hpx::this_thread::get_id());
-    HPX_TEST(hpx::parallel::execution::sync_execute(
-                 exec2, &thread_id_test, 42) == hpx::this_thread::get_id());
-
-    id1 = hpx::parallel::execution::sync_execute(lexec1, &thread_id_test, 42);
-    id2 = hpx::this_thread::get_id();
-    std::cout << "hpx::parallel::execution::sync_execute - limiting exec Ids "
-              << id1 << " : " << id2 << std::endl;
-
-    HPX_TEST(hpx::parallel::execution::sync_execute(
-                 lexec1, &thread_id_test, 42) == hpx::this_thread::get_id());
-    HPX_TEST(hpx::parallel::execution::sync_execute(
-                 lexec2, &thread_id_test, 42) == hpx::this_thread::get_id());
-}
-
-void test_async()
-{
-    auto exec1 = hpx::parallel::execution::parallel_executor(
-        hpx::threads::thread_stacksize_small);
-    auto exec2 = hpx::parallel::execution::default_executor(
-        hpx::threads::thread_stacksize_large);
-
-    hpx::threads::executors::limiting_executor<decltype(exec1)> lexec1(
[…]
