### 2020-08-19 1309167
Aggregate initialize CPOs
