+++ libs/executors/include/hpx/executors/limiting_executor.hpp
@@ -28,140 +28,144 @@
 //
 // --------------------------------------------------------------------
 namespace hpx { namespace parallel { namespace execution {
+    namespace experimental {
 
-    // by convention the title is 7 chars (for alignment)
-    using print_on = debug::enable_print<false>;
-    static constexpr print_on lim_debug("LIMEXEC");
+        // by convention the title is 7 chars (for alignment)
+        using print_on = debug::enable_print<false>;
+        static constexpr print_on lim_debug("LIMEXEC");
 
-    ///////////////////////////////////////////////////////////////////////////
-    namespace detail {
-        HPX_HAS_MEMBER_XXX_TRAIT_DEF(in_flight_estimate)
-    }    // namespace detail
+        ///////////////////////////////////////////////////////////////////////////
+        namespace detail {
+            HPX_HAS_MEMBER_XXX_TRAIT_DEF(in_flight_estimate)
+        }    // namespace detail
 
-    template <typename BaseExecutor>
-    struct limiting_executor
-    {
-        // --------------------------------------------------------------------
-        // RAII wrapper for counting task completions (count_down)
-        // count_up is done in the executor when the task is first scheduled
-        // This object is destructed when the task has completed
-        // --------------------------------------------------------------------
-        struct on_exit
+        template <typename BaseExecutor>
+        struct limiting_executor
         {
-            on_exit(limiting_executor const& this_e)
-              : executor_(this_e)
+            // --------------------------------------------------------------------
+            // RAII wrapper for counting task completions (count_down)
+            // count_up is done in the executor when the task is first scheduled
+            // This object is destructed when the task has completed
+            // --------------------------------------------------------------------
+            struct on_exit
             {
-            }
-            ~on_exit()
-            {
-                lim_debug.debug(hpx::debug::str<>("Count Down"));
-                executor_.count_down();
-            }
-            limiting_executor const& executor_;
-        };
-
-        // --------------------------------------------------------------------
-        // this is the default wrapper struct that invokes count up / down
-        // and uses the limiting executor counter to control throttling
-        //
-        // Note that we have to add a dummy template parameter B (same as BaseName)
-        // to inner struct, to allow template deduction to SFINAE properly
-        // and not give us an incomplete type on the 'in flight' specialization
-        // --------------------------------------------------------------------
-        template <typename F, typename B = BaseExecutor, typename Enable = void>
-        struct throttling_wrapper
-        {
-            throttling_wrapper(
-                limiting_executor& lim, BaseExecutor const& base, F&& f)
-              : limiting_(lim)
-              , f_(std::forward<F>(f))
-            {
-                limiting_.count_up();
-                if (exceeds_upper())
+                on_exit(limiting_executor const& this_e)
+                  : executor_(this_e)
                 {
-                    lim_debug.debug(hpx::debug::str<>("Exceeds_upper"));
-                    hpx::util::yield_while([&]() { return exceeds_lower(); });
-                    lim_debug.debug(hpx::debug::str<>("Below_lower"));
                 }
-            }
-
-            // when task completes, on_exit destructor calls count_down
-            template <typename... Ts>
-            decltype(auto) operator()(Ts&&... ts)
-            {
-                on_exit _{limiting_};
-                return hpx::util::invoke(f_, std::forward<Ts>(ts)...);
-            }
-
-            // returns true if too many tasks would be in flight
-            // NB. we use ">" because we count up right before testing
-            bool exceeds_upper()
-            {
-                return (limiting_.count_ > limiting_.upper_threshold_);
-            }
-            // returns true if we have not yet reached the lower threshold
-            bool exceeds_lower()
-            {
[…]
