### 2020-08-19 2bab386
Add alias targets with namespace used for exporting
