+++ cmake/HPX_GeneratePackage.cmake
@@ -41,6 +41,15 @@ export(
   FILE "${CMAKE_CURRENT_BINARY_DIR}/lib/cmake/${HPX_PACKAGE_NAME}/HPXTargets.cmake"
 )
 
+# Add aliases with the namespace for use within HPX
+foreach(export_target ${HPX_EXPORT_TARGETS})
+  add_library(HPX::${export_target} ALIAS ${export_target})
+endforeach()
+
+foreach(export_target ${HPX_EXPORT_INTERNAL_TARGETS})
+  add_library(HPXInternal::${export_target} ALIAS ${export_target})
+endforeach()
+
 # Export HPXTargets in the install directory
 install(
   EXPORT HPXTargets
