+++ libs/executors/tests/unit/limiting_executor.cpp
@@ -65,9 +65,8 @@ void test_limit()
     std::vector<hpx::future<void>> futures;
     // scope block for executor lifetime (block on destruction)
     {
-        hpx::parallel::execution::experimental::limiting_executor<decltype(
-            exec1)>
-            lexec1(exec1, max1 / 2, max1, block_on_exit);
+        hpx::execution::experimental::limiting_executor<decltype(exec1)> lexec1(
+            exec1, max1 / 2, max1, block_on_exit);
 
         // run this loop for N seconds and launch as many tasks as we can
         // then check that there were never more than N on a given executor
