### 2020-08-20 f520874
Moving more action related files to actions modules
- flyby: remove unused manage_object_action
- flyby: remove obsolete header reference from build system
