+++ libs/actions_base/include/hpx/actions_base/continuation_fwd.hpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2007-2015 Hartmut Kaiser
+//  Copyright (c) 2007-2020 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -8,14 +8,15 @@
 
 #include <hpx/config.hpp>
 
-namespace hpx { namespace actions
-{
+namespace hpx { namespace actions {
+
     class HPX_EXPORT continuation;
 
     template <typename Result, typename RemoteResult = Result>
     struct typed_continuation;
 
-    template <typename Result, typename RemoteResult, typename F, typename ...Ts>
-    void trigger(typed_continuation<Result, RemoteResult>&& cont, F&& f, Ts&&... vs);
-}}
-
+    template <typename Result, typename RemoteResult, typename F,
+        typename... Ts>
+    void trigger(
+        typed_continuation<Result, RemoteResult>&& cont, F&& f, Ts&&... vs);
+}}    // namespace hpx::actions
