+++ libs/actions/include/hpx/actions/transfer_continuation_action.hpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2007-2016 Hartmut Kaiser
+//  Copyright (c) 2007-2020 Hartmut Kaiser
 //  Copyright (c)      2011 Bryce Lelbach
 //  Copyright (c)      2011 Thomas Heller
 //
@@ -13,14 +13,14 @@
 #include <hpx/config.hpp>
 
 #if defined(HPX_HAVE_NETWORKING)
-#include <hpx/actions/action_support.hpp>
 #include <hpx/actions/transfer_base_action.hpp>
+#include <hpx/actions_base/actions_base_support.hpp>
 #include <hpx/async_distributed/applier/apply_helper.hpp>
 #include <hpx/runtime/actions/continuation.hpp>
-#include <hpx/runtime/parcelset/detail/per_action_data_counter_registry.hpp>
 #include <hpx/serialization/input_archive.hpp>
 #include <hpx/serialization/output_archive.hpp>
 #include <hpx/serialization/serialization_fwd.hpp>
+#include <hpx/serialization/traits/needs_automatic_registration.hpp>
 #include <hpx/threading_base/thread_helpers.hpp>
 #include <hpx/threading_base/thread_init_data.hpp>
 #include <hpx/type_support/pack.hpp>
@@ -31,8 +31,8 @@
 
 #include <hpx/config/warnings_prefix.hpp>
 
-namespace hpx { namespace actions
-{
+namespace hpx { namespace actions {
+
     /// \cond NOINTERNAL
 
     ///////////////////////////////////////////////////////////////////////////
@@ -51,13 +51,13 @@ namespace hpx { namespace actions
         transfer_continuation_action() = default;
 
         // construct an action from its arguments
-        template <typename ...Ts>
-        explicit transfer_continuation_action(continuation_type&& cont, Ts&&... vs);
+        template <typename... Ts>
+        explicit transfer_continuation_action(
+            continuation_type&& cont, Ts&&... vs);
 
-        template <typename ...Ts>
-        transfer_continuation_action(
-            threads::thread_priority priority, continuation_type&& cont,
-            Ts&&... vs);
+        template <typename... Ts>
+        transfer_continuation_action(threads::thread_priority priority,
+            continuation_type&& cont, Ts&&... vs);
 
         bool has_continuation() const override;
 
@@ -74,15 +74,14 @@ namespace hpx { namespace actions
         /// \note This \a get_thread_function will be invoked to retrieve the
         ///       thread function for an action which has to be invoked without
         ///       continuations.
-        template <std::size_t ...Is>
-        threads::thread_function_type
-        get_thread_function(util::index_pack<Is...>,
-            naming::id_type&& target, naming::address::address_type lva,
+        template <std::size_t... Is>
+        threads::thread_function_type get_thread_function(
+            util::index_pack<Is...>, naming::id_type&& target,
+            naming::address::address_type lva,
             naming::address::component_type comptype);
 
-        threads::thread_function_type
-        get_thread_function(naming::id_type&& target,
-            naming::address::address_type lva,
+        threads::thread_function_type get_thread_function(
+            naming::id_type&& target, naming::address::address_type lva,
             naming::address::component_type comptype) override;
 
         template <std::size_t... Is>
@@ -99,10 +98,10 @@ namespace hpx { namespace actions
 
         // serialization support
         // loading ...
-        void load(hpx::serialization::input_archive & ar) override;
+        void load(hpx::serialization::input_archive& ar) override;
 
         // saving ...
-        void save(hpx::serialization::output_archive & ar) override;
+        void save(hpx::serialization::output_archive& ar) override;
 
         void load_schedule(serialization::input_archive& ar,
             naming::gid_type&& target, naming::address_type lva,
[…]
