### 2020-08-20 26295fb
Move chrono functionality to hpx::chrono namespace
