### 2020-08-20 #2352 @davidbeckingsale
Formatting issue in the paper:

> Sketch of HPX’s architecture with all the components and their interactions. section sketches HPX’s architectures. The components of HPX and their references are listed below:

I think "section" should be a link somewhere else. Also, there is no sketch of the components.

Overall, I think the paper is okay. The first few paragraphs of the summary are good and the example code is excellent. However, the section listing out the components is too complex, and even the first few paragraphs do not (in my opinion) provide a clear description for a __diverse, non-specialist audience__.  I think the text can be reworded in a way that makes it easier to understand for someone not familiar with HPC terminology and the concept of AMTs (@bhatele and I are both intimately familiar with these areas). Additionally, the components section uses undefined terms like 'PGAS'.

The paragraph listing applications using hpx is good, but needs reformatting to make it easier to read rather than just being a list of references. Perhaps grouping the applications into some categories would work.
