### 2020-08-20 c7e589e
Add CPOs for all BinaryOperation transform overloads
