+++ libs/full/segmented_algorithms/include/hpx/parallel/segmented_algorithms/detail/dispatch.hpp
@@ -265,8 +265,10 @@ namespace hpx { namespace parallel { inline namespace v1 { namespace detail {
                         typename traits2::local_iterator,
                         typename traits3::local_iterator> {
                     auto&& p = f.get();
-                    return util::in_in_out_result<typename traits1::local_iterator,
-                        typename traits2::local_iterator, typename traits3::local_iterator>
+                    return util::in_in_out_result<
+                        typename traits1::local_iterator,
+                        typename traits2::local_iterator,
+                        typename traits3::local_iterator>
                         {traits1::remote(p.in1),
                             traits2::remote(p.in2), traits3::remote(p.out)};
                 });

[…]
