### 2020-08-20 f67a092
Attempting to fix memory leak in create_topology()
- avoid double constructing reinitialize_static data
