+++ libs/topology/CMakeLists.txt
@@ -35,6 +35,7 @@ add_hpx_module(
     hpx_errors
     hpx_format
     hpx_logging
+    hpx_static_reinit
     hpx_type_support
     hpx_util
     Hwloc::hwloc
