### 2020-08-20 76392f9
Attempting to fix memory leak in create_topology()
- avoid double constructing reinitialize_static data
