+++ libs/core/topology/CMakeLists.txt
@@ -40,6 +40,7 @@ add_hpx_module(
     hpx_errors
     hpx_format
     hpx_logging
+    hpx_static_reinit
     hpx_type_support
     hpx_util
   DEPENDENCIES Hwloc::hwloc
