### 2020-08-20 #2352 @diehlpk
@davidbeckingsale

`I think "section" should be a link somewhere else. Also, there is no sketch of the components.`

I am sorry but in the latest version there was an issue with the Figure. In the latest PDF there is Figure 1 and there is a reference to the figure.

`The paragraph listing applications using hpx is good, but needs reformatting to make it easier to read rather than just being a list of references. Perhaps grouping the applications into some categories would work.`

Thanks for this suggestion. We grouped the application into Scientific Computing and Libraries in the revised version.

@hkaiser Will work on your other suggestions.
