### 2020-08-20 94f1c11
Make cuda event pool dynamic instead of fixed size
