### 2020-08-20 4aac386
Fix comment abiut fixed/dynamic size
