### 2020-08-21 e648991
Fix a problem with undefined HPX_HAVE_DEPRECATION_WARNINGS
