+++ .jenkins/cscs/env-clang-apex.sh
@@ -34,7 +34,7 @@ configure_extra_options+=" -DHPX_WITH_DEPRECATION_WARNINGS=OFF"
 configure_extra_options+=" -DHPX_WITH_COMPILER_WARNINGS=ON"
 configure_extra_options+=" -DHPX_WITH_COMPILER_WARNINGS_AS_ERRORS=ON"
 configure_extra_options+=" -DHPX_WITH_SPINLOCK_DEADLOCK_DETECTION=ON"
-configure_extra_options+=" -DHPX_WITH_LIBCDS=ON"
+#configure_extra_options+=" -DHPX_WITH_LIBCDS=ON"
 configure_extra_options+=" -DHPX_WITH_APEX=ON"
 configure_extra_options+=" -DHPX_WITH_PAPI=ON"
 configure_extra_options+=" -DAPEX_WITH_PAPI=ON"
