+++ CMakeLists.txt
@@ -959,26 +959,28 @@ if(HPX_WITH_APEX)
 endif()
 
 # LibCDS option
-hpx_option(
-  HPX_WITH_LIBCDS BOOL "Enable LibCDS support (experimental)." OFF
-  CATEGORY "Thread Manager" ADVANCED
-)
-if(HPX_WITH_LIBCDS)
-  hpx_option(
-    HPX_WITH_LIBCDS_GIT_REPOSITORY STRING
-    "Define the LibCDS git repository to use."
-    https://github.com/STEllAR-GROUP/libcds CATEGORY "Thread Manager" ADVANCED
-  )
-  hpx_option(
-    HPX_WITH_LIBCDS_GIT_TAG STRING "Define the LibCDS git tag to use." hpx-1.5
-    CATEGORY "Thread Manager" ADVANCED
-  )
-  include(HPX_SetupLibCDS)
-  if(NOT libcds_POPULATED)
-    hpx_error("HPX_WITH_LIBCDS was set to ON, but HPX failed to fetch LibCDS")
-  endif()
-  hpx_add_config_define(HPX_HAVE_LIBCDS) # tell HPX that we use LibCDS
-endif()
+# NOTE: The libcds option is disabled for the 1.5.0 release as it is not ready
+# for public consumption yet.
+# hpx_option(
+#   HPX_WITH_LIBCDS BOOL "Enable LibCDS support (experimental)." OFF
+#   CATEGORY "Thread Manager" ADVANCED
+# )
+# if(HPX_WITH_LIBCDS)
+#   hpx_option(
+#     HPX_WITH_LIBCDS_GIT_REPOSITORY STRING
+#     "Define the LibCDS git repository to use."
+#     https://github.com/STEllAR-GROUP/libcds CATEGORY "Thread Manager" ADVANCED
+#   )
+#   hpx_option(
+#     HPX_WITH_LIBCDS_GIT_TAG STRING "Define the LibCDS git tag to use." hpx-1.5
+#     CATEGORY "Thread Manager" ADVANCED
+#   )
+#   include(HPX_SetupLibCDS)
+#   if(NOT libcds_POPULATED)
+#     hpx_error("HPX_WITH_LIBCDS was set to ON, but HPX failed to fetch LibCDS")
+#   endif()
+#   hpx_add_config_define(HPX_HAVE_LIBCDS) # tell HPX that we use LibCDS
+# endif()
 
 hpx_option(
   HPX_WITH_PAPI BOOL "Enable the PAPI based performance counter." OFF
