### 2020-08-21 87984b8
Update the release notes for the release candidate 3
