### 2020-08-21 de67f5b
More fixes to reinitialize_static
