### 2020-08-21 95ff1ad
More fixes to reinitialize_static
