+++ cmake/HPX_AddConfigTest.cmake
@@ -319,8 +319,8 @@ endfunction()
 # ##############################################################################
 function(hpx_check_for_c11_aligned_alloc)
   add_hpx_config_test(
-    HPX_WITH_C11_ALIGNED_ALLOC
-    SOURCE cmake/tests/c11_aligned_alloc.cpp FILE ${ARGN}
+    HPX_WITH_C11_ALIGNED_ALLOC SOURCE cmake/tests/c11_aligned_alloc.cpp FILE
+                                      ${ARGN}
   )
 endfunction()
 
