### 2020-08-22 fd24a21
Fix transform docs, add deprecation warning and fix is_segmented trait check
