+++ libs/parallelism/algorithms/include/hpx/parallel/container_algorithms/transform.hpp
@@ -9,24 +9,10 @@
 
 #pragma once
 
-#include <hpx/config.hpp>
-#include <hpx/concepts/concepts.hpp>
-#include <hpx/iterator_support/range.hpp>
-#include <hpx/iterator_support/traits/is_iterator.hpp>
-#include <hpx/iterator_support/traits/is_range.hpp>
-#include <hpx/parallel/util/result_types.hpp>
-#include <hpx/parallel/util/tagged_pair.hpp>
-#include <hpx/parallel/util/tagged_tuple.hpp>
+#if defined(DOXYGEN)
+namespace hpx {
+    // clang-format off
 
-#include <hpx/algorithms/traits/projected_range.hpp>
-#include <hpx/parallel/algorithms/transform.hpp>
-#include <hpx/parallel/tagspec.hpp>
-#include <hpx/parallel/util/projection_identity.hpp>
-
-#include <type_traits>
-#include <utility>
-
-namespace hpx { namespace parallel { inline namespace v1 {
     /// Applies the given function \a f to the given range \a rng and stores
     /// the result in another range, beginning at dest.
     ///
@@ -65,7 +51,7 @@ namespace hpx { namespace parallel { inline namespace v1 {
     ///                     \endcode \n
     ///                     The signature does not need to have const&.
     ///                     The type \a Type must be such that an object of
-    ///                     type \a InIter can be dereferenced and then
+    ///                     type \a range_iterator<Rng>::type can be dereferenced and then
     ///                     implicitly converted to \a Type. The type \a Ret
     ///                     must be such that an object of type \a OutIter can
     ///                     be dereferenced and assigned a value of type
@@ -86,9 +72,9 @@ namespace hpx { namespace parallel { inline namespace v1 {
     /// within each thread.
     ///
     /// \returns  The \a transform algorithm returns a
-    ///           \a hpx::future<tagged_pair<tag::in(InIter), tag::out(OutIter)> >
+    ///           \a hpx::future<ranges::unary_transform_result<range_iterator<Rng>::type, OutIter> >
     ///           if the execution policy is of type \a parallel_task_policy
-    ///           and returns \a tagged_pair<tag::in(InIter), tag::out(OutIter)>
+    ///           and returns \a ranges::unary_transform_result<range_iterator<Rng>::type, OutIter>
     ///           otherwise.
     ///           The \a transform algorithm returns a tuple holding an iterator
     ///           referring to the first element after the input sequence and
@@ -97,22 +83,95 @@ namespace hpx { namespace parallel { inline namespace v1 {
     ///           copied.
     ///
     template <typename ExPolicy, typename Rng, typename OutIter, typename F,
-        typename Proj = util::projection_identity,
-        HPX_CONCEPT_REQUIRES_(hpx::is_execution_policy<ExPolicy>::value&&
-                hpx::traits::is_range<Rng>::value&& hpx::traits::is_iterator<
-                    OutIter>::value&& traits::is_projected_range<Proj,
-                    Rng>::value&& traits::is_indirect_callable<ExPolicy, F,
-                    traits::projected_range<Proj, Rng>>::value)>
+        typename Proj = util::projection_identity>
     typename util::detail::algorithm_result<ExPolicy,
-        util::in_out_result<typename hpx::traits::range_iterator<Rng>::type,
+        ranges::unary_transform_result<
+            typename hpx::traits::range_iterator<Rng>::type,
             OutIter>>::type
-    transform(
-        ExPolicy&& policy, Rng&& rng, OutIter dest, F&& f, Proj&& proj = Proj())
-    {
-        return transform(std::forward<ExPolicy>(policy), hpx::util::begin(rng),
[…]
