### 2020-08-24 b543db8
Don't acquire lock if not needed
