+++ cmake/templates/hpxrun.py.in
@@ -221,10 +221,10 @@ def build_cmd(options, args):
     if options.localities > 1:
         # Selecting the parcelport for hpx via hpx ini configuration
         select_parcelport = (lambda pp:
-            ['--hpx:ini=hpx.parcel.verbs.enable=1'] if pp == 'verbs'
-            else ['--hpx:ini=hpx.parcel.ipc.enable=1'] if pp == 'ipc'
-            else ['--hpx:ini=hpx.parcel.mpi.enable=1', '--hpx:ini=hpx.parcel.bootstrap=mpi'] if pp == 'mpi'
-            else ['--hpx:ini=hpx.parcel.tcp.enable=1'] if pp == 'tcp'
+            ['--hpx:ini=hpx.parcel.verbs.priority=1000', '--hpx:ini=hpx.parcel.verbs.enable=1'] if pp == 'verbs'
+            else ['--hpx:ini=hpx.parcel.ipc.priority=1000', '--hpx:ini=hpx.parcel.ipc.enable=1'] if pp == 'ipc'
+            else ['--hpx:ini=hpx.parcel.mpi.priority=1000', '--hpx:ini=hpx.parcel.mpi.enable=1', '--hpx:ini=hpx.parcel.bootstrap=mpi'] if pp == 'mpi'
+            else ['--hpx:ini=hpx.parcel.tcp.priority=1000', '--hpx:ini=hpx.parcel.tcp.enable=1'] if pp == 'tcp'
             else [])
         cmd += select_parcelport(options.parcelport)
 
