### 2020-08-24 6dcab4e
Update rostam jenkins environments
