### 2020-08-24 f0bbedc
Update rostam jenkins environments
