+++ .jenkins/lsu/env-clang-7.sh
@@ -8,7 +8,7 @@ module purge
 module load cmake
 module load clang/7
 module load boost/1.65.0-${build_type,,}
-module load hwloc/1
+module load hwloc
 module load openmpi
 
 export HPXRUN_RUNWRAPPER=srun
