### 2020-08-25 60308e8
Making transform_reduce conforming to C++20
