+++ libs/include/include/hpx/include/parallel_transform_reduce.hpp
@@ -9,7 +9,7 @@
 
 #include <hpx/config.hpp>
 #include <hpx/parallel/algorithms/transform_reduce.hpp>
-#include <hpx/parallel/algorithms/transform_reduce_binary.hpp>
+#include <hpx/parallel/container_algorithms/transform_reduce.hpp>
 
 #if defined(HPX_HAVE_DISTRIBUTED_RUNTIME)
 #include <hpx/parallel/segmented_algorithms/transform_reduce.hpp>
