### 2020-08-25 fdd4eff
Refactor synchronization primitives
- reduce compile time by moving code to implementation files
- fix small performance regression allowing to avoid acquiring lock under
  limited circumstances
- fix #include issues
