+++ libs/core/synchronization/src/detail/condition_variable.cpp
@@ -5,14 +5,13 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#include <hpx/synchronization/condition_variable.hpp>
-
 #include <hpx/assert.hpp>
 #include <hpx/execution_base/this_thread.hpp>
 #include <hpx/modules/errors.hpp>
 #include <hpx/modules/logging.hpp>
 #include <hpx/modules/memory.hpp>
 #include <hpx/runtime/threads/thread_helpers.hpp>
+#include <hpx/synchronization/detail/condition_variable.hpp>
 #include <hpx/synchronization/no_mutex.hpp>
 #include <hpx/synchronization/spinlock.hpp>
 #include <hpx/thread_support/unlock_guard.hpp>
@@ -24,6 +23,7 @@
 #include <utility>
 
 namespace hpx { namespace lcos { namespace local { namespace detail {
+
     ///////////////////////////////////////////////////////////////////////////
     condition_variable::condition_variable() {}
 
