### 2020-08-25 #2352 @davidbeckingsale
I like the refactored applications piece - can you make it a section (or sub section) so it stands out?

Other changes look good.
