### 2020-08-26 #2352 @diehlpk
@davidbeckingsale I moved the applications to a new section. Please let me know if this change is sufficient.
