### 2020-08-26 5b2572c
Update HPX_VERSION_TAG for the release candidate 3
