### 2020-08-26 7e0c472
Add HPX_NODISCARD to enable_user_polling structs
Avoid mistakes where user polling is enabled and disabled immediately,
instead of at the end of a scope.
