+++ libs/async_cuda/include/hpx/async_cuda/cuda_future.hpp
@@ -309,7 +309,7 @@ namespace hpx { namespace cuda { namespace experimental {
 
     // -----------------------------------------------------------------
     // This RAII helper class enables polling for a scoped block
-    struct enable_user_polling
+    struct HPX_NODISCARD enable_user_polling
     {
         enable_user_polling(std::string const& pool_name = "")
           : pool_name_(pool_name)
