### 2020-08-26 #2352 @danielskatz
@diehlpk  - while we are waiting for @davidbeckingsale , I took at a look at the paper.  It would be good to fix the casing on the references - for example
> Kaiser, H., Brodowicz, M., & Sterling, T. (2009). ParalleX: An Advanced Parallel Execution Model for Scaling-impaired Applications. In 2009 international conference on parallel processing workshops (pp. 394–401). IEEE. doi:10.1109/icppw.2009.14

should be
> Kaiser, H., Brodowicz, M., & Sterling, T. (2009). ParalleX: An Advanced Parallel Execution Model for Scaling-impaired Applications. In 2009 International Conference on Parallel Processing Workshops (pp. 394–401). IEEE. doi:10.1109/icppw.2009.14
