### 2020-08-26 #2352 @diehlpk
@danielskatz The BibTeX entries are corrected in the version above.
