### 2020-08-26 #2352 @davidbeckingsale
@diehlpk - it looks like moving the applications to a section has messed up the formatting of the list.
