### 2020-08-26 #2352 @davidbeckingsale
Looks good. I second @danielskatz's request - please take another look at your references to fix up capitalization, e.g.:

> Wagle, B., Kellar, S., Serio, A., & Kaiser, H. (2018). Methodology for Adaptive ActiveMessage Coalescing in Task Based Runtime Systems. In2018 ieee international parallel and distributed processing symposium workshops (ipdpsw)
