### 2020-08-26 #2352 @danielskatz
> @danielskatz The BibTeX entries are corrected in the version above.

No, sorry, not all of them - for example
* Supercomputing frontiers and innovations
* Proceedings of the eighth annual conference on Object-oriented programming systems, languages, and applications
* 2018 ieee international parallel and distributed processing symposium workshops (ipdpsw)
