### 2020-08-26 #2352 @davidbeckingsale
> @davidbeckingsale - please let me know when you have checked off the last time in your list. At that point, I'll do a final copyediting read of the paper, and we'll move on to processing it

I think this is good to go. Thanks @diehlpk & @hkaiser for addressing all the issues promptly.
