### 2020-08-26 6b6c26c
Add final version of the JOSS paper
