### 2020-08-26 1edb13b
Relax condition on `notified_` flag.
- flyby: fix potential deadlock situation where thread would be suspended even
  if the counter has already reached zero
