### 2020-08-26 #2352 @diehlpk
@danielskatz We currently are working on a version of HPX, which includes all the changes and the paper. This will be done by next week, so I will work on your checklist next week.
