### 2020-08-26 46eb6ab
Partially revert changes to handling the `latch::notify_` flag
