### 2020-08-27 0e10231
Adapt the reserve to NMAX * 8 cf nested for loops
