### 2020-08-28 3d33e77
Small documentation formatting fixes
