### 2020-08-28 dea4e73
Fixing compiler warnings
