### 2020-08-29 3530c2e
Adding communications set
- A communication_set is a set of connected communicator_node objects that
  are connected using a binomial communication pattern. A communication_set
  can be created once and reused with any of the collective operations.
  This not only removes the overhead of establishing the connections
  for each instance of a collective operation, it also switches from
  the linear communication model to a tree based one.
