+++ libs/full/collectives/tests/unit/communication_set.cpp
@@ -1,0 +1,278 @@
+//  Copyright (c) 2020 Hartmut Kaiser
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include <hpx/hpx_init.hpp>
+#include <hpx/modules/async_distributed.hpp>
+#include <hpx/modules/collectives.hpp>
+#include <hpx/modules/format.hpp>
+#include <hpx/modules/pack_traversal.hpp>
+#include <hpx/modules/testing.hpp>
+
+#include <hpx/collectives/detail/communication_set_node.hpp>
+
+#include <cstddef>
+#include <cstdint>
+#include <memory>
+#include <mutex>
+#include <string>
+#include <utility>
+#include <vector>
+
+namespace test {
+
+    ///////////////////////////////////////////////////////////////////////////
+    struct communication_set_tag
+    {
+    };
+
+    ///////////////////////////////////////////////////////////////////////////
+    // tag used to extract node the given node of the communicating set is
+    // connected to
+    struct get_connected_to
+    {
+    };
+
+    // tag used to extract the number of the given node of the communication set
+    struct get_site_number
+    {
+    };
+
+    // tag used to verify all nodes are connected
+    struct get_connected_to_zero
+    {
+    };
+
+    ///////////////////////////////////////////////////////////////////////////
+    using get_connected_to_action =
+        typename hpx::lcos::detail::communication_set_node::template get_action<
+            communication_set_tag, std::size_t, get_connected_to>;
+
+    using get_site_number_action =
+        typename hpx::lcos::detail::communication_set_node::template get_action<
+            communication_set_tag, std::size_t, get_site_number>;
+
+    using get_connected_to_zero_action =
+        typename hpx::lcos::detail::communication_set_node::template get_action<
+            communication_set_tag, hpx::future<void>, get_connected_to_zero>;
+}    // namespace test
+
+namespace hpx { namespace traits {
+
+    template <typename Communicator>
+    struct communication_operation<Communicator, test::communication_set_tag>
+      : std::enable_shared_from_this<
+            communication_operation<Communicator, test::communication_set_tag>>
+    {
+        communication_operation(Communicator& comm)
+          : communicator_(comm)
+        {
+        }
+
+        template <typename Result>
+        Result get(std::size_t which, test::get_connected_to)
+        {
+            return communicator_.connect_to_;
+        }
+
+        template <typename Result>
+        Result get(std::size_t which, test::get_site_number)
+        {
+            return communicator_.site_;
+        }
+
+        template <typename Result>
+        Result get(std::size_t, test::get_connected_to_zero)
+        {
+            // first forward request to parent
+            if (communicator_.connect_to_ != communicator_.site_)
+            {
+                hpx::sync(test::get_connected_to_zero_action{},
+                    communicator_.connected_node_.get(), communicator_.site_,
+                    test::get_connected_to_zero{});
+            }
+
+            // now, handle request ourselves
+            using mutex_type = typename Communicator::mutex_type;
+
+            auto this_ = this->shared_from_this();
+            auto on_ready = [this_ = std::move(this_)](
+                                shared_future<void>&& f) {
+                f.get();    // propagate any exceptions
+
[…]
