### 2020-08-31 6a0c58c
Add copyright and license to joss paper file
