+++ docs/joss_paper/paper.md
@@ -1,4 +1,11 @@
-﻿---
+﻿<!-- Copyright (c) 2020 Patrick Diehl                                        -->
+<!--                                                                         -->
+<!-- SPDX-License-Identifier: BSL-1.0                                        -->
+<!-- Distributed under the Boost Software License, Version 1.0. (See         -->
+<!-- accompanying file LICENSE_1_0.txt or copy at                            -->
+<!-- http://www.boost.org/LICENSE_1_0.txt)                                   -->
+
+---
 title: 'HPX - The C++ Standard Library for Parallelism and Concurrency'
 tags:
  - concurrency
