### 2020-08-31 40f3e3a
Fix typo in the joss paper markdown
