### 2020-08-31 026d22f
Small docs formatting fixes
