+++ docs/sphinx/contributing/docker_image.rst
@@ -5,7 +5,7 @@
     Distributed under the Boost Software License, Version 1.0. (See accompanying
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-.. _docker:
+.. _using_docker:
 
 ============================
 Using docker for development
