### 2020-08-31 5e6bbd2
Update the list of PR and issues + final release notes
