+++ docs/sphinx/releases/whats_new_1_5_0.rst
@@ -37,9 +37,10 @@ and many bug fixes.
   corresponding algorithms fully conformant to C++20 as well as to make
   algorithm extension easier for the user. ``all_of``/``any_of``/``none_of``,
   ``copy``, ``count``, ``destroy``, ``equal``, ``fill``, ``find``, ``for_each``,
-  ``generate``, ``mismatch``, ``move``, ``reduce`` are using those CPOs (all in
-  namespace ``hpx``).  We also have adapted their corresponding ``hpx::ranges``
-  versions to be conforming to C++20 in this release.
+  ``generate``, ``mismatch``, ``move``, ``reduce``, ``transform_reduce`` are
+  using those CPOs (all in namespace ``hpx``).  We also have adapted their
+  corresponding ``hpx::ranges`` versions to be conforming to C++20 in this
+  release.
 * We have adapted support for ``co_await`` to C++20, in addition to
   ``hpx::future`` it now also supports ``hpx::shared_future``. We have also
   added allocator support for futures returned by ``co_return``. It is no longer
@@ -154,15 +155,19 @@ Breaking changes
 Closed issues
 =============
 
+* :hpx-issue:`4918` - Rename distributed_executors module
 * :hpx-issue:`4900` - Adding JOSS status badge to README
 * :hpx-issue:`4897` - Compiler warning, deprecated header used by HPX itself
 * :hpx-issue:`4886` - A future bound to an action executing on a different locality doesn't capture exception state
+* :hpx-issue:`4880` - Undefined reference to main build error when HPX_WITH_DYNAMIC_HPX_MAIN=OFF
 * :hpx-issue:`4877` - hpx_main might not able to start hpx runtime properly
 * :hpx-issue:`4850` - Issues creating templated component
+* :hpx-issue:`4829` - Spack package & HPX_WITH_GENERIC_CONTEXT_COROUTINES
 * :hpx-issue:`4820` - PAPI counters don't work
 * :hpx-issue:`4818` - HPX can't be used with IO pool turned off
 * :hpx-issue:`4816` - Build of HPX fails when find_package(Boost) is called before FetchContent_MakeAvailable(hpx)
 * :hpx-issue:`4813` - HPX MPI Future failed
+* :hpx-issue:`4811` - Remove HPX::hpx_no_wrap_main target before 1.5.0 release
 * :hpx-issue:`4810` - In hpx::for_each::invoke_projected the hpx::util::decay is misguided
 * :hpx-issue:`4787` - `transform_inclusive_scan` gives incorrect results for non-commutative operator
 * :hpx-issue:`4786` - transform_inclusive_scan tries to implicitly convert between types, instead of using the provided `conv` function
@@ -172,11 +177,13 @@ Closed issues
 * :hpx-issue:`4758` - CMake profiling results
 * :hpx-issue:`4755` - Building HPX with support for PAPI fails
 * :hpx-issue:`4754` - CMake cache creation breaks when using HPX with mimalloc
+* :hpx-issue:`4752` - HPX MPI Future build failed
 * :hpx-issue:`4746` - Memory leak when using dataflow icw components
 * :hpx-issue:`4731` - Bug in stencil example, calculation of locality IDs
 * :hpx-issue:`4723` - Build fail with NETWORKING OFF
 * :hpx-issue:`4720` - Add compatibility headers for modules that had their module headers implicitly generated in 1.4.1
 * :hpx-issue:`4719` - Undeprecate some module headers
+* :hpx-issue:`4712` - Rename HPX_MPI_WITH_FUTURES option
[…]
