+++ cmake/packaging/rpm/Changelog.txt
@@ -1,3 +1,6 @@
+* Wed Sep 02 2020 STE||AR Group <contact@stellar-group.org> 1.5.0
+- HPX Release 1.5.0 (https://hpx-docs.stellar-group.org/tags/1.5.0/html/releases/whats_new_1_5_0.html)
+
 * Wed Feb 12 2020 STE||AR Group <contact@stellar-group.org> 1.4.1-1
 - HPX Release 1.4.1 (https://hpx-docs.stellar-group.org/tags/1.4.1/html/releases/whats_new_1_4_1.html)
 
