+++ docs/sphinx/releases/whats_new_1_5_0.rst
@@ -9,7 +9,7 @@
 .. _hpx_1_5_0:
 
 ===========================
-|hpx| V1.5.0 (?)
+|hpx| V1.5.0 (Sep 02, 2020)
 ===========================
 
 General changes
