+++ docs/sphinx/quickstart.rst
@@ -33,7 +33,7 @@ It is also recommended that you check out the latest stable tag:
 
 .. code-block:: sh
 
-    git checkout 1.4.1
+    git checkout 1.5.0
 
 |hpx| dependencies
 ==================
