+++ libs/parallelism/algorithms/tests/unit/algorithms/transform_reduce_binary_bad_alloc.cpp
@@ -127,20 +127,18 @@ void test_transform_reduce_binary_bad_alloc_async(ExPolicy&& p, IteratorTag)
 template <typename IteratorTag>
 void test_transform_reduce_binary_bad_alloc()
 {
-    using namespace hpx::parallel;
+    using namespace hpx::execution;
 
     test_transform_reduce_binary_bad_alloc(IteratorTag());
 
     // If the execution policy object is of type vector_execution_policy,
     // std::terminate shall be called. therefore we do not test exceptions
     // with a vector execution policy
-    test_transform_reduce_binary_bad_alloc(execution::seq, IteratorTag());
-    test_transform_reduce_binary_bad_alloc(execution::par, IteratorTag());
+    test_transform_reduce_binary_bad_alloc(seq, IteratorTag());
+    test_transform_reduce_binary_bad_alloc(par, IteratorTag());
 
-    test_transform_reduce_binary_bad_alloc_async(
-        execution::seq(execution::task), IteratorTag());
-    test_transform_reduce_binary_bad_alloc_async(
-        execution::par(execution::task), IteratorTag());
+    test_transform_reduce_binary_bad_alloc_async(seq(task), IteratorTag());
+    test_transform_reduce_binary_bad_alloc_async(par(task), IteratorTag());
 }
 
 void transform_reduce_binary_bad_alloc_test()
