+++ libs/parallelism/algorithms/tests/unit/algorithms/exclusive_scan_validate.cpp
@@ -48,7 +48,7 @@ template <typename ExPolicy>
 void test_exclusive_scan_validate(
     ExPolicy p, std::vector<int>& a, std::vector<int>& b)
 {
-    using namespace hpx::parallel;
+    using namespace hpx::execution;
     typedef std::vector<int>::iterator Iter;
 
     // test 1, fill array with numbers counting from 0, then run scan algorithm
@@ -159,17 +159,17 @@ void exclusive_scan_validate()
     std::vector<int> a, b;
     // test scan algorithms using separate array for output
     DEBUG_OUT("\nValidating separate arrays sequential");
-    test_exclusive_scan_validate(hpx::parallel::execution::seq, a, b);
+    test_exclusive_scan_validate(hpx::execution::seq, a, b);
 
     DEBUG_OUT("\nValidating separate arrays parallel");
-    test_exclusive_scan_validate(hpx::parallel::execution::par, a, b);
+    test_exclusive_scan_validate(hpx::execution::par, a, b);
 
     // test scan algorithms using same array for input and output
     DEBUG_OUT("\nValidating in_place arrays sequential ");
-    test_exclusive_scan_validate(hpx::parallel::execution::seq, a, a);
+    test_exclusive_scan_validate(hpx::execution::seq, a, a);
 
     DEBUG_OUT("\nValidating in_place arrays parallel ");
-    test_exclusive_scan_validate(hpx::parallel::execution::par, a, a);
+    test_exclusive_scan_validate(hpx::execution::par, a, a);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
