### 2020-09-01 88cd425
Separate begin from end iterator in transform
