### 2020-09-02 585b29d
Update the release procedure with the new hpx website
