### 2020-09-02 3774894
Fix CUDA architecture flag on jenkins rostam configuration
