+++ .jenkins/lsu/env-gcc-9-cuda-11.sh
@@ -22,4 +22,4 @@ configure_extra_options+=" -DHPX_WITH_NETWORKING=OFF"
 configure_extra_options+=" -DHPX_WITH_DISTRIBUTED_RUNTIME=OFF"
 configure_extra_options+=" -DHPX_WITH_LOGGING=OFF"
 configure_extra_options+=" -DHPX_WITH_ASYNC_MPI=ON"
-configure_extra_options+=" -DNVCC_FLAGS=-arch=sm_35"
+configure_extra_options+=" -DCUDA_NVCC_FLAGS=-arch=sm_35"
