### 2020-09-02 #2352 @diehlpk
Release number 1.5.0

https://github.com/STEllAR-GROUP/hpx/tree/1.5.0
