### 2020-09-02 #2352 @diehlpk
@danielskatz Please find above the requested information, since we released the latest HPX version today.
