### 2020-09-02 #2352 @diehlpk
> 👋 @diehlpk - which paper is the right one? In the branch or in master?  Note the one in master didn't build...

@danielskatz Both should be the same, since i did a pull request from the branch JOSS to master. I will check soon why the pull request failed.
