### 2020-09-02 #2352 @hkaiser
@danielskatz the paper on the JOSS branch might not be the latest version, we'll remove the comment on the master branch.
