### 2020-09-02 6887589
Removing license comments, preventing inspect from looking at the paper
