### 2020-09-02 #2352 @danielskatz
I'm suggesting a number of small changes in https://github.com/STEllAR-GROUP/hpx/pull/4939 - please either merge this or let me know what you disagree with.
