### 2020-09-02 a703f64
Add basic cmake HIP handling
