+++ cmake/HPX_AddLibrary.cmake
@@ -213,7 +213,10 @@ function(add_hpx_library name)
     endforeach()
   endif()
 
-  if(HPX_WITH_CUDA_COMPUTE AND NOT HPX_WITH_CUDA_CLANG)
+  if(HPX_WITH_CUDA_COMPUTE
+     AND NOT HPX_WITH_CUDA_CLANG
+     AND NOT HPX_WITH_HIP
+  )
     cuda_add_library(
       ${name} ${${name}_linktype} ${exclude_from_all} ${${name}_SOURCES}
       ${${name}_HEADERS} ${${name}_AUXILIARY}
