+++ libs/full/async_cuda/CMakeLists.txt
@@ -10,10 +10,10 @@ if(NOT HPX_WITH_ASYNC_CUDA)
   return()
 endif()
 
-if(NOT TARGET Cuda::cuda)
+if(NOT TARGET Cuda::cuda AND NOT HPX_WITH_HIP)
   message(
     FATAL_ERROR
-      "Cuda future support has been enabled, but the Cuda installation is not setup"
+      "CUDA/HIP future support has been enabled, but the CUDA or HIP installation is not setup"
   )
 endif()
 
@@ -47,6 +47,14 @@ if(HPX_WITH_DISTRIBUTED_RUNTIME)
   )
 endif()
 
+if(HPX_WITH_HIP AND TARGET roc::hipblas)
+  set(async_cuda_extra_deps ${async_cuda_extra_deps} roc::hipblas)
+elseif(HPX_WITH_CUDA AND TARGET Cuda::cuda)
+  set(async_cuda_extra_deps ${async_cuda_extra_deps} Cuda::cuda
+                            ${CUDA_CUBLAS_LIBRARIES}
+  )
+endif()
+
 include(HPX_AddModule)
 add_hpx_module(
   full async_cuda
@@ -58,6 +66,6 @@ add_hpx_module(
   COMPAT_HEADERS ${async_cuda_compat_headers}
   MODULE_DEPENDENCIES hpx_program_options hpx_runtime_configuration
                       hpx_runtime_local ${async_cuda_extra_module_dependencies}
-  DEPENDENCIES hpx_core hpx_parallelism Cuda::cuda
+  DEPENDENCIES hpx_core hpx_parallelism ${async_cuda_extra_deps}
   CMAKE_SUBDIRS examples tests
 )
